Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051 351155.1250
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203    634.3320
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579   1318.7560
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644   1179.0695
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491   1844.4353
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284   2010.4285
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150   1274.3594
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357    275.7474
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416   1204.8629
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316  0.00714    475.3120
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178    800.1933
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230   1056.5876
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175   1122.9120
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241    585.6182
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427    380.9472
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542   3614.7334
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304   1569.7313
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395   2011.3270
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596    842.0994
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520   3390.5708
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522   1355.7436
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380    601.0843
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756    353.9086
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200   1305.6691
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075   2170.2249
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010  0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341  0.00171   4425.5410
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341   1363.4000
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291    293.4231
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291    487.3438
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389  14516.2002
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476   1275.9446
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538    265.4677
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445    355.5078
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650   1133.1501
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333    422.8840
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341  0.04431    315.9084
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853   1354.6001
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943    371.5514
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223   1056.3701
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997    230.8641
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282    561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148  10771.1963
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066    306.0146
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566  12862.4922
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805   7489.7417
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012    305.3624
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169    649.9975
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109    915.4989
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156    758.4840
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983