Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261    323.8441
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041    772.9746
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250   1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410    0.00118   0.00761    928.6858
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491   1844.4353
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550  -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416    947.4878
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378    440.9104
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392   1255.0884
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261    554.9243
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710    843.7961
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450    873.0448
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294  -0.00096    520.5570
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866  29195.6016
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613    289.6113
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307 147470.5312
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917  -0.01993  10308.6406
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310   2833.9001
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048  -0.00332   1418.6622
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485   2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664   1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961   1086.6963
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180   3904.5254
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242    507.9263
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245   6353.3179
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860    695.2302
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132   1580.1621
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511  -0.01801   1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739    789.7573
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003   0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875    597.2893
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732   4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905   6487.0010
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210    585.8992
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189    556.5211
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204    853.2902
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493   2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333    422.8840
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476  11992.5273
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943    371.5514
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790    209.6952
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083   1309.2347
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099   1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307   0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312    593.7773
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654   1576.2065
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852   0.00248   1432.4019
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576   1045.1074
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129   0.00491    261.7371
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805   7489.7417
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858    298.4490
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157    766.0220
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253   1237.0536
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235