Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976    301.2347
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813   1130.5791
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555    430.4604
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005    769.3426
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403   1787.3024
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345   1120.6929
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501   3889.7292
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382    560.8309
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491  0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696   4143.4019
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898    588.3761
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950    950.3138
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083    994.7876
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642    613.6390
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651    419.0927
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887    366.8537
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096    520.5570
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025   4425.5151
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907   1943.4092
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903   2360.1536
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279    930.0871
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252   5761.3486
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181  11655.4316
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409   2312.8413
#> 37416 713282.5 9560692 1850.61 30.2590471  0.00736  0.01684   2556.8049
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262   1760.2162
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809   1102.9956
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240   1201.4724
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849   1120.5961
#> 34887 714662.5 9557392 2609.02 51.0207457 -0.01878 -0.04351   1442.9247
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372   4154.5786
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261   3353.5422
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396    704.7076
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778   4594.4351
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483    343.1384
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610   1872.9276
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400   1982.7782
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870   2099.7930
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320    911.5911
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105   4432.8760
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363    389.2691
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445    355.5078
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077   8759.9961
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399   1601.4852
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790    209.6952
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413  0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196    576.1450
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268    252.0072
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257    549.3252
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204    618.2027
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308    422.5022
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577    423.5598
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836    250.5111
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050   1471.4346
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282   1659.6211
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368   1071.8734
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177    184.1042
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169    649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491