Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235