Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958    0.02642    671.1807
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921   -0.12438    753.6304
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199    0.00659   3553.7166
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305   -0.00555    430.4604
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753    0.01487   1077.7585
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046    0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240   -0.00250   1413.3312
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065    0.00005    769.3426
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059   -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118    0.00761    928.6858
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310    0.01440   1088.8639
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951   -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284    0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854    0.01264   5057.1431
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223    0.01783    377.8918
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234    849.0547
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123    0.01164    895.6606
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270   -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070   -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602    0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131    0.00501   3889.7292
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430   -0.06070 138869.3906
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677   1675.4679
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151    0.00899   1394.2274
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059   -0.00628    556.0121
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077   -0.00284   2010.4285
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108   -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651   -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521   -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987    0.02107   1392.5880
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733    598.7983
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205    0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804   -0.01596   5296.9014
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275    0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193   -0.00077   3559.4204
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743    0.01313    783.4548
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606   -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904   -0.00866  29195.6016
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752   -0.00517   1000.7119
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066   -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174    0.00025   4425.5151
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093   -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843    0.04407    344.8395
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552   -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661    0.01199    196.4080
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290    0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749    0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538    807.8237
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283    0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409    0.02771    674.3859
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478    0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175    0.00236   2715.4138
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953   -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681   -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160    0.01690   5121.6924
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161   -0.00409   2312.8413
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807   -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548    0.05543    352.1558
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034    0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062   -0.00262   1760.2162
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550   -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554    0.03304   1569.7313
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078    0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238   -0.00851   1012.5963
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604    0.00354   1147.7384
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024   -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627   -0.05393   8301.6094
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188    0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469    0.02940    987.8085
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865    0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458   -0.14813 287033.4688
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332   -0.00058   1817.2371
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489    0.00241   1752.9242
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299    0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474    0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175    0.04195   1901.6743
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109    0.00298   1533.7925
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250   -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124   -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258    0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410   -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102   -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281    0.04759    158.9370
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693    0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364   -0.00094   1392.6008
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110   -0.02450   2589.9910
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538    807.8237
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902    0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020    0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675   -0.01905   6487.0010
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166    0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195   -0.00205   1112.2626
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585   -0.02665   1904.4769
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251   -0.00389  14516.2002
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989   -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378    0.02382    100.0000
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159   -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196    0.00336   2806.2644
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872    0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298   -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162   -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336    0.01403    241.2991
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112    0.03538    265.4677
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197    0.00463    851.6306
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724    0.01464   4029.5947
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065    0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846    0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069   -0.00900    614.5898
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180    0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193   -0.00267  31588.1953
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292    0.00462    716.5762
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185    0.01655    560.8554
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338    0.00518   1137.0179
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068   -0.00312    593.7773
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462    0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978    0.01282   2322.7944
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310   -0.07650    766.6254
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858    0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923    0.05547    128.2613
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980   -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972    0.00308    422.5022
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523    0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360    0.00750   1892.8978
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074   -0.02904    608.3561
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070   -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375    0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502    0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598    0.02208    842.4825
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129    0.00491    261.7371
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108   -0.00948   1538.1741
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508   -0.00177    184.1042
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298    0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811   -0.02281    492.4785
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366    0.00525    162.0592
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379   -0.00249    148.1857
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023   -0.00227    841.2025
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456   -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771    0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491