Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235