Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976     301.2347
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555     430.4604
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324     861.3537
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057    1630.3259
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403    1787.3024
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340     604.1032
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949    6217.8213
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328    2235.5259
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 18005 714832.5 9557912 2660.76 31.9120303  0.00191 -0.00261     554.9243
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917     116.1986
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205     629.4165
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051     779.0555
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651     419.0927
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752 -0.00517    1000.7119
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270    4057.5637
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279     930.0871
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516     576.0335
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181   11655.4316
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427     380.9472
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409    2312.8413
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166     650.6246
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304    1569.7313
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240    1201.4724
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627 -0.05393    8301.6094
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911     173.2806
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321    1170.5904
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981    3975.3608
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778    4594.4351
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077    3056.9475
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075    2170.2249
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119    6589.5796
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610     454.8103
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166    1813.7841
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632     620.0229
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559     866.5775
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085    2229.5105
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715     520.7355
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875    1165.3953
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028    2394.7261
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167  0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697     412.1185
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483    1097.3082
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714  480535.5000
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650     766.6254
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547     128.2613
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710     472.0236
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750    1892.8978
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047    2826.7700
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907    1281.5044
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875    2156.9358
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841     756.0055
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013     616.4128
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157     766.0220
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448     941.8503
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603