Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919   -0.04051 351155.1250
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056   -0.06976    301.2347
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221    0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305   -0.00555    430.4604
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479    0.01261    323.8441
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046    0.00324    861.3537
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891   -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884    0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645    0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491   1844.4353
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542    0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123    0.01164    895.6606
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720    0.01250   3686.1240
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626   4316.0977
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189   2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051    0.00889    362.8363
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490   -0.00450    873.0448
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521   -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917    116.1986
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593    939.0466
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950    950.3138
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996    0.00626   1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276   9614.6924
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275    0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211    0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765    0.00114    150.1498
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696    0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545    0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047   -0.00887    366.8537
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294   -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846   -0.00034   4495.6929
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509    0.00499    885.0064
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066   -0.00036    516.4709
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661    0.01199    196.4080
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613    0.00903   2360.1536
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585   -0.00974   2136.6939
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008    0.00102    483.4184
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399   -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588    749.2502
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063    576.2819
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473   2930.2791
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062   -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126   -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114    0.03026    247.0894
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941   1290.7709
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972    0.03512  13178.3223
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784    0.00596    842.0994
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240   1201.4724
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448   1258.8196
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133    0.00367    806.0966
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169   -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394    587.8840
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278   -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400    608.6523
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491    0.01061   5744.7583
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338    0.00728   1137.2450
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299    0.00141   4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070   -0.02661   2508.5374
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583   1898.0105
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442   -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396    704.7076
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401   -0.00391   1794.7964
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110   -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526    706.1463
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451   -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216   -0.02606    919.4241
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205   1112.2626
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367   1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320    911.5911
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427    866.9398
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534    0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009    0.00109    605.1908
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226    0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284    0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872    0.05419    761.5427
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162   -0.00517   1060.5494
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812   -0.02208    951.3401
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220   -0.03290   2481.5632
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650   1133.1501
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253   -0.00333    422.8840
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014    271.4219
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001    0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399   1601.4852
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084   -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236    0.00483   1097.3082
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490   -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170    253.5129
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674   -0.00376    629.8236
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423   -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292    0.00462    716.5762
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307    0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338    0.00518   1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322   -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379   -0.01018    937.6911
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978    0.01282   2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305   -0.00705    968.9738
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608    0.00148  10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523    0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719    0.00079   1647.6451
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148    481.8292
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011   -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366    0.00525    162.0592
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379   -0.00249    148.1857
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247    0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208    0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491