Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813    1130.5791
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043    1806.4552
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487    1077.7585
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644    1179.0695
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257     357.1436
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804     628.5153
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542     674.0147
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387    1989.3012
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281    1325.6138
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581  0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104    1641.4587
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665   25318.3555
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032     994.6740
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 22566 712832.5 9559842 1897.49 33.0602377  0.00672  0.04398     267.2659
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491  0.00581    1855.1630
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678    1891.6584
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357     275.7474
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114     150.1498
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346     543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355   14919.6260
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167    1938.5389
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373   16031.7500
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752 -0.00517    1000.7119
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316  0.00714     475.3120
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796    4729.6382
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613     289.6113
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063   18733.8848
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058     431.4362
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199     196.4080
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072   54831.6289
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102     483.4184
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427     380.9472
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055     773.0359
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620     379.0855
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367     806.0966
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220    8936.2666
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574     183.3220
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236    9755.3027
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195    1901.6743
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981    3975.3608
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380     601.0843
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915    2036.0378
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294     734.9908
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483     343.1384
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450    2589.9910
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226     338.3474
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000  162178.1562
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400    1982.7782
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603    1562.0503
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067    1123.7318
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476    1275.9446
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177     997.3089
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116    1817.5222
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363     389.2691
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199     836.4196
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077    8759.9961
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399    1601.4852
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083    1309.2347
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035    1471.9783
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036     476.9504
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759    4131.5356
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099     909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308     422.5022
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960     884.2003
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624    1870.8307
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151  143213.1562
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270   35950.5508
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087 -0.00573    1490.0312
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448     941.8503
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156     758.4840
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403     609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931    1812.3820
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250