Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029    5577.3916
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357    1752.0375
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969     368.5606
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896     276.0310
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433    1475.3191
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406    2421.1145
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388    1051.4222
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328    2235.5259
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096    3210.5781
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899    1394.2274
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889     362.8363
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773     179.3077
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694    2123.4473
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651     419.0927
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146    1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124    2114.2253
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796    4729.6382
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025    4425.5151
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974    2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102     483.4184
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197     711.1091
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299     688.0706
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304    1569.7313
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316     454.9674
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240    1201.4724
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469    1021.5407
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389   47022.9688
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213     328.7922
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078  622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812    1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363     565.5207
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756     353.9086
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133    2397.2681
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610     454.8103
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000  162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253     528.1924
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655     130.9468
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150     503.5590
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208     951.3401
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363     389.2691
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290    2481.5632
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077    8759.9961
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083    1309.2347
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654    1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948    1222.5433
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099     909.0560
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875    2156.9358
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119     792.3544
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585    5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458     485.7418
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281     492.4785
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020     673.2861
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235