Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491