Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029   5577.3916
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969    368.5606
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043   1806.4552
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266    0.02896    276.0310
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661   -0.01339   1002.5877
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677   1675.4679
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679    550.0344
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542    674.0147
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378    440.9104
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903    0.01447    331.4833
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272   4955.5586
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530   9000.4356
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430   -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044 129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203  10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189   2317.1819
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314   2360.1501
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437   1155.1469
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114    150.1498
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651    419.0927
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338  10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714    475.3120
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940    0.00679    550.0344
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230   1056.5876
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395    0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530    937.1306
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181   2406.1252
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161   -0.00409   2312.8413
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739    365.7626
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550   -0.00941   1290.7709
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845    0.00454   3476.0818
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972    0.03512  13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078    0.00219    420.7032
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224    942.0322
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242    507.9263
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567   -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813 287033.4688
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491    0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509   -0.01211  49347.5547
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431   4079.0366
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261   1627.8989
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372   4154.5786
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079    0.00321   1170.5904
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801   1071.6141
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380    601.0843
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292   -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767   -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483    343.1384
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000 162178.1562
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230   1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465   -0.01036   2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063   2430.2278
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189    556.5211
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684    209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585   -0.02665   1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389  14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000    0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427    866.9398
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331    699.3024
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232    977.7320
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105   4432.8760
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463   1050.7473
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064    683.2795
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548    532.8210
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527   -0.02077   8759.9961
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065    0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790    209.6952
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697    412.1185
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780    825.4451
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098    164.0802
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487   1769.3392
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060    383.3547
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555   1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165    805.1403
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635   -0.00105    868.7802
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262   -0.00502    520.5218
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532  16771.3066
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875   2156.9358
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233    968.7780
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788   -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119    792.3544
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585   5714.3267
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803   1058.5630
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573   1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305   -0.01135   1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931   1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057    722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775