Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051 351155.1250
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813   1130.5791
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712   3636.5303
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483    566.7674
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041    772.9746
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766 140591.4062
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083    994.7876
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 41037 715532.5 9558792 2529.58 32.84652 -0.00585 -0.00974   2136.6939
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373    527.3528
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405 -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401   1722.6481
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684   2556.8049
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553   1750.4175
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472 -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448   1258.8196
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 20368 715252.5 9560622 1865.83 33.57246  0.00025 -0.01245   6353.3179
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420
#> 21683 714792.5 9561052 1807.84 55.73275 -0.02070 -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.52758 -0.00193  0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546  0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915   2036.0378
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367   1966.2112
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316 -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159 -0.01559    866.5775
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419    761.5427
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 25140 713332.5 9560692 1809.95 52.25719  0.00075 -0.01945    259.7421
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990  0.02790    209.6952
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662   1269.2650
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011    201.1272
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997    230.8641
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392
#> 1281  715102.5 9560032 2120.64 34.64905 -0.00930 -0.00360   4288.7329
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577    423.5598
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292
#> 1434  714672.5 9559902 2194.92 42.45216 -0.01196 -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585   5714.3267
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124  25826.6914
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607