Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555    430.4604
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097   0.01483    566.7674
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090   0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292   1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403   1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264   5057.1431
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214   0.00466    724.0713
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024    394.8751
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096   3210.5781
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382   0.00672   1114.8801
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284   2010.4285
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898    588.3761
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150   1274.3594
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051    779.0555
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743   0.01313    783.4548
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991   1403.0193
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321   0.01169    350.1799
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307 147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974   2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964    571.1364
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602    345.8933
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020   1563.5549
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599   1378.5760
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594  16587.9980
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190    640.8842
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299   0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169   0.01441    424.2680
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970   1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044    288.5468
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152    850.4633
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405  -0.05655    130.9468
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842    246.5161
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109    605.1908
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570   0.02230    714.7432
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298  -0.00712   1010.5097
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738   0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493   2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452    578.5847
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132    737.5982
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923    811.5580
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759   1880.4403
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223   1056.3701
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307   0.02997    230.8641
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038   2017.3547
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955    713.4463
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868    235.8511
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417   1520.6630
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273  -0.01047   2826.7700
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532  16771.3066
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549  -0.00169    649.9975
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803   1058.5630
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157    766.0220
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491