Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919   -0.04051 351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123    0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056   -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347    0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199    0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633   -0.01813   1130.5791
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292    0.00712   3636.5303
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917    0.00333    258.6537
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046    0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498   -0.00752   1321.5997
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273   -0.02403   1787.3024
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922    0.01088    266.4355
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311    586.7354
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021   -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952    0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189    287.5425
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017   -0.00032    994.6740
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126   -0.02044 129412.3984
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856    0.01114   4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993   -0.00413    364.9788
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356    0.00774    527.1068
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916   2009.1686
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416   1204.8629
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696    0.02514    297.7842
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904   -0.00866  29195.6016
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509    0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338  10787.6299
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666   -0.00504   1020.2766
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843    0.04407    344.8395
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552   -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735   -0.00925    665.6116
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708   -0.02538    807.8237
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855   -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478    0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175    0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917   -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762    0.01252   5761.3486
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542    860.8840
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548    0.05543    352.1558
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099    0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102  35685.0000
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238   -0.00851   1012.5963
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448   1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594  16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601   -0.08389  47022.9688
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370   -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458   -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524  10861.9482
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489    0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417    0.00516    0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911    173.2806
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212  46310.0195
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878   -0.04351   1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070   -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263    0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444   -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218   -0.00372   4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833    369.3771
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247    0.01123    364.2903
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789   -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124   -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070    0.00010   3467.2690
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102   -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018    0.02281    0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495   -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705   -0.00915   2036.0378
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864    0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143    0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445    0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181   -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610    454.8103
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166   1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354    0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675   -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488   -0.00632    620.0229
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234    849.0547
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427    0.02403    474.8613
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585   -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320    911.5911
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159   -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849    0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325   -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298   -0.00712   1010.5097
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812   -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328   1066.7134
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409   1575.9425
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846    0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176    0.02973    215.5101
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527   -0.02077   8759.9961
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193   -0.00267  31588.1953
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423   -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687   -0.00223   1056.3701
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338    0.00518   1137.0179
#> 330   714232.5 9560362 2165.00 18.060776    0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233    0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635   -0.00105    868.7802
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502    520.5218
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719    0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.634509    0.00082    0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438   1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852    0.00248   1432.4019
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375    0.00215    917.5296
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607   -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182   -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523    888.2811
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263   -0.02012    305.3624
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161    0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309    0.01011    501.7867
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014   -0.00884    672.2466
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021    0.00109    915.4989
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933   -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281    492.4785
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253   1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366    0.00525    162.0592
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379   -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607