Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491