Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642    671.1807
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969    368.5606
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741    746.4155
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041    772.9746
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046    0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644   1179.0695
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073    0.00174   2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440   1088.8639
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.971087  0.00183    0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088    266.4355
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466    724.0713
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378    440.9104
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898    458.6938
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261    554.9243
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710    843.7961
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077   -0.00284   2010.4285
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108   -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629 360379.1875
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422    0.02357    275.7474
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077   3559.4204
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651    419.0927
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545    0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887    366.8537
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904   -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063   5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376   2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499    885.0064
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666   -0.00504   1020.2766
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925    665.6116
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949   1182.1506
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319   3457.7519
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161   -0.00409   2312.8413
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262   1760.2162
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553   1750.4175
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809   1102.9956
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961   1086.6963
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131    0.00201    788.0621
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555   4439.0967
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664   1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188    0.00832   3022.3462
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017    0.01213    328.7922
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205   1112.2626
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332   -0.00058   1817.2371
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241   1752.9242
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281   1178.4420
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141   4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384    0.00175   1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372   4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722   1373.4147
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484    495.0292
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292   -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401   -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705    413.5702
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538    807.8237
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477   2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157   -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606    919.4241
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152   1530.5657
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632    620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744    226.7911
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870   2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063   2430.2278
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684    209.1778
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265   1148.6487
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158    837.3730
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712   1010.5097
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463   1050.7473
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875   1165.3953
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028   2394.7261
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833    300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431    315.9084
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133    828.8604
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367   3002.8828
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697    412.1185
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514    897.9290
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423   -0.00983    270.7949
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655    560.8554
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196    576.1450
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479    0.02251    168.1733
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282    561.2366
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361   1011.0248
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750   1892.8978
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438   1276.5448
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215    917.5296
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841    756.0055
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177    184.1042
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732    0.01858    298.4490
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799   1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109    915.4989
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253   1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020    673.2861
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403    609.8665
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456   -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491