Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642     671.1807
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813    1130.5791
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333     258.6537
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340     604.1032
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257     357.1436
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804     628.5153
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416     947.4878
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526     539.8119
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766  140591.4062
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164     895.6606
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501    3889.7292
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312    2657.6936
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774     527.1068
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118    2359.5059
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616  0.01276    9614.6924
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146    1347.6224
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407     344.8395
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279     930.0871
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560     977.4719
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825     162.6688
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409    2312.8413
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026     247.0894
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927 -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643     670.2289
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219     420.7032
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240    1201.4724
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367     806.0966
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822     472.8748
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534     344.1447
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337     486.7551
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574     183.3220
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372    4154.5786
#> 28312 715712.5 9558032 2783.23 24.2527305 -0.00007 -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722    1373.4147
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778    4594.4351
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133    2397.2681
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185     953.8614
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875     597.2893
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921     525.3245
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152    1530.5657
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655     130.9468
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150     503.5590
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009  0.00109     605.1908
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736     477.8591
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403     241.2991
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328    1066.7134
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875    1165.3953
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409    1575.9425
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273   72123.6172
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167  0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853    1354.6001
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790     209.6952
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362     283.7090
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035    1471.9783
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462     716.5762
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312     593.7773
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714  480535.5000
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130     107.5831
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099     909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960     884.2003
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260    1484.6593
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148     481.8292
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948    1538.1741
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029     894.2672
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585    5714.3267
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403     609.8665
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769