Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043   1806.4552
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339   1002.5877
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387   1989.3012
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378    440.9104
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530   9000.4356
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382  0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710    843.7961
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739    558.2286
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150   1274.3594
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642    613.6390
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355  14919.6260
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606 -0.00556    660.3690
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613    289.6113
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907   1943.4092
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319   3457.7519
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166    650.6246
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304   1569.7313
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927 -0.00316    454.9674
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013   4229.5708
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604  0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594  16587.9980
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017  0.01213    328.7922
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860    695.2302
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583   1898.0105
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957    495.4453
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292 -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610   1872.9276
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000 162178.1562
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632    620.0229
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291    487.3438
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389  14516.2002
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331    699.3024
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517   1060.5494
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463   1050.7473
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538    265.4677
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945    259.7421
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833    300.0125
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948   1222.5433
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513    161.9878
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868    235.8511
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204    618.2027
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555   1882.0179
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577    423.5598
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499    617.3855
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566  12862.4922
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368   1071.8734
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172 111973.8906
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012    305.3624
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057    722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775