Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051 351155.1250
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579   1318.7560
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041    772.9746
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644   1179.0695
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005    769.3426
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403   1787.3024
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264   5057.1431
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677   1343.5006
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104   1641.4587
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518    854.4269
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710    843.7961
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413    364.9788
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917    116.1986
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357    275.7474
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051    779.0555
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416   1204.8629
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313    783.4548
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124   2114.2253
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499    885.0064
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796   4729.6382
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230   1056.5876
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117   1018.2141
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169    495.8840
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927 -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643    670.2289
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596    842.0994
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879   1200.4210
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196   1052.5341
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957    495.4453
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426   1315.5334
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075   2170.2249
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325    595.0426
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477   2945.2100
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015    790.2421
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870   2099.7930
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116   1817.5222
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363    389.2691
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945    259.7421
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598    186.3527
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218  0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282   2322.7944
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251    168.1733
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312    254.0656
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417   1520.6630
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960    884.2003
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050   1471.4346
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119    792.3544
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585   5714.3267
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799   1036.6606
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769