Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250    1413.3312
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491    1844.4353
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311     586.7354
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665   25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070  138869.3906
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899    1394.2274
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973     514.5488
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450     873.0448
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629  360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773     179.3077
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205     629.4165
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077    3559.4204
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270    4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128  -0.00855   -0.02175    1122.9120
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542    3614.7334
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169     495.8840
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219     420.7032
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367     806.0966
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213     328.7922
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171     788.0004
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241    1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661    2508.5374
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118   19557.6621
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778    4594.4351
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001    1184.7952
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171    4425.5410
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606     919.4241
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384     947.2075
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230    1044.6443
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306    1718.3652
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996    1599.3491
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559     866.5775
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158     837.3730
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875    1165.3953
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409    1575.9425
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028    2394.7261
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476   11992.5273
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697     412.1185
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.085050  -0.00196   -0.00175     725.1355
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983     270.7949
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 20135 714772.5 9560232 2107.52 37.850738  -0.00782   -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759    4131.5356
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868     235.8511
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502     520.5218
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964     543.0025
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907    1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849     764.8193
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282    1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491     261.7371
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377    1727.6055
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448     941.8503
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775