Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183   -0.09203    634.3320
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221    0.00969    368.5606
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659    0.01041    772.9746
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503    0.02433   1475.3191
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104   -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453    0.00257    357.1436
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068    0.00588   1295.1610
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769   -0.00281   1325.6138
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430   -0.06070 138869.3906
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491    0.00581   1855.1630
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077   -0.00284   2010.4285
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993   -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490   -0.00450    873.0448
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651   -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206    0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987    0.02107   1392.5880
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598    0.01212    561.1948
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452   -0.01437   1155.1469
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646   -0.04086   6042.5757
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440   -0.02651    419.0927
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566   -0.03355  14919.6260
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353    0.00063   5035.0000
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316    0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666   -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679    550.0344
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093   -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060    0.00270   4057.5637
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034    0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552   -0.02058    431.4362
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613    0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149    0.01949   1182.1506
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290    0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008    0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395    0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286   -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749    0.04779   2972.0193
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399   -0.00731   1384.7959
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002    0.00588    749.2502
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017    0.00197    711.1091
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160    0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221    0.00121   4159.6738
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856   -0.01625   1063.9150
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114    0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448   -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554    0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099    0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406   -0.00055    773.0359
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104    0.01664   1898.5588
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078    0.00219    420.7032
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904   -0.02013   4229.5708
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627   -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093    0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743   -0.05242    507.9263
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509   -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556   -0.11694 102301.4609
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489    0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589    0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890   -0.00911    173.2806
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474    0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171   -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175    0.04195   1901.6743
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354    0.00245    629.5983
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079    0.00321   1170.5904
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109    0.00298   1533.7925
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669   -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102   -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281    0.04759    158.9370
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970    0.00200   1305.6691
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705   -0.00915   2036.0378
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010    0.00140   1066.3957
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110   -0.02450   2589.9910
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730   -0.02610   1872.9276
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143    0.00477   2945.2100
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124    0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360   -0.03000 162178.1562
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458    0.04291    293.4231
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010   -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319    0.00291    487.3438
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427    0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996    0.04684    209.1778
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585   -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166   -0.01265   1148.6487
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316   -0.00427    866.9398
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378    0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332    0.01842    246.5161
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849    0.00150    503.5590
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570    0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759    0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284    0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872    0.05419    761.5427
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533    0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812   -0.02208    951.3401
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112    0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815   -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458    0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197    0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480    0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345   -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306   -0.00626    450.7346
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712    0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031   -0.00199    836.4196
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544   -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263    0.03598    186.3527
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001    0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973   -0.00943    371.5514
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069   -0.00900    614.5898
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737    0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196   -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844    0.00514    897.9290
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218    0.00902    218.7467
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410    0.00400   4126.7812
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978    0.01282   2322.7944
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950   -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118    0.00162    290.2031
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154   -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233    0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070   -0.00060    383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972    0.00308    422.5022
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511    0.01499    617.3855
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082    0.02578   1258.8865
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552    0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375    0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105   -0.01576   1045.1074
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598    0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361   -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137   -0.01523    888.2811
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572    0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284   -0.00886  43249.3945
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278   -0.01172 111973.8906
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137   -0.01523    888.2811
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305    0.00585   5714.3267
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282    0.00458    485.7418
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716   -0.03234   1707.3130
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166    0.00156    758.4840
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379   -0.00249    148.1857
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023   -0.00227    841.2025
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901