Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491