Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051  351155.1250
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203     634.3320
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357    1752.0375
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813    1130.5791
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475    1001.0861
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333     258.6537
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005     769.3426
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403    1787.3024
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440    1088.8639
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588    1295.1610
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547     519.4228
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149   12632.7793
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378     440.9104
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530    9000.4356
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312    2657.6936
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203   10263.2734
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581    1855.1630
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696    4143.4019
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678    1891.6584
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212     561.1948
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452 -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205     629.4165
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596    5296.9014
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916    2009.1686
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089 -0.00051     779.0555
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146    1347.6224
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096     520.5570
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167    1938.5389
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169     350.1799
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442 -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940  0.00679     550.0344
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175     804.5679
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949    1182.1506
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530     937.1306
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993   10308.6406
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197     711.1091
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409    2312.8413
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166     650.6246
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485    2240.4185
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512   13178.3223
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469    1021.5407
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472 -0.00228    2262.3796
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822     472.8748
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412    1350.3611
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061    5744.7583
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860     695.2302
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132    1580.1621
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161    3065.3694
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441     424.2680
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190     264.6404
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426    1315.5334
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391    1794.7964
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121    1783.0658
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538     807.8237
#> 28260 715312.5 9558302 2681.75 34.48404 -0.01125 -0.02685    2568.4631
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606     919.4241
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.99779  0.01354  0.00245     629.5983
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210     585.8992
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476    1404.4454
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609    4275.3457
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331     699.3024
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505     351.4078
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109     605.1908
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465 -0.00105    4432.8760
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204     853.2902
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208     951.3401
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132     737.5982
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290    2481.5632
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199     836.4196
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133     828.8604
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759    1880.4403
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990  0.02790     209.6952
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298     436.1459
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223    1056.3701
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011     201.1272
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997     230.8641
#> 29107 713172.5 9559272 2067.58 28.26401  0.02413  0.00528     355.3475
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705     968.9738
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863    1098.3462
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740     909.4791
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417    1520.6630
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260    1484.6593
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578    1258.8865
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264  0.00624    1870.8307
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368    1071.8734
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172  111973.8906
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169     649.9975
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858     298.4490
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458     485.7418
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172     351.5812
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388     180.4414
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149    4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607