Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919   -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921   -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021   -0.01579   1318.7560
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896    276.0310
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046    0.00324    861.3537
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073    0.00174   2310.4871
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979   -0.01542    674.0147
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223    0.01783    377.8918
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102    0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831   -0.00149  12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769   -0.00281   1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269    491.8017
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024    394.8751
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272   4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412    0.00898    458.6938
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501   3889.7292
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672    0.04398    267.2659
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189   2317.1819
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993   -0.00413    364.9788
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664    0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382    0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804   -0.01596   5296.9014
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646   -0.04086   6042.5757
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224    0.00177    291.5937
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353    0.00063   5035.0000
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993    0.01373  16031.7500
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796   4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178    800.1933
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452    437.3965
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492    0.00072  54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359   -0.00279    930.0871
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779   2972.0193
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357    473.4709
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658    0.04602    345.8933
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458   -0.00401   1722.6481
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032    0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625   1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126   -0.00485   2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102  35685.0000
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410    0.03620    379.0855
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078    0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201    788.0621
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240   1201.4724
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320    0.00180   3904.5254
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538    0.00112    745.3315
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188    0.00832   3022.3462
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017    0.01213    328.7922
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246    0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337    486.7551
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129    0.00171    788.0004
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332   -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491    0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781  29995.1543
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212  46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236   9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878   -0.04351   1442.9247
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195   1901.6743
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722   1373.4147
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789   -0.00261   3353.5422
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190    264.6404
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296    0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774   -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109   2047.9031
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185    953.8614
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125   -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166   1813.7841
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118   -0.00252   3631.7017
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004   -0.02384    947.2075
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400   -0.01210    585.8992
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319    0.00291    487.3438
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367   1966.2112
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084   -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382    100.0000
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759    0.00158    837.3730
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328   1066.7134
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075   8141.7549
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028   2394.7261
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263    0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476  11992.5273
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001    0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973   -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196    0.01734    335.0773
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601   -0.00011    712.5228
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292    0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041   -0.00011    201.1272
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413    0.00528    355.3475
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035    0.00955    713.4463
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950   -0.05130    107.5831
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858    0.00312    254.0656
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523    0.02247    282.1076
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907   1281.5044
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361   -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129    0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452    0.00368   1071.8734
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108   -0.00948   1538.1741
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196   -0.01935    903.6397
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161    0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523    888.2811
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858    298.4490
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627   -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222    991.8300
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716   -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962   -0.00172    351.5812
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288    0.00448    941.8503
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135   1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379   -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403    609.8665
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456   -0.01586    941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603