Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123  0.00393   2081.0056
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021 -0.01579   1318.7560
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712   3636.5303
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483    566.7674
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073  0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642    613.6390
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495    266.5888
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330   1456.9246
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620    379.0855
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078  0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013   4229.5708
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728   1137.2450
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583   1898.0105
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610   1872.9276
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856   1789.1596
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867   2247.8748
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465 -0.01036   2505.8369
#> 43536 713782.5 9558732 2281.50 32.74912  0.00840  0.00189    556.5211
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177    997.3089
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226  0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284  0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273  72123.6172
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399   1601.4852
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216   1412.2295
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645    234.0114
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060    383.3547
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635 -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577    423.5598
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.45216 -0.01196 -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011    501.7867
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249    148.1857
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806