Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041  0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041    772.9746
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264   5057.1431
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021 -0.09328   2235.5259
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629 360379.1875
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356  0.00774    527.1068
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156    816.6036
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270   4057.5637
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307 147470.5312
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373    527.3528
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330   1456.9246
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427    380.9472
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004
#> 20368 715252.5 9560622 1865.83 33.57246  0.00025 -0.01245   6353.3179
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739    789.7573
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384    947.2075
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159 -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002 -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419    761.5427
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715    520.7355
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132    737.5982
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171    756.8818
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216   1412.2295
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483   1097.3082
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662   1269.2650
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 20135 714772.5 9560232 2107.52 37.85074 -0.00782 -0.00048   1550.6704
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060    383.3547
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.73999  0.01354  0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172 111973.8906
#> 1434  714672.5 9559902 2194.92 42.45216 -0.01196 -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288  0.00448    941.8503
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769