Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921   -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056   -0.06976     301.2347
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479    0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753    0.01487    1077.7585
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041     772.9746
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373   -0.00057    1630.3259
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431    4079.0366
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891   -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453    0.00257     357.1436
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284    0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035    2697.5974
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226    2325.2053
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010    1527.1028
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160    9344.2930
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501    3889.7292
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312    2657.6936
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440    0.00710     843.7961
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043    0.00973     514.5488
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898     588.3761
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593     939.0466
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996    0.00626    1014.7159
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916    2009.1686
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646   -0.04086    6042.5757
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154    0.00124    2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566   -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294   -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846   -0.00034    4495.6929
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888    0.00862     874.5408
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752   -0.00517    1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169     350.1799
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149    0.01949    1182.1506
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359   -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290    0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964     571.1364
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530     937.1306
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405   -0.01565    2232.6133
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140   -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458   -0.00401    1722.6481
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090   -0.00299     688.0706
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126   -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554    0.03304    1569.7313
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.860876    0.00282    0.00109     767.4097
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240    1201.4724
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629   -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320    0.00180    3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448    1258.8196
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627   -0.05393    8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664    1807.2659
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188    0.00832    3022.3462
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335     195.0293
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278   -0.00412    1350.3611
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556   -0.11694  102301.4609
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195    1901.6743
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263    0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444   -0.00017     759.9937
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439   -0.01161    3065.3694
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441     424.2680
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970    1443.9200
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102   -0.01778    4594.4351
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605    0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133    2397.2681
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526     706.1463
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125   -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181   -0.00341    1363.4000
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358     802.7622
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675   -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144   -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312   -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893   -0.00867    2247.8748
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770   -0.03400    1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454    0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253     528.1924
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319    0.00291     487.3438
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205    1112.2626
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655     130.9468
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251   -0.00389   14516.2002
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382     100.0000
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325   -0.00505     351.4078
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165    0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226    0.00736     477.8591
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156   -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403     241.2991
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815   -0.03345  665381.5625
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075    8141.7549
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028    2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846    0.00794    1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502    1428.8546
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527   -0.02077    8759.9961
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001    0.00759    1880.4403
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601   -0.00011     712.5228
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706   -0.03035    1471.9783
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292    0.00462     716.5762
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586    0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948    1222.5433
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066    0.05257     549.3252
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070   -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972    0.00308     422.5022
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420    0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944    0.00836     250.5111
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044   -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273   -0.01047    2826.7700
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598    0.02208     842.4825
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452    0.00368    1071.8734
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108   -0.00948    1538.1741
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523     888.2811
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181   -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627   -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021    0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222     991.8300
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135    1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379   -0.00249     148.1857
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806