Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919   -0.04051 351155.1250
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347    0.02357   1752.0375
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498   -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514   -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333   -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677   1675.4679
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068    0.00588   1295.1610
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407    0.00547    519.4228
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160    0.00311    586.7354
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414   -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234    849.0547
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070   -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720    0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017   -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746    0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430   -0.06070 138869.3906
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382    0.00672   1114.8801
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151    0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426   -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856    0.01114   4510.8833
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733    598.7983
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490   -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643    0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108   -0.00678   1891.6584
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521   -0.00739    558.2286
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669   -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225    0.00325    616.8652
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626   1014.7159
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275    0.00605    798.6775
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765    0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440   -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743    0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154    0.00124   2114.2253
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566   -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224    0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547    0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606   -0.00556    660.3690
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229   -0.00511 361979.2812
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458    0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316    0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856    0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666   -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679    550.0344
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174    0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288    0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060    0.00270   4057.5637
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613    0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149    0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492    0.00072  54831.6289
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538    807.8237
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449   -0.00181   2406.1252
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917   -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504    0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405   -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017    0.00197    711.1091
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329   -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211    0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708    0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282    0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658    0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221    0.00121   4159.6738
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161   -0.00409   2312.8413
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529    0.02009   1247.7061
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126   -0.00485   2240.4185
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099    0.00240   1201.4724
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478   -0.07132  42970.7891
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771   -0.00599   1378.5760
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024   -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093    0.00664   1807.2659
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246    0.00534    344.1447
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491    0.01061   5744.7583
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521   -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890   -0.00911    173.2806
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310   -0.00860    695.2302
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666   -0.03583   1898.0105
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079    0.00321   1170.5904
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213   -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253    0.00753   1337.7854
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970    0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774   -0.00224   1137.3363
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153   -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693    0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106   -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010    0.00140   1066.3957
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212    0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902    0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118   -0.00252   3631.7017
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115    0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400   -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770   -0.03400   1982.7782
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983    0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010   -0.00870   2099.7930
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405   -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025   -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783   -0.00367   1966.2112
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953   -0.01797   7557.7866
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084   -0.02996   1599.3491
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159   -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196    0.00336   2806.2644
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002   -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009    0.00109    605.1908
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909   -0.01181   1286.3949
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162   -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156   -0.00116   1817.5222
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345   -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306   -0.00626    450.7346
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724    0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712    0.00923    811.5580
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818    0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263    0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167    0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987   -0.00507    504.3085
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196    0.01734    335.0773
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846    0.00216   1412.2295
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490   -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452   -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218    0.00902    218.7467
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185    0.01655    560.8554
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782   -0.00048   1550.6704
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342    0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066    0.05257    549.3252
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233    0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229    0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608    0.00148  10771.1963
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360    0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719    0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236   -0.00066    306.0146
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343    0.01177   1558.2231
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502    0.00282   1659.6211
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129    0.00491    261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293    0.00186   2751.4626
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253   -0.01233    968.7780
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014   -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595    0.03064   1485.5201
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986   -0.05013    616.4128
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933   -0.00157    766.0220
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110   -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250