Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438    753.6304
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976    301.2347
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917   0.00333    258.6537
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174   2310.4871
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345   1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088    266.4355
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102   0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526    539.8119
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382   0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889    362.8363
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593    939.0466
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598   0.01212    561.1948
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346    543.0749
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323    183.1836
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672    565.1107
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613    289.6113
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917  -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401   1722.6481
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102  35685.0000
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316    454.9674
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393   8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367    806.0966
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556  -0.04394    587.8840
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556  -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520   3390.5708
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123    364.2903
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298   1533.7925
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003   0.01196   1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401  -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133   2397.2681
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325    595.0426
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119   6589.5796
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000 162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384    947.2075
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400   1982.7782
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291    487.3438
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609   4275.3457
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105   4432.8760
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419    761.5427
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738   0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208    951.3401
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028   2394.7261
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069  -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362    283.7090
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170    253.5129
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152   0.00098    164.0802
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032    994.9224
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599  12873.8018
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820  -0.00260   1484.6593
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347   1893.3645
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215    917.5296
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948   1538.1741
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012    305.3624
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052    451.7209
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253   1237.0536
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448    941.8503
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235