Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659   3553.7166
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555    430.4604
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741    746.4155
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767   3321.6113
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292   1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403   1787.3024
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 125   713992.5 9561062 1878.28 49.971087  0.00183    0.03387   1989.3012
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160    0.00311    586.7354
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164    895.6606
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272   4955.5586
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530   9000.4356
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312   2657.6936
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051    0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314   2360.1501
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257    0.00593    939.0466
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083    994.7876
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325    616.8652
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996    0.00626   1014.7159
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323    183.1836
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063   5035.0000
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373  16031.7500
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063  18733.8848
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560    977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197    711.1091
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282    0.00542    860.8840
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057    0.02427    380.9472
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485   2240.4185
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406   -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596    842.0994
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851   1012.5963
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771   -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399   3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594  16587.9980
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453    0.01337    486.7551
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376   2048.3198
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038   -0.04212  46310.0195
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334   4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384    0.00175   1026.6818
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801   1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213   -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970   1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010   3467.2690
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778   4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317    0.00077   3056.9475
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200   1305.6691
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292   -0.00109   2047.9031
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185    953.8614
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226    338.3474
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015    790.2421
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642   1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559    866.5775
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232    977.7320
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162   -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403    241.2991
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493   2212.0525
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363    389.2691
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075   8141.7549
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923    811.5580
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171    756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794   1363.5321
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853   1354.6001
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216   1412.2295
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033   3446.5520
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035   1471.9783
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223   1056.3701
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462    716.5762
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048   1550.6704
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018    937.6911
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452   -0.00948   1222.5433
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257    549.3252
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099    909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360   4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308    422.5022
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578   1258.8865
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607   -0.04566  12862.4922
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805   7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278   -0.01172 111973.8906
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340    0.01809   1158.0426
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523    888.2811
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013    616.4128
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716   -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253   1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156    758.4840
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250