Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373    527.3528
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775