Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769