Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250