Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261    554.9243
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899   1394.2274
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284   2010.4285
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276   9614.6924
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903   2360.1536
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974   2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102    483.4184
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332   1418.6622
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542   3614.7334
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656   3036.7383
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196   1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957    495.4453
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224   1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133   2397.2681
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610   1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477   2945.2100
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973    215.5101
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018    937.6911
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260   1484.6593
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532  16771.3066
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052    451.7209
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388    180.4414
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336