Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261    323.8441
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333    258.6537
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324    861.3537
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403   1787.3024
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835   2425.3115
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378    440.9104
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769   1032.7177
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096   3210.5781
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593    939.0466
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672    565.1107
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452    437.3965
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409   2312.8413
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102  35685.0000
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596    842.0994
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393   8301.6094
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335    195.0293
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224   1137.3363
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875    597.2893
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341   1363.4000
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063   2430.2278
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118   2273.4758
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476  11992.5273
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216   1412.2295
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376    629.8236
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 33666 713632.5 9561012 1775.88 21.178812  -0.00950   -0.05130    107.5831
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165    805.1403
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960    884.2003
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849    764.8193
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148    481.8292
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156    758.4840
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946