Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041    772.9746
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579   2246.7725
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174   2310.4871
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345   1120.6929
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526    539.8119
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024    394.8751
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160   9344.2930
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367    793.9665
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083    994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325    616.8652
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077   3559.4204
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651    419.0927
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887    366.8537
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167   1938.5389
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373  16031.7500
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613    289.6113
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779   2972.0193
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236   2715.4138
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542   3614.7334
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026    247.0894
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102  35685.0000
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558   1504.9717
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478 -0.07132  42970.7891
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940    987.8085
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394    587.8840
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522   1355.7436
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011  0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705    413.5702
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279  39495.1797
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400   1982.7782
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036   2505.8369
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655    130.9468
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367   1966.2112
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609   4275.3457
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382    100.0000
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642   1688.1782
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476   1275.9446
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831    454.9113
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995   1026.7981
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973    215.5101
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476  11992.5273
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077   8759.9961
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654   1576.2065
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513    161.9878
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250   2225.4426
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547    128.2613
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247    282.1076
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964    543.0025
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282   1659.6211
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972    331.7462
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161  0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119    792.3544
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253   1237.0536
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389