Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603