Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501 -0.00649   1399.2329
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357   1752.0375
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345   1120.6929
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387   1989.3012
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426 -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605    798.6775
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077    300.7915
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063   5035.0000
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661  0.01199    196.4080
#> 26567 714022.5 9558872 2327.42 46.66569  0.00290  0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102    483.4184
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739    365.7626
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809   1102.9956
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793    864.9472
#> 10467 714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849   1120.5961
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334
#> 24723 712942.5 9560202 1983.18 33.07170  0.00285 -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143  0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384    947.2075
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291    487.3438
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158    837.3730
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284  0.01085   2229.5105
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007 -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409   1575.9425
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367   3002.8828
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216   1412.2295
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099   1146.2140
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645    234.0114
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257    549.3252
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250   2225.4426
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260   1484.6593
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172 111973.8906
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803   1058.5630
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288  0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491