Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029    5577.3916
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475    1001.0861
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896     276.0310
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579    2246.7725
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440    1088.8639
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035    2697.5974
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311     586.7354
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281    1325.6138
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665   25318.3555
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392    1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382     560.8309
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889     362.8363
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450     873.0448
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898     588.3761
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088    2322.1746
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950     950.3138
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276    9614.6924
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924    8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743   0.01313     783.4548
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034    4495.6929
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672     565.1107
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174   0.00025    4425.5151
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063   18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925     665.6116
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974    2136.6939
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197     711.1091
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319    3457.7519
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409    2312.8413
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195  -0.00656    3036.7383
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620     379.0855
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354    1147.7384
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556  -0.04394     587.8840
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245    6353.3179
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556  -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781   29995.1543
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236    9755.3027
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322    4605.1411
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439  -0.01161    3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123     364.2903
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169   0.01441     424.2680
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484     495.0292
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380     601.0843
#> 16071 714562.5 9560062 2166.69 28.696018    0.02281   0.04759     158.9370
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401  -0.00391    1794.7964
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171    4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094    1392.6008
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875     597.2893
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252    3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015     790.2421
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400    1982.7782
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345    2417.1389
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797    7557.7866
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505     351.4078
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831     454.9113
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328    1066.7134
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466    2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064     683.2795
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445     355.5078
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014     271.4219
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833     300.0125
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853    1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759    1880.4403
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514     897.9290
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422     967.9738
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292   0.00462     716.5762
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036     476.9504
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645     234.0114
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196     576.1450
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130     107.5831
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868     235.8511
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599   12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555    1882.0179
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964     543.0025
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952  -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852   0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347    1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532   16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576    1045.1074
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805    7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809    1158.0426
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585    5714.3267
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858     298.4490
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087  -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234    1707.3130
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181  -0.01149    4242.3579
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603