Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083     994.7876
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491