Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555    430.4604
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250   1413.3312
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104   1641.4587
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898    458.6938
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710    843.7961
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916   2009.1686
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416   1204.8629
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034   4495.6929
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949   1182.1506
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731   1384.7959
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.9235105 -0.01364 -0.03357    473.4709
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401   1722.6481
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512  13178.3223
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961   1086.6963
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219    420.7032
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627 -0.05393   8301.6094
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299  0.00141   4611.0718
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161   3065.3694
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511 -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196   1052.5341
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380    601.0843
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010  0.00140   1066.3957
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705    413.5702
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477   2945.2100
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152   1530.5657
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856   1789.1596
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609   4275.3457
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067   1123.7318
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476   1275.9446
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177    997.3089
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712   1010.5097
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983    270.7949
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518   1137.0179
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948   1222.5433
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552   2515.0752
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204    618.2027
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836    250.5111
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948   1538.1741
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012    305.3624
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109    915.4989
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491