Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983