Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203     634.3320
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292    1233.9299
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104   -0.02046    1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730   -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035    2697.5974
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102    0.00388    1051.4222
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466     724.0713
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842     868.6137
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530    9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518     854.4269
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430   -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398     267.2659
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257    0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071    0.00150    1274.3594
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275    0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642     613.6390
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606   -0.00556     660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904   -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063    5035.0000
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714     475.3120
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025    4425.5151
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102     483.4184
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117    1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855   -0.02175    1122.9120
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092     412.5849
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241     585.6182
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160    0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304    1569.7313
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664    1898.5588
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240    1201.4724
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224     942.0322
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771   -0.00599    1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354    1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813  287033.4688
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236    9755.3027
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017     759.9937
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801    1071.6141
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317    0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200    1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767   -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133    2397.2681
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705     413.5702
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157   -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610     454.8103
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000  162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020    0.00279   39495.1797
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189     556.5211
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655     130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000    0.00320     911.5911
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067    1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476    1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109     605.1908
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226    0.00736     477.8591
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875    1165.3953
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945     259.7421
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538   -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853    1354.6001
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084   -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601   -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196   -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740     909.4791
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148   10771.1963
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282    1659.6211
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491     261.7371
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233     968.7780
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124   25826.6914
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253    1237.0536
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607