Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029   5577.3916
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357   1752.0375
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483    566.7674
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041  0.01741    746.4155
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073  0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835   2425.3115
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838    238.6685
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452 -0.01437   1155.1469
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159
#> 26567 714022.5 9558872 2327.42 46.66569  0.00290  0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602    345.8933
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401   1722.6481
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739    365.7626
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454   3476.0818
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656   3036.7383
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395   2011.3270
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961   1086.6963
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555   4439.0967
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474   2343.2100
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246  0.00534    344.1447
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911    173.2806
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017    759.9937
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363    565.5207
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410 -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915   2036.0378
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610   1872.9276
#> 28260 715312.5 9558302 2681.75 34.48404 -0.01125 -0.02685   2568.4631
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118 -0.00252   3631.7017
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603   1562.0503
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109    605.1908
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132    737.5982
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445    355.5078
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333    422.8840
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011    201.1272
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577    423.5598
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491    261.7371
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884    672.2466
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250