Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603