Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309   -0.00681  -0.00029    5577.3916
#> 4965  715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634     500.5027
#> 28223 712742.5 9560482 1927.69 34.6679573   -0.00292   0.00712    3636.5303
#> 25090 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628     556.0121
#> 46178 714892.5 9559312 2379.54 32.7668833   -0.00266   0.02896     276.0310
#> 20799 714792.5 9561002 1857.29 23.1291603    0.00090   0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.1804123   -0.01059  -0.07431    4079.0366
#> 27973 715022.5 9559332 2308.28 20.0632631    0.00940   0.00679     550.0344
#> 32622 714972.5 9557762 2687.38 30.5564122   -0.00730  -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.2839150   -0.00284   0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.9691094   -0.00295   0.00035    2697.5974
#> 23817 714842.5 9557832 2677.58 26.6574344    0.00922   0.01088     266.4355
#> 24018 714292.5 9559282 2334.12 32.8654321    0.00102   0.00388    1051.4222
#> 42598 713542.5 9559972 2184.39 35.2414881   -0.00841  -0.00010    1527.1028
#> 30800 715182.5 9557582 2772.37 39.2040642   -0.02475   0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732    0.03696   0.02514     297.7842
#> 18304 714582.5 9558932 2463.25 22.1946661   -0.02131   0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032     994.6740
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746   0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126  -0.02044  129412.3984
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203   10263.2734
#> 26215 714332.5 9560262 2138.08 25.7212213   -0.00491   0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713  -0.00173    8967.1924
#> 32821 715022.5 9559952 2208.33 43.4164499   -0.00643   0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.5289733   -0.03613  -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.5428498    0.03206   0.02773     179.3077
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598   0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937   0.00083     994.7876
#> 1635  713102.5 9559052 2172.87 31.4840308   -0.00064  -0.00916    2009.1686
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028   0.01318     289.0063
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545   0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.5927252    0.00804   0.00346     543.0749
#> 5728  713502.5 9559662 2140.27 34.3075669   -0.00154   0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887     366.8537
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547   0.03323     183.1836
#> 4605  714852.5 9557892 2677.78 26.9336000    0.00321   0.01169     350.1799
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063   18733.8848
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903    2360.1536
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585  -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.6656935    0.00290   0.00230    1056.5876
#> 23029 713502.5 9559102 2174.07 56.4575423    0.03395   0.01495     266.5888
#> 33240 714972.5 9557652 2741.04 39.4343932   -0.00286  -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749   0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283   0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626    0.02409   0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.7968588   -0.00449  -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399  -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.7531604   -0.00119  -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.1320321    0.00002   0.00588     749.2502
#> 15766 715562.5 9558592 2646.16 29.0208853    0.00194   0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260  -0.03310    2833.9001
#> 19887 714662.5 9559632 2266.47 45.1794410   -0.02797  -0.00063     576.2819
#> 32909 712812.5 9560452 1883.85 38.2512354   -0.00221   0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.3468842   -0.00161  -0.00409    2312.8413
#> 13029 713542.5 9560242 2019.80 29.5451417    0.00576  -0.00166     650.6246
#> 47100 712752.5 9560462 1910.00 41.9118627   -0.00032   0.00542    3614.7334
#> 37416 713282.5 9560692 1850.61 30.2590471    0.00736   0.01684    2556.8049
#> 7668  713372.5 9559062 2164.82 37.9349627   -0.01807  -0.01883     616.1993
#> 23213 715592.5 9558562 2654.69 33.2195200   -0.00034   0.00294     739.5768
#> 46783 713592.5 9560762 1832.60 36.0103974    0.00448  -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170   -0.02554   0.03304    1569.7313
#> 31561 713302.5 9559192 2092.11 36.5277783    0.00406  -0.00055     773.0359
#> 27695 714142.5 9558452 2449.28 20.4901167   -0.00539   0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.0591409   -0.00410   0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.2791006    0.00645  -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.6244125   -0.01972   0.03512   13178.3223
#> 20601 715542.5 9558182 2773.29 32.7554242   -0.00604   0.00354    1147.7384
#> 41721 715182.5 9557562 2784.85 31.7899903   -0.00588   0.01448    1258.8196
#> 40216 714602.5 9560592 2015.55 34.3688733   -0.00093   0.00664    1807.2659
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133   0.00367     806.0966
#> 12772 715172.5 9557642 2759.42 34.8054672   -0.00567  -0.03793     864.9472
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370  -0.00190     640.8842
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278  -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.9778994    0.01453   0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.2237194   -0.03986  -0.02524   10861.9482
#> 10933 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376    2048.3198
#> 18440 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781   29995.1543
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521  -0.00870    1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890  -0.00911     173.2806
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084  -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.0207457   -0.01878  -0.04351    1442.9247
#> 46138 715202.5 9559582 2194.40 18.8600517   -0.02143  -0.05078  622888.2500
#> 30933 714672.5 9561082 1798.37 54.8102886    0.00542  -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.9813582    0.05379   0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.5275756   -0.00193   0.02812    1423.7578
#> 47244 714262.5 9559302 2350.47 17.0592454    0.02213  -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.8039510   -0.00442  -0.00118   19557.6621
#> 37439 714602.5 9560672 1982.99 18.9064613   -0.00070   0.00010    3467.2690
#> 10617 713232.5 9560722 1879.95 26.4970062    0.00258   0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003   0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.6893384   -0.01102  -0.01778    4594.4351
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605   0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.0463191   -0.08153  -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.5663341   -0.00693   0.01133    2397.2681
#> 44612 714932.5 9558822 2578.80 38.2552461    0.01257   0.00044     288.5468
#> 22847 714582.5 9560382 2040.49 25.3361937    0.00181  -0.00341    1363.4000
#> 30626 713302.5 9559172 2106.91 35.9977924    0.01354   0.00245     629.5983
#> 17674 713882.5 9558562 2339.20 41.0363832   -0.03675  -0.01905    6487.0010
#> 14779 712502.5 9560202 2004.61 27.5226006    0.00115   0.01015     790.2421
#> 30378 714852.5 9557752 2651.70 41.2982249   -0.00010  -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.8077801   -0.00486   0.02306    1718.3652
#> 34718 714802.5 9558802 2599.22 44.6500280    0.01996   0.04684     209.1778
#> 39466 714942.5 9557772 2680.35 16.7532859    0.00166  -0.01265    1148.6487
#> 13072 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781   29995.1543
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196   0.00336    2806.2644
#> 25212 714272.5 9559012 2433.45 37.6714021   -0.01909  -0.01181    1286.3949
#> 15558 714732.5 9559502 2283.66 53.4317521    0.00390  -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124   0.00204     853.2902
#> 20067 714242.5 9560442 2125.23 38.6402737    0.00363  -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.6350750   -0.00242   0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.7659191   -0.00458   0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.7174203   -0.00494   0.01064     683.2795
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306  -0.00626     450.7346
#> 25140 713332.5 9560692 1809.95 52.2571887    0.00075  -0.01945     259.7421
#> 13512 715152.5 9557992 2778.81 43.3895209   -0.00724   0.01464    4029.5947
#> 43243 715682.5 9558462 2693.23 27.4062902   -0.00031  -0.00199     836.4196
#> 47005 714942.5 9559192 2333.66 29.2208476   -0.04937  -0.02273   72123.6172
#> 44396 715512.5 9558102 2845.31 26.5984197    0.01263   0.03598     186.3527
#> 18627 714292.5 9559252 2330.78 18.4962235   -0.00538  -0.00052    3508.5591
#> 36126 715052.5 9558292 2759.47 29.0621382    0.00196   0.01734     335.0773
#> 20135 714772.5 9560232 2107.52 37.8507379   -0.00782  -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.2147477   -0.01338   0.00518    1137.0179
#> 12044 713612.5 9559292 2247.76 23.0747293   -0.00112  -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972   -0.03004   0.00714  480535.5000
#> 12519 713982.5 9560912 1956.98 28.8249337   -0.00452  -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.0930780   -0.00305  -0.00705     968.9738
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118   0.00162     290.2031
#> 35038 715092.5 9558152 2715.69 40.1838220    0.00270  -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.3936922    0.03479   0.02251     168.1733
#> 36889 714302.5 9559802 2243.60 21.1662069    0.01923   0.05547     128.2613
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229   0.00099     909.0560
#> 34532 713812.5 9559502 2328.80 44.9376528   -0.00070  -0.00060     383.3547
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479    0.00972   0.00308     422.5022
#> 3868  714542.5 9560622 2024.15 29.4471659    0.00360   0.00750    1892.8978
#> 48184 714202.5 9558222 2489.17 36.1135298    0.02236  -0.00066     306.0146
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757  -0.00003 5426606.0000
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.7737308    0.00452   0.00368    1071.8734
#> 1434  714672.5 9559902 2194.92 42.4521619   -0.01196  -0.01935     903.6397
#> 23808 715172.5 9558702 2533.50 22.3459270    0.00811  -0.02281     492.4785
#> 126   715042.5 9561262 1733.49 30.1106510   -0.00288   0.00448     941.8503
#> 13548 715812.5 9558122 2821.65 15.5082486    0.02751   0.00388     180.4414
#> 28978 713982.5 9560462 2133.10  5.2293858    0.00366   0.00525     162.0592
#> 6206  714122.5 9560992 1917.28 37.9699131   -0.01023  -0.00227     841.2025
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181  -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.6213801    0.00208   0.00492     772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769