Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029    5577.3916
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393    2081.0056
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555     430.4604
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339    1002.5877
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644    1179.0695
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491    1844.4353
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466     724.0713
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104    1641.4587
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392    1255.0884
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518     854.4269
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189    2317.1819
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314    2360.1501
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088    2322.1746
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357     275.7474
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605     798.6775
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034    4495.6929
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452     437.3965
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270    4057.5637
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063   18733.8848
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102     483.4184
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530     937.1306
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588     749.2502
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427     380.9472
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166     650.6246
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542    3614.7334
#> 37416 713282.5 9560692 1850.61 30.2590471  0.00736  0.01684    2556.8049
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262    1760.2162
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941    1290.7709
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656    3036.7383
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596     842.0994
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940     987.8085
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822     472.8748
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017  0.01213     328.7922
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337     486.7551
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241    1752.9242
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281    1178.4420
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322    4605.1411
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123     364.2903
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011  0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753    1337.7854
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981    3975.3608
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077    3056.9475
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015     790.2421
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265    1148.6487
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842     246.5161
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559     866.5775
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831     454.9113
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493    2212.0525
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132     737.5982
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064     683.2795
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650    1133.1501
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464    4029.5947
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273   72123.6172
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973     215.5101
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943     371.5514
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983     270.7949
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099    1146.2140
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759    4131.5356
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705     968.9738
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257     549.3252
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282    1659.6211
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972     331.7462
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368    1071.8734
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948    1538.1741
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177     184.1042
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169     649.9975
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253    1237.0536
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448     941.8503
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135    1293.4855
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388     180.4414
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249     148.1857
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931    1812.3820
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983