Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 4965  715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027 33.905923
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642    671.1807 41.600174
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021 -0.01579   1318.7560 36.259061
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121 43.531614
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861 39.335272
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113 30.966650
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073  0.00174   2310.4871 34.082394
#> 43327 713962.5 9561082 1839.11 63.36168 -0.04951 -0.00529   3282.1186 29.058700
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974 23.540544
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147 44.795559
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228 35.252374
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088    266.4355 21.611395
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354 20.133737
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104 24.774122
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269    491.8017 35.138929
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425 22.269151
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547 35.780068
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833 25.321297
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930 23.454600
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842 38.961703
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586  8.761098
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326 19.942369
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518 29.513629
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177 21.398255
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356 28.852436
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243 24.905902
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906 37.628430
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398    267.2659 24.586765
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488 36.829154
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924 28.565957
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501 29.483262
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746 33.621163
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422  0.02357    275.7474 18.834269
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838 27.189139
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063 15.648623
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077    300.7915 14.659698
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570 28.960152
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647 23.193905
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817 29.935899
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063   5035.0000 19.231901
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408 27.955757
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107 30.108359
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151 29.071306
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092 34.789424
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679 27.054494
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116 31.451945
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506 45.524362
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779   2972.0193 27.702509
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237 42.761559
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117   1018.2141 33.134722
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959 36.650391
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719 36.732897
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316 33.773570
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332   1418.6622 34.900005
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519 30.439529
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738 28.379173
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436   2602.7390 28.952703
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166    650.6246 34.187819
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150 33.241865
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000 35.453482
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674 28.246819
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063 15.648623
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555   4439.0967 25.229624
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474   2343.2100 27.588491
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085 29.818443
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840 44.596170
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611 29.738228
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551 24.941426
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198 34.677125
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004 25.761328
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583 32.323414
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399 26.645975
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728   1137.2450 41.010600
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302 36.555853
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002 42.773591
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261   1627.8989 33.574754
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454 33.813677
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491 30.648085
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876 26.322827
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608 45.093497
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410 -0.00380    601.0843 33.646374
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404 25.746431
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475 23.160673
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756    353.9086 46.649651
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952 33.994159
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334 31.517262
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249 26.778328
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908 25.370571
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027 33.905923
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008 34.424450
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384 21.185114
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893 21.668691
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841 16.579107
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152   1530.5657 29.730780
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732   4237.6934 34.909745
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622 22.717777
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010 30.482501
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210    585.8992 35.487287
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547 35.780068
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465 -0.01036   2505.8369 26.906098
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454 33.813677
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491 41.558348
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590 20.891760
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002 -0.01232    977.7320 25.804873
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078 14.818407
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494 36.186868
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902 24.350706
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991 32.295339
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677 21.429767
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666 42.063124
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445    355.5078 18.572427
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209 24.344977
#> 9665  713362.5 9558812 2275.28 31.49492  0.01306 -0.00626    450.7346 34.081249
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171    756.8818 31.480020
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794   1363.5321 32.361229
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273  72123.6172 36.135302
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084 25.838678
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403 33.143317
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661 37.338514
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090 18.628004
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129 33.722577
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218  0.00902    218.7467 38.595583
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400   4126.7812 25.853575
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655    560.8554 24.875536
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065 36.122124
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656 21.265329
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018 22.878205
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060    383.3547 31.206719
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248 25.587149
#> 1281  715102.5 9560032 2120.64 34.64905 -0.00930 -0.00360   4288.7329 35.393895
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022 26.907817
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635 -0.00105    868.7802 43.676573
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978 26.147502
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146 30.956337
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491    261.7371 18.619409
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292 24.257314
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741 40.603800
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780 25.089822
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426 22.852422
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055 31.814627
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672 29.121153
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606 38.910710
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128 42.407471
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458    485.7418 16.607182
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300 29.050106
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785 37.533892
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312 22.237065
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536 31.394649
#>       distroad slides distdeforest distslidespast log.carea
#> 4965    300.00   TRUE       300.00            100  2.699406
#> 23178   300.00   TRUE       300.00             21  2.826839
#> 47223   300.00   TRUE       300.00            100  3.120164
#> 25090   300.00   TRUE       300.00             26  2.745084
#> 40756   300.00   TRUE       300.00            100  3.000471
#> 48187   279.54   TRUE        21.35             46  3.521349
#> 44952   300.00   TRUE       300.00            100  3.363704
#> 43327   173.40   TRUE        33.81            100  3.516154
#> 49372   213.54   TRUE        20.21             68  3.430977
#> 20151   300.00   TRUE       300.00            100  2.828669
#> 16811   300.00   TRUE       300.00             59  2.715521
#> 23817   300.00   TRUE       300.00              4  2.425592
#> 11463   300.00   TRUE       300.00            100  2.768442
#> 10899   300.00   TRUE       300.00             20  2.644350
#> 3835    300.00   TRUE       300.00             39  2.691790
#> 21703   300.00   TRUE       300.00             34  2.458702
#> 8019    300.00   TRUE       300.00             61  2.928936
#> 2990    300.00   TRUE       123.56              5  2.520462
#> 45869   300.00   TRUE       300.00             12  3.970546
#> 43007   300.00   TRUE       300.00             37  2.473902
#> 38410    69.52   TRUE        47.61             65  3.695093
#> 22283   256.35   TRUE         0.00            100  2.772566
#> 40307    77.82   TRUE        76.41             54  3.524078
#> 16843   300.00   TRUE       300.00             12  3.013982
#> 29420   253.02   TRUE         0.00            100  3.954264
#> 18005   300.00   TRUE       300.00              0  2.744234
#> 38279   300.00   TRUE       300.00             24  5.142607
#> 23375   300.00   TRUE       195.00              2  3.224136
#> 22566   300.00   TRUE         7.49              0  2.426944
#> 34275   300.00   TRUE       300.00            100  2.711427
#> 6808    300.00   TRUE       300.00             40  3.952656
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 34280   300.00   TRUE         0.00            100  3.365895
#> 3133    300.00   TRUE       300.00            100  2.440511
#> 17004   300.00   TRUE       300.00            100  3.257531
#> 34642   300.00   TRUE       300.00              2  2.460907
#> 19128   300.00   TRUE       300.00             15  2.478266
#> 40738   300.00   TRUE       300.00            100  2.716468
#> 11687   300.00   TRUE       300.00             68  3.186466
#> 48434   300.00   TRUE       300.00              6  2.847438
#> 17594   138.35   TRUE         0.00             60  3.701999
#> 19784   300.00   TRUE       300.00            100  2.941780
#> 46235   300.00   TRUE        41.37              0  2.752133
#> 46171   300.00   TRUE       270.22            100  3.645964
#> 13699   300.00   TRUE        53.95             43  3.288564
#> 47884   300.00   TRUE       300.00            100  2.905563
#> 37341   300.00   TRUE       300.00              1  2.823221
#> 651     300.00   TRUE       300.00            100  3.072673
#> 34707   165.13   TRUE        39.56            100  3.473052
#> 4031    300.00   TRUE       300.00             31  2.907317
#> 39599   300.00   TRUE       300.00             25  3.007839
#> 835     300.00   TRUE       300.00            100  3.141386
#> 2718    300.00   TRUE       300.00            100  2.990104
#> 48808   300.00   TRUE       300.00            100  4.066528
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 2595    300.00   TRUE       300.00            100  3.538794
#> 32909   118.07   TRUE        98.15              0  3.619059
#> 33350    10.00   TRUE       166.26              2  3.415431
#> 13029   300.00   TRUE        96.65              4  2.813330
#> 29801   300.00   TRUE       300.00              2  3.026907
#> 3418    300.00   TRUE       300.00             73  4.552486
#> 16355   300.00   TRUE        58.52            100  2.657980
#> 45592   300.00   TRUE       300.00              2  2.460907
#> 26010    69.52   TRUE       166.56              2  3.647295
#> 12137   300.00   TRUE       300.00            100  3.369811
#> 39166   300.00   TRUE         1.90             90  2.994673
#> 6845    300.00   TRUE       300.00             29  2.769292
#> 32879   300.00   TRUE       300.00              6  3.130450
#> 46461   288.47   TRUE         0.00            100  2.687311
#> 10933   264.46   TRUE        18.17             56  3.311398
#> 10073   300.00   TRUE       300.00            100  2.896526
#> 47718   300.00   TRUE        10.00              1  3.759272
#> 31117   300.00   TRUE       291.23            100  3.277641
#> 18348   300.00   TRUE       300.00            100  3.055854
#> 41125   300.00   TRUE         0.00            100  2.842129
#> 37720   300.00   TRUE       300.00            100  3.690657
#> 27915   300.00   TRUE       300.00              6  3.211627
#> 22907   300.00   TRUE       205.28              1  3.147505
#> 10689   300.00   TRUE       300.00             61  2.237416
#> 17058   300.00   TRUE       300.00              8  2.751808
#> 10363   300.00   TRUE       297.43              6  3.599377
#> 10692   300.00   TRUE        60.19             11  2.778935
#> 2801    300.00   TRUE       300.00             26  2.422656
#> 13259   203.85   TRUE        10.45             74  3.485288
#> 22297   300.00   TRUE       300.00             11  2.548891
#> 49378    22.68   TRUE       134.63             25  3.073643
#> 4667     24.78   TRUE       172.24             38  3.119102
#> 23626   300.00   TRUE       300.00            100  3.336505
#> 7409    300.00   TRUE       300.00            100  2.866282
#> 34984   300.00   TRUE       300.00            100  2.699406
#> 48574    96.21   TRUE       101.32             10  3.143827
#> 14494   300.00   TRUE       300.00             28  2.535469
#> 10456    24.34   TRUE         0.00             18  2.776185
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 46038   300.00   TRUE       243.86            100  3.184852
#> 42640   300.00   TRUE       300.00            100  3.627130
#> 18498   300.00   TRUE       300.00             83  2.904587
#> 17674   300.00   TRUE       300.00              1  3.812044
#> 14140   300.00   TRUE       300.00             65  2.767823
#> 48461   300.00   TRUE       300.00             61  2.928936
#> 33142   300.00   TRUE       300.00            100  3.398953
#> 41408   300.00   TRUE       205.28              1  3.147505
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 23176   300.00   TRUE       300.00            100  3.203943
#> 29431   300.00  FALSE       300.00             42  2.702050
#> 36561   300.00  FALSE        70.00            100  2.990220
#> 46165   300.00  FALSE       300.00            100  2.545811
#> 40459   300.00  FALSE       216.66             10  3.025531
#> 18286   300.00  FALSE       300.00             42  2.931097
#> 27965   300.00  FALSE       300.00            100  2.382556
#> 20458   300.00  FALSE       300.00             28  2.424012
#> 26963   300.00  FALSE       300.00            100  3.420610
#> 19493   300.00  FALSE       291.23            100  2.550849
#> 12716   300.00  FALSE       300.00             25  2.731444
#> 9665    300.00  FALSE       300.00             64  2.653921
#> 39193   300.00  FALSE       117.12            100  2.879028
#> 49835   300.00  FALSE         0.00            100  3.134665
#> 47005   300.00  FALSE       300.00            100  4.858077
#> 13286   300.00  FALSE       300.00             76  2.499561
#> 40139   300.00  FALSE         1.11             21  3.274260
#> 12974   300.00  FALSE       300.00            100  2.711947
#> 29944   300.00  FALSE       300.00             48  2.452873
#> 31201   300.00  FALSE       300.00            100  2.404000
#> 33956   300.00  FALSE       300.00            100  2.339942
#> 36245   110.19  FALSE        81.50             20  3.615611
#> 48869   138.98  FALSE         0.00             63  2.748851
#> 26389   300.00  FALSE        76.38             26  3.197613
#> 43222   300.00  FALSE       300.00            100  2.404946
#> 40909   234.82  FALSE         0.00             43  4.109707
#> 34532   300.00  FALSE       300.00             92  2.583601
#> 41612   300.00  FALSE       300.00            100  3.004762
#> 1281    300.00  FALSE       300.00            100  3.632329
#> 15008   300.00  FALSE       300.00             48  2.625829
#> 1479    300.00  FALSE       300.00            100  2.938910
#> 3868    300.00  FALSE       300.00            100  3.277127
#> 48184   300.00  FALSE       300.00             75  2.485742
#> 10127   300.00  FALSE       300.00             49  2.417865
#> 37783   300.00  FALSE       300.00             18  2.682893
#> 11394    46.04  FALSE        55.00            100  3.187005
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 43972   300.00  FALSE       300.00            100  3.063725
#> 28526   300.00  FALSE        46.28            100  3.237445
#> 22628   300.00  FALSE       300.00            100  2.951467
#> 46148   300.00  FALSE       300.00            100  3.015637
#> 6577    300.00  FALSE       300.00             70  2.789872
#> 11570   300.00  FALSE       300.00            100  2.686405
#> 24962   300.00  FALSE       300.00            100  2.996437
#> 23808   300.00  FALSE       300.00             72  2.692387
#> 42990   205.00  FALSE         0.00            100  3.173195
#> 26635   300.00  FALSE       300.00            100  3.092389