Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043    1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292    1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073    0.00174    2310.4871
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529    3282.1186
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183    0.03387    1989.3012
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416     947.4878
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104    1641.4587
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665   25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160    9344.2930
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905   -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382    0.00672    1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856    0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581    1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973     514.5488
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108   -0.00678    1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629  360379.1875
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694    2123.4473
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252    0.01118    2359.5059
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051     779.0555
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646   -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346     543.0749
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365    1536.2647
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063    5035.0000
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169     350.1799
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117    1018.2141
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565    2232.6133
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550   -0.00941    1290.7709
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656    3036.7383
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809    1102.9956
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596     842.0994
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474    2343.2100
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188    0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190     640.8842
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524   10861.9482
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261    1627.8989
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193    0.02812    1423.7578
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739     789.7573
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422    1159.2100
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317    0.00077    3056.9475
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767   -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121    1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526     706.1463
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341    1363.4000
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252    3631.7017
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867    2247.8748
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063    2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486    0.02306    1718.3652
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367    1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519    0.00609    4275.3457
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842     246.5161
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109     605.1908
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226    0.00736     477.8591
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085    2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116    1817.5222
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363     389.2691
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650    1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309   -0.03409    1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548     532.8210
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333     422.8840
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833     300.0125
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538   -0.00052    3508.5591
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236    0.00483    1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780     825.4451
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048    1550.6704
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312     593.7773
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462    0.00268     252.0072
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740     909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980   -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577     423.5598
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511    0.01499     617.3855
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875    2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105   -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598    0.02208     842.4825
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886   43249.3945
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340    0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709    1181.2555
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124   25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253    1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227     841.2025
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901