Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983