Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491