Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea   cslope
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375 23.81728
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606 20.47808
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552 34.11906
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585 28.28349
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310 20.93129
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537 24.43780
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679 34.63988
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191 23.93245
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639 29.11027
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028 20.06899
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353 30.17138
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228 35.25237
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355 21.61140
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878 34.91204
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713 38.77034
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259 14.89690
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425 22.26915
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833 25.32130
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518 29.51363
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240 33.82571
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309 40.73558
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243 24.90590
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679 34.63988
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801 28.39865
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274 27.97524
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622 22.71778
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746 33.62116
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652 30.06023
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469 20.57090
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555 28.20385
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204 34.56196
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757 34.39122
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749 34.25199
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016 39.28772
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817 29.93590
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812 34.29783
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198 34.67712
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766 41.11946
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344 30.00981
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679 27.05449
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362 37.64390
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536 28.61065
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289 39.36048
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364 29.35492
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502 10.71374
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133 41.84941
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246 28.79056
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536 43.10075
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413 29.16814
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334 29.64598
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791 33.75466
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150 33.24187
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768 25.46053
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894 29.02146
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709 37.36258
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359 31.44278
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549 38.16014
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717 24.96320
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674 28.24682
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855 33.98556
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963 32.21570
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002 28.18609
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445 28.34823
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840 44.59617
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293 19.32587
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688 33.18743
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611 29.73823
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961 33.41318
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818 34.18954
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105 36.63378
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621 41.64257
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454 33.81368
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147 12.84342
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680 39.49054
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854 25.33619
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292 30.11638
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453 23.51934
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957 27.05220
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231 43.59750
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010 30.48250
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992 35.48729
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875 31.93151
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924 20.82071
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369 26.90610
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626 17.29530
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613 25.48516
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911 28.09327
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866 38.66835
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398 29.40477
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676 15.76207
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775 38.37385
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908 15.49736
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372 38.54860
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097 35.04611
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991 32.29534
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981 36.44642
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425 18.66983
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947 35.93477
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818 31.48002
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261 23.44772
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840 23.94505
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172 36.13530
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546 25.38891
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527 21.44237
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961 28.07550
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403 33.14332
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185 32.55546
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347 24.13757
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082 39.88989
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355 25.00732
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290 36.08431
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451 28.89312
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129 33.72258
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738 28.75732
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272 38.49703
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554 24.87554
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504 40.96018
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179 22.10013
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475 25.23077
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944 33.42177
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252 39.77129
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878 26.66832
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462 28.70175
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027 22.60720
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366 36.10723
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236 32.28216
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802 43.67657
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218 32.69354
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076 25.72695
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978 26.14750
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855 17.04836
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111 22.86044
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146 30.95634
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700 36.31979
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865 22.75846
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448 32.45118
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019 32.00313
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346 26.06500
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825 31.36314
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780 25.08982
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544 27.56958
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201 29.18590
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490 17.04435
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630 33.35818
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606 38.91071
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220 36.13072
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312 22.23706
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861 36.81999
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592  4.43985
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857 33.48595
#>       distroad slides distdeforest distslidespast log.carea
#> 31357   158.92   TRUE         0.00              5  3.243543
#> 42632   300.00   TRUE       300.00             10  2.566509
#> 13748   300.00   TRUE       300.00              2  3.256827
#> 29391   300.00   TRUE         0.56            100  3.032521
#> 46178   300.00   TRUE       300.00             31  2.440958
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 34512   300.00   TRUE       195.00              2  3.224136
#> 21939   300.00   TRUE       300.00            100  3.168886
#> 8869    300.00   TRUE       300.00            100  3.036974
#> 14456   300.00   TRUE       247.02            100  3.183868
#> 32622   300.00   TRUE       300.00            100  3.265863
#> 16811   300.00   TRUE       300.00             59  2.715521
#> 23817   300.00   TRUE       300.00              4  2.425592
#> 22814   300.00   TRUE       300.00              6  2.976574
#> 27375   300.00   TRUE       300.00            100  2.859781
#> 38287    41.43   TRUE         1.90            100  3.349380
#> 21703   300.00   TRUE       300.00             34  2.458702
#> 2990    300.00   TRUE       123.56              5  2.520462
#> 40307    77.82   TRUE        76.41             54  3.524078
#> 48055   300.00   TRUE         1.67              0  3.566570
#> 30125   300.00   TRUE       300.00            100  2.748832
#> 18005   300.00   TRUE       300.00              0  2.744234
#> 23375   300.00   TRUE       195.00              2  3.224136
#> 14522   300.00   TRUE       300.00            100  3.047228
#> 15340   300.00   TRUE       300.00            100  3.144334
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 34280   300.00   TRUE         0.00            100  3.365895
#> 27957   300.00   TRUE       300.00             75  2.790190
#> 13847   300.00   TRUE       300.00            100  3.062637
#> 2950    300.00   TRUE       300.00            100  2.891568
#> 4329    300.00   TRUE        16.35              2  3.551379
#> 28577   300.00   TRUE         0.56              0  3.781222
#> 29667   300.00   TRUE       300.00             27  2.734860
#> 37348   300.00   TRUE       300.00            100  4.465317
#> 48434   300.00   TRUE       300.00              6  2.847438
#> 24100   300.00   TRUE       300.00            100  5.558684
#> 20001   264.46   TRUE        18.17             56  3.311398
#> 6967    300.00   TRUE       300.00             85  3.008718
#> 31316   300.00   TRUE       300.00             85  2.740390
#> 47884   300.00   TRUE       300.00            100  2.905563
#> 13134   300.00   TRUE       300.00             67  2.634917
#> 18394   300.00   TRUE       300.00             57  3.372940
#> 3944    300.00   TRUE       300.00            100  4.739031
#> 33240   300.00   TRUE       300.00            100  2.756740
#> 421     182.02   TRUE         0.00             83  2.874627
#> 49569    25.00   TRUE        35.00            100  3.348814
#> 7466    255.23   TRUE         0.00             60  3.163437
#> 30406   300.00   TRUE       300.00             85  2.770965
#> 39732   300.00   TRUE       300.00             25  3.364146
#> 47100    82.80   TRUE       129.73             36  3.558076
#> 31462   300.00   TRUE       300.00              1  3.466909
#> 29801   300.00   TRUE       300.00              2  3.026907
#> 23213   300.00   TRUE       300.00            100  2.868983
#> 44825   300.00   TRUE       300.00            100  2.392854
#> 44650   300.00   TRUE       300.00            100  3.110849
#> 31561   300.00   TRUE       300.00              0  2.888200
#> 31121   300.00   TRUE       300.00             45  3.194113
#> 21013   300.00   TRUE        13.47             64  3.177528
#> 16355   300.00   TRUE        58.52            100  2.657980
#> 11618   300.00   TRUE       300.00              1  2.578737
#> 2552    300.00   TRUE       300.00            100  3.005436
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 21680   104.12   TRUE       110.76              4  3.591103
#> 6845    300.00   TRUE       300.00             29  2.769292
#> 1973    300.00   TRUE       300.00            100  2.290100
#> 8125    300.00   TRUE       300.00             48  5.457933
#> 32879   300.00   TRUE       300.00              6  3.130450
#> 16009   300.00   TRUE       300.00            100  3.049449
#> 34615    17.57   TRUE       142.95             16  3.011436
#> 19832   300.00   TRUE       300.00            100  3.278299
#> 28312   300.00   TRUE       300.00            100  3.198702
#> 22907   300.00   TRUE       205.28              1  3.147505
#> 30933    99.29   TRUE         0.00              6  3.137802
#> 7485    300.00   TRUE       300.00            100  2.627640
#> 47298   300.00   TRUE       300.00            100  3.126386
#> 19756   300.00   TRUE       300.00            100  2.694631
#> 49133   300.00   TRUE       300.00             38  2.694996
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 17227   300.00   TRUE       300.00            100  2.467494
#> 17674   300.00   TRUE       300.00              1  3.812044
#> 14140   300.00   TRUE       300.00             65  2.767823
#> 6212    300.00   TRUE       300.00             47  5.501416
#> 9864    300.00   TRUE       300.00            100  2.722792
#> 33142   300.00   TRUE       300.00            100  3.398953
#> 41310   300.00   TRUE       300.00             25  3.046207
#> 23832   300.00   TRUE       300.00              8  2.676567
#> 6470    300.00   TRUE       300.00              2  2.959800
#> 31253   300.00   TRUE       122.87            100  3.878395
#> 36744   300.00   TRUE       300.00            100  2.937989
#> 32405   300.00  FALSE       300.00            100  2.530158
#> 16208   256.60  FALSE         0.00             67  2.937807
#> 27884   300.00  FALSE       300.00            100  2.781892
#> 15186    60.00  FALSE         0.00            100  3.149477
#> 16464   300.00  FALSE       300.00            100  3.004540
#> 27965   300.00  FALSE       300.00            100  2.382556
#> 37832   300.00  FALSE       207.18              2  3.011485
#> 28162   130.00  FALSE        75.00            100  3.197540
#> 13512   300.00  FALSE       300.00            100  3.605261
#> 39193   300.00  FALSE       117.12            100  2.879028
#> 45166   300.00  FALSE       300.00              1  3.379256
#> 29125   300.00  FALSE       300.00            100  2.626221
#> 47005   300.00  FALSE       300.00            100  4.858077
#> 16878   300.00  FALSE       300.00             18  3.154988
#> 44396   300.00  FALSE       300.00            100  2.270336
#> 36833   187.46  FALSE        18.49             45  3.942504
#> 40139   300.00  FALSE         1.11             21  3.274260
#> 8831    300.00  FALSE       300.00             24  2.615022
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 8413     80.73  FALSE         0.00            100  3.040329
#> 41753   300.00  FALSE       300.00            100  2.860419
#> 38689   300.00  FALSE       300.00             18  2.953242
#> 37029   300.00  FALSE       237.21            100  2.916688
#> 31201   300.00  FALSE       300.00            100  2.404000
#> 6143    300.00  FALSE       185.58            100  2.985864
#> 24009   300.00  FALSE       300.00            100  2.303471
#> 48869   138.98  FALSE         0.00             63  2.748851
#> 13737   300.00  FALSE       300.00            100  2.678473
#> 3733    300.00  FALSE       300.00            100  3.055767
#> 29107   300.00  FALSE       300.00            100  2.550653
#> 45486   300.00  FALSE         0.00             68  3.366011
#> 7553    300.00  FALSE       300.00            100  2.739830
#> 36721   300.00  FALSE       300.00            100  2.209482
#> 41855   234.40  FALSE        47.05            100  3.040739
#> 42424   275.21  FALSE        93.50             70  2.791131
#> 44823   300.00  FALSE       300.00            100  2.749146
#> 26768   300.00  FALSE       207.22            100  2.673964
#> 1479    300.00  FALSE       300.00            100  2.938910
#> 29716   300.00  FALSE       300.00            100  2.716439
#> 3724    300.00  FALSE       300.00            100  2.450415
#> 3868    300.00  FALSE       300.00            100  3.277127
#> 28286   300.00  FALSE       300.00            100  2.790556
#> 32430   300.00  FALSE       300.00            100  2.398827
#> 48184   300.00  FALSE       300.00             75  2.485742
#> 44434   300.00  FALSE       300.00            100  3.451290
#> 974      80.00  FALSE        90.05            100  3.099987
#> 34336   300.00  FALSE       300.00            100  3.106036
#> 7986    275.06  FALSE         0.00             65  3.156065
#> 44216   300.00  FALSE       300.00            100  3.167741
#> 24446   300.00  FALSE       300.00            100  2.925561
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 9526    300.00  FALSE       300.00            100  2.898919
#> 33881   300.00  FALSE       300.00            100  3.171879
#> 39883   233.81  FALSE       219.23            100  2.474870
#> 49812   300.00  FALSE       300.00             65  3.024717
#> 46148   300.00  FALSE       300.00            100  3.015637
#> 30403   300.00  FALSE       209.91            100  2.884241
#> 42990   205.00  FALSE         0.00            100  3.173195
#> 11025   300.00  FALSE       300.00             50  2.828200
#> 28978   300.00  FALSE        10.00            100  2.209674
#> 39234   300.00  FALSE         0.00            100  2.170806