Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.7898858   -0.04921   -0.12438     753.6304
#> 13748 715232.5 9559542 2212.12 26.1234377    0.00023   -0.00043    1806.4552
#> 40756 714022.5 9558862 2331.20 45.0854759   -0.00075    0.00475    1001.0861
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753    0.01487    1077.7585
#> 1300  714042.5 9558482 2408.27 24.1484522    0.00659    0.01041     772.9746
#> 48187 712722.5 9560152 1835.29 29.3279907   -0.00333   -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835   -0.02191   -0.00579    2246.7725
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073    0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.4019303    0.00310    0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.6924129   -0.00645    0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.5564122   -0.00730   -0.01491    1844.4353
#> 7349  714032.5 9558502 2402.95 31.3528235    0.01924    0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223    0.01783     377.8918
#> 22814 713862.5 9558582 2327.48 48.6120312   -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423    0.00214    0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160    0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021   -0.09328    2235.5259
#> 10899 713852.5 9559652 2282.82 28.8140475    0.00952    0.01378     440.9104
#> 43007 713372.5 9559192 2120.47 46.0050732    0.03696    0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.0103974    0.00448   -0.00328     592.3326
#> 25081 713392.5 9558452 2242.24 46.5384969   -0.02716    0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.2449118    0.00051    0.00889     362.8363
#> 32821 715022.5 9559952 2208.33 43.4164499   -0.00643    0.00314    2360.1501
#> 18391 715112.5 9559432 2235.02 22.2576915   -0.02651   -0.03629  360379.1875
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356    0.00774     527.1068
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598    0.01212     561.1948
#> 7879  714722.5 9559232 2356.19 36.8039440   -0.01097    0.00156     816.6036
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743    0.01313     783.4548
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566   -0.03355   14919.6260
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547    0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.4454260    0.00606   -0.00556     660.3690
#> 24100 715132.5 9558992 2420.18 29.2139721   -0.03229   -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.5357857    0.00140   -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.2913143    0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.9336000    0.00321    0.01169     350.1799
#> 13699 712902.5 9559902 1917.21 30.4876572   -0.00093   -0.00907    1943.4092
#> 17036 712832.5 9560582 1951.38 27.6457866    0.00060    0.00270    4057.5637
#> 17423 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294     739.5768
#> 22332 715482.5 9558922 2475.28 19.9366394   -0.02617    0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585   -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.5469237    0.03480    0.02530     937.1306
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855   -0.02175    1122.9120
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762    0.01252    5761.3486
#> 375   715622.5 9558922 2510.87 37.3665885   -0.01017    0.00197     711.1091
#> 42941 713252.5 9559742 1990.10 28.9544222   -0.04708    0.00408   76318.0156
#> 34005 714922.5 9558842 2559.23 52.8312924    0.03658    0.04602     345.8933
#> 44825 714322.5 9557882 2534.80 40.6885978    0.00114    0.03026     247.0894
#> 664   714452.5 9559852 2174.83 44.8608765    0.00282    0.00109     767.4097
#> 21013 712742.5 9560032 1845.53 29.3526278    0.00568   -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.2357795    0.00927    0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180   -0.00104    0.01664    1898.5588
#> 38197 712522.5 9560322 1942.04 62.7772667   -0.01611    0.02961    1086.6963
#> 2552  714942.5 9557652 2719.32 34.8736492   -0.01238   -0.00851    1012.5963
#> 41721 715182.5 9557562 2784.85 31.7899903   -0.00588    0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.0239387    0.00538    0.00112     745.3315
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133    0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.0344196    0.00188    0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469    0.02940     987.8085
#> 7041  713472.5 9559092 2137.03 18.3489734   -0.10372   -0.00008  825875.6875
#> 46461 712682.5 9560032 1892.30 22.9778994    0.01453    0.01337     486.7551
#> 27244 714162.5 9558622 2316.00 36.7512319   -0.02556   -0.11694  102301.4609
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521   -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.6704448    0.01489    0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.0186426   -0.00589    0.00849    1120.5961
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338    0.00728    1137.2450
#> 43201 715712.5 9557922 2852.82 31.4622584   -0.00949    0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084   -0.02236    9755.3027
#> 46138 715202.5 9559582 2194.40 18.8600517   -0.02143   -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506   -0.02070   -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175    0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444   -0.00017     759.9937
#> 30933 714672.5 9561082 1798.37 54.8102886    0.00542   -0.00722    1373.4147
#> 26101 715562.5 9558682 2594.87 32.4110766    0.00079    0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.5275756   -0.00193    0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.3528235    0.01924    0.02606     564.6876
#> 47244 714262.5 9559302 2350.47 17.0592454    0.02213   -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.8039510   -0.00442   -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.7928761   -0.00246    0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.9064613   -0.00070    0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.8127340    0.00546    0.01484     495.0292
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317    0.00077    3056.9475
#> 14494 715332.5 9558792 2557.22 20.7204457    0.01267    0.00483     343.1384
#> 1223  713632.5 9559632 2226.22 50.8087514   -0.01881    0.00121    1783.0658
#> 9606  714392.5 9559172 2353.05 30.7833035   -0.00185   -0.01185     953.8614
#> 16065 715242.5 9560622 1871.61 34.3064209   -0.00019    0.00119    6589.5796
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124    0.01166    1813.7841
#> 13278 712752.5 9560452 1901.25 42.1135439   -0.00118   -0.00252    3631.7017
#> 49346 715322.5 9558782 2553.74 25.3952084    0.00319    0.00291     487.3438
#> 41310 714452.5 9560402 2092.27 12.4629780   -0.00195   -0.00205    1112.2626
#> 34718 714802.5 9558802 2599.22 44.6500280    0.01996    0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025   -0.05345    2417.1389
#> 28253 714102.5 9559892 2288.88  2.9576081    0.00378    0.02382     100.0000
#> 36561 715292.5 9560802 1844.64 31.9372405    0.00002   -0.01232     977.7320
#> 40459 712542.5 9559192 1992.60 39.8171290   -0.00162   -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521    0.00390   -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124    0.00204     853.2902
#> 26023 715832.5 9557632 3097.56 41.7101816    0.00133   -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.6402737    0.00363   -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.6350750   -0.00242    0.01132     737.5982
#> 49363 712772.5 9559552 1935.91 29.0369281   -0.03220   -0.03290    2481.5632
#> 12716 713962.5 9557762 2402.60 18.9322444    0.03480    0.01500     538.8209
#> 42329 715852.5 9558702 2651.92 22.8707563    0.00321    0.00548     532.8210
#> 39193 715332.5 9560852 1872.90 30.2905598   -0.00161    0.00171     756.8818
#> 13286 713662.5 9559702 2268.14 41.9966605   -0.01341    0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405    0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.5669070   -0.00978    0.00367    3002.8828
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001    0.00759    1880.4403
#> 31884 715782.5 9558612 2632.60 30.2796736    0.00084   -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.2478257    0.00737    0.01033    3446.5520
#> 33175 715822.5 9558832 2563.13 36.6228893   -0.00674   -0.00376     629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126   -0.00452   -0.01298     436.1459
#> 1691  714222.5 9559022 2448.69 38.2168579   -0.01292    0.00462     716.5762
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586    0.00654    1576.2065
#> 36721 714092.5 9559752 2301.81 28.9332227    0.03418    0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.1838220    0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858    0.00312     254.0656
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788    0.00119    0.00361    1011.0248
#> 24819 714032.5 9560882 1981.86 33.2842642    0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635   -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.2945705    0.00537   -0.00577     423.5598
#> 32430 715832.5 9558112 2824.39 23.6940967    0.00944    0.00836     250.5111
#> 48184 714202.5 9558222 2489.17 36.1135298    0.02236   -0.00066     306.0146
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354    0.00907    1281.5044
#> 31098 714352.5 9559042 2410.15 30.3363964    0.00537   -0.00347    1893.3645
#> 2999  714592.5 9558492 2515.83 47.0318136    0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757   -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.0041019    0.00745    0.00875    2156.9358
#> 15138 713702.5 9561122 1780.68 49.2216583    0.00607   -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.7218013   -0.00598    0.02208     842.4825
#> 48050 714022.5 9560572 2099.82 18.7139475    0.01182   -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137   -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129    0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327    0.00293    0.00186    2751.4626
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865   -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.5674870   -0.03284   -0.00886   43249.3945
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137   -0.01523     888.2811
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992   -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.6626120   -0.00983    0.00803    1058.5630
#> 47245 715372.5 9560792 1900.65 34.4708598   -0.00011   -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868    0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463    0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305   -0.01135    1293.4855
#> 46567 714472.5 9560352 2080.13 21.3816390    0.01336    0.01403     609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131   -0.01023   -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.6210169   -0.02110   -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607