Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393    2081.0056
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628     556.0121
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741     746.4155
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065   0.00005     769.3426
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340     604.1032
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440    1088.8639
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406    2421.1145
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226    2325.2053
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842     868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526     539.8119
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189     287.5425
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234     849.0547
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261     554.9243
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899    1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114    4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696    4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889     362.8363
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450     873.0448
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917     116.1986
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437    1155.1469
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051     779.0555
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924    8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651     419.0927
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146    1347.6224
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556     660.3690
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862     874.5408
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321   0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504    1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036     516.4709
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907    1943.4092
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903    2360.1536
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964     571.1364
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480   0.02530     937.1306
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504   0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310    2833.9001
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009    1247.7061
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081     940.6002
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472  -0.00228    2262.3796
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242     507.9263
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335     195.0293
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911     173.2806
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833     369.3771
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123     364.2903
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812    1423.7578
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879    1200.4210
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109    2047.9031
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044     288.5468
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610    1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226     338.3474
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875     597.2893
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279   39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856    1789.1596
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895   0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015     790.2421
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144  317260.6875
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603    1562.0503
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265    1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609    4275.3457
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842     246.5161
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232     977.7320
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109     605.1908
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570   0.02230     714.7432
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116    1817.5222
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363  -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466    2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875    1165.3953
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.302976    0.00540   0.00650    1133.1501
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028    2394.7261
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014     271.4219
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598     186.3527
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167   0.01133     828.8604
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943     371.5514
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069  -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983     270.7949
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292   0.00462     716.5762
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036     476.9504
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645     234.0114
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282    2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955     713.4463
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863    1098.3462
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204     618.2027
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555    1882.0179
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960     884.2003
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820  -0.00260    1484.6593
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952  -0.00438    1276.5448
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282    1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238     787.0053
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064    1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585    5714.3267
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052     451.7209
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109     915.4989
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419     323.0790
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403     609.8665
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983