Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261    323.8441
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090   0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403   1787.3024
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226   2325.2053
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518    854.4269
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070 138869.3906
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398    267.2659
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838    238.6685
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284   2010.4285
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898    588.3761
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917    116.1986
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156    816.6036
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193  -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346    543.0749
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174   0.00025   4425.5151
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060   0.00270   4057.5637
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279    930.0871
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964    571.1364
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310   2833.9001
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602    345.8933
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739    365.7626
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684   2556.8049
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102  35685.0000
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512  13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219    420.7032
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180   3904.5254
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367    806.0966
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474   0.01334   4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321   1170.5904
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511  -0.01801   1071.6141
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003   0.01196   1052.5341
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075   2170.2249
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185  -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526    706.1463
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732   4237.6934
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384    947.2075
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253    528.1924
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298  -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403    241.2991
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132    737.5982
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290   2481.5632
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199    836.4196
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598    186.3527
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362    283.7090
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223   1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292   0.00462    716.5762
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705    968.9738
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312    254.0656
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417   1520.6630
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273  -0.01047   2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952  -0.00438   1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050   1471.4346
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282   1659.6211
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809   1158.0426
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172  -0.00119    792.3544
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064   1485.5201
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281    492.4785
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983