Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969    368.5606
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005    769.3426
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403   1787.3024
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491   1844.4353
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783    377.8918
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388   1051.4222
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 22283 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032    994.6740
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939    703.7817
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613    289.6113
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036    516.4709
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199    196.4080
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427    380.9472
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739    365.7626
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473   2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328    592.3326
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664   1898.5588
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728   1137.2450
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351   1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879   1200.4210
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109   2047.9031
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450   2589.9910
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367   1966.2112
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665   1904.4769
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230    714.7432
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171    756.8818
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223   1056.3701
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312    593.7773
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487   1769.3392
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270  35950.5508
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491