Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603