Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123  0.00393    2081.0056
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438     753.6304
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483     566.7674
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333     258.6537
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579    2246.7725
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292    1233.9299
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403    1787.3024
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345    1120.6929
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264    5057.1431
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120     512.9740
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311     586.7354
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250    3686.1240
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261     554.9243
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398     267.2659
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151  0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114    4510.8833
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450     873.0448
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150    1274.3594
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452 -0.01437    1155.1469
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642     613.6390
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924    8023.8657
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514     297.7842
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034    4495.6929
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167    1938.5389
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862     874.5408
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517    1000.7119
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169     350.1799
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527  0.04613     289.6113
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072   54831.6289
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449 -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560     977.4719
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993   10308.6406
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690    5121.6924
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684    2556.8049
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941    1290.7709
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055     773.0359
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620     379.0855
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555    4439.0967
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389   47022.9688
#> 30268 714172.5 9558612 2328.84 54.95066  0.03743 -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394     587.8840
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246  0.00534     344.1447
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332 -0.00058    1817.2371
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520    3390.5708
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141    4611.0718
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017     759.9937
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833     369.3771
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739     789.7573
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118   19557.6621
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759     158.9370
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001    1184.7952
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634     500.5027
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226     338.3474
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000  162178.1562
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118 -0.00252    3631.7017
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744     226.7911
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253     528.1924
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655     130.9468
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093     118.1402
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476    1275.9446
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181    1286.3949
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204     853.2902
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715     520.7355
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064     683.2795
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409    1575.9425
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028    2394.7261
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833     300.0125
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367    3002.8828
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007 -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475     515.1661
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655     560.8554
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032     994.9224
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645     234.0114
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268     252.0072
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868     235.8511
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060     383.3547
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361    1011.0248
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148   10771.1963
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491     261.7371
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523     888.2811
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458     485.7418
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305 -0.01135    1293.4855
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769