Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051 351155.1250
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969    368.5606
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487   1077.7585
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041    772.9746
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057   1630.3259
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403   1787.3024
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440   1088.8639
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406   2421.1145
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783    377.8918
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842    868.6137
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528    0.01392   1255.0884
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518    854.4269
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044 129412.3984
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856    0.01114   4510.8833
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413    364.9788
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521   -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088   2322.1746
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325    616.8652
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276   9614.6924
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642    613.6390
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646   -0.04086   6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924   8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453    0.01077    300.7915
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177    291.5937
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714    475.3120
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407    344.8395
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949   1182.1506
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779   2972.0193
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731   1384.7959
#> 421   714892.5 9561022 1849.01 26.132032  0.00002    0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993  10308.6406
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565   2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364   -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797   -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009   1247.7061
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625   1063.9150
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550   -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845    0.00454   3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656   3036.7383
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809   1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240   1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132  42970.7891
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228   2262.3796
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469    0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190    640.8842
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524  10861.9482
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556   -0.11694 102301.4609
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236   9755.3027
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334   4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261   1627.8989
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522   1355.7436
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970   1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010   3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484    495.0292
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042    0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756    353.9086
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200   1305.6691
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133   2397.2681
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610   1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538    807.8237
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157   -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732   4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675   -0.01905   6487.0010
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234    849.0547
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189    556.5211
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345   2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585   -0.02665   1904.4769
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158    837.3730
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403    241.2991
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738    0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493   2212.0525
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479    604.0975
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464   4029.5947
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431    315.9084
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734    335.0773
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362    283.7090
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422    967.9738
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223   1056.3701
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038   2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452   -0.00948   1222.5433
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257    549.3252
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650    766.6254
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599  12873.8018
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308    422.5022
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420    0.01960    884.2003
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600   -0.01849    764.8193
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598    0.02208    842.4825
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732    0.01858    298.4490
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281    492.4785
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403    609.8665
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456   -0.01586    941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603