Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183   -0.09203    634.3320
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199    0.00659   3553.7166
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292    0.00712   3636.5303
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240   -0.00250   1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005    769.3426
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090    0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951   -0.00529   3282.1186
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387   1989.3012
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607    0.00677   1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894   -0.03416    947.4878
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903    0.01447    331.4833
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696    0.02514    297.7842
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270   -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501   3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471   -0.01382    560.8309
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746    0.00096   3210.5781
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832    0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151    0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426   -0.00367    793.9665
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490   -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643    0.00314   2360.1501
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382    0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205    629.4165
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211    0.00642    613.6390
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646   -0.04086   6042.5757
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743    0.01313    783.4548
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323    183.1836
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353    0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511 361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376   2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509    0.00499    885.0064
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666   -0.00504   1020.2766
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395    0.01495    266.5888
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283    0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409    0.02771    674.3859
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107    0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399   -0.00731   1384.7959
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708    0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283    590.1536
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306   -0.00436   2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090   -0.00299    688.0706
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294    739.5768
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845    0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800   -0.06020   1563.5549
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410    0.03620    379.0855
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201    788.0621
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478   -0.07132  42970.7891
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629   -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320    0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354   1147.7384
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169   -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335    195.0293
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400    608.6523
#> 7     714952.5 9559862 2284.19 16.106417    0.00516    0.01574    183.3220
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860    695.2302
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878   -0.04351   1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143   -0.05078 622888.2500
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354    0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722   1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439   -0.01161   3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247    0.01123    364.2903
#> 332   714852.5 9561102 1787.11 45.358204    0.00540    0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546    0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003    0.01196   1052.5341
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713   6379.7974
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881    0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285   -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708   -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864    0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143    0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125   -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445    0.00875    597.2893
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152   1530.5657
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458    0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144   -0.00856   1789.1596
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744    226.7911
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306   1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427    0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996    0.04684    209.1778
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084   -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067   1123.7318
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336   2806.2644
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165    0.00565   1410.8372
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156   -0.00116   1817.5222
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533    0.01493   2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815   -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064    683.2795
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548    532.8210
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263    0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167    0.01133    828.8604
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196    0.01734    335.0773
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990    0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069   -0.00900    614.5898
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098    164.0802
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041   -0.00011    201.1272
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305   -0.00705    968.9738
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342    0.02552   2515.0752
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233    0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555   1882.0179
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420    0.01960    884.2003
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070   -0.00050   1471.4346
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264    0.00624   1870.8307
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598    0.02208    842.4825
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182   -0.00972    331.7462
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253   -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263   -0.02012    305.3624
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858    298.4490
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933   -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281    492.4785
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247    0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607