Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029   5577.3916
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357   1752.0375
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741    746.4155
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770    474.5059
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378    440.9104
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096   3210.5781
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108  -0.00678   1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150   1274.3594
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596   5296.9014
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924   8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511 361979.2812
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613    289.6113
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407    344.8395
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771    674.3859
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009   1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543    352.1558
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539   0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643    670.2289
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512  13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596    842.0994
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081    940.6002
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474   2343.2100
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367    806.0966
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190    640.8842
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728   1137.2450
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439  -0.01161   3065.3694
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169   0.01441    424.2680
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957    495.4453
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915   2036.0378
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921    525.3245
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856   1789.1596
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189    556.5211
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405  -0.05655    130.9468
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000   0.00320    911.5911
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109    605.1908
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181   1286.3949
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452    578.5847
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945    259.7421
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014    271.4219
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697    412.1185
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759   4131.5356
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196    576.1450
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513    161.9878
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250   2225.4426
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165    805.1403
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105    868.7802
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820  -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236  -0.00066    306.0146
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074  -0.02904    608.3561
#> 974   713402.5 9561062 1854.56 23.634509    0.00082   0.02578   1258.8865
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270  35950.5508
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809   1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064   1485.5201
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281    492.4785
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419    323.0790
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305  -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181  -0.01149   4242.3579
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603