Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813   1130.5791 35.101113
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305 -0.00555    430.4604 46.301865
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043   1806.4552 34.119064
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121 43.531614
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877 32.738808
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752   1321.5997 39.598259
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259 30.998736
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426 24.128972
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032  6.084812
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639 29.110267
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387   1989.3012 29.596708
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918 16.677656
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053 47.548622
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326 19.942369
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240 33.825709
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292 44.908432
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309 40.735580
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274 27.975237
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285 34.621548
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629 360379.1875 34.507529
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356  0.00774    527.1068 16.455921
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950    950.3138 37.901731
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212    561.1948 20.380109
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083    994.7876 36.075715
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165 26.741086
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089 -0.00051    779.0555 28.203847
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498 27.153043
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224 37.588323
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412 34.148858
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937 32.097669
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570 28.960152
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690 18.763795
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299 23.826450
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517   1000.7119 21.672702
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151 29.071306
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395 32.094804
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768 25.460526
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307 147470.5312 27.068818
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102    483.4184 26.882034
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530    937.1306 38.832788
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859 40.091576
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875 33.218947
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959 36.650391
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588    749.2502 10.713738
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406 31.563099
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335 23.337144
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001  9.812475
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332   1418.6622 34.900005
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401   1722.6481 31.445643
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706 27.323211
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334 29.645982
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326 19.942369
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553   1750.4175 30.389681
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941   1290.7709 37.362578
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359 31.442778
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621 19.591146
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963 32.215698
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322 42.353040
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445 28.348233
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448   1258.8196 29.979444
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659 26.364080
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688 37.425030
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793    864.9472 37.920066
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842 29.557747
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611 29.738228
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220 13.220428
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866 38.668349
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878 -0.04351   1442.9247 42.935738
#> 21683 714792.5 9561052 1807.84 55.73275 -0.02070 -0.02661   2508.5374 19.745272
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132   1580.1621 41.642573
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771 26.597847
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722   1373.4147 12.843422
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261   3353.5422 23.386991
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351 28.987654
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404 25.746431
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759    158.9370 29.621345
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952 33.994159
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915   2036.0378 33.615434
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964 33.294577
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908 25.370571
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910 27.720271
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237 42.761559
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893 21.668691
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.802973
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560  0.00610    454.8103 27.601096
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841 16.579107
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152   1530.5657 29.730780
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856   1789.1596 23.285578
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867   2247.8748 31.598622
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924 20.820713
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465 -0.01036   2505.8369 26.906098
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652 31.665086
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778 34.409553
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389 33.245303
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093    118.1402  3.946533
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457 43.145441
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024 30.834870
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230    714.7432 36.366204
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494 36.186868
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118   2273.4758 36.362194
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525 33.383959
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677 21.429767
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691 29.823599
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625 22.984775
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875   1165.3953 31.920625
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975 25.270304
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445    355.5078 18.572427
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261 23.447725
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101 34.567690
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546 25.388906
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273 34.261157
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085 27.805069
#> 18627 714292.5 9559252 2330.78 18.49622 -0.00538 -0.00052   3508.5591 31.257140
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403 33.143317
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734    335.0773 19.295500
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347 24.137566
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355 25.007316
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783 32.072458
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218  0.00902    218.7467 38.595583
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738 28.757325
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949 38.077629
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802 23.684356
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356 28.853009
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911 44.395062
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547 32.236897
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268    252.0072 19.737823
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392 29.837923
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511 13.572224
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204    618.2027 22.607196
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613 19.645004
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148  10771.1963 27.612555
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978 26.147502
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146 30.956337
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578   1258.8865 22.758457
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346 26.064996
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296 41.510792
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282   1659.6211  7.987605
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562 29.166417
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417 22.882788
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270  35950.5508 32.711879
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053 39.523775
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042 11.038605
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119    792.3544 27.569583
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124  25826.6914 27.306023
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220 36.130719
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249    148.1857 33.485945
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579 31.663367
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855    229.6061 21.518003
#>       distroad slides distdeforest distslidespast log.carea
#> 44302   300.00   TRUE       300.00            100  3.053301
#> 15762   300.00   TRUE       300.00            100  2.633933
#> 13748   300.00   TRUE       300.00              2  3.256827
#> 25090   300.00   TRUE       300.00             26  2.745084
#> 24051   300.00   TRUE       300.00             11  3.001122
#> 8598    300.00   TRUE       300.00            100  3.121100
#> 11088   300.00   TRUE       300.00            100  3.212274
#> 34512   300.00   TRUE       195.00              2  3.224136
#> 24129   300.00   TRUE       300.00             89  2.886120
#> 39885   210.57   TRUE         0.00            100  2.781111
#> 8869    300.00   TRUE       300.00            100  3.036974
#> 125     207.04   TRUE        21.03            100  3.298701
#> 31023   135.00   TRUE         0.00              2  2.577367
#> 39629   300.00   TRUE       300.00             85  3.366461
#> 22283   256.35   TRUE         0.00            100  2.772566
#> 48055   300.00   TRUE         1.67              0  3.566570
#> 18304   300.00   TRUE       300.00             57  3.589919
#> 30125   300.00   TRUE       300.00            100  2.748832
#> 15340   300.00   TRUE       300.00            100  3.144334
#> 6676    300.00   TRUE       300.00            100  3.303289
#> 18391   300.00   TRUE       300.00            100  5.556760
#> 10320   300.00   TRUE       300.00             30  2.721899
#> 28230   300.00   TRUE       300.00            100  2.977867
#> 34565   300.00   TRUE        68.13            100  2.749114
#> 8073    138.71   TRUE        76.41             35  2.997730
#> 23031   300.00   TRUE       300.00              2  2.798938
#> 2950    300.00   TRUE       300.00            100  2.891568
#> 35142   300.00   TRUE       300.00            100  2.176525
#> 12868   119.34   TRUE        79.16             61  3.129568
#> 15990   300.00   TRUE         0.00              0  3.754443
#> 47429   300.00   TRUE       300.00            100  2.464778
#> 40738   300.00   TRUE       300.00            100  2.716468
#> 37828   300.00   TRUE       300.00              2  2.819787
#> 6310    300.00   TRUE       300.00            100  4.032926
#> 27084   300.00   TRUE       300.00             81  3.000309
#> 46171   300.00   TRUE       270.22            100  3.645964
#> 29926   300.00   TRUE       300.00            100  2.537617
#> 17423   300.00   TRUE       300.00            100  2.868983
#> 22332   300.00   TRUE       300.00            100  5.168705
#> 2976    300.00   TRUE       300.00            100  2.684323
#> 16015   152.09   TRUE        90.23             45  2.971800
#> 21952   300.00   TRUE       300.00             64  2.828908
#> 25244   300.00   TRUE       300.00              6  5.916915
#> 835     300.00   TRUE       300.00            100  3.141386
#> 421     182.02   TRUE         0.00             83  2.874627
#> 36120   300.00   TRUE       300.00             94  4.013201
#> 15766   300.00   TRUE       300.00            100  2.760448
#> 462      61.17   TRUE        57.16             74  3.452385
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 18265   300.00   TRUE       300.00            100  3.236197
#> 6317    300.00   TRUE       300.00            100  2.837633
#> 47100    82.80   TRUE       129.73             36  3.558076
#> 46783   256.35   TRUE         0.00            100  2.772566
#> 20065   300.00   TRUE       300.00              5  3.243142
#> 44650   300.00   TRUE       300.00            100  3.110849
#> 31561   300.00   TRUE       300.00              0  2.888200
#> 28046   300.00   TRUE       300.00             15  2.896560
#> 2552    300.00   TRUE       300.00            100  3.005436
#> 42765   300.00   TRUE       300.00             12  2.974066
#> 21680   104.12   TRUE       110.76              4  3.591103
#> 41721   300.00   TRUE       300.00            100  3.099963
#> 40216   300.00   TRUE       300.00            100  3.257022
#> 37769   300.00   TRUE       300.00            100  4.672310
#> 12772   300.00   TRUE       300.00            100  2.936990
#> 43859   300.00   TRUE       300.00            100  2.806780
#> 32879   300.00   TRUE       300.00              6  3.130450
#> 7       300.00   TRUE       300.00              0  2.263215
#> 24135   300.00   TRUE       122.87            100  3.878395
#> 34887   300.00   TRUE       300.00            100  3.159244
#> 21683   130.93   TRUE         0.00             21  3.399421
#> 28312   300.00   TRUE       300.00            100  3.198702
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 30933    99.29   TRUE         0.00              6  3.137802
#> 47086   300.00   TRUE       300.00            100  3.525504
#> 49072   300.00   TRUE       300.00            100  3.662232
#> 2801    300.00   TRUE       300.00             26  2.422656
#> 16071   300.00   TRUE       300.00            100  2.201225
#> 49378    22.68   TRUE       134.63             25  3.073643
#> 22301   300.00   TRUE         0.00            100  3.308786
#> 39624   300.00   TRUE       300.00              2  3.254015
#> 7409    300.00   TRUE       300.00            100  2.866282
#> 39405   300.00   TRUE       300.00            100  3.413298
#> 38103   300.00   TRUE       300.00             31  2.907317
#> 10456    24.34   TRUE         0.00             18  2.776185
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 46792   300.00   TRUE       300.00             18  2.657830
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 46038   300.00   TRUE       243.86            100  3.184852
#> 12074   300.00   TRUE       300.00             63  3.252649
#> 10213   300.00   TRUE       300.00            100  3.351772
#> 9864    300.00   TRUE       300.00            100  2.722792
#> 33142   300.00   TRUE       300.00            100  3.398953
#> 17322    75.77   TRUE       138.35              1  3.235115
#> 34718   300.00   TRUE       300.00            100  2.320516
#> 10785   300.00   TRUE       300.00            100  3.383302
#> 9984    300.00   TRUE        36.46             76  2.072398
#> 17374   300.00   TRUE       300.00            100  3.630971
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 15211   300.00  FALSE       300.00            100  2.844665
#> 30809   181.84  FALSE         0.00            100  2.854150
#> 40459   300.00  FALSE       216.66             10  3.025531
#> 35852   300.00  FALSE       300.00            100  3.356690
#> 11111   300.00  FALSE       300.00            100  3.344795
#> 20458   300.00  FALSE       300.00             28  2.424012
#> 20067   300.00  FALSE       268.03            100  2.590250
#> 7517    144.37  FALSE        60.16            100  5.823071
#> 46930   300.00  FALSE         0.00            100  3.066473
#> 30267   300.00  FALSE       300.00            100  2.781107
#> 19493   300.00  FALSE       291.23            100  2.550849
#> 45166   300.00  FALSE       300.00              1  3.379256
#> 39796   300.00  FALSE       300.00             71  2.333468
#> 16878   300.00  FALSE       300.00             18  3.154988
#> 20547   300.00  FALSE        42.56            100  4.078911
#> 45071   264.04  FALSE         0.00            100  2.702696
#> 18627   300.00  FALSE       300.00             20  3.545129
#> 40139   300.00  FALSE         1.11             21  3.274260
#> 36126   300.00  FALSE       300.00            100  2.525145
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 41753   300.00  FALSE       300.00            100  2.860419
#> 20319   300.00  FALSE       300.00            100  3.167901
#> 33956   300.00  FALSE       300.00            100  2.339942
#> 6143    300.00  FALSE       185.58            100  2.985864
#> 8895    300.00  FALSE       300.00             90  2.432641
#> 22545   300.00  FALSE       300.00             96  2.215056
#> 29835   300.00  FALSE       300.00             45  3.616112
#> 34345   300.00  FALSE       300.00            100  2.972060
#> 12044   300.00  FALSE       300.00            100  3.304782
#> 44267   300.00  FALSE       226.30            100  2.401413
#> 49831   300.00  FALSE       300.00            100  3.247811
#> 330     300.00  FALSE       275.17            100  2.372638
#> 42424   275.21  FALSE        93.50             70  2.791131
#> 36889   300.00  FALSE       300.00             48  2.108096
#> 5861    300.00  FALSE       300.00            100  4.032264
#> 3868    300.00  FALSE       300.00            100  3.277127
#> 48184   300.00  FALSE       300.00             75  2.485742
#> 974      80.00  FALSE        90.05            100  3.099987
#> 44216   300.00  FALSE       300.00            100  3.167741
#> 26283   300.00  FALSE       300.00            100  2.962620
#> 18246    18.49  FALSE         0.00             29  3.220009
#> 10981   300.00  FALSE       300.00            100  5.155983
#> 16845   300.00  FALSE        25.55             33  3.874467
#> 20950   300.00  FALSE       300.00             81  4.555706
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 41873   300.00  FALSE       300.00            100  2.265064
#> 9526    300.00  FALSE       300.00            100  2.898919
#> 36141   300.00  FALSE       300.00             85  4.412069
#> 30403   300.00  FALSE       209.91            100  2.884241
#> 39234   300.00  FALSE         0.00            100  2.170806
#> 47629   300.00  FALSE       300.00            100  3.627607
#> 40852   300.00  FALSE       300.00            100  2.360983