Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309  -0.00681   -0.00029    5577.3916
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261     323.8441
#> 30059 713832.5 9559662 2288.02 18.2779266   0.01917    0.00333     258.6537
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 39885 715062.5 9561022 1840.94 38.4357278   0.00250   -0.01340     604.1032
#> 8869  715062.5 9559392 2278.56 41.4019303   0.00310    0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.8450293   0.00884    0.01345    1120.6929
#> 14456 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010    1527.1028
#> 43327 713962.5 9561082 1839.11 63.3616837  -0.04951   -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.3415951  -0.00979   -0.01542     674.0147
#> 31023 714712.5 9561042 1838.66 44.1200421  -0.00223    0.01783     377.8918
#> 38993 714202.5 9558402 2475.75 33.1593594  -0.01607    0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321   0.00102    0.00388    1051.4222
#> 23925 715202.5 9557652 2782.99 41.9571264  -0.00542    0.01842     868.6137
#> 33792 714652.5 9558172 2598.36 59.4586952  -0.03525   -0.11024     394.8751
#> 38410 713712.5 9561172 1776.08 36.3192217  -0.02678   -0.04272    4955.5586
#> 16843 714882.5 9558512 2701.78 26.6379538  -0.00270   -0.00769    1032.7177
#> 2107  714142.5 9558562 2382.67 35.5537501   0.01412    0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.9269342  -0.00070   -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.1946661  -0.02131    0.00501    3889.7292
#> 20290 714812.5 9561092 1775.73 34.6267043  -0.01537   -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295  -0.01440    0.00710     843.7961
#> 25081 713392.5 9558452 2242.24 46.5384969  -0.02716    0.03696    4143.4019
#> 47022 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.2343959   0.00993   -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314    2360.1501
#> 18391 715112.5 9559432 2235.02 22.2576915  -0.02651   -0.03629  360379.1875
#> 3133  715042.5 9559252 2344.13 29.0208853   0.00422    0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.4804185   0.00252    0.01118    2359.5059
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866  -0.00211    0.00642     613.6390
#> 22735 714492.5 9559772 2170.30 39.6887865  -0.00440   -0.02651     419.0927
#> 5728  713502.5 9559662 2140.27 34.3075669  -0.00154    0.00124    2114.2253
#> 37348 715632.5 9558182 2743.16 13.6203527  -0.00904   -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.7021461  -0.01771   -0.01939     703.7817
#> 36931 712852.5 9558892 2102.80 30.5357857   0.00140   -0.00070    2562.7559
#> 34563 715022.5 9558162 2675.70 18.8898455  -0.01993    0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255   0.01458    0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.2913143   0.00316    0.00714     475.3120
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613     289.6113
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907    1943.4092
#> 3944  715262.5 9558352 2654.20 33.5552733  -0.04492    0.00072   54831.6289
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779    2972.0193
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993   10308.6406
#> 49569 713232.5 9561122 1868.73 37.3121575  -0.00405   -0.01565    2232.6133
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.9544222  -0.04708    0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.4028314  -0.00140   -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542     860.8840
#> 27875 714832.5 9561032 1838.22 41.2122812  -0.01160    0.01690    5121.6924
#> 10507 714892.5 9559282 2374.45 42.8423462  -0.01057    0.02427     380.9472
#> 18265 715522.5 9558782 2539.23 35.8562718  -0.00458   -0.00401    1722.6481
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.2590471   0.00736    0.01684    2556.8049
#> 35083 714612.5 9558152 2596.12 60.5983082   0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625    1063.9150
#> 3418  715372.5 9558492 2601.00 22.6272492  -0.01168   -0.04102   35685.0000
#> 21013 712742.5 9560032 1845.53 29.3526278   0.00568   -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.5260665   0.01927   -0.00316     454.9674
#> 5103  714612.5 9559892 2220.36 49.0079450  -0.00078    0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.1089251   0.01784    0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.3058340  -0.00131    0.00201     788.0621
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.9781091   0.00165    0.00555    4439.0967
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.0344196   0.00188    0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386   0.01469    0.02940     987.8085
#> 11612 715212.5 9557752 2820.51 39.1622383   0.00017    0.01213     328.7922
#> 32879 713912.5 9559672 2238.00 31.9767745  -0.00278   -0.00412    1350.3611
#> 20368 715252.5 9560622 1865.83 33.5724620   0.00025   -0.01245    6353.3179
#> 27244 714162.5 9558622 2316.00 36.7512319  -0.02556   -0.11694  102301.4609
#> 7     714952.5 9559862 2284.19 16.1064166   0.00516    0.01574     183.3220
#> 7086  712762.5 9560962 2022.00 44.9135886  -0.01259   -0.00281    1178.4420
#> 27915 715232.5 9559552 2208.93 25.4805791   0.00171   -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195    1901.6743
#> 29228 712782.5 9560422 1870.56 32.9061121  -0.00218   -0.00372    4154.5786
#> 5911  713952.5 9561282 1801.16 33.7896130   0.00466   -0.01776    1021.8267
#> 10689 713542.5 9559132 2211.95 31.9813582   0.05379    0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.3295700   0.02247    0.01123     364.2903
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981    3975.3608
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957     495.4453
#> 34984 715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.6578678  -0.01341    0.00171    4425.5410
#> 9606  714392.5 9559172 2353.05 30.7833035  -0.00185   -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.4323251   0.04864    0.04226     338.3474
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875     597.2893
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921     525.3245
#> 47722 712732.5 9561022 1999.03 37.4926392   0.00004   -0.02384     947.2075
#> 9864  714952.5 9557592 2752.40 31.9893796  -0.00983    0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063    2430.2278
#> 41310 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205    1112.2626
#> 36765 715302.5 9560262 1952.73 27.2034632  -0.00585   -0.02665    1904.4769
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996    1599.3491
#> 31983 714302.5 9559912 2204.77 23.4649135  -0.01332    0.01842     246.5161
#> 10145 713782.5 9560932 1865.62 23.5955479  -0.00203    0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.3476599  -0.02196    0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.9372405   0.00002   -0.01232     977.7320
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736     477.8591
#> 25212 714272.5 9559012 2433.45 37.6714021  -0.01909   -0.01181    1286.3949
#> 26023 715832.5 9557632 3097.56 41.7101816   0.00133   -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363     389.2691
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075    8141.7549
#> 13512 715152.5 9557992 2778.81 43.3895209  -0.00724    0.01464    4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923     811.5580
#> 49835 713272.5 9560372 1898.68 32.8258980   0.00846    0.00794    1363.5321
#> 13286 713662.5 9559702 2268.14 41.9966605  -0.01341    0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.7881179   0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507     504.3085
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853    1354.6001
#> 36126 715052.5 9558292 2759.47 29.0621382   0.00196    0.01734     335.0773
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267   31588.1953
#> 29527 714332.5 9557502 2567.02 48.9048126  -0.00452   -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.6842028   0.02218    0.00902     218.7467
#> 47059 713822.5 9558272 2413.20 38.1750320  -0.00687   -0.00223    1056.3701
#> 34105 715432.5 9558592 2621.73 18.9563087  -0.00301   -0.01099    1146.2140
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 29107 713172.5 9559272 2067.58 28.2640080   0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645     234.0114
#> 17746 714752.5 9558322 2648.52 47.7365517  -0.00068   -0.00312     593.7773
#> 44267 712972.5 9559042 2191.01 22.2691506   0.01462    0.00268     252.0072
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.0192155   0.03310   -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.1782951   0.00667    0.00863    1098.3462
#> 44823 714202.5 9557442 2548.13 43.2302386  -0.00233    0.01282     561.2366
#> 48988 714682.5 9560852 1942.97 20.2907274   0.00480    0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497  -0.00930   -0.00360    4288.7329
#> 17628 714032.5 9560252 2160.03 30.2945705   0.00537   -0.00577     423.5598
#> 23499 714272.5 9558232 2508.35 28.9676002  -0.00820   -0.00260    1484.6593
#> 38362 714402.5 9558042 2499.31 56.9010116   0.03044   -0.00964     543.0025
#> 974   713402.5 9561062 1854.56 23.6345090   0.00082    0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.1722787  -0.03757   -0.00003 5426606.0000
#> 4943  712452.5 9559172 1927.46 34.4651303   0.00105   -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208     842.4825
#> 16845 712772.5 9560022 1831.76 21.3999736  -0.00865   -0.00805    7489.7417
#> 18504 713372.5 9560672 1786.89  6.4738501  -0.00278   -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.0021874   0.00203    0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.4521619  -0.01196   -0.01935     903.6397
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.6572738  -0.01305    0.00585    5714.3267
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803    1058.5630
#> 15880 715892.5 9557572 3113.36 53.0432868   0.00962   -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.1106510  -0.00288    0.00448     941.8503
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982