Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775