Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466    724.0713
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281   1325.6138
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096   3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593    939.0466
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279    930.0871
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553   1750.4175
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809   1102.9956
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961   1086.6963
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393   8301.6094
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161   3065.3694
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957    495.4453
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856   1789.1596
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075   8141.7549
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273  72123.6172
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077   8759.9961
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035   1471.9783
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223   1056.3701
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759   4131.5356
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836    250.5111
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849    764.8193
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875   2156.9358
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177    184.1042
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249    148.1857
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607