Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606 20.478085
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259 30.998736
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695 36.517465
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761    928.6858 28.902283
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032  6.084812
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436 15.987814
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406   2421.1145 31.035978
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876 26.322827
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388   1051.4222 29.305072
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416    947.4878 34.912037
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021 -0.09328   2235.5259 14.896903
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104 24.774122
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138 25.504070
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425 22.269151
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930 23.454600
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606 28.757325
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842 38.961703
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356 28.852436
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269 22.635271
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936 18.648630
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906 37.628430
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721 35.008294
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398    267.2659 24.586765
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838    238.6685 14.328529
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274 27.975237
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019 34.875368
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501 29.483262
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761 47.067337
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746 33.621163
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165 26.741086
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642    613.6390 38.478700
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498 27.153043
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193 27.419468
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355  14919.6260 34.878806
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937 32.097669
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570 28.960152
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299 23.826450
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517   1000.7119 21.672702
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799 17.635641
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709 38.075910
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965 25.741848
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092 34.789424
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395 32.094804
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679 27.054494
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362 37.643900
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964    571.1364 29.354920
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237 42.761559
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120 32.175591
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859 40.091576
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849 21.719111
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001  9.812475
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091 27.974091
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211  0.00241    585.6182 34.249698
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156 39.063117
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332   1418.6622 34.900005
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840 35.894087
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283    590.1536 43.100750
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602    345.8933 41.733673
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924 21.267047
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739    365.7626 29.842507
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334 29.645982
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791 33.754663
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768 25.460526
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000 35.453482
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097 30.842318
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289 32.530252
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588 33.576473
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596    842.0994 27.152470
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621 19.591146
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724 29.132612
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963 32.215698
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891 25.325308
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445 28.348233
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448   1258.8196 29.979444
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094 31.788271
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966 26.541697
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822    472.8748 29.989757
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842 29.557747
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246  0.00534    344.1447 45.664163
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399 26.645975
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302 36.555853
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583   1898.0105 36.633776
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132   1580.1621 41.642573
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491 30.648085
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261   3353.5422 23.386991
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970   1443.9200 32.180748
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341 32.922728
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475 23.160673
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756    353.9086 46.649651
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691 40.586611
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417  34908.6562 33.408596
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681 25.235926
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008 34.424450
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384 21.185114
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121   1783.0658 40.458269
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614 33.823990
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610   1872.9276 34.498362
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143  0.00477   2945.2100 30.733456
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000 24.747766
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560  0.00610    454.8103 27.601096
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622 22.717777
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856   1789.1596 23.285578
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421 16.521238
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230   1044.6443 16.784226
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164 28.512099
#> 43536 713782.5 9558732 2281.50 32.74912  0.00840  0.00189    556.5211 29.094797
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613 25.485163
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457 43.145441
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318 26.486120
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024 30.834870
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676 15.762069
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002 -0.01232    977.7320 25.804873
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158    837.3730 38.132633
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222 31.737851
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677 21.429767
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666 42.063124
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132    737.5982 14.600111
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795 23.527366
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445    355.5078 18.572427
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501 25.113313
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724  0.01464   4029.5947 35.934767
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580 21.112349
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261 23.447725
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273  72123.6172 36.135302
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833    300.0125 25.894828
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604 47.873489
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090 18.628004
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520 31.248545
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129 33.722577
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236 34.783695
#> 20135 714772.5 9560232 2107.52 37.85074 -0.00782 -0.00048   1550.6704 32.332581
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433 30.485938
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392 29.837923
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656 21.265329
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613 19.645004
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248 25.587149
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451 32.145797
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700 36.319795
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296 41.510792
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972    331.7462 23.361208
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780 25.089822
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172 111973.8906 23.095929
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055 31.814627
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209 24.035007
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989 37.619836
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305 -0.01135   1293.4855 39.382827
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665 16.948092
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691 33.526625
#>       distroad slides distdeforest distslidespast log.carea
#> 42632   300.00   TRUE       300.00             10  2.566509
#> 11088   300.00   TRUE       300.00            100  3.212274
#> 37265   300.00   TRUE         2.45              0  3.071539
#> 34512   300.00   TRUE       195.00              2  3.224136
#> 28293   300.00   TRUE       300.00              4  2.967869
#> 39885   210.57   TRUE         0.00            100  2.781111
#> 26531   300.00   TRUE       300.00              2  2.552843
#> 38827   300.00   TRUE        52.52             39  3.384015
#> 7349    300.00   TRUE       300.00              8  2.751808
#> 24018   300.00   TRUE       300.00              5  3.021777
#> 22814   300.00   TRUE       300.00              6  2.976574
#> 38287    41.43   TRUE         1.90            100  3.349380
#> 10899   300.00   TRUE       300.00             20  2.644350
#> 44328   300.00   TRUE       300.00            100  3.122417
#> 21703   300.00   TRUE       300.00             34  2.458702
#> 45869   300.00   TRUE       300.00             12  3.970546
#> 43964   300.00   TRUE       300.00             46  2.952143
#> 43007   300.00   TRUE       300.00             37  2.473902
#> 29420   253.02   TRUE         0.00            100  3.954264
#> 15109   300.00   TRUE       300.00             40  2.931675
#> 29472   111.09   TRUE         0.00             25  3.424505
#> 38279   300.00   TRUE       300.00             24  5.142607
#> 38032   300.00   TRUE       300.00             60  2.828706
#> 22566   300.00   TRUE         7.49              0  2.426944
#> 27978   300.00   TRUE       300.00             12  2.377795
#> 15340   300.00   TRUE       300.00            100  3.144334
#> 25081   300.00   TRUE       300.00            100  3.617357
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 13585   300.00   TRUE       300.00            100  2.769655
#> 34280   300.00   TRUE         0.00            100  3.365895
#> 23031   300.00   TRUE       300.00              2  2.798938
#> 19730   300.00   TRUE       300.00            100  2.787913
#> 35142   300.00   TRUE       300.00            100  2.176525
#> 6557    300.00   TRUE       300.00              5  3.147064
#> 47351   300.00   TRUE       300.00            100  4.173758
#> 47429   300.00   TRUE       300.00            100  2.464778
#> 40738   300.00   TRUE       300.00            100  2.716468
#> 6310    300.00   TRUE       300.00            100  4.032926
#> 27084   300.00   TRUE       300.00             81  3.000309
#> 4605    300.00   TRUE       300.00             10  2.544291
#> 18751   300.00   TRUE       300.00            100  2.713046
#> 5578    300.00   TRUE       300.00            100  2.640875
#> 13699   300.00   TRUE        53.95             43  3.288564
#> 29926   300.00   TRUE       300.00            100  2.537617
#> 47884   300.00   TRUE       300.00            100  2.905563
#> 13134   300.00   TRUE       300.00             67  2.634917
#> 33240   300.00   TRUE       300.00            100  2.756740
#> 4031    300.00   TRUE       300.00             31  2.907317
#> 20220   300.00   TRUE       300.00              8  3.050346
#> 21952   300.00   TRUE       300.00             64  2.828908
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 462      61.17   TRUE        57.16             74  3.452385
#> 375     300.00   TRUE       300.00            100  2.851936
#> 26698   300.00   TRUE       300.00              7  2.767615
#> 42941   300.00   TRUE       300.00             17  4.882627
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 39520   300.00   TRUE       300.00            100  2.934945
#> 30406   300.00   TRUE       300.00             85  2.770965
#> 34005   300.00   TRUE       300.00            100  2.538942
#> 27875   157.93   TRUE         0.00             57  3.709413
#> 28845   300.00   TRUE       300.00             54  2.563199
#> 47100    82.80   TRUE       129.73             36  3.558076
#> 31462   300.00   TRUE       300.00              1  3.466909
#> 23213   300.00   TRUE       300.00            100  2.868983
#> 3418    300.00   TRUE       300.00             73  4.552486
#> 664     300.00   TRUE       300.00             12  2.885027
#> 35014   300.00   TRUE       300.00             57  2.826223
#> 13146   300.00   TRUE       300.00              2  3.278424
#> 24501   300.00   TRUE       300.00              0  2.925363
#> 28046   300.00   TRUE       300.00             15  2.896560
#> 7064    300.00   TRUE       300.00            100  3.079714
#> 2552    300.00   TRUE       300.00            100  3.005436
#> 17431   300.00   TRUE         1.11              5  4.633173
#> 21680   104.12   TRUE       110.76              4  3.591103
#> 41721   300.00   TRUE       300.00            100  3.099963
#> 18897   300.00   TRUE        20.00              5  3.919162
#> 39722   300.00   TRUE       300.00            100  2.906387
#> 18958   300.00   TRUE       300.00              5  2.674746
#> 43859   300.00   TRUE       300.00            100  2.806780
#> 47320   300.00   TRUE       300.00             43  2.536741
#> 18440   300.00   TRUE       300.00            100  4.477051
#> 31117   300.00   TRUE       291.23            100  3.277641
#> 41125   300.00   TRUE         0.00            100  2.842129
#> 4588    300.00   TRUE       300.00              1  3.610558
#> 19832   300.00   TRUE       300.00            100  3.278299
#> 28312   300.00   TRUE       300.00            100  3.198702
#> 10689   300.00   TRUE       300.00             61  2.237416
#> 47086   300.00   TRUE       300.00            100  3.525504
#> 12629   300.00   TRUE       300.00            100  3.159543
#> 38668   300.00   TRUE       300.00            100  3.022236
#> 13259   203.85   TRUE        10.45             74  3.485288
#> 22297   300.00   TRUE       300.00             11  2.548891
#> 14192   300.00   TRUE       300.00            100  3.115833
#> 20426   300.00   TRUE       300.00             76  4.542933
#> 1997    168.72   TRUE         0.00              5  3.379717
#> 48574    96.21   TRUE       101.32             10  3.143827
#> 14494   300.00   TRUE       300.00             28  2.535469
#> 1223    300.00   TRUE       300.00              4  3.251167
#> 9606    300.00   TRUE       300.00              0  2.979485
#> 26058   300.00   TRUE        76.02             39  3.272521
#> 17825    73.66   TRUE       137.63             18  3.469116
#> 22847   300.00   TRUE       300.00             41  3.134623
#> 46792   300.00   TRUE       300.00             18  2.657830
#> 18498   300.00   TRUE       300.00             83  2.904587
#> 12074   300.00   TRUE       300.00             63  3.252649
#> 14779    55.08   TRUE         0.00              1  2.897760
#> 46971   300.00   TRUE       262.51            100  3.018968
#> 43834    20.21   TRUE       150.14              2  3.331390
#> 43536   300.00   TRUE       300.00              8  2.745482
#> 23832   300.00   TRUE       300.00              8  2.676567
#> 17374   300.00   TRUE       300.00            100  3.630971
#> 4785    300.00  FALSE         0.00            100  3.050663
#> 15211   300.00  FALSE       300.00            100  2.844665
#> 32405   300.00  FALSE       300.00            100  2.530158
#> 36561   300.00  FALSE        70.00            100  2.990220
#> 748     300.00  FALSE       300.00             12  2.922919
#> 31071   300.00  FALSE       300.00             76  3.259480
#> 20458   300.00  FALSE       300.00             28  2.424012
#> 26963   300.00  FALSE       300.00            100  3.420610
#> 32869   300.00  FALSE       300.00             78  2.867820
#> 8066    300.00  FALSE       300.00            100  2.834598
#> 19493   300.00  FALSE       291.23            100  2.550849
#> 441     294.31  FALSE       152.09            100  3.054287
#> 13512   300.00  FALSE       300.00            100  3.605261
#> 16542   300.00  FALSE       300.00            100  2.909320
#> 45166   300.00  FALSE       300.00              1  3.379256
#> 47005   300.00  FALSE       300.00            100  4.858077
#> 1085    300.00  FALSE       300.00            100  2.477139
#> 39685   300.00  FALSE       300.00            100  2.918481
#> 29944   300.00  FALSE       300.00             48  2.452873
#> 21925   300.00  FALSE       300.00             95  3.537385
#> 31201   300.00  FALSE       300.00            100  2.404000
#> 33175   300.00  FALSE       300.00            100  2.799219
#> 20135   300.00  FALSE       300.00            100  3.190519
#> 12519   300.00  FALSE       120.00            100  3.087264
#> 49831   300.00  FALSE       300.00            100  3.247811
#> 43222   300.00  FALSE       300.00            100  2.404946
#> 36889   300.00  FALSE       300.00             48  2.108096
#> 41612   300.00  FALSE       300.00            100  3.004762
#> 484     300.00  FALSE        55.63            100  3.216864
#> 44434   300.00  FALSE       300.00            100  3.451290
#> 26283   300.00  FALSE       300.00            100  2.962620
#> 48050   300.00  FALSE        72.23            100  2.520806
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 18504   144.96  FALSE         4.48             89  5.049117
#> 28526   300.00  FALSE        46.28            100  3.237445
#> 31440   300.00  FALSE       300.00             96  2.654870
#> 32342   300.00  FALSE       300.00             81  2.961658
#> 32618   300.00  FALSE       300.00             56  3.111762
#> 46567   300.00  FALSE       300.00             78  2.785235
#> 48157    57.15  FALSE       157.06             45  2.887769