Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59291   300.00   TRUE       300.00            100  3.318273
#> 31357 23.81728   158.92   TRUE         0.00              5  3.243543
#> 14254 15.66524   300.00   TRUE       300.00             10  2.510336
#> 30059 20.19218   300.00   TRUE       300.00             41  2.412719
#> 24051 32.73881   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59826   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99874   300.00   TRUE       300.00            100  3.212274
#> 48187 30.96665   279.54   TRUE        21.35             46  3.521349
#> 21939 23.93245   300.00   TRUE       300.00            100  3.168886
#> 8869  29.11027   300.00   TRUE       300.00            100  3.036974
#> 32622 30.17138   300.00   TRUE       300.00            100  3.265863
#> 16811 35.25237   300.00   TRUE       300.00             59  2.715521
#> 39629 47.54862   300.00   TRUE       300.00             85  3.366461
#> 10899 24.77412   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13893   300.00   TRUE       300.00             39  2.691790
#> 33792 51.20810   300.00   TRUE       300.00              5  2.596460
#> 30800 33.24817   300.00   TRUE       300.00            100  4.403435
#> 43007 38.96170   300.00   TRUE       300.00             37  2.473902
#> 40307 29.51363    77.82   TRUE        76.41             54  3.524078
#> 29420 28.85244   253.02   TRUE         0.00            100  3.954264
#> 22933 33.06310    87.56   TRUE       115.40              2  3.506583
#> 22566 24.58676   300.00   TRUE         7.49              0  2.426944
#> 15340 27.97524   300.00   TRUE       300.00            100  3.144334
#> 26215 23.83161   300.00   TRUE       300.00             10  3.268382
#> 34275 36.82915   300.00   TRUE       300.00            100  2.711427
#> 6808  28.56596   300.00   TRUE       300.00             40  3.952656
#> 40828 25.59975   300.00   TRUE       300.00            100  2.941037
#> 18391 34.50753   300.00   TRUE       300.00            100  5.556760
#> 20194 28.11733   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07590   300.00   TRUE       300.00            100  2.972687
#> 34565 20.38011   300.00   TRUE        68.13            100  2.749114
#> 2950  28.20385   300.00   TRUE       300.00            100  2.891568
#> 19128 14.65970   300.00   TRUE       300.00             15  2.478266
#> 15990 34.14886   300.00   TRUE         0.00              0  3.754443
#> 5728  40.40842   300.00   TRUE       300.00              2  3.325151
#> 47011 44.53429   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379   300.00   TRUE       300.00              2  2.819787
#> 19784 27.95576   300.00   TRUE       300.00            100  2.941780
#> 46201 22.18378   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946   300.00   TRUE       300.00             85  3.008718
#> 13699 34.78942   300.00   TRUE        53.95             43  3.288564
#> 17036 24.92825    25.33   TRUE       208.61             15  3.608265
#> 3944  39.36048   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57055   300.00   TRUE       300.00            100  3.329742
#> 39599 33.13472   300.00   TRUE       300.00             25  3.007839
#> 21952 40.09158   300.00   TRUE       300.00             64  2.828908
#> 5084  31.59518   300.00   TRUE       300.00             18  2.722101
#> 22299 21.71911   300.00   TRUE       127.54              2  2.615513
#> 375   27.97409   300.00   TRUE       300.00            100  2.851936
#> 42941 39.06312   300.00   TRUE       300.00             17  4.882627
#> 39520 35.89409   300.00   TRUE       300.00            100  2.934945
#> 27875 21.26705   157.93   TRUE         0.00             57  3.709413
#> 37416 28.77165    72.23   TRUE        87.56             31  3.407698
#> 19966 29.22887   300.00   TRUE       300.00             23  3.350329
#> 46783 19.94237   256.35   TRUE         0.00            100  2.772566
#> 10364 26.09307   179.84   TRUE         0.00              0  3.541090
#> 46065 29.77089   300.00   TRUE       251.11            100  3.482407
#> 11618 33.98556   300.00   TRUE       300.00              1  2.578737
#> 38197 34.34252   105.52   TRUE        65.32             50  3.036108
#> 24501 27.15247   300.00   TRUE       300.00              0  2.925363
#> 45500 34.95616   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21570   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862   300.00   TRUE       300.00              2  2.460907
#> 37714 30.73174    13.57   TRUE       170.00              4  3.480344
#> 46461 24.94143   288.47   TRUE         0.00            100  2.687311
#> 33910 32.67349   300.00   TRUE       300.00              6  6.024389
#> 20368 30.34442   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02642   300.00   TRUE       300.00             92  3.259412
#> 18440 34.94470   300.00   TRUE       300.00            100  4.477051
#> 6217  30.97181   300.00   TRUE       300.00              1  3.243763
#> 7     13.22043   300.00   TRUE       300.00              0  2.263215
#> 43201 41.82821   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451   300.00   TRUE       300.00              1  3.610558
#> 21683 19.74527   130.93   TRUE         0.00             21  3.399421
#> 33875 27.56958   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76764   300.00   TRUE       300.00            100  2.880810
#> 15981 30.00351   300.00   TRUE       300.00              0  2.799064
#> 26101 28.64617   300.00   TRUE       300.00            100  3.068405
#> 43074 33.59481   300.00   TRUE       205.02              1  3.153436
#> 47086 23.38699   300.00   TRUE       300.00            100  3.525504
#> 12629 32.18075   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88375   300.00   TRUE       300.00             59  3.185096
#> 13259 23.16067   203.85   TRUE        10.45             74  3.485288
#> 23626 26.77833   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710   300.00   TRUE         0.00              0  3.804807
#> 7409  25.37057   300.00   TRUE       300.00            100  2.866282
#> 34984 33.90592   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13016   300.00   TRUE       300.00            100  3.645966
#> 17825 30.73346    73.66   TRUE       137.63             18  3.469116
#> 30363 29.80297   300.00   TRUE       300.00            100  2.929656
#> 17227 43.59750   300.00   TRUE       300.00            100  2.467494
#> 24532 32.54629   300.00   TRUE         7.49            100  6.276647
#> 10213 31.59862   300.00   TRUE       300.00            100  3.351772
#> 14779 16.52124    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48729   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47316   300.00   TRUE       300.00            100  3.297274
#> 31491 27.55870   300.00   TRUE       300.00            100  3.385647
#> 49346 23.34574   300.00   TRUE       300.00             29  2.687835
#> 41310 17.29530   300.00   TRUE       300.00             25  3.046207
#> 31253 38.66835   300.00   TRUE       122.87            100  3.878395
#> 36744 29.40477   300.00   TRUE       300.00            100  2.937989
#> 4785  26.48612   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603   276.16  FALSE        91.33            100  3.227418
#> 29431 20.89176   300.00  FALSE       300.00             42  2.702050
#> 27884 15.49736   300.00  FALSE       300.00            100  2.781892
#> 25212 38.98290   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89274   300.00  FALSE       300.00            100  3.348210
#> 40459 36.18687   300.00  FALSE       216.66             10  3.025531
#> 8635  27.68188   300.00  FALSE       300.00             23  2.716617
#> 35852 36.36219   300.00  FALSE       300.00            100  3.356690
#> 18746 39.13646     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42977   300.00  FALSE       300.00             28  2.424012
#> 32869 14.60011   300.00  FALSE       300.00             78  2.867820
#> 37832 36.44642   300.00  FALSE       207.18              2  3.011485
#> 28162 18.66983   130.00  FALSE        75.00            100  3.197540
#> 42329 19.22789   300.00  FALSE       300.00            100  2.726581
#> 43243 27.98612   300.00  FALSE       300.00            100  2.922424
#> 47005 36.13530   300.00  FALSE       300.00            100  4.858077
#> 18627 31.25714   300.00  FALSE       300.00             20  3.545129
#> 31884 24.13757   300.00  FALSE       300.00            100  3.117018
#> 44668 30.10263     9.15  FALSE       163.30              6  2.852799
#> 38689 36.08431   300.00  FALSE       300.00             18  2.953242
#> 37029 28.89312   300.00  FALSE       237.21            100  2.916688
#> 8895  38.07763   300.00  FALSE       300.00             90  2.432641
#> 22545 23.68436   300.00  FALSE       300.00             96  2.215056
#> 34810 15.55466   300.00  FALSE       300.00            100  2.997789
#> 12519 30.48594   300.00  FALSE       120.00            100  3.087264
#> 35038 40.20388   300.00  FALSE       300.00            100  3.347416
#> 34532 31.20672   300.00  FALSE       300.00             92  2.583601
#> 41612 25.58715   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755   300.00  FALSE       185.65            100  3.274624
#> 24819 29.93704   300.00  FALSE       150.00            100  2.905872
#> 5861  27.61256   300.00  FALSE       300.00            100  4.032264
#> 29716 32.69354   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695   300.00  FALSE       300.00            100  2.450415
#> 48184 30.95634   300.00  FALSE       300.00             75  2.485742
#> 31098 34.65764   300.00  FALSE       300.00             45  3.277234
#> 42638 23.42137    86.41  FALSE       172.46            100  6.734528
#> 40750 32.09996   300.00  FALSE       300.00            100  3.272034
#> 16845 22.88279   300.00  FALSE        25.55             33  3.874467
#> 9320  18.13927   300.00  FALSE       300.00             93  2.700519
#> 33881 29.18590   300.00  FALSE       300.00            100  3.171879
#> 31440 24.03501   300.00  FALSE       300.00             96  2.654870
#> 49812 33.35818   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91071   300.00  FALSE       300.00            100  3.015637
#> 23808 37.53389   300.00  FALSE       300.00             72  2.692387
#> 13548 16.97846   300.00  FALSE       300.00            100  2.256336