Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634    500.5027
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023   -0.00043   1806.4552
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741    746.4155
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514   -0.00644   1179.0695
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453    0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440   1088.8639
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284    0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854    0.01264   5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770    474.5059
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607    0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102    0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189    287.5425
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272   4955.5586
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530   9000.4356
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203  10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426   -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043    0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993   -0.00413    364.9788
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987    0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664    0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593    939.0466
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276   9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916   2009.1686
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440   -0.02651    419.0927
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543   -0.05146   1347.6224
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255   -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904   -0.00866  29195.6016
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376   2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714    475.3120
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735   -0.00925    665.6116
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149    0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492    0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290    0.00230   1056.5876
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181   2406.1252
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175    0.00236   2715.4138
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708    0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283    590.1536
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458   -0.00401   1722.6481
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739    365.7626
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800   -0.06020   1563.5549
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201    788.0621
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165    0.00555   4439.0967
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469    0.02940    987.8085
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509   -0.01211  49347.5547
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860    695.2302
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439   -0.01161   3065.3694
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801   1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739    789.7573
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253    0.00753   1337.7854
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317    0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495   -0.03756    353.9086
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401   -0.00391   1794.7964
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153   -0.02417  34908.6562
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185    953.8614
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708   -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864    0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019    0.00119   6589.5796
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610    454.8103
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166   1813.7841
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675   -0.01905   6487.0010
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758   -0.00288   2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319    0.00291    487.3438
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345   2417.1389
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203    0.00642   1688.1782
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002   -0.01232    977.7320
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284    0.01085   2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162   -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156   -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463   1050.7473
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165    0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479    604.0975
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075   8141.7549
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171    756.8818
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014    271.4219
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327    0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476  11992.5273
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527   -0.02077   8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367   3002.8828
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990    0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846    0.00216   1412.2295
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844    0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170    253.5129
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674   -0.00376    629.8236
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422    967.9738
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292    0.00462    716.5762
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782   -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338    0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785    0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424    0.00196    576.1450
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112   -0.00038   2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305   -0.00705    968.9738
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118    0.00162    290.2031
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066    0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070   -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555   1882.0179
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750   1892.8978
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070   -0.00050   1471.4346
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361   -0.01841    756.0055
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148    481.8292
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108   -0.00948   1538.1741
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177    184.1042
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992   -0.01052    451.7209
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013    616.4128
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222    991.8300
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288    0.00448    941.8503
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751    0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366    0.00525    162.0592
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983