Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507 159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046    0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065    0.00005    769.3426
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250   -0.01340    604.1032
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529   3282.1186
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730    0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842    868.6137
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129    0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898    458.6938
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261    554.9243
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151    0.00899   1394.2274
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051    0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773    179.3077
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257    0.00593    939.0466
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950    950.3138
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077   3559.4204
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545    0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177    291.5937
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846   -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606   -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904   -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888    0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070   2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714    475.3120
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063  18733.8848
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492    0.00072  54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359   -0.00279    930.0871
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708   -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560    977.4719
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797   -0.00063    576.2819
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330   1456.9246
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436   2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299    688.0706
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883    616.1993
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262   1760.2162
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304   1569.7313
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406   -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800   -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558   1504.9717
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972    0.03512  13178.3223
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448   1258.8196
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205   1112.2626
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849   1120.5961
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384    0.00175   1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372   4154.5786
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161   3065.3694
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801   1071.6141
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879   1200.4210
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778   4594.4351
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044    288.5468
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166   1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902    0.01358    802.7622
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856   1789.1596
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867   2247.8748
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230   1044.6443
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403    474.8613
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519    0.00609   4275.3457
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559    866.5775
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085   2229.5105
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309   -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945    259.7421
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028   2394.7261
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014    271.4219
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527   -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065    704.4843
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267  31588.1953
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462    716.5762
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997    230.8641
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759   4131.5356
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018    937.6911
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552   2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282    561.2366
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511    0.01499    617.3855
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215    917.5296
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186   2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177    184.1042
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788   -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884    672.2466
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052    451.7209
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799   1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716   -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931   1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607