Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607