Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393   2081.0056
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261    323.8441
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339   1002.5877
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065   0.00005    769.3426
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345   1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547    519.4228
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783    377.8918
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269    491.8017
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382    560.8309
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029  -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382   0.00672   1114.8801
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973    514.5488
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108  -0.00678   1891.6584
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325    616.8652
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276   9614.6924
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086   6042.5757
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294  -0.00096    520.5570
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556    660.3690
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492   0.00072  54831.6289
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395   0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964    571.1364
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560    977.4719
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063    576.2819
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543    352.1558
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454   3476.0818
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596    842.0994
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240   1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555   4439.0967
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242    507.9263
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118  19557.6621
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484    495.0292
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001   1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153  -0.02417  34908.6562
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325    595.0426
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226    338.3474
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291    487.3438
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842    246.5161
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298  -0.00712   1010.5097
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715    520.7355
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875   1165.3953
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409   1575.9425
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199    836.4196
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697    412.1185
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011    712.5228
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298    436.1459
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282   2322.7944
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342   0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513    161.9878
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105    868.7802
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820  -0.00260   1484.6593
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964    543.0025
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282   1659.6211
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233    968.7780
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884    672.2466
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803   1058.5630
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087  -0.00573   1490.0312
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156    758.4840
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983