Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813   1130.5791
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507 159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005    769.3426
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588   1295.1610
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311    586.7354
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281   1325.6138
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312   2657.6936
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114   4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605    798.6775
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338  10787.6299
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565   2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436   2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299    688.0706
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026    247.0894
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512  13178.3223
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596    842.0994
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220   8936.2666
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813 287033.4688
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957    495.4453
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867   2247.8748
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403    241.2991
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 41753 715722.5 9558192 2771.39 26.085050  -0.00196   -0.00175    725.1355
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098    164.0802
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011    201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518   1137.0179
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947    842.1514
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499    617.3855
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282   1659.6211
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148    481.8292
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458    485.7418
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674