Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393    2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438     753.6304
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507  159065.0781
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057    1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949    6217.8213
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491    1844.4353
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088     266.4355
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311     586.7354
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898     458.6938
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131   0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385     674.0721
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899    1394.2274
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773     179.3077
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156     816.6036
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924    8023.8657
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651     419.0927
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556     660.3690
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511  361979.2812
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321   0.01169     350.1799
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907    1943.4092
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307  147470.5312
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964     571.1364
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117    1018.2141
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241     585.6182
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542    3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684    2556.8049
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543     352.1558
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109     767.4097
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316     454.9674
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512   13178.3223
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596     842.0994
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851    1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081     940.6002
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940     987.8085
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822     472.8748
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213     328.7922
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400     608.6523
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061    5744.7583
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728    1137.2450
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175    1026.6818
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801     172.7491
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298    1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456    1531.4274
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915    2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109    2047.9031
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294     734.9908
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171    4425.5410
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291     293.4231
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036    2505.8369
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603    1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403     474.8613
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067    1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331     699.3024
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493    2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328    1066.7134
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466    2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479     604.0975
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463     851.6306
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500     538.8209
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598     186.3527
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077    8759.9961
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035    1471.9783
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400    4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152   0.00098     164.0802
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018     937.6911
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955     713.4463
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250    2225.4426
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547     128.2613
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555    1882.0179
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577     423.5598
#> 28286 713702.5 9559842 2245.07 27.968935    0.00511   0.01499     617.3855
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208     842.4825
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972     331.7462
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809    1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011     501.7867
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172     351.5812
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241   0.00020     673.2861
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403     609.8665
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661    5642.7544
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592   300.00   TRUE       300.00            100  2.699406
#> 15412 37.59291   300.00   TRUE       300.00            100  3.318273
#> 36052 53.74516   300.00   TRUE       300.00            100  2.877158
#> 40756 39.33527   300.00   TRUE       300.00            100  3.000471
#> 5999  27.14216   300.00   TRUE       300.00            100  5.201575
#> 11088 30.99874   300.00   TRUE       300.00            100  3.212274
#> 34512 34.63988   300.00   TRUE       195.00              2  3.224136
#> 42917 38.80357   300.00   TRUE       300.00            100  3.010367
#> 38411 32.75829   279.54   TRUE         0.00            100  3.793638
#> 32622 30.17138   300.00   TRUE       300.00            100  3.265863
#> 125   29.59671   207.04   TRUE        21.03            100  3.298701
#> 16811 35.25237   300.00   TRUE       300.00             59  2.715521
#> 23817 21.61140   300.00   TRUE       300.00              4  2.425592
#> 7349  26.32283   300.00   TRUE       300.00              8  2.751808
#> 11463 20.13374   300.00   TRUE       300.00            100  2.768442
#> 43007 38.96170   300.00   TRUE       300.00             37  2.473902
#> 22283 19.94237   256.35   TRUE         0.00            100  2.772566
#> 2107  35.42369   300.00   TRUE       300.00             45  2.661523
#> 15109 22.63527   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843   300.00   TRUE       300.00             57  3.589919
#> 30125 40.73558   300.00   TRUE       300.00            100  2.748832
#> 22933 33.06310    87.56   TRUE       115.40              2  3.506583
#> 38032 35.00829   300.00   TRUE       300.00             60  2.828706
#> 20290 15.45668    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19965   300.00   TRUE         0.00            100  2.926238
#> 15340 27.97524   300.00   TRUE       300.00            100  3.144334
#> 27708 25.71091   300.00   TRUE       300.00            100  2.253599
#> 7879  28.58085   300.00   TRUE       300.00            100  2.912011
#> 12084 34.59576   300.00   TRUE         0.00              0  3.904384
#> 22735 36.51689   300.00   TRUE       300.00             90  2.622310
#> 47429 32.09767   300.00   TRUE       300.00            100  2.464778
#> 37828 18.76379   300.00   TRUE       300.00              2  2.819787
#> 37348 39.28772   300.00   TRUE       300.00            100  4.465317
#> 24100 34.29783   300.00   TRUE       300.00            100  5.558684
#> 4605  17.63564   300.00   TRUE       300.00             10  2.544291
#> 13699 34.78942   300.00   TRUE        53.95             43  3.288564
#> 22332 27.06882   300.00   TRUE       300.00            100  5.168705
#> 33240 29.35492   300.00   TRUE       300.00            100  2.756740
#> 39599 33.13472   300.00   TRUE       300.00             25  3.007839
#> 26698 34.24970   300.00   TRUE       300.00              7  2.767615
#> 30406 43.10075   300.00   TRUE       300.00             85  2.770965
#> 2595  30.43953   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917   118.07   TRUE        98.15              0  3.619059
#> 47100 29.64598    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77165    72.23   TRUE        87.56             31  3.407698
#> 7668  36.32037   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078   300.00   TRUE       300.00             23  2.546735
#> 32280 27.83142   126.66   TRUE         0.00              5  3.195825
#> 664   30.84232   300.00   TRUE       300.00             12  2.885027
#> 16355 28.24682   300.00   TRUE        58.52            100  2.657980
#> 13146 33.57647   300.00   TRUE       300.00              2  3.278424
#> 16370 31.15859   300.00   TRUE       300.00              8  4.119860
#> 24501 27.15247   300.00   TRUE       300.00              0  2.925363
#> 45500 34.95616   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21570   300.00   TRUE       300.00            100  3.005436
#> 41783 28.18609   300.00   TRUE       300.00            100  2.973405
#> 7385  28.71034   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73174    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844   300.00   TRUE         1.90             90  2.994673
#> 18958 29.98976   300.00   TRUE       300.00              5  2.674746
#> 11612 35.16700   300.00   TRUE       300.00            100  2.516921
#> 47320 45.66416   300.00   TRUE       300.00             43  2.536741
#> 32879 29.73823   300.00   TRUE       300.00              6  3.130450
#> 10467 38.99608   300.00   TRUE       300.00             90  2.784369
#> 19892 27.02642   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341   300.00   TRUE        10.00              1  3.759272
#> 30287 19.84840   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060   300.00   TRUE       300.00            100  3.055854
#> 34615 34.18954    17.57   TRUE       142.95             16  3.011436
#> 15981 30.00351   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81368   300.00   TRUE       205.28              1  3.147505
#> 10689 30.64809   300.00   TRUE       300.00             61  2.237416
#> 28009 27.01782   300.00   TRUE       300.00             29  3.185767
#> 47086 23.38699   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99816   135.89   TRUE        82.11              4  4.291317
#> 12629 32.18075   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88375   300.00   TRUE       300.00             59  3.185096
#> 2801  25.74643   300.00   TRUE       300.00             26  2.422656
#> 22297 46.64965   300.00   TRUE       300.00             11  2.548891
#> 22301 33.61543   300.00   TRUE         0.00            100  3.308786
#> 1916  29.27356   300.00   TRUE       300.00             21  3.311309
#> 7409  25.37057   300.00   TRUE       300.00            100  2.866282
#> 34984 33.90592   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13016   300.00   TRUE       300.00            100  3.645966
#> 42640 34.90975   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00351   300.00   TRUE       300.00              0  2.799064
#> 17227 43.59750   300.00   TRUE       300.00            100  2.467494
#> 10213 31.59862   300.00   TRUE       300.00            100  3.351772
#> 14140 35.48729   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78423   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90610   300.00   TRUE       300.00            100  3.398953
#> 11332 25.68169   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516   300.00   TRUE       300.00              8  2.676567
#> 4785  26.48612   300.00  FALSE         0.00            100  3.050663
#> 15211 30.83487   300.00  FALSE       300.00            100  2.844665
#> 15186 38.54860    60.00  FALSE         0.00            100  3.149477
#> 24836 25.58371   300.00  FALSE       300.00             95  2.679300
#> 2664  39.29115   300.00  FALSE       300.00            100  2.881694
#> 40459 36.18687   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29534   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38396   300.00  FALSE       300.00            100  3.344795
#> 14717 50.85631   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13646     4.48  FALSE       205.00             27  3.028048
#> 26963 42.06312   300.00  FALSE       300.00            100  3.420610
#> 32869 14.60011   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030   300.00  FALSE       300.00            100  2.781107
#> 1488  29.92157   300.00  FALSE       300.00            100  2.930251
#> 12716 24.34498   300.00  FALSE       300.00             25  2.731444
#> 16878 25.38891   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237   300.00  FALSE       300.00            100  2.270336
#> 36833 28.07550   187.46  FALSE        18.49             45  3.942504
#> 31884 24.13757   300.00  FALSE       300.00            100  3.117018
#> 20319 32.07246   300.00  FALSE       300.00            100  3.167901
#> 36245 25.85357   110.19  FALSE        81.50             20  3.615611
#> 22545 23.68436   300.00  FALSE       300.00             96  2.215056
#> 34345 44.39506   300.00  FALSE       300.00            100  2.972060
#> 5656  31.27261   300.00  FALSE        94.43            100  2.853361
#> 36721 26.66832   300.00  FALSE       300.00            100  2.209482
#> 35038 40.20388   300.00  FALSE       300.00            100  3.347416
#> 36889 19.64500   300.00  FALSE       300.00             48  2.108096
#> 35826 44.22776   298.23  FALSE        16.35             96  2.958591
#> 48845 31.55909   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58715   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755   300.00  FALSE       185.65            100  3.274624
#> 5861  27.61256   300.00  FALSE       300.00            100  4.032264
#> 14671 37.88225   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204   300.00  FALSE       205.59            100  2.626915
#> 28286 17.04836   300.00  FALSE       300.00            100  2.790556
#> 31098 34.65764   300.00  FALSE       300.00             45  3.277234
#> 42638 23.42137    86.41  FALSE       172.46            100  6.734528
#> 24446 31.36314   300.00  FALSE       300.00            100  2.925561
#> 48050 23.36121   300.00  FALSE        72.23            100  2.520806
#> 7132  34.04916   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16642   300.00  FALSE       300.00            100  5.155983
#> 20950 32.71188   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53826   300.00  FALSE       300.00             75  2.484816
#> 41873 11.03860   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242   300.00  FALSE       300.00            100  3.063725
#> 9320  18.13927   300.00  FALSE       300.00             93  2.700519
#> 23808 37.53389   300.00  FALSE       300.00             72  2.692387
#> 15880 48.89679   300.00  FALSE       300.00            100  2.546026
#> 11025 36.81999   300.00  FALSE       300.00             50  2.828200
#> 46567 16.94809   300.00  FALSE       300.00             78  2.785235
#> 40852 21.51800   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25448   300.00  FALSE       122.13            100  3.751491