Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309  -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.2471896  -0.00501   -0.00649   1399.2329
#> 37912 712802.5 9559952 1838.40 52.1013441   0.00183   -0.09203    634.3320
#> 31357 714752.5 9561022 1848.22 33.4464113  -0.00347    0.02357   1752.0375
#> 28223 712742.5 9560482 1927.69 34.6679573  -0.00292    0.00712   3636.5303
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487   1077.7585
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896    276.0310
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.8482293   0.00046    0.00324    861.3537
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767   3321.6113
#> 38411 713552.5 9560652 1851.63 46.0050732  -0.00891   -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.6924129  -0.00645    0.00835   2425.3115
#> 37270 714852.5 9561032 1838.96 35.1589821  -0.00854    0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088    266.4355
#> 38993 714202.5 9558402 2475.75 33.1593594  -0.01607    0.00677   1343.5006
#> 38410 713712.5 9561172 1776.08 36.3192217  -0.02678   -0.04272   4955.5586
#> 40307 713302.5 9560702 1845.07 39.0046749   0.00129    0.02510   3342.5518
#> 15109 713872.5 9559602 2288.64 25.1855058   0.00602    0.01518    854.4269
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677   1675.4679
#> 31576 713942.5 9558492 2387.12 46.1431560  -0.05634   -0.02626   4316.0977
#> 35942 712862.5 9558902 2109.97 31.4003790   0.00151    0.00189   2317.1819
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733    598.7983
#> 3453  713482.5 9559642 2145.11 35.1257506   0.00108   -0.00678   1891.6584
#> 27708 715222.5 9557782 2844.65 29.5428498   0.03206    0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.0224015  -0.00987    0.02107   1392.5880
#> 38077 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733    598.7983
#> 28230 715182.5 9558102 2791.83 39.0717109  -0.00669   -0.00950    950.3138
#> 17004 713102.5 9559032 2186.79 36.5335079  -0.01383    0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308  -0.00064   -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.2317267  -0.00193   -0.00077   3559.4204
#> 12868 712532.5 9560332 1913.92 61.9373106  -0.03543   -0.05146   1347.6224
#> 6967  715342.5 9560192 1995.36 44.8815030  -0.00666   -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.8331648  -0.00442   -0.00178    800.1933
#> 29926 714702.5 9557622 2637.00 44.6425796   0.01843    0.04407    344.8395
#> 37341 714412.5 9559142 2377.24 40.5997894  -0.00735   -0.00925    665.6116
#> 33240 714972.5 9557652 2741.04 39.4343932  -0.00286   -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.5469237   0.03480    0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.9680266   0.00107    0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.7968588  -0.00449   -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.7099578   0.00478    0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560    977.4719
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.9506071  -0.00762    0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.3121575  -0.00405   -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.7170711  -0.03260   -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.3665885  -0.01017    0.00197    711.1091
#> 10507 714892.5 9559282 2374.45 42.8423462  -0.01057    0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.8562718  -0.00458   -0.00401   1722.6481
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009   1247.7061
#> 31462 714082.5 9558542 2367.84 34.7269720  -0.02027   -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.9349627  -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.2195200  -0.00034    0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170  -0.02554    0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.2748872  -0.00845    0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.5277783   0.00406   -0.00055    773.0359
#> 11618 713852.5 9558402 2432.14 54.0591409  -0.00410    0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.2791006   0.00645   -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.0079450  -0.00078    0.00219    420.7032
#> 26010 712852.5 9560542 1928.25 30.9781091   0.00165    0.00555   4439.0967
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.9506633   0.03743   -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.2403562  -0.01169   -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536  -0.01556   -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.6771246  -0.00178   -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.1293630   0.00370   -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.9767745  -0.00278   -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.2237194  -0.03986   -0.02524  10861.9482
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781  29995.1543
#> 18348 714042.5 9558882 2341.41 44.0638922  -0.00338    0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.9135886  -0.01259   -0.00281   1178.4420
#> 28312 715712.5 9558032 2783.23 24.2527305  -0.00007   -0.02132   1580.1621
#> 10689 713542.5 9559132 2211.95 31.9813582   0.05379    0.00801    172.7491
#> 47086 713692.5 9559972 2175.51 25.0457041  -0.00789   -0.00261   3353.5422
#> 37439 714602.5 9560672 1982.99 18.9064613  -0.00070    0.00010   3467.2690
#> 19756 715572.5 9558972 2485.95 38.8127340   0.00546    0.01484    495.0292
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981   3975.3608
#> 10692 712712.5 9561022 2005.85 41.4334430   0.01410   -0.00380    601.0843
#> 48574 712692.5 9560412 1901.97 41.9783258   0.01364   -0.00094   1392.6008
#> 14950 713382.5 9559172 2111.91 48.1066824   0.00915   -0.02705    413.5702
#> 28260 715312.5 9558302 2681.75 34.4840379  -0.01125   -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875    597.2893
#> 49459 714692.5 9557342 2639.74 42.0470808   0.00216   -0.02606    919.4241
#> 30626 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245    629.5983
#> 14140 714672.5 9559262 2326.36 45.7925059  -0.01400   -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.9548275  -0.00770   -0.03400   1982.7782
#> 33142 715652.5 9558212 2740.94 14.3812407  -0.00465   -0.01036   2505.8369
#> 41310 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205   1112.2626
#> 28253 714102.5 9559892 2288.88  2.9576081   0.00378    0.02382    100.0000
#> 11288 712822.5 9560442 1873.42 39.6097183  -0.00465   -0.00105   4432.8760
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736    477.8591
#> 2664  713342.5 9558572 2275.86 56.9754961   0.00872    0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.4317521   0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.1532385   0.00156   -0.00116   1817.5222
#> 14694 715032.5 9559482 2307.44 44.1733271  -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761   0.02295    0.02715    520.7355
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538    265.4677
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.7174203  -0.00494    0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.0369281  -0.03220   -0.03290   2481.5632
#> 441   714312.5 9560972 1916.59 31.3029762   0.00540    0.00650   1133.1501
#> 42329 715852.5 9558702 2651.92 22.8707563   0.00321    0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923    811.5580
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171    756.8818
#> 44396 715512.5 9558102 2845.31 26.5984197   0.01263    0.03598    186.3527
#> 4007  714132.5 9558412 2464.44 43.7201175   0.00174   -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759   1880.4403
#> 33175 715822.5 9558832 2563.13 36.6228893  -0.00674   -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126  -0.00452   -0.01298    436.1459
#> 1691  714222.5 9559022 2448.69 38.2168579  -0.01292    0.00462    716.5762
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.6559462   0.00379   -0.01018    937.6911
#> 41855 713182.5 9560452 1851.53 45.1782951   0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.9887513  -0.00270 1018.51947    842.1514
#> 48988 714682.5 9560852 1942.97 20.2907274   0.00480    0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479   0.00972    0.00308    422.5022
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836    250.5111
#> 484   713812.5 9560592 2016.41 44.3010967  -0.00719    0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.1135298   0.02236   -0.00066    306.0146
#> 12300 712372.5 9560062 1903.64 50.0369772   0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.4446854  -0.00070   -0.00050   1471.4346
#> 40750 712962.5 9558702 2126.06 34.7040537  -0.00264    0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875   2156.9358
#> 4269  714162.5 9560572 2087.45  7.3344327   0.00293    0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.8113784  -0.04770    0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.4521619  -0.01196   -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474   0.00340    0.01809   1158.0426
#> 22628 714942.5 9557732 2696.19 26.2116732   0.00788   -0.01029    894.2672
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803   1058.5630
#> 23808 715172.5 9558702 2533.50 22.3459270   0.00811   -0.02281    492.4785
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253   1237.0536
#> 30296 713322.5 9559782 2042.96 52.6210169  -0.02110   -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250