Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501   -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642     671.1807
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921   -0.12438     753.6304
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266    0.02896     276.0310
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257     357.1436
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542     674.0147
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088     266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766  140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164     895.6606
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129    0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905   -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151    0.00899    1394.2274
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083     994.7876
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276    9614.6924
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275    0.00605     798.6775
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904   -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939     703.7817
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229   -0.00511  361979.2812
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169     350.1799
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907    1943.4092
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925     665.6116
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194    0.00516     576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436    2602.7390
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845    0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406   -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927   -0.00316     454.9674
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240    1201.4724
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081     940.6002
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399    3900.3445
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448    1258.8196
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393    8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469    0.02940     987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567   -0.03793     864.9472
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534     344.1447
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129    0.00171     788.0004
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574     183.3220
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797    7557.7866
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334    4905.2002
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007   -0.02132    1580.1621
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123     364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484     495.0292
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181   -0.00001    1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094    1392.6008
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019    0.00119    6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675   -0.01905    6487.0010
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758   -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000    0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427     866.9398
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842     246.5161
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642    1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150     503.5590
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085    2229.5105
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738    0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493    2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466    2633.9666
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463     851.6306
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290    2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500     538.8209
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306   -0.00626     450.7346
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794    1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502    1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338   -0.00399    1601.4852
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236    0.00483    1097.3082
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098     164.0802
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130     107.5831
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257     549.3252
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599   12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740     909.4791
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980   -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635   -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577     423.5598
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066     306.0146
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598    0.02208     842.4825
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948    1538.1741
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177     184.1042
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011     501.7867
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157     766.0220
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456   -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491