Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976     301.2347
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555     430.4604
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043    1806.4552
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896     276.0310
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503   0.02433    1475.3191
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046    1024.1573
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804     628.5153
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783     377.8918
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416     947.4878
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665   25318.3555
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530    9000.4356
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626    4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114    4510.8833
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314    2360.1501
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694    2123.4473
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437    1155.1469
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916    2009.1686
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743   0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696   0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566  -0.03355   14919.6260
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866   29195.6016
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373   16031.7500
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174   0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907    1943.4092
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279     930.0871
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779    2972.0193
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400     608.6523
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092     412.5849
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241     585.6182
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262    1760.2162
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109     767.4097
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620     379.0855
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240    1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851    1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599    1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354    1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594   16587.9980
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242     507.9263
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190     640.8842
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860     695.2302
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195    1901.6743
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801     172.7491
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169   0.01441     424.2680
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118   19557.6621
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.696018    0.02281   0.04759     158.9370
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401  -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153  -0.02417   34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294     734.9908
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121    1783.0658
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143   0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181  -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632     620.0229
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036    2505.8369
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403     474.8613
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265    1148.6487
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842     246.5161
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105    4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736     477.8591
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403     241.2991
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452     578.5847
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290    2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445     355.5078
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945     259.7421
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923     811.5580
#> 47005 714942.5 9559192 2333.66 29.220848   -0.04937  -0.02273   72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973     215.5101
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833     300.0125
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507     504.3085
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065     704.4843
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175     725.1355
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292   0.00462     716.5762
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759    4131.5356
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714  480535.5000
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705     968.9738
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513     161.9878
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099     909.0560
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282    1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368    1071.8734
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233     968.7780
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709    1181.2555
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052     451.7209
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627  -0.00124   25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087  -0.00573    1490.0312
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388     180.4414
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181  -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607