Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644   1179.0695
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890   2301.0454
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491   1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378    440.9104
#> 22283 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355  14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096    520.5570
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063   5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565   2232.6133
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739    365.7626
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809   1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213    328.7922
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337    486.7551
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728   1137.2450
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141   4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372   4154.5786
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450   2589.9910
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144 317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559    866.5775
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500    538.8209
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923    811.5580
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 21925 714542.5 9559232 2314.33 50.247826   0.00737    0.01033   3446.5520
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267  31588.1953
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513    161.9878
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105    868.7802
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499    617.3855
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578   1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875   2156.9358
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805   7489.7417
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119    792.3544
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491