Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555     430.4604
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250    1413.3312
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503    0.02433    1475.3191
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102    0.00388    1051.4222
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979   -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530    9000.4356
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312    2657.6936
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398     267.2659
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710     843.7961
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973     514.5488
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108   -0.00678    1891.6584
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774     527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257    0.00593     939.0466
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422    0.02357     275.7474
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275    0.00605     798.6775
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991    1403.0193
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517    1000.7119
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613     289.6113
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949    1182.1506
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230    1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560     977.4719
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252    5761.3486
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797   -0.00063     576.2819
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057    0.02427     380.9472
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262    1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026     247.0894
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099    0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845    0.00454    3476.0818
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620     379.0855
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596     842.0994
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013    4229.5708
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354    1147.7384
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664    1807.2659
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394     587.8840
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017    0.01213     328.7922
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534     344.1447
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129    0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491    0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281    1178.4420
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384    0.00175    1026.6818
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739     789.7573
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484     495.0292
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778    4594.4351
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181   -0.00001    1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401   -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075    2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171    4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044     288.5468
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606     919.4241
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675   -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015     790.2421
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758   -0.00288    2144.8164
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655     130.9468
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265    1148.6487
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642    1688.1782
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565    1410.8372
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403     241.2991
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328    1066.7134
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875    1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064     683.2795
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995    1026.7981
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853    1354.6001
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338   -0.00399    1601.4852
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697     412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236    0.00483    1097.3082
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223    1056.3701
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048    1550.6704
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654    1576.2065
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955     713.4463
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257     549.3252
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650     766.6254
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511    0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607   -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340    0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709    1181.2555
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585    5714.3267
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419     323.0790
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388     180.4414
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249     148.1857
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775