Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393    2081.0056
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339    1002.5877
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644    1179.0695
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046    1024.1573
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035    2697.5974
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378     440.9104
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164     895.6606
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530    9000.4356
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501    3889.7292
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899    1394.2274
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114    4510.8833
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450     873.0448
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774     527.1068
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616  0.01276    9614.6924
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499     885.0064
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025    4425.5151
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063   18733.8848
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072   54831.6289
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230    1056.5876
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236    2715.4138
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542     860.8840
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427     380.9472
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090 -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883     616.1993
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558    1504.9717
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224     942.0322
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469    1021.5407
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180    3904.5254
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112     745.3315
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394     587.8840
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241    1752.9242
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860     695.2302
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372    4154.5786
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776    1021.8267
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812    1423.7578
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363     565.5207
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422    1159.2100
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380     601.0843
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417   34908.6562
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294     734.9908
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705     413.5702
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685    2568.4631
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606     919.4241
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.5225656 -0.01454  0.00144  317260.6875
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609    4275.3457
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336    2806.2644
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105    4432.8760
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181    1286.3949
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493    2212.0525
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466    2633.9666
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479     604.0975
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463     851.6306
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923     811.5580
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833     300.0125
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035    1471.9783
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997     230.8641
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032     994.9224
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955     713.4463
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247     282.1076
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499     617.3855
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050    1471.4346
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532   16771.3066
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215     917.5296
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624    1870.8307
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875    2156.9358
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148     481.8292
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935     903.6397
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377    1727.6055
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029     894.2672
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119     792.3544
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052     451.7209
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234    1707.3130
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135    1293.4855
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403     609.8665
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235