Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501  -0.00649   1399.2329
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097   0.01483    566.7674
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741    746.4155
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588   1295.1610
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088    266.4355
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024    394.8751
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769   1032.7177
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131   0.00501   3889.7292
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203  10263.2734
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889    362.8363
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437   1155.1469
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193   1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346    543.0749
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862    874.5408
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060   0.00270   4057.5637
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949   1182.1506
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964    571.1364
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917  -0.01993  10308.6406
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330   1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409   2312.8413
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473   2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026    247.0894
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454   3476.0818
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599   1378.5760
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940    987.8085
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728   1137.2450
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878  -0.04351   1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261   1627.8989
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484    495.0292
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957    495.4453
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109   2047.9031
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267   0.00483    343.1384
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867   2247.8748
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063   2430.2278
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306   1718.3652
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000   0.00320    911.5911
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599   0.00301    338.9676
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505    351.4078
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433  -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105   4432.8760
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403    241.2991
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538    265.4677
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132    737.5982
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367   3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011    201.1272
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038   2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452  -0.00948   1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705    968.9738
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599  12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952  -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805   7489.7417
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377   1727.6055
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172  -0.00119    792.3544
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052    451.7209
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172    351.5812
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569   300.00   TRUE       300.00             21  3.145890
#> 36052 53.74516   300.00   TRUE       300.00            100  2.877158
#> 44302 35.10111   300.00   TRUE       300.00            100  3.053301
#> 28223 29.02260    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18304   300.00   TRUE       300.00            100  2.753405
#> 29391 28.28349   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292   300.00   TRUE       300.00            100  2.872981
#> 34512 34.63988   300.00   TRUE       195.00              2  3.224136
#> 42917 38.80357   300.00   TRUE       300.00            100  3.010367
#> 38411 32.75829   279.54   TRUE         0.00            100  3.793638
#> 27973 30.00981   300.00   TRUE       300.00             85  2.740390
#> 49372 23.54054   213.54   TRUE        20.21             68  3.430977
#> 47255 19.67766   300.00   TRUE       260.03            100  3.112324
#> 23817 21.61140   300.00   TRUE       300.00              4  2.425592
#> 38287 14.89690    41.43   TRUE         1.90            100  3.349380
#> 25690 21.74776   300.00   TRUE       300.00            100  2.732242
#> 21703 22.26915   300.00   TRUE       300.00             34  2.458702
#> 33792 51.20810   300.00   TRUE       300.00              5  2.596460
#> 40307 29.51363    77.82   TRUE        76.41             54  3.524078
#> 16843 21.39825   300.00   TRUE       300.00             12  3.013982
#> 18304 44.90843   300.00   TRUE       300.00             57  3.589919
#> 38032 35.00829   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676   300.00   TRUE         7.49              0  2.426944
#> 31576 26.93532   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668    95.53   TRUE         0.00             49  4.011286
#> 4827  21.07396   300.00   TRUE       300.00              0  2.559711
#> 4477  32.16986   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733   300.00   TRUE       300.00             41  3.143823
#> 13847 20.57090   300.00   TRUE       300.00            100  3.062637
#> 17004 27.18914   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29171   300.00   TRUE       300.00            100  2.902371
#> 35142 27.15304   300.00   TRUE       300.00            100  2.176525
#> 22735 36.51689   300.00   TRUE       300.00             90  2.622310
#> 19128 14.65970   300.00   TRUE       300.00             15  2.478266
#> 12868 37.58832   119.34   TRUE        79.16             61  3.129568
#> 29667 34.25199   300.00   TRUE       300.00             27  2.734860
#> 47011 44.53429   300.00   TRUE       300.00             10  3.652797
#> 19784 27.95576   300.00   TRUE       300.00            100  2.941780
#> 27084 21.67270   300.00   TRUE       300.00             81  3.000309
#> 18564 21.70020   300.00   TRUE       300.00            100  2.676979
#> 17036 24.92825    25.33   TRUE       208.61             15  3.608265
#> 651   45.52436   300.00   TRUE       300.00            100  3.072673
#> 33240 29.35492   300.00   TRUE       300.00            100  2.756740
#> 5084  31.59518   300.00   TRUE       300.00             18  2.722101
#> 835   36.65039   300.00   TRUE       300.00            100  3.141386
#> 36120 31.56310   300.00   TRUE       300.00             94  4.013201
#> 18892 27.06309   235.02   TRUE        27.24              0  2.675293
#> 42941 39.06312   300.00   TRUE       300.00             17  4.882627
#> 7466  28.79056   255.23   TRUE         0.00             60  3.163437
#> 27875 21.26705   157.93   TRUE         0.00             57  3.709413
#> 32909 28.37917   118.07   TRUE        98.15              0  3.619059
#> 39732 29.16814   300.00   TRUE       300.00             25  3.364146
#> 31462 33.75466   300.00   TRUE       300.00              1  3.466909
#> 23213 25.46053   300.00   TRUE       300.00            100  2.868983
#> 44825 29.02146   300.00   TRUE       300.00            100  2.392854
#> 32280 27.83142   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307   179.84   TRUE         0.00              0  3.541090
#> 664   30.84232   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44278   300.00   TRUE       300.00              0  2.888200
#> 17728 28.55221    30.14   TRUE       140.15              2  3.303483
#> 5103  34.91318   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15247   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59115   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95616   300.00   TRUE         0.00              2  3.626296
#> 40270 27.59766   300.00   TRUE       300.00              5  3.139431
#> 41721 29.97944   300.00   TRUE       300.00            100  3.099963
#> 39166 29.81844   300.00   TRUE         1.90             90  2.994673
#> 46461 24.94143   288.47   TRUE         0.00            100  2.687311
#> 10073 25.76133   300.00   TRUE       300.00            100  2.896526
#> 17345 29.82131   300.00   TRUE       300.00              2  4.693266
#> 16009 33.41318   300.00   TRUE       300.00            100  3.049449
#> 18348 41.01060   300.00   TRUE       300.00            100  3.055854
#> 4588  31.66451   300.00   TRUE       300.00              1  3.610558
#> 34887 42.93574   300.00   TRUE       300.00            100  3.159244
#> 21683 19.74527   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475   300.00   TRUE       300.00              6  3.211627
#> 19832 36.63378   300.00   TRUE       300.00            100  3.278299
#> 17058 26.32283   300.00   TRUE       300.00              8  2.751808
#> 47086 23.38699   300.00   TRUE       300.00            100  3.525504
#> 28173 23.88375   300.00   TRUE       300.00             59  3.185096
#> 19756 30.11638   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09350   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79821    17.89   TRUE       145.83              4  3.064162
#> 2801  25.74643   300.00   TRUE       300.00             26  2.422656
#> 49133 23.51934   300.00   TRUE       300.00             38  2.694996
#> 4667  31.51726    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27356   300.00   TRUE       300.00             21  3.311309
#> 14494 21.18511   300.00   TRUE       300.00             28  2.535469
#> 30626 30.00351   300.00   TRUE       300.00              0  2.799064
#> 10213 31.59862   300.00   TRUE       300.00            100  3.351772
#> 35458 37.47316   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151   300.00   TRUE       300.00             47  5.501416
#> 31491 27.55870   300.00   TRUE       300.00            100  3.385647
#> 49346 23.34574   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81368   300.00   TRUE       205.28              1  3.147505
#> 17322 31.66509    75.77   TRUE       138.35              1  3.235115
#> 6470  28.09327   300.00   TRUE       300.00              2  2.959800
#> 36744 29.40477   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94470   300.00   TRUE       300.00            100  4.477051
#> 32405 15.76207   300.00  FALSE       300.00            100  2.530158
#> 46165 14.81841   300.00  FALSE       300.00            100  2.545811
#> 48601 35.85799   300.00  FALSE       245.19            100  2.998830
#> 11288 28.53731   132.07  FALSE        85.50              1  3.646686
#> 15558 50.34408   300.00  FALSE       300.00             45  2.657927
#> 18286 24.35071   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29534   300.00  FALSE       300.00            100  2.382556
#> 26023 45.07287   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42977   300.00  FALSE       300.00             28  2.424012
#> 32869 14.60011   300.00  FALSE       300.00             78  2.867820
#> 28162 18.66983   130.00  FALSE        75.00            100  3.197540
#> 13512 35.93477   300.00  FALSE       300.00            100  3.605261
#> 34347 31.72009   300.00  FALSE       300.00            100  3.477538
#> 12974 37.33851   300.00  FALSE       300.00            100  2.711947
#> 38689 36.08431   300.00  FALSE       300.00             18  2.953242
#> 29527 36.72659   300.00  FALSE       300.00            100  2.639632
#> 33956 38.59558   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732   300.00  FALSE       185.58            100  2.985864
#> 34105 20.07701   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703   300.00  FALSE       300.00            100  2.303471
#> 29107 25.23077   300.00  FALSE       300.00            100  2.550653
#> 12044 32.23690   300.00  FALSE       300.00            100  3.304782
#> 12519 30.48594   300.00  FALSE       120.00            100  3.087264
#> 33410 32.87976   300.00  FALSE       300.00            100  2.986312
#> 7983  38.50620   300.00  FALSE        57.15             22  2.884583
#> 49831 29.83792   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70175   234.40  FALSE        47.05            100  3.040739
#> 36889 19.64500   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820   234.82  FALSE         0.00             43  4.109707
#> 48988 18.95860   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55909   169.72  FALSE         0.00            100  2.030594
#> 17628 26.99204   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695   300.00  FALSE       300.00            100  2.450415
#> 34336 32.45118   300.00  FALSE       300.00            100  3.106036
#> 14763 10.14479   110.04  FALSE        60.35            100  3.107720
#> 280   29.32341    60.55  FALSE       127.54             29  3.333837
#> 7132  34.04916   300.00  FALSE       225.26             12  3.030143
#> 4269  19.11674   300.00  FALSE       194.53            100  3.439564
#> 16845 22.88279   300.00  FALSE        25.55             33  3.874467
#> 28526 31.81463   300.00  FALSE        46.28            100  3.237445
#> 9526  27.56958   300.00  FALSE       300.00            100  2.898919
#> 31440 24.03501   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91071   300.00  FALSE       300.00            100  3.015637
#> 24962 29.05011   300.00  FALSE       300.00            100  2.996437
#> 23808 37.53389   300.00  FALSE       300.00             72  2.692387
#> 14645 45.21325   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679   300.00  FALSE       300.00            100  2.546026
#> 13548 16.97846   300.00  FALSE       300.00            100  2.256336
#> 28978  4.43985   300.00  FALSE        10.00            100  2.209674