Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483    566.7674
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073  0.00174   2310.4871
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088    266.4355
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151  0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426 -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276   9614.6924
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.17307  0.00359 -0.00279    930.0871
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427    380.9472
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009   1247.7061
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469   1021.5407
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911    173.2806
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708
#> 21683 714792.5 9561052 1807.84 55.73275 -0.02070 -0.02661   2508.5374
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583   1898.0105
#> 29228 712782.5 9560422 1870.56 32.90611 -0.00218 -0.00372   4154.5786
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756    353.9086
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957
#> 44612 714932.5 9558822 2578.80 38.25525  0.01257  0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526    706.1463
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451 -0.03921    525.3245
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118 -0.00252   3631.7017
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778
#> 6470  713742.5 9558802 2260.79 30.68017  0.00000  0.00320    911.5911
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642   1688.1782
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159 -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230    714.7432
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007 -0.00463   1050.7473
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328   1066.7134
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724  0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032    994.9224
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650    766.6254
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863   1098.3462
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499    617.3855
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.90101  0.03044 -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438   1276.5448
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284 -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119    792.3544
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172    351.5812
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023 -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983