Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203    634.3320
#> 16435 713512.5 9559092 2166.13 55.97397  0.02056 -0.06976    301.2347
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403   1787.3024
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264   5057.1431
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114   4510.8833
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678   1891.6584
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605    798.6775
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527  0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270   4057.5637
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117   1018.2141
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401   1722.6481
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809   1102.9956
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078  0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555   4439.0967
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354   1147.7384
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332 -0.00058   1817.2371
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322   4605.1411
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546  0.01484    495.0292
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560  0.00610    454.8103
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210    585.8992
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465 -0.01036   2505.8369
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077   8759.9961
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367   3002.8828
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216   1412.2295
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655    560.8554
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.85074 -0.00782 -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282   2322.7944
#> 33666 713632.5 9561012 1775.88 21.17881 -0.00950 -0.05130    107.5831
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863   1098.3462
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262 -0.00502    520.5218
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438   1276.5448
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264  0.00624   1870.8307
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208    842.4825
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491    261.7371
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270  35950.5508
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011    501.7867
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585   5714.3267
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057    722.3952
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603