Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203    634.3320
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305 -0.00555    430.4604
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835   2425.3115
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416    947.4878
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203  10263.2734
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083    994.7876
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276   9614.6924
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313    783.4548
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796   4729.6382
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661  0.01199    196.4080
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102    483.4184
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373    527.3528
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236   2715.4138
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941   1290.7709
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512  13178.3223
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472 -0.00228   2262.3796
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220   8936.2666
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261   1627.8989
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261   3353.5422
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970   1443.9200
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526    706.1463
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560  0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451 -0.03921    525.3245
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732   4237.6934
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230   1044.6443
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063   2430.2278
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177    997.3089
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328   1066.7134
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875   1165.3953
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 25140 713332.5 9560692 1809.95 52.25719  0.00075 -0.01945    259.7421
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011    201.1272
#> 29107 713172.5 9559272 2067.58 28.26401  0.02413  0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312    593.7773
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 33666 713632.5 9561012 1775.88 21.17881 -0.00950 -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257    549.3252
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417   1520.6630
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578   1258.8865
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172 111973.8906
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336