Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.
partition_cv(data, coords = c("x", "y"), nfold = 10, repetition = 1,
  seed1 = NULL, return_factor = FALSE)

Arguments

data 	data.frame containing at least the columns specified by coords

coords 	(ignored by partition_cv)

nfold 	number of partitions (folds) in nfold-fold cross-validation partitioning

repetition 	numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1 	seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor 	if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value
If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition) resampling objects. Each of these resampling objects is a list of length nfold corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).
Details
This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.
See also
sperrorest, represampling
Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[['1']][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx , ]

#>               x       y     dem      slope    hcurv      vcurv        carea
#> 31358  712882.5 9560002 1911.52 25.5642309 -0.00681   -0.00029    5577.3916
#> 36710  715232.5 9559582 2198.66 23.2471896 -0.00501   -0.00649    1399.2329
#> 37912  712802.5 9559952 1838.40 52.1013441  0.00183   -0.09203     634.3320
#> 36052  714842.5 9558892 2483.27 68.7898858 -0.04921   -0.12438     753.6304
#> 16435  713512.5 9559092 2166.13 55.9739659  0.02056   -0.06976     301.2347
#> 44302  714932.5 9557982 2649.56 37.3006984  0.01633   -0.01813    1130.5791
#> 47223  715302.5 9557472 2857.45 39.5335149 -0.01021   -0.01579    1318.7560
#> 42632  714852.5 9557902 2674.80 30.6526691  0.00221    0.00969     368.5606
#> 16749  715072.5 9558242 2745.27 49.0537816  0.00097    0.01483     566.7674
#> 29391  712842.5 9560152 1930.16 28.0955584  0.00753    0.01487    1077.7585
#> 1300   714042.5 9558482 2408.27 24.1484522  0.00659    0.01041     772.9746
#> 11088  715532.5 9558762 2549.28 38.0501272 -0.00373   -0.00057    1630.3259
#> 34512  713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677    1675.4679
#> 21939  713642.5 9558712 2299.09 33.7506519 -0.00503    0.02433    1475.3191
#> 9669   713852.5 9558612 2308.97 52.1804123 -0.01059   -0.07431    4079.0366
#> 28293  713762.5 9558812 2250.98 33.2854102  0.00118    0.00761     928.6858
#> 27973  715022.5 9559332 2308.28 20.0632631  0.00940    0.00679     550.0344
#> 26531  714462.5 9560452 2104.74 19.5241098  0.00453    0.00257     357.1436
#> 1768   713912.5 9558552 2357.19 38.6924129 -0.00645    0.00835    2425.3115
#> 43327  713962.5 9561082 1839.11 63.3616837 -0.04951   -0.00529    3282.1186
#> 43669  712392.5 9560162 2001.90 47.2839150 -0.00284    0.01804     628.5153
#> 125    713992.5 9561062 1878.28 49.9710871  0.00183    0.03387    1989.3012
#> 27375  714292.5 9558982 2458.96 43.6009423  0.00214    0.00466     724.0713
#> 11463  713112.5 9558902 2244.08 21.8915714  0.00160    0.00311     586.7354
#> 3835   714142.5 9558552 2388.42 32.9003825  0.00979   -0.00269     491.8017
#> 33792  714652.5 9558172 2598.36 59.4586952 -0.03525   -0.11024     394.8751
#> 43964  714802.5 9558452 2684.33 37.9825182 -0.02123    0.01164     895.6606
#> 40307  713302.5 9560702 1845.07 39.0046749  0.00129    0.02510    3342.5518
#> 38279  714192.5 9558612 2325.14 28.2525489 -0.04430   -0.06070  138869.3906
#> 27978  713552.5 9558532 2373.05 17.0449214  0.00832    0.00838     238.6685
#> 38070  715212.5 9559222 2342.41 55.6771737  0.00426   -0.00367     793.9665
#> 47210  714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733     598.7983
#> 18391  715112.5 9559432 2235.02 22.2576915 -0.02651   -0.03629  360379.1875
#> 13585  714552.5 9560092 2123.69 57.5289733 -0.03613   -0.01898     588.3761
#> 34280  713652.5 9560702 1896.37 41.2735877  0.00382    0.00088    2322.1746
#> 28230  715182.5 9558102 2791.83 39.0717109 -0.00669   -0.00950     950.3138
#> 6763   714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626    1014.7159
#> 7879   714722.5 9559232 2356.19 36.8039440 -0.01097    0.00156     816.6036
#> 23141  715632.5 9558942 2495.99 34.3768948  0.01275    0.00605     798.6775
#> 34642  714962.5 9559862 2282.08 22.1608616 -0.00028    0.01318     289.0063
#> 19128  714852.5 9557872 2681.49 20.7576880  0.00453    0.01077     300.7915
#> 12868  712532.5 9560332 1913.92 61.9373106 -0.03543   -0.05146    1347.6224
#> 36931  712852.5 9558892 2102.80 30.5357857  0.00140   -0.00070    2562.7559
#> 47091  713382.5 9559182 2122.71 48.0379275  0.04527    0.04613     289.6113
#> 31316  715022.5 9559332 2308.28 20.0632631  0.00940    0.00679     550.0344
#> 18751  714882.5 9559182 2381.81 39.2991752  0.00066   -0.00036     516.4709
#> 46171  712822.5 9558932 2089.30 29.8545389 -0.01174    0.00025    4425.5151
#> 15766  715562.5 9558592 2646.16 29.0208853  0.00194    0.00516     576.0335
#> 21053  712862.5 9559992 1901.01 30.9506071 -0.00762    0.01252    5761.3486
#> 462    713712.5 9561182 1785.08 47.7170711 -0.03260   -0.03310    2833.9001
#> 42941  713252.5 9559742 1990.10 28.9544222 -0.04708    0.00408   76318.0156
#> 7466   712962.5 9560392 1819.03 18.4028314 -0.00140   -0.02330    1456.9246
#> 39520  715342.5 9557622 2889.48 41.6488751 -0.01282    0.00542     860.8840
#> 30406  715312.5 9560152 2029.10 34.1373984  0.00953   -0.02283     590.1536
#> 34005  714922.5 9558842 2559.23 52.8312924  0.03658    0.04602     345.8933
#> 18265  715522.5 9558782 2539.23 35.8562718 -0.00458   -0.00401    1722.6481
#> 25700  715172.5 9559272 2324.49 35.5348425 -0.00529    0.02009    1247.7061
#> 31462  714082.5 9558542 2367.84 34.7269720 -0.02027   -0.01473    2930.2791
#> 664    714452.5 9559852 2174.83 44.8608765  0.00282    0.00109     767.4097
#> 17728  713092.5 9560652 1900.14 30.2791006  0.00645   -0.00395    2011.3270
#> 16370  713852.5 9558632 2296.97 34.6244125 -0.01972    0.03512   13178.3223
#> 37769  714412.5 9558562 2382.08 39.8767166 -0.04601   -0.08389   47022.9688
#> 30268  714172.5 9558612 2328.84 54.9506633  0.03743   -0.05242     507.9263
#> 39166  712832.5 9560142 1922.46 39.6601386  0.01469    0.02940     987.8085
#> 43859  714892.5 9557602 2722.70 28.1293630  0.00370   -0.00190     640.8842
#> 11612  715212.5 9557752 2820.51 39.1622383  0.00017    0.01213     328.7922
#> 32879  713912.5 9559672 2238.00 31.9767745 -0.00278   -0.00412    1350.3611
#> 46461  712682.5 9560032 1892.30 22.9778994  0.01453    0.01337     486.7551
#> 17345  713922.5 9559682 2227.44 34.4909134 -0.06509   -0.01211   49347.5547
#> 31117  712892.5 9558922 2129.84 30.5506826  0.00521   -0.00870    1895.1399
#> 6217   714442.5 9559932 2151.30 36.6704448  0.01489    0.00241    1752.9242
#> 7      714952.5 9559862 2284.19 16.1064166  0.00516    0.01574     183.3220
#> 16009  715122.5 9559402 2264.81 50.0186426 -0.00589    0.00849    1120.5961
#> 7086   712762.5 9560962 2022.00 44.9135886 -0.01259   -0.00281    1178.4420
#> 47702  714692.5 9559712 2200.51 33.0986259 -0.13038   -0.04212   46310.0195
#> 41125  712772.5 9559412 2021.88 38.8763323 -0.00310   -0.00860     695.2302
#>             cslope distroad slides distdeforest distslidespast log.carea
#> 31358  34.42788799   300.00   TRUE        15.00           9.00  3.746431
#> 36710  30.71569444   300.00   TRUE       300.00          21.00  3.145890
#> 37912  30.29457046   300.00   TRUE         9.15           2.00  2.802317
#> 36052  53.74516006   300.00   TRUE       300.00         100.00  2.877158
#> 16435  49.44396589   300.00   TRUE       300.00          41.00  2.478905
#> 44302  35.10111340   300.00   TRUE       300.00         100.00  3.053301
#> 47223  36.25906111   300.00   TRUE       300.00         100.00  3.120164
#> 42632  20.47808456   300.00   TRUE       300.00          10.00  2.566509
#> 16749  32.18303935   300.00   TRUE       300.00         100.00  2.753405
#> 29391  28.28348860   300.00   TRUE         0.56         100.00  3.032521
#> 1300   27.53062206   300.00   TRUE       300.00           6.00  2.888165
#> 11088  30.99873559   300.00   TRUE       300.00         100.00  3.212274
#> 34512  34.63988238   300.00   TRUE       195.00           2.00  3.224136
#> 21939  23.93244710   300.00   TRUE       300.00         100.00  3.168886
#> 9669   31.66451255   300.00   TRUE       300.00           1.00  3.610558
#> 28293  28.90228302   300.00   TRUE       300.00           4.00  2.967869
#> 27973  30.00981044   300.00   TRUE       300.00          85.00  2.740390
#> 26531  15.98781432   300.00   TRUE       300.00           2.00  2.552843
#> 1768   28.18551282   300.00   TRUE       300.00           7.00  3.384768
#> 43327  29.05870050   173.40   TRUE        33.81         100.00  3.516154
#> 43669  28.47657538    24.22   TRUE         0.00          89.00  2.798316
#> 125    29.59670787   207.04   TRUE        21.03         100.00  3.298701
#> 27375  38.77033512   300.00   TRUE       300.00         100.00  2.859781
#> 11463  20.13373692   300.00   TRUE       300.00         100.00  2.768442
#> 3835   35.13892862   300.00   TRUE       300.00          39.00  2.691790
#> 33792  51.20810294   300.00   TRUE       300.00           5.00  2.596460
#> 43964  28.75732470   300.00   TRUE       300.00          46.00  2.952143
#> 40307  29.51362898    77.82   TRUE        76.41          54.00  3.524078
#> 38279  37.62843024   300.00   TRUE       300.00          24.00  5.142607
#> 27978  14.32852854   300.00   TRUE       300.00          12.00  2.377795
#> 38070  34.23308234   300.00   TRUE       300.00         100.00  2.899802
#> 47210  26.91526538   300.00   TRUE       300.00         100.00  2.777281
#> 18391  34.50752913   300.00   TRUE       300.00         100.00  5.556760
#> 13585  47.06733695   300.00   TRUE       300.00         100.00  2.769655
#> 34280  33.62116342   300.00   TRUE         0.00         100.00  3.365895
#> 28230  37.90173111   300.00   TRUE       300.00         100.00  2.977867
#> 6763   16.21814335   125.47   TRUE         0.00           6.00  3.006344
#> 7879   28.58085369   300.00   TRUE       300.00         100.00  2.912011
#> 23141  30.29170567   300.00   TRUE       300.00         100.00  2.902371
#> 34642  15.64862330   300.00   TRUE       300.00           2.00  2.460907
#> 19128  14.65969815   300.00   TRUE       300.00          15.00  2.478266
#> 12868  37.58832319   119.34   TRUE        79.16          61.00  3.129568
#> 36931  27.20919273   300.00   TRUE       300.00         100.00  3.408707
#> 47091  40.15173637   300.00   TRUE       300.00          25.00  2.461815
#> 31316  30.00981044   300.00   TRUE       300.00          85.00  2.740390
#> 18751  38.07591028   300.00   TRUE       300.00         100.00  2.713046
#> 46171  29.07130557   300.00   TRUE       270.22         100.00  3.645964
#> 15766  23.33714395   300.00   TRUE       300.00         100.00  2.760448
#> 21053  34.25141699   300.00   TRUE         0.00           2.00  3.760524
#> 462     9.81247520    61.17   TRUE        57.16          74.00  3.452385
#> 42941  39.06311656   300.00   TRUE       300.00          17.00  4.882627
#> 7466   28.79055625   255.23   TRUE         0.00          60.00  3.163437
#> 39520  35.89408699   300.00   TRUE       300.00         100.00  2.934945
#> 30406  43.10075014   300.00   TRUE       300.00          85.00  2.770965
#> 34005  41.73367284   300.00   TRUE       300.00         100.00  2.538942
#> 18265  31.44564267   300.00   TRUE       300.00         100.00  3.236197
#> 25700  32.54113797   300.00   TRUE       300.00         100.00  3.096112
#> 31462  33.75466258   300.00   TRUE       300.00           1.00  3.466909
#> 664    30.84231811   300.00   TRUE       300.00          12.00  2.885027
#> 17728  28.55220580    30.14   TRUE       140.15           2.00  3.303483
#> 16370  31.15859081   300.00   TRUE       300.00           8.00  4.119860
#> 37769  37.42503022   300.00   TRUE       300.00         100.00  4.672310
#> 30268  47.27245585   300.00   TRUE       300.00          41.00  2.705801
#> 39166  29.81844253   300.00   TRUE         1.90          90.00  2.994673
#> 43859  29.55774674   300.00   TRUE       300.00         100.00  2.806780
#> 11612  35.16700355   300.00   TRUE       300.00         100.00  2.516921
#> 32879  29.73822844   300.00   TRUE       300.00           6.00  3.130450
#> 46461  24.94142578   288.47   TRUE         0.00         100.00  2.687311
#> 17345  29.82130732   300.00   TRUE       300.00           2.00  4.693266
#> 31117  26.64597522   300.00   TRUE       291.23         100.00  3.277641
#> 6217   30.97180657   300.00   TRUE       300.00           1.00  3.243763
#> 7      13.22042816   300.00   TRUE       300.00           0.00  2.263215
#> 16009  33.41317974   300.00   TRUE       300.00         100.00  3.049449
#> 7086   34.20443445   300.00   TRUE       128.61          63.00  3.071308
#> 47702  33.66642708   300.00   TRUE       300.00          86.00  4.665675
#> 41125  36.55585324   300.00   TRUE         0.00         100.00  2.842129
#>  [ reached getOption("max.print") -- omitted 324 rows ]