Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769