Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917     116.1986
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491