Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 44302 714932.5 9557982 2649.56 37.3006984   0.01633   -0.01813   1130.5791
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896    276.0310
#> 19963 714612.5 9560552 2031.70 26.8482293   0.00046    0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519  -0.00503    0.02433   1475.3191
#> 28293 713762.5 9558812 2250.98 33.2854102   0.00118    0.00761    928.6858
#> 1768  713912.5 9558552 2357.19 38.6924129  -0.00645    0.00835   2425.3115
#> 47255 713552.5 9559962 2191.77 33.5787645  -0.00068    0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.6567008   0.00730    0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088    266.4355
#> 10899 713852.5 9559652 2282.82 28.8140475   0.00952    0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.1555304  -0.00769   -0.00281   1325.6138
#> 38410 713712.5 9561172 1776.08 36.3192217  -0.02678   -0.04272   4955.5586
#> 38279 714192.5 9558612 2325.14 28.2525489  -0.04430   -0.06070 138869.3906
#> 20290 714812.5 9561092 1775.73 34.6267043  -0.01537   -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838    238.6685
#> 38070 715212.5 9559222 2342.41 55.6771737   0.00426   -0.00367    793.9665
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889    362.8363
#> 8288  714842.5 9557712 2679.04 28.9395253   0.00902    0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.2343959   0.00993   -0.00413    364.9788
#> 3453  713482.5 9559642 2145.11 35.1257506   0.00108   -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.2576915  -0.02651   -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.5289733  -0.03613   -0.01898    588.3761
#> 27957 715012.5 9559332 2306.35 20.4803764   0.00225    0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.0208853   0.00422    0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.4804185   0.00252    0.01118   2359.5059
#> 2950  715042.5 9557712 2740.72 26.5250811  -0.00089   -0.00051    779.0555
#> 5728  713502.5 9559662 2140.27 34.3075669  -0.00154    0.00124   2114.2253
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.6203527  -0.00904   -0.00866  29195.6016
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862    874.5408
#> 34563 715022.5 9558162 2675.70 18.8898455  -0.01993    0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255   0.01458    0.00672    565.1107
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.0632631   0.00940    0.00679    550.0344
#> 29926 714702.5 9557622 2637.00 44.6425796   0.01843    0.04407    344.8395
#> 17423 715592.5 9558562 2654.69 33.2195200  -0.00034    0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.5998104  -0.03552   -0.02058    431.4362
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903   2360.1536
#> 41037 715532.5 9558792 2529.58 32.8465245  -0.00585   -0.00974   2136.6939
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117   1018.2141
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.7252601  -0.01399   -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.9949135   0.00175    0.00236   2715.4138
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.7218224  -0.00048   -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542    860.8840
#> 18265 715522.5 9558782 2539.23 35.8562718  -0.00458   -0.00401   1722.6481
#> 39732 715362.5 9559582 2230.97 23.3468842  -0.00161   -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.8738238   0.00090   -0.00299    688.0706
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625   1063.9150
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361  -0.00099    0.02169    495.8840
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.3526278   0.00568   -0.01558   1504.9717
#> 17728 713092.5 9560652 1900.14 30.2791006   0.00645   -0.00395   2011.3270
#> 38197 712522.5 9560322 1942.04 62.7772667  -0.01611    0.02961   1086.6963
#> 45500 712812.5 9560052 1852.16 56.7703772   0.00904   -0.02013   4229.5708
#> 26010 712852.5 9560542 1928.25 30.9781091   0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.4082118   0.00291   -0.00081    940.6002
#> 47366 714892.5 9557612 2718.78 28.0239387   0.00538    0.00112    745.3315
#> 8125  713452.5 9559132 2120.44 42.1404729  -0.05458   -0.14813 287033.4688
#> 10467 714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400    608.6523
#> 27244 714162.5 9558622 2316.00 36.7512319  -0.02556   -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.5506826   0.00521   -0.00870   1895.1399
#> 42201 714642.5 9557452 2592.29 13.2255848  -0.02084   -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.8600517  -0.02143   -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506  -0.02070   -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.5936474   0.00384    0.00175   1026.6818
#> 28312 715712.5 9558032 2783.23 24.2527305  -0.00007   -0.02132   1580.1621
#> 34164 714212.5 9558492 2421.45 51.3295700   0.02247    0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.5275756  -0.00193    0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.9750697   0.00511   -0.01801   1071.6141
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441    424.2680
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.0256786   0.04250   -0.00970   1443.9200
#> 10617 713232.5 9560722 1879.95 26.4970062   0.00258    0.00422   1159.2100
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391   1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576  -0.00767   -0.03713   6379.7974
#> 26058 713402.5 9560442 1880.12 31.3774607   0.00730   -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.4323251   0.04864    0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209  -0.00019    0.00119   6589.5796
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.4975862  -0.00622    0.00152   1530.5657
#> 17227 714812.5 9558892 2545.43 54.0717460   0.00458    0.04291    293.4231
#> 10213 715542.5 9558782 2532.10 35.4133754  -0.00893   -0.00867   2247.8748
#> 41408 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403    474.8613
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781  29995.1543
#> 36561 715292.5 9560802 1844.64 31.9372405   0.00002   -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908   0.00534    0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230    714.7432
#> 35852 715112.5 9559962 2173.61 39.3157909  -0.00738    0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.5904544  -0.00812   -0.02208    951.3401
#> 8066  715842.5 9558732 2636.38 33.7174203  -0.00494    0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.0369281  -0.03220   -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.9322444   0.03480    0.01500    538.8209
#> 43243 715682.5 9558462 2693.23 27.4062902  -0.00031   -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.8258980   0.00846    0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.6658331   0.01253   -0.00333    422.8840
#> 7538  714722.5 9558582 2660.25 15.9946898   0.00544   -0.01014    271.4219
#> 20547 713732.5 9560822 1851.67 36.9683829  -0.01405    0.00476  11992.5273
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.7201175   0.00174   -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.9416565  -0.00065    0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.9979810   0.00990    0.02790    209.6952
#> 8831  714372.5 9559852 2200.77 48.7604272  -0.00457    0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362    283.7090
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267  31588.1953
#> 33175 715822.5 9558832 2563.13 36.6228893  -0.00674   -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126  -0.00452   -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.3642967   0.03423   -0.00983    270.7949
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.4571579  -0.03224   -0.02036    476.9504
#> 46509 712512.5 9559532 1860.09 89.9887513  -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.3936922   0.03479    0.02251    168.1733
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479   0.00972    0.00308    422.5022
#> 5861  713882.5 9559182 2350.43 27.3936851  -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705   0.00537   -0.00577    423.5598
#> 28286 713702.5 9559842 2245.07 27.9689348   0.00511    0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836    250.5111
#> 48184 714202.5 9558222 2489.17 36.1135298   0.02236   -0.00066    306.0146
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894   0.01354    0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.4446854  -0.00070   -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.0318136   0.01600   -0.01849    764.8193
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875   2156.9358
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.7247292   0.00361   -0.01841    756.0055
#> 16845 712772.5 9560022 1831.76 21.3999736  -0.00865   -0.00805   7489.7417
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948   1538.1741
#> 43972 713642.5 9558702 2304.16 24.0235474   0.00340    0.01809   1158.0426
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.1977476  -0.00595    0.03064   1485.5201
#> 32342 714862.5 9559622 2294.74 36.9672369   0.01021    0.00109    915.4989
#> 47245 715372.5 9560792 1900.65 34.4708598  -0.00011   -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.5327393  -0.00305   -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.5082486   0.02751    0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.7319259   0.07379   -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.2726234  -0.01181   -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607