Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607