Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.8578160   -0.01106   -0.04634    500.5027
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183   -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.9523260   -0.00123    0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056   -0.06976    301.2347
#> 24051 715462.5 9559372 2319.46 32.3279977   -0.00661   -0.01339   1002.5877
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498   -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907   -0.00333   -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835   -0.02191   -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.8152355   -0.00857    0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.5848574    0.00065    0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519   -0.00503    0.02433   1475.3191
#> 27973 715022.5 9559332 2308.28 20.0632631    0.00940    0.00679    550.0344
#> 14456 713542.5 9559972 2184.39 35.2414881   -0.00841   -0.00010   1527.1028
#> 37270 714852.5 9561032 1838.96 35.1589821   -0.00854    0.01264   5057.1431
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223    0.01783    377.8918
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021   -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542    0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.9003825    0.00979   -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581    0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.3814924   -0.01966   -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.6234412   -0.02169   -0.00160   9344.2930
#> 48498 714972.5 9557972 2681.66 38.3469193    0.00528    0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.6379538   -0.00270   -0.00769   1032.7177
#> 48055 712862.5 9559972 1902.48 29.0168746   -0.00720    0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.1855058    0.00602    0.01518    854.4269
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017   -0.00032    994.6740
#> 27978 713552.5 9558532 2373.05 17.0449214    0.00832    0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295   -0.01440    0.00710    843.7961
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713   -0.00173   8967.1924
#> 3453  713482.5 9559642 2145.11 35.1257506    0.00108   -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.1282451   -0.00033   -0.03917    116.1986
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.4803764    0.00225    0.00325    616.8652
#> 13847 715052.5 9559232 2337.87 26.1148433   -0.01452   -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.4960559    0.01205    0.02205    629.4165
#> 1635  713102.5 9559052 2172.87 31.4840308   -0.00064   -0.00916   2009.1686
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646   -0.04086   6042.5757
#> 46756 713372.5 9559192 2120.47 46.0050732    0.03696    0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.9373106   -0.03543   -0.05146   1347.6224
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047   -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.6343274   -0.00669   -0.00991   1403.0193
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566   -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.0575616    0.02224    0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547    0.03323    183.1836
#> 46201 714832.5 9561022 1845.94 34.5739922   -0.00856    0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000    0.00321    0.01169    350.1799
#> 18751 714882.5 9559182 2381.81 39.2991752    0.00066   -0.00036    516.4709
#> 29926 714702.5 9557622 2637.00 44.6425796    0.01843    0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267   -0.05063  18733.8848
#> 22332 715482.5 9558922 2475.28 19.9366394   -0.02617    0.00307 147470.5312
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283    0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855   -0.02175   1122.9120
#> 22299 713532.5 9560362 1975.58 27.7099578    0.00478    0.01092    412.5849
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504    0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.9235105   -0.01364   -0.03357    473.4709
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140   -0.02330   1456.9246
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090   -0.00299    688.0706
#> 37416 713282.5 9560692 1850.61 30.2590471    0.00736    0.01684   2556.8049
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126   -0.00485   2240.4185
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099    0.02169    495.8840
#> 664   714452.5 9559852 2174.83 44.8608765    0.00282    0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.2645529   -0.00195   -0.00656   3036.7383
#> 38197 712522.5 9560322 1942.04 62.7772667   -0.01611    0.02961   1086.6963
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078    0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.1089251    0.01784    0.00596    842.0994
#> 17431 712802.5 9560022 1832.71 41.5251162   -0.00478   -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.0111801    0.00394   -0.00224    942.0322
#> 20601 715542.5 9558182 2773.29 32.7554242   -0.00604    0.00354   1147.7384
#> 37769 714412.5 9558562 2382.08 39.8767166   -0.04601   -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133    0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633    0.03743   -0.05242    507.9263
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370   -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.1622383    0.00017    0.01213    328.7922
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509   -0.01211  49347.5547
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338    0.00728   1137.2450
#> 24135 712612.5 9559492 1903.67 46.7567938   -0.06953   -0.01797   7557.7866
#> 20797 713012.5 9560392 1842.25 36.6498183    0.01299    0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.0763086   -0.01474    0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384    0.00175   1026.6818
#> 19832 715182.5 9557662 2761.27 38.4328630   -0.00666   -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.9977924    0.01354    0.00245    629.5983
#> 5911  713952.5 9561282 1801.16 33.7896130    0.00466   -0.01776   1021.8267
#> 26101 715562.5 9558682 2594.87 32.4110766    0.00079    0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.3295700    0.02247    0.01123    364.2903
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109    0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.2390497    0.01011    0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.0592454    0.02213   -0.01363    565.5207
#> 332   714852.5 9561102 1787.11 45.3582039    0.00540    0.00879   1200.4210
#> 39624 715222.5 9559552 2205.54 23.8974967    0.00401   -0.00391   1794.7964
#> 14494 715332.5 9558792 2557.22 20.7204457    0.01267    0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.2552461    0.01257    0.00044    288.5468
#> 24723 712942.5 9560202 1983.18 33.0716969    0.00285   -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.3774607    0.00730   -0.02610   1872.9276
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157   -0.00152    850.4633
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124    0.01166   1813.7841
#> 11057 714962.5 9557642 2737.48 42.6859287   -0.00488   -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.4926392    0.00004   -0.02384    947.2075
#> 31491 712832.5 9558892 2091.39 29.7004132    0.00186    0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.8077801   -0.00486    0.02306   1718.3652
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025   -0.05345   2417.1389
#> 28253 714102.5 9559892 2288.88  2.9576081    0.00378    0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.4649135   -0.01332    0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203    0.00642   1688.1782
#> 48601 713692.5 9560152 2058.14 59.5572439   -0.01433   -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.6097183   -0.00465   -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.0797253    0.00165    0.00565   1410.8372
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284    0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.9754961    0.00872    0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831   -0.00298   -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.1532385    0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124    0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.1973563    0.04336    0.01403    241.2991
#> 18746 713292.5 9560952 1890.51 41.8855067   -0.00402   -0.01328   1066.7134
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174   -0.01466   2633.9666
#> 30267 714442.5 9557512 2612.30 28.7659191   -0.00458    0.00479    604.0975
#> 49363 712772.5 9559552 1935.91 29.0369281   -0.03220   -0.03290   2481.5632
#> 45166 715262.5 9558722 2531.59 13.5533166   -0.00502   -0.00028   2394.7261
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544   -0.01014    271.4219
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405    0.00476  11992.5273
#> 8831  714372.5 9559852 2200.77 48.7604272   -0.00457    0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.1945125    0.00236    0.00483   1097.3082
#> 14800 712952.5 9558682 2120.89 38.2735807   -0.00022    0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.6228893   -0.00674   -0.00376    629.8236
#> 34345 715852.5 9557902 2874.08 43.6559462    0.00379   -0.01018    937.6911
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118    0.00162    290.2031
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154   -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858    0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402    0.00009   -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.2302386   -0.00233    0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.9376528   -0.00070   -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479    0.00972    0.00308    422.5022
#> 17628 714032.5 9560252 2160.03 30.2945705    0.00537   -0.00577    423.5598
#> 28286 713702.5 9559842 2245.07 27.9689348    0.00511    0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.6940967    0.00944    0.00836    250.5111
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074   -0.02904    608.3561
#> 4943  712452.5 9559172 1927.46 34.4651303    0.00105   -0.01576   1045.1074
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865   -0.00805   7489.7417
#> 34334 714642.5 9557912 2582.25 30.9912871    0.00253   -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.0021874    0.00203    0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.5342555   -0.00163   -0.00377   1727.6055
#> 37254 713482.5 9560892 1823.73 39.8824462    0.01014   -0.00884    672.2466
#> 23808 715172.5 9558702 2533.50 22.3459270    0.00811   -0.02281    492.4785
#> 41373 715332.5 9560862 1873.10 29.1245906   -0.00166    0.00156    758.4840
#> 15363 714742.5 9559212 2373.13 33.5833482   -0.00247    0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.6213801    0.00208    0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.8520794    0.04456   -0.01586    941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603