Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715    520.7355
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607