Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976     301.2347
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226    2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388    1051.4222
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044  129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889     362.8363
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450     873.0448
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338   10787.6299
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613     289.6113
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949    1182.1506
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373     527.3528
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252    5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454    3476.0818
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081     940.6002
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337     486.7551
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870    1895.1399
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812    1423.7578
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484     495.0292
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778    4594.4351
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152     850.4633
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252    3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230    1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831     454.9113
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345  665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064     683.2795
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171     756.8818
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077    8759.9961
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011     712.5228
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223    1056.3701
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948    1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955     713.4463
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513     161.9878
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099     909.0560
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836     250.5111
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972     331.7462
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270   35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 40653 714262.5 9561262 1791.94 12.222336   0.00161    0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884     672.2466
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172     351.5812
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227     841.2025
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491