Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393   2081.0056
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333    258.6537
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041    772.9746
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250   1413.3312
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083    994.7876
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916   2009.1686
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034   4495.6929
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169    350.1799
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279    930.0871
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964    571.1364
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941   1290.7709
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020   1563.5549
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664   1898.5588
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393   8301.6094
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534    344.1447
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441    424.2680
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196   1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713   6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384    947.2075
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291    487.3438
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559    866.5775
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517   1060.5494
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445    355.5078
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099    909.0560
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886  43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884    672.2466
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573   1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491