Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501   -0.00649   1399.2329
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199    0.00659   3553.7166
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305   -0.00555    430.4604
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097    0.01483    566.7674
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075    0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479    0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753    0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741    746.4155
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104   -0.02046   1024.1573
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440   1088.8639
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491   1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284    0.01804    628.5153
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407    0.00547    519.4228
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214    0.00466    724.0713
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831   -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021   -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952    0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769   -0.00281   1325.6138
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160   9344.2930
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602    0.01518    854.4269
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017   -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191   -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746    0.00096   3210.5781
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581   1855.1630
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521   -0.00739    558.2286
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422    0.02357    275.7474
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097    0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193   1809.3838
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184    0.03924   8023.8657
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543   -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545    0.02405   5681.2412
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294   -0.00096    520.5570
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167   1938.5389
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338  10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752   -0.00517   1000.7119
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066   -0.00036    516.4709
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093   -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060    0.00270   4057.5637
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661    0.01199    196.4080
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617    0.00307 147470.5312
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708   -0.02538    807.8237
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283    0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409    0.02771    674.3859
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917   -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658    0.04602    345.8933
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090   -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739    365.7626
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554    0.03304   1569.7313
#> 664   714452.5 9559852 2174.83 44.860876    0.00282    0.00109    767.4097
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539    0.00809   1102.9956
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395   2011.3270
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629   -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771   -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472   -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594  16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627   -0.05393   8301.6094
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567   -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458   -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278   -0.00412   1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332   -0.00058   1817.2371
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781  29995.1543
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299    0.00141   4611.0718
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444   -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218   -0.00372   4154.5786
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801    172.7491
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247    0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428   -0.00522   1355.7436
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441    424.2680
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970   1443.9200
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042    0.00957    495.4453
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140   1066.3957
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110   -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285   -0.01325    595.0426
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451   -0.03921    525.3245
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758   -0.00288   2144.8164
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683   -0.00603   1562.0503
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382    100.0000
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849    0.00150    503.5590
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465   -0.00105   4432.8760
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298   -0.00712   1010.5097
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403    241.2991
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165    0.00875   1165.3953
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220   -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500    538.8209
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199    836.4196
#> 47005 714942.5 9559192 2333.66 29.220848   -0.04937   -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014    271.4219
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476  11992.5273
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846    0.00216   1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084   -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236    0.00483   1097.3082
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193   -0.00267  31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366    267.5057
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322   -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785    0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112   -0.00038   2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268    252.0072
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342    0.02552   2515.0752
#> 330   714232.5 9560362 2165.00 18.060776    0.01142    0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858    0.00312    254.0656
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229    0.00099    909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119    0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360   4288.7329
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980   -0.00710    472.0236
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750   1892.8978
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852    0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070   -0.00050   1471.4346
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607   -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598    0.02208    842.4825
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770    0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108   -0.00948   1538.1741
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716   -0.03234   1707.3130
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366    0.00525    162.0592
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674