Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393   2081.0056
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438    753.6304
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333    258.6537
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677   1343.5006
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024    394.8751
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510   3342.5518
#> 18005 714832.5 9557912 2660.76 31.9120303  0.00191 -0.00261    554.9243
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382  0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710    843.7961
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107   1392.5880
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193   1809.3838
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114    150.1498
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146   1347.6224
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167   1938.5389
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511 361979.2812
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613    289.6113
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825    162.6688
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319   3457.7519
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543    352.1558
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485   2240.4185
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055    773.0359
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620    379.0855
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555   4439.0967
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448   1258.8196
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860    695.2302
#> 34887 714662.5 9557392 2609.02 51.0207457 -0.01878 -0.04351   1442.9247
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833    369.3771
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812   1423.7578
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196   1052.5341
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778   4594.4351
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417  34908.6562
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483    343.1384
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610   1872.9276
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875    597.2893
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166   1813.7841
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732   4237.6934
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384    947.2075
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288   2144.8164
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870   2099.7930
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476   1275.9446
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419    761.5427
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116   1817.5222
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328   1066.7134
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466   2633.9666
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445    355.5078
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945    259.7421
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333    422.8840
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001  0.00759   1880.4403
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376    629.8236
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.3155672  0.04152  0.00098    164.0802
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036    476.9504
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413  0.00528    355.3475
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257    549.3252
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750   1892.8978
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177   1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875   2156.9358
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012    305.3624
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234   1707.3130
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769