Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438     753.6304
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741     746.4155
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896     276.0310
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949    6217.8213
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842     868.6137
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766  140591.4062
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096    3210.5781
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899    1394.2274
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413     364.9788
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917     116.1986
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150    1274.3594
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357     275.7474
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077    3559.4204
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924    8023.8657
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124    2114.2253
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613     289.6113
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102     483.4184
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560     977.4719
#> 18892 712682.5 9560202 1839.08 40.9235105 -0.01364 -0.03357     473.4709
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160  0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306 -0.00436    2602.7390
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409    2312.8413
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026     247.0894
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169     495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102   35685.0000
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454    3476.0818
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558    1504.9717
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643     670.2289
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081     940.6002
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399    3900.3445
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594   16587.9980
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242     507.9263
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822     472.8748
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534     344.1447
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694  102301.4609
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017     759.9937
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426    1315.5334
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133    2397.2681
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483     343.1384
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044     288.5468
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152    1530.5657
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996  0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609    4275.3457
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565    1410.8372
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118    2273.4758
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064     683.2795
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995    1026.7981
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945     259.7421
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978  0.00367    3002.8828
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033    3446.5520
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983     270.7949
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462     716.5762
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413  0.00528     355.3475
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714  480535.5000
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955     713.4463
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257     549.3252
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204     618.2027
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099     909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308     422.5022
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960     884.2003
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499     617.3855
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050    1471.4346
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282    1659.6211
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523     888.2811
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886   43249.3945
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161  0.00709    1181.2555
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884     672.2466
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253    1237.0536
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249     148.1857
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983