Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261    323.8441
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507 159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292   1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403   1787.3024
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770    474.5059
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355  14919.6260
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070   2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672    565.1107
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949   1182.1506
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299    688.0706
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553   1750.4175
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813 287033.4688
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241   1752.9242
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261   3353.5422
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957    495.4453
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001   1184.7952
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875    597.2893
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856   1789.1596
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603   1562.0503
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301    338.9676
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367   3002.8828
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943    371.5514
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983    270.7949
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099   1146.2140
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312    593.7773
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308    422.5022
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805   7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419    323.0790
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156    758.4840
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491