Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501   -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919   -0.04051  351155.1250
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958    0.02642     671.1807
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199    0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021   -0.01579    1318.7560
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917    0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896     276.0310
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273   -0.02403    1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104   -0.02046    1024.1573
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491    1844.4353
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588    1295.1610
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024     394.8751
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475    0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696    0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272    4955.5586
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510    3342.5518
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602    0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471   -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746    0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626    4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426   -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856    0.01114    4510.8833
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643    0.00314    2360.1501
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206    0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917     116.1986
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252    0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452   -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205     629.4165
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097    0.00156     816.6036
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916    2009.1686
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184    0.03924    8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440   -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743    0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543   -0.05146    1347.6224
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255   -0.00365    1536.2647
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904   -0.00866   29195.6016
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511  361979.2812
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338   10787.6299
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796    4729.6382
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025    4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843    0.04407     344.8395
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492    0.00072   54831.6289
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530     937.1306
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008  825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478    0.01092     412.5849
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588     749.2502
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194    0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762    0.01252    5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658    0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458   -0.00401    1722.6481
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807   -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548    0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062   -0.00262    1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941    1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102   35685.0000
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927   -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013    4229.5708
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567   -0.03793     864.9472
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017    0.01213     328.7922
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129    0.00171     788.0004
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722    1373.4147
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247    0.01123     364.2903
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801    1071.6141
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970    0.00200    1305.6691
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153   -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140    1066.3957
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143    0.00477    2945.2100
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610     454.8103
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360   -0.03000  162178.1562
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744     226.7911
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465   -0.01036    2505.8369
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683   -0.00603    1562.0503
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585   -0.02665    1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427     866.9398
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332    0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325   -0.00505     351.4078
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570    0.02230     714.7432
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284    0.01085    2229.5105
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156   -0.00116    1817.5222
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715     520.7355
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650    1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945     259.7421
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464    4029.5947
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199     836.4196
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327    0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341    0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167    0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527   -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987   -0.00507     504.3085
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001    0.00759    1880.4403
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457    0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084   -0.00083    1309.2347
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175     725.1355
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490   -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662    1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452   -0.01298     436.1459
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250    2225.4426
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009   -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229    0.00099     909.0560
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608    0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523    0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420    0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236   -0.00066     306.0146
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600   -0.01849     764.8193
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552    0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264    0.00624    1870.8307
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282    1659.6211
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129    0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148     481.8292
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284   -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108   -0.00948    1538.1741
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983    0.00803    1058.5630
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458     485.7418
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253    1237.0536
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751    0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403     609.8665
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235