Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398    267.2659
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114   4510.8833
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325    616.8652
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276   9614.6924
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034   4495.6929
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796   4729.6382
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.95061 -0.00762  0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405 -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330   1456.9246
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684   2556.8049
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026    247.0894
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656   3036.7383
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620    379.0855
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354   1147.7384
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332 -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728   1137.2450
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.52758 -0.00193  0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759    158.9370
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210    585.8992
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291    487.3438
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655    130.9468
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226  0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118   2273.4758
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409   1575.9425
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833    300.0125
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 18627 714292.5 9559252 2330.78 18.49622 -0.00538 -0.00052   3508.5591
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943    371.5514
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990  0.02790    209.6952
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218  0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650    766.6254
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250   2225.4426
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251    168.1733
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060    383.3547
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438   1276.5448
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858    298.4490
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536
#> 41373 715332.5 9560862 1873.10 29.12459 -0.00166  0.00156    758.4840
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235