Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607