Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305   -0.00555    430.4604
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753    0.01487   1077.7585
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046    0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373   -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433   1475.3191
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884    0.01345   1120.6929
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223    0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214    0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311    586.7354
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021   -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189    287.5425
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475    0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160   9344.2930
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530   9000.4356
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312   2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203  10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426   -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581   1855.1630
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643    0.00314   2360.1501
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593    939.0466
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225    0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422    0.02357    275.7474
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916   2009.1686
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077    300.7915
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743    0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545    0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154    0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047   -0.00887    366.8537
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904   -0.00866  29195.6016
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458    0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796   4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178    800.1933
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025   4425.5151
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060    0.00270   4057.5637
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585   -0.00974   2136.6939
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855   -0.02175   1122.9120
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175    0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825    162.6688
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542    860.8840
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576   -0.00166    650.6246
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739    365.7626
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625   1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553   1750.4175
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800   -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539    0.00809   1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410    0.03620    379.0855
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318    289.0063
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320    0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472   -0.00228   2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538    0.00112    745.3315
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664   1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133    0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242    507.9263
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335    195.0293
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205   1112.2626
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589    0.00849   1120.5961
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212  46310.0195
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143   -0.05078 622888.2500
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218   -0.00372   4154.5786
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466   -0.01776   1021.8267
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789   -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253    0.00753   1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970   1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070    0.00010   3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546    0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003    0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018    0.02281    0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317    0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495   -0.03756    353.9086
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296    0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153   -0.02417  34908.6562
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267    0.00483    343.1384
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004   -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400   -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529    0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253    528.1924
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996    0.04684    209.1778
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166   -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251   -0.00389  14516.2002
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336   2806.2644
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118   2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815   -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064    683.2795
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306   -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409   1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548    532.8210
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502   1428.8546
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527   -0.02077   8759.9961
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475    515.1661
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084   -0.00083   1309.2347
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170    253.5129
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452   -0.01298    436.1459
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366    267.5057
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379   -0.01018    937.6911
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035    0.00955    713.4463
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097    0.01487   1769.3392
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360   4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608    0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502    520.5218
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273   -0.01047   2826.7700
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347   1893.3645
#> 280   713182.5 9560632 1862.83 39.004102    0.00745    0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309    0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673    0.00788   -0.01029    894.2672
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573   1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135   1293.4855
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762