Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501   -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919   -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106   -0.04634     500.5027
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056   -0.06976     301.2347
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199    0.00659    3553.7166
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633   -0.01813    1130.5791
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292    0.00712    3636.5303
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059   -0.00628     556.0121
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065    0.00005     769.3426
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059   -0.07431    4079.0366
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679     550.0344
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645    0.00835    2425.3115
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414   -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525   -0.11024     394.8751
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169   -0.00160    9344.2930
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602    0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029   -0.02312    2657.6936
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537   -0.00203   10263.2734
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225    0.00325     616.8652
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616    0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275    0.00605     798.6775
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646   -0.04086    6042.5757
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440   -0.02651     419.0927
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804    0.00346     543.0749
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255   -0.00365    1536.2647
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904   -0.00866   29195.6016
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888    0.00862     874.5408
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509    0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458    0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321    0.01169     350.1799
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174    0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288    0.01452     437.3965
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735   -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617    0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149    0.01949    1182.1506
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008    0.00102     483.4184
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480    0.02530     937.1306
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175    0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194    0.00516     576.0335
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953   -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681   -0.01319    3457.7519
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576   -0.00166     650.6246
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032    0.00542    3614.7334
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034    0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062   -0.00262    1760.2162
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406   -0.00055     773.0359
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927   -0.00316     454.9674
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238   -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028    0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478   -0.07132   42970.7891
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627   -0.05393    8301.6094
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093    0.00664    1807.2659
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370   -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865    0.00335     195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246    0.00534     344.1447
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986   -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129    0.00171     788.0004
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491    0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989   -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489    0.00241    1752.9242
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171   -0.00261    1627.8989
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379    0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079    0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428   -0.00522    1355.7436
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924    0.02606     564.6876
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011    0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789   -0.00261    3353.5422
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213   -0.01363     565.5207
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546    0.01484     495.0292
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281    0.04759     158.9370
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292   -0.00109    2047.9031
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153   -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693    0.01133    2397.2681
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010    0.00140    1066.3957
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864    0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019    0.00119    6589.5796
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125   -0.02685    2568.4631
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622    0.00152    1530.5657
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118   -0.00252    3631.7017
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004   -0.02384     947.2075
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400   -0.01210     585.8992
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465   -0.01036    2505.8369
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427    0.02403     474.8613
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378    0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332    0.01842     246.5161
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159   -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196    0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325   -0.00505     351.4078
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759    0.00158     837.3730
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738    0.00118    2273.4758
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112    0.03538     265.4677
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825    0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480    0.01500     538.8209
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075   -0.01945     259.7421
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253   -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544   -0.01014     271.4219
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007   -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174   -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001    0.00759    1880.4403
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065    0.00475     515.1661
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084   -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737    0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196   -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844    0.00514     897.9290
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490   -0.00780     825.4451
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674   -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218    0.00902     218.7467
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393    0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068   -0.00312     593.7773
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112   -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586    0.00654    1576.2065
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066    0.05257     549.3252
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233    0.01282     561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972    0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804   -0.00165     805.1403
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537   -0.00577     423.5598
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820   -0.00260    1484.6593
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044   -0.00964     543.0025
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375    0.00215     917.5296
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182   -0.00972     331.7462
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572    0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909   -0.01151  143213.1562
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108   -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253   -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278   -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263   -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340    0.01809    1158.0426
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549   -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992   -0.01052     451.7209
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282    0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021    0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933   -0.00157     766.0220
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011   -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962   -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288    0.00448     941.8503
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456   -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505    0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491