Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357    1752.0375
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057    1630.3259
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174    2310.4871
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387    1989.3012
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466     724.0713
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281    1325.6138
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024     394.8751
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766  140591.4062
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530    9000.4356
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710     843.7961
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678    1891.6584
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774     527.1068
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616  0.01276    9614.6924
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167    1938.5389
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752 -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796    4729.6382
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613     289.6113
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907    1943.4092
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949    1182.1506
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102     483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530     937.1306
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560     977.4719
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241     585.6182
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408   76318.0156
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166     650.6246
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596     842.0994
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478 -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081     940.6002
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180    3904.5254
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112     745.3315
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664    1807.2659
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389   47022.9688
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236    9755.3027
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583    1898.0105
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713    6379.7974
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094    1392.6008
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705     413.5702
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610    1872.9276
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875     597.2893
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152    1530.5657
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427  0.02403     474.8613
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265    1148.6487
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476    1275.9446
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712    1010.5097
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403     241.2991
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409    1575.9425
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273   72123.6172
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167  0.01133     828.8604
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065     704.4843
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399    1601.4852
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170     253.5129
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400    4126.7812
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099    1146.2140
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655     560.8554
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036     476.9504
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247     282.1076
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964     543.0025
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600 -0.01849     764.8193
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972     331.7462
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886   43249.3945
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377    1727.6055
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234    1707.3130
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135    1293.4855
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525     162.0592
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769