Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507 159065.0781
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890   2301.0454
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761    928.6858
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466    724.0713
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626   4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593    939.0466
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605    798.6775
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672    565.1107
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025   4425.5151
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199    196.4080
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330   1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427    380.9472
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009   1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473   2930.2791
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485   2240.4185
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809   1102.9956
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393   8301.6094
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220   8936.2666
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241   1752.9242
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261   3353.5422
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981   3975.3608
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856   1789.1596
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011    201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714 480535.5000
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513    161.9878
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250   2225.4426
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099    909.0560
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308    422.5022
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532  16771.3066
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282   1659.6211
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064   1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156    758.4840
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946