Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041    772.9746
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769