Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958   0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393   2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357   1752.0375
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969    368.5606
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917   0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487   1077.7585
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503   0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174   2310.4871
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491   1844.4353
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088    266.4355
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269    491.8017
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518    854.4269
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710    843.7961
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973    514.5488
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108  -0.00678   1891.6584
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150   1274.3594
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051    779.0555
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924   8023.8657
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405   5681.2412
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294  -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373  16031.7500
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321   0.01169    350.1799
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907   1943.4092
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310   2833.9001
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063    576.2819
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401   1722.6481
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299    688.0706
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201    788.0621
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132  42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594  16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664   1807.2659
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245   6353.3179
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911    173.2806
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439  -0.01161   3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123    364.2903
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739    789.7573
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879   1200.4210
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450   2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143   0.00477   2945.2100
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216  -0.02606    919.4241
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632    620.0229
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116   1817.5222
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738   0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208    951.3401
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463    851.6306
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476  11992.5273
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790    209.6952
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022   0.00662   1269.2650
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099   1146.2140
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645    234.0114
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868    235.8511
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361   1011.0248
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305  -0.01135   1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250