Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183   -0.09203    634.3320
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199    0.00659   3553.7166
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659    0.01041    772.9746
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498   -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333   -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065    0.00005    769.3426
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059   -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118    0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250   -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891   -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884    0.01345   1120.6929
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175   -0.08226   2325.2053
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160    0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841   -0.00010   1527.1028
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123    0.01164    895.6606
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131    0.00501   3889.7292
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471   -0.01382    560.8309
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672    0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537   -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151    0.00899   1394.2274
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716    0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051    0.00889    362.8363
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993   -0.00413    364.9788
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356    0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664    0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382    0.00088   2322.1746
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669   -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598    0.01212    561.1948
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452   -0.01437   1155.1469
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211    0.00642    613.6390
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414    0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184    0.03924   8023.8657
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545    0.02405   5681.2412
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547    0.03323    183.1836
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379    0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458    0.00672    565.1107
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752   -0.00517   1000.7119
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527    0.04613    289.6113
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034    0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552   -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735   -0.00925    665.6116
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008    0.00102    483.4184
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286   -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749    0.04779   2972.0193
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372   -0.00008 825875.6875
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449   -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399   -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119   -0.01560    977.4719
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504    0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017    0.00197    711.1091
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658    0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681   -0.01319   3457.7519
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027   -0.01473   2930.2791
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410    0.03620    379.0855
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028    0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478   -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165    0.00555   4439.0967
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588    0.01448   1258.8196
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627   -0.05393   8301.6094
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188    0.00832   3022.3462
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246    0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986   -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509   -0.01211  49347.5547
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589    0.00849   1120.5961
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474    0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384    0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444   -0.00017    759.9937
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707    0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379    0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924    0.02606    564.6876
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011    0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442   -0.00118  19557.6621
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546    0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102   -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281    0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495   -0.03756    353.9086
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296    0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153   -0.02417  34908.6562
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267    0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881    0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185   -0.01185    953.8614
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019    0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445    0.00875    597.2893
#> 30626 713302.5 9559172 2106.91 35.99779  0.01354    0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902    0.01358    802.7622
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115    0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770   -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966   -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465   -0.01036   2505.8369
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614   -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166   -0.01265   1148.6487
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332    0.01842    246.5161
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534    0.01476   1275.9446
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226    0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284    0.01085   2229.5105
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390   -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336    0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007   -0.00463   1050.7473
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112    0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815   -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494    0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220   -0.03290   2481.5632
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161    0.00171    756.8818
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846    0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176    0.02973    215.5101
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167    0.01133    828.8604
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978    0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.49622 -0.00538   -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174   -0.03065    704.4843
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737    0.01033   3446.5520
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452   -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358    0.02422    967.9738
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185    0.01655    560.8554
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785    0.00645    234.0114
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978    0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035    0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118    0.00162    290.2031
#> 330   714232.5 9560362 2165.00 18.06078  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923    0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009   -0.00599  12873.8018
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600   -0.01849    764.8193
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264    0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.00410  0.00745    0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502    0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598    0.02208    842.4825
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182   -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452    0.00368   1071.8734
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253   -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203    0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163   -0.00377   1727.6055
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788   -0.01029    894.2672
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992   -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983    0.00803   1058.5630
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627   -0.00124  25826.6914
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087   -0.00573   1490.0312
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023   -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181   -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607