Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.9523260   -0.00123   0.00393    2081.0056
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813    1130.5791
#> 14254 714852.5 9557882 2680.00 23.3354251    0.00479   0.01261     323.8441
#> 30059 713832.5 9559662 2288.02 18.2779266    0.01917   0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753   0.01487    1077.7585
#> 42733 715112.5 9559062 2399.70 36.0545152   -0.00041   0.01741     746.4155
#> 14456 713542.5 9559972 2184.39 35.2414881   -0.00841  -0.00010    1527.1028
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951  -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150   -0.00284   0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.9691094   -0.00295   0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979  -0.01542     674.0147
#> 125   713992.5 9561062 1878.28 49.9710871    0.00183   0.03387    1989.3012
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223   0.01783     377.8918
#> 22814 713862.5 9558582 2327.48 48.6120312   -0.02894  -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423    0.00214   0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160   0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.6648828   -0.00831  -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542   0.01842     868.6137
#> 3835  714142.5 9558552 2388.42 32.9003825    0.00979  -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581   0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110    0.00414  -0.00104    1641.4587
#> 30800 715182.5 9557582 2772.37 39.2040642   -0.02475   0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.9825182   -0.02123   0.01164     895.6606
#> 16843 714882.5 9558512 2701.78 26.6379538   -0.00270  -0.00769    1032.7177
#> 15109 713872.5 9559602 2288.64 25.1855058    0.00602   0.01518     854.4269
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471  -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746   0.00096    3210.5781
#> 22566 712832.5 9559842 1897.49 33.0602377    0.00672   0.04398     267.2659
#> 6676  715092.5 9559912 2217.44 41.7078897   -0.00077  -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494   -0.00493   0.00733     598.7983
#> 45575 715432.5 9559452 2305.28 24.2343959    0.00993  -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192    0.00490  -0.00450     873.0448
#> 47220 715052.5 9557752 2726.93 32.7216197   -0.00664   0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.0720602    0.01257   0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.2735877    0.00382   0.00088    2322.1746
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937   0.00083     994.7876
#> 13847 715052.5 9559232 2337.87 26.1148433   -0.01452  -0.01437    1155.1469
#> 3466  713232.5 9560612 1812.72 51.5662016   -0.01804  -0.01596    5296.9014
#> 4329  712882.5 9559982 1911.46 26.2317267   -0.00193  -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.6644985    0.06765   0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.6887865   -0.00440  -0.02651     419.0927
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545   0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.3075669   -0.00154   0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.6343274   -0.00669  -0.00991    1403.0193
#> 17594 714832.5 9561052 1818.04 46.1672203   -0.01353   0.00063    5035.0000
#> 38686 714292.5 9557832 2496.22 38.4809914   -0.00509   0.00499     885.0064
#> 47884 715032.5 9557712 2738.31 28.2892182    0.00006   0.00175     804.5679
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661   0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903    2360.1536
#> 2976  715122.5 9559142 2390.05 33.1679538    0.01008   0.00102     483.4184
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.1431350   -0.00759  -0.00400     608.6523
#> 25244 713472.5 9559092 2137.03 18.3489734   -0.10372  -0.00008  825875.6875
#> 421   714892.5 9561022 1849.01 26.1320321    0.00002   0.00588     749.2502
#> 40733 712862.5 9558892 2108.60 29.9949135    0.00175   0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.9863262   -0.02917  -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.0208853    0.00194   0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504   0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.3121575   -0.00405  -0.01565    2232.6133
#> 42941 713252.5 9559742 1990.10 28.9544222   -0.04708   0.00408   76318.0156
#> 32909 712812.5 9560452 1883.85 38.2512354   -0.00221   0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.8423462   -0.01057   0.02427     380.9472
#> 31462 714082.5 9558542 2367.84 34.7269720   -0.02027  -0.01473    2930.2791
#> 20065 713772.5 9558852 2221.21 28.0199280   -0.00568  -0.02553    1750.4175
#> 46065 713072.5 9559162 2109.39 16.2645529   -0.00195  -0.00656    3036.7383
#> 16370 713852.5 9558632 2296.97 34.6244125   -0.01972   0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.7772667   -0.01611   0.02961    1086.6963
#> 28046 714882.5 9558492 2711.77 28.3058340   -0.00131   0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.7703772    0.00904  -0.02013    4229.5708
#> 20601 715542.5 9558182 2773.29 32.7554242   -0.00604   0.00354    1147.7384
#> 7385  713752.5 9560122 2064.61 31.7710827   -0.02605  -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.0239387    0.00538   0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024  -0.00474    2343.2100
#> 8136  713242.5 9560602 1798.18 41.2403562   -0.01169  -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556  -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178  -0.00822     472.8748
#> 33910 715202.5 9559572 2195.66 13.7916671   -0.04376  -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.9386311   -0.01491   0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781   29995.1543
#> 47702 714692.5 9559712 2200.51 33.0986259   -0.13038  -0.04212   46310.0195
#> 43201 715712.5 9557922 2852.82 31.4622584   -0.00949   0.00520    3390.5708
#> 27915 715232.5 9559552 2208.93 25.4805791    0.00171  -0.00261    1627.8989
#> 15981 713302.5 9559172 2106.91 35.9977924    0.01354   0.00245     629.5983
#> 10689 713542.5 9559132 2211.95 31.9813582    0.05379   0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.9597815   -0.01439  -0.01161    3065.3694
#> 39325 714142.5 9558592 2357.72 40.7928761   -0.00246   0.00396     704.7076
#> 332   714852.5 9561102 1787.11 45.3582039    0.00540   0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.8127340    0.00546   0.01484     495.0292
#> 39624 715222.5 9559552 2205.54 23.8974967    0.00401  -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576   -0.00767  -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016  -0.00294     734.9908
#> 1223  713632.5 9559632 2226.22 50.8087514   -0.01881   0.00121    1783.0658
#> 39405 713122.5 9559052 2169.30 37.6461919   -0.02110  -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.0716969    0.00285  -0.01325     595.0426
#> 37115 715612.5 9559202 2358.15  9.7276774   -0.01020   0.00279   39495.1797
#> 48461 714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234     849.0547
#> 9984  712742.5 9560092 1848.56 23.6127367    0.00614  -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.7532859    0.00166  -0.01265    1148.6487
#> 23176 714012.5 9558892 2308.59 43.0291304   -0.01084  -0.02996    1599.3491
#> 31983 714302.5 9559912 2204.77 23.4649135   -0.01332   0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.2968763   -0.00079  -0.00331     699.3024
#> 25584 715062.5 9558532 2675.93 37.6885908    0.00534   0.01476    1275.9446
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009   0.00109     605.1908
#> 24836 714202.5 9558272 2507.87 38.4890128   -0.00226   0.00736     477.8591
#> 14694 715032.5 9559482 2307.44 44.1733271   -0.00007  -0.00463    1050.7473
#> 14717 715382.5 9558062 2799.05 50.5904544   -0.00812  -0.02208     951.3401
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174  -0.01466    2633.9666
#> 19493 714252.5 9560372 2156.93 24.8795463    0.00825   0.00445     355.5078
#> 13512 715152.5 9557992 2778.81 43.3895209   -0.00724   0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.5533166   -0.00502  -0.00028    2394.7261
#> 29125 714992.5 9558792 2584.41 24.6658331    0.01253  -0.00333     422.8840
#> 39796 714802.5 9560372 2046.48 40.1557471    0.00176   0.02973     215.5101
#> 13286 713662.5 9559702 2268.14 41.9966605   -0.01341   0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167   0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527  -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.4734869    0.00987  -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235   -0.00538  -0.00052    3508.5591
#> 48183 713642.5 9558262 2314.44 50.0232262    0.01338  -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.0621382    0.00196   0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.9416565   -0.00065   0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069  -0.00900     614.5898
#> 8413  713242.5 9561232 1967.45 38.1945125    0.00236   0.00483    1097.3082
#> 37029 713022.5 9559082 2163.08 36.0321698    0.00490  -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.8637272    0.01180   0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.9048126   -0.00452  -0.01298     436.1459
#> 3733  713762.5 9559632 2297.63 34.2147477   -0.01338   0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645     234.0114
#> 13908 715022.5 9559372 2290.81 27.2429972   -0.03004   0.00714  480535.5000
#> 36889 714302.5 9559802 2243.60 21.1662069    0.01923   0.05547     128.2613
#> 34532 713812.5 9559502 2328.80 44.9376528   -0.00070  -0.00060     383.3547
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209    0.00980  -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.3936851   -0.02608   0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.5974835    0.02523   0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.4471659    0.00360   0.00750    1892.8978
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719   0.00079    1647.6451
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074  -0.02904     608.3561
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354   0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.0318136    0.01600  -0.01849     764.8193
#> 19767 714272.5 9558052 2431.67 55.7247292    0.00361  -0.01841     756.0055
#> 34334 714642.5 9557912 2582.25 30.9912871    0.00253  -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.0021874    0.00203   0.00238     787.0053
#> 33881 712992.5 9558822 2173.26 33.1977476   -0.00595   0.03064    1485.5201
#> 35257 713822.5 9557982 2350.80 29.3079371    0.01549  -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992  -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678    0.00732   0.01858     298.4490
#> 47245 715372.5 9560792 1900.65 34.4708598   -0.00011  -0.00419     323.0790
#> 41373 715332.5 9560862 1873.10 29.1245906   -0.00166   0.00156     758.4840
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946