Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969    368.5606
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387   1989.3012
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466    724.0713
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447    331.4833
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250   3686.1240
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382    560.8309
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899   1394.2274
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150   1274.3594
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325    616.8652
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511 361979.2812
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338  10787.6299
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178    800.1933
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307 147470.5312
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236   2715.4138
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306 -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090 -0.00299    688.0706
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543    352.1558
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109    767.4097
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664   1898.5588
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081    940.6002
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627 -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664   1807.2659
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241   1752.9242
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299  0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583   1898.0105
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722   1373.4147
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484    495.0292
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200   1305.6691
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010  0.00140   1066.3957
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705    413.5702
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226    338.3474
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875    597.2893
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000 162178.1562
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063   2430.2278
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320    911.5911
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105   4432.8760
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116   1817.5222
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363    389.2691
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500    538.8209
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548    532.8210
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333    422.8840
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014    271.4219
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167  0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.3155672  0.04152  0.00098    164.0802
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268    252.0072
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948   1222.5433
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868    235.8511
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308    422.5022
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960    884.2003
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836    250.5111
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368   1071.8734
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585   5714.3267
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172    351.5812
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983