Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea   cslope
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056 37.59291
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861 39.33527
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310 20.93129
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537 24.43780
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312 26.97714
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679 34.63988
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871 34.08239
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436 15.98781
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028 20.06899
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355 21.61140
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354 20.13374
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104 24.77412
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751 51.20810
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062 28.52642
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518 29.51363
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356 28.85244
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665 34.23308
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121 43.53161
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363 21.07396
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622 22.71778
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448 25.59975
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584 40.08470
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473 36.48022
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474 18.83427
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059 29.32226
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165 26.74109
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014 26.79494
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657 34.59576
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927 36.51689
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915 14.65970
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548 17.99259
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937 32.09767
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690 18.76379
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389 27.41603
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016 39.28772
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119 21.67270
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382 22.18378
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933 39.47851
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536 28.61065
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871 24.82626
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876 40.43707
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193 27.70251
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306 38.83279
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141 33.13472
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120 32.17559
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875 33.21895
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849 21.71911
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502 10.71374
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091 27.97409
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622 34.90001
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246 28.79056
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933 41.73367
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924 21.26705
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481 31.44564
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390 28.95270
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413 29.16814
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791 33.75466
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993 36.32037
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150 33.24187
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768 25.46053
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326 19.94237
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000 35.45348
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818 26.09307
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359 31.44278
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674 28.24682
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621 19.59115
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967 25.22962
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002 28.18609
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445 28.34823
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254 24.66526
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840 44.59617
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748 29.98976
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179 30.34442
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242 30.97181
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961 33.41318
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420 34.20443
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195 33.66643
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247 42.93574
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818 34.18954
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454 33.81368
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771 26.59785
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076 37.44852
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608 45.09350
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351 28.98765
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031 29.27356
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249 26.77833
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468 34.69145
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426 28.17348
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463 31.23193
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000 24.74777
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841 16.57911
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562 31.62555
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983 30.00351
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075 33.11581
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782 37.47316
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278 27.55870
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438 23.34574
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487 27.25159
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866 38.66835
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782 27.41603
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775 38.37385
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908 15.49736
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097 35.04611
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222 31.73785
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473 43.51156
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525 33.38396
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538    265.4677 21.42977
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795 23.52737
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549 35.46265
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421 14.48380
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661 37.33851
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347 24.13757
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082 39.88989
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520 31.24855
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355 25.00732
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650 31.52299
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236 34.78369
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812 25.85357
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802 23.68436
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504 40.96018
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475 25.23077
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114 32.70615
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911 44.39506
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547 32.23690
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738 32.87976
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426 40.20388
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613 19.64500
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099    909.0560 44.22776
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598 26.99204
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593 28.51496
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561 30.96436
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193 45.58395
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074 37.77167
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626 19.11674
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292 24.25731
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741 40.60380
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906 23.09593
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053 39.52377
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055 31.81463
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811 26.85625
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867 18.13927
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201 29.18590
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267 33.56329
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975 30.88185
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490 17.04435
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606 38.91071
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300 29.05011
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312 22.23706
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414 16.97846
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820 40.15861
#>       distroad slides distdeforest distslidespast log.carea
#> 15412   300.00   TRUE       300.00            100  3.318273
#> 40756   300.00   TRUE       300.00            100  3.000471
#> 46178   300.00   TRUE       300.00             31  2.440958
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 8059    300.00   TRUE       300.00              4  3.150244
#> 34512   300.00   TRUE       195.00              2  3.224136
#> 44952   300.00   TRUE       300.00            100  3.363704
#> 26531   300.00   TRUE       300.00              2  2.552843
#> 14456   300.00   TRUE       247.02            100  3.183868
#> 23817   300.00   TRUE       300.00              4  2.425592
#> 11463   300.00   TRUE       300.00            100  2.768442
#> 10899   300.00   TRUE       300.00             20  2.644350
#> 33792   300.00   TRUE       300.00              5  2.596460
#> 18843   300.00   TRUE       300.00            100  5.147959
#> 40307    77.82   TRUE        76.41             54  3.524078
#> 29420   253.02   TRUE         0.00            100  3.954264
#> 38070   300.00   TRUE       300.00            100  2.899802
#> 47022   300.00   TRUE       300.00             26  2.745084
#> 4827    300.00   TRUE       300.00              0  2.559711
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 40828   300.00   TRUE       300.00            100  2.941037
#> 3453    300.00   TRUE       300.00              2  3.276843
#> 47220   300.00   TRUE       300.00            100  3.327041
#> 3133    300.00   TRUE       300.00            100  2.440511
#> 15390    60.00   TRUE       118.92              2  3.372821
#> 23031   300.00   TRUE       300.00              2  2.798938
#> 3466    101.32   TRUE        95.02             19  3.724022
#> 12084   300.00   TRUE         0.00              0  3.904384
#> 22735   300.00   TRUE       300.00             90  2.622310
#> 19128   300.00   TRUE       300.00             15  2.478266
#> 45104   300.00   TRUE       300.00            100  2.894014
#> 47429   300.00   TRUE       300.00            100  2.464778
#> 37828   300.00   TRUE       300.00              2  2.819787
#> 34359   300.00   TRUE       294.31            100  3.287475
#> 37348   300.00   TRUE       300.00            100  4.465317
#> 27084   300.00   TRUE       300.00             81  3.000309
#> 46201   167.75   TRUE         0.00             55  3.674828
#> 43092   300.00   TRUE       300.00            100  2.903195
#> 18394   300.00   TRUE       300.00             57  3.372940
#> 44550   300.00   TRUE       300.00            100  2.968524
#> 26567   300.00   TRUE       300.00            100  3.023906
#> 34707   165.13   TRUE        39.56            100  3.473052
#> 16015   152.09   TRUE        90.23             45  2.971800
#> 39599   300.00   TRUE       300.00             25  3.007839
#> 20220   300.00   TRUE       300.00              8  3.050346
#> 25244   300.00   TRUE       300.00              6  5.916915
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 421     182.02   TRUE         0.00             83  2.874627
#> 375     300.00   TRUE       300.00            100  2.851936
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 7466    255.23   TRUE         0.00             60  3.163437
#> 34005   300.00   TRUE       300.00            100  2.538942
#> 27875   157.93   TRUE         0.00             57  3.709413
#> 18265   300.00   TRUE       300.00            100  3.236197
#> 33350    10.00   TRUE       166.26              2  3.415431
#> 39732   300.00   TRUE       300.00             25  3.364146
#> 31462   300.00   TRUE       300.00              1  3.466909
#> 7668    300.00   TRUE       300.00              5  2.789721
#> 29801   300.00   TRUE       300.00              2  3.026907
#> 23213   300.00   TRUE       300.00            100  2.868983
#> 46783   256.35   TRUE         0.00            100  2.772566
#> 3418    300.00   TRUE       300.00             73  4.552486
#> 10364   179.84   TRUE         0.00              0  3.541090
#> 31561   300.00   TRUE       300.00              0  2.888200
#> 16355   300.00   TRUE        58.52            100  2.657980
#> 28046   300.00   TRUE       300.00             15  2.896560
#> 26010    69.52   TRUE       166.56              2  3.647295
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 21680   104.12   TRUE       110.76              4  3.591103
#> 5239    197.29   TRUE         4.67             57  3.591568
#> 6845    300.00   TRUE       300.00             29  2.769292
#> 18958   300.00   TRUE       300.00              5  2.674746
#> 20368   300.00   TRUE       166.26            100  3.803001
#> 6217    300.00   TRUE       300.00              1  3.243763
#> 16009   300.00   TRUE       300.00            100  3.049449
#> 7086    300.00   TRUE       128.61             63  3.071308
#> 47702   300.00   TRUE       300.00             86  4.665675
#> 34887   300.00   TRUE       300.00            100  3.159244
#> 34615    17.57   TRUE       142.95             16  3.011436
#> 22907   300.00   TRUE       205.28              1  3.147505
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 39325   300.00   TRUE       300.00             60  2.848009
#> 10363   300.00   TRUE       297.43              6  3.599377
#> 49072   300.00   TRUE       300.00            100  3.662232
#> 1916    300.00   TRUE       300.00             21  3.311309
#> 23626   300.00   TRUE       300.00            100  3.336505
#> 44612   300.00   TRUE       300.00            100  2.460216
#> 24723   300.00   TRUE        36.63            100  2.774548
#> 42617   300.00   TRUE       300.00            100  2.848895
#> 22847   300.00   TRUE       300.00             41  3.134623
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 28588   300.00   TRUE       300.00             78  5.209992
#> 30626   300.00   TRUE       300.00              0  2.799064
#> 47722   300.00   TRUE        65.02              1  2.976445
#> 35458   300.00   TRUE       300.00            100  3.297274
#> 31491   300.00   TRUE       300.00            100  3.385647
#> 49346   300.00   TRUE       300.00             29  2.687835
#> 39466   300.00   TRUE       300.00            100  3.060187
#> 31253   300.00   TRUE       122.87            100  3.878395
#> 10145   276.16  FALSE        91.33            100  3.227418
#> 16208   256.60  FALSE         0.00             67  2.937807
#> 27884   300.00  FALSE       300.00            100  2.781892
#> 16464   300.00  FALSE       300.00            100  3.004540
#> 31071   300.00  FALSE       300.00             76  3.259480
#> 14694   300.00  FALSE       300.00            100  3.021498
#> 11111   300.00  FALSE       300.00            100  3.344795
#> 20458   300.00  FALSE       300.00             28  2.424012
#> 8066    300.00  FALSE       300.00            100  2.834598
#> 43426   300.00  FALSE        82.11            100  3.910718
#> 25140   104.12  FALSE        46.28             74  2.414542
#> 12974   300.00  FALSE       300.00            100  2.711947
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 8413     80.73  FALSE         0.00            100  3.040329
#> 21925   300.00  FALSE       300.00             95  3.537385
#> 41753   300.00  FALSE       300.00            100  2.860419
#> 14800   300.00  FALSE       300.00            100  3.103552
#> 33175   300.00  FALSE       300.00            100  2.799219
#> 36245   110.19  FALSE        81.50             20  3.615611
#> 22545   300.00  FALSE       300.00             96  2.215056
#> 13737   300.00  FALSE       300.00            100  2.678473
#> 29107   300.00  FALSE       300.00            100  2.550653
#> 39440   300.00  FALSE       300.00             92  2.369237
#> 34345   300.00  FALSE       300.00            100  2.972060
#> 12044   300.00  FALSE       300.00            100  3.304782
#> 33410   300.00  FALSE       300.00            100  2.986312
#> 35038   300.00  FALSE       300.00            100  3.347416
#> 36889   300.00  FALSE       300.00             48  2.108096
#> 35826   298.23  FALSE        16.35             96  2.958591
#> 17628   300.00  FALSE       205.59            100  2.626915
#> 23499   300.00  FALSE       300.00             50  3.171627
#> 9007    300.00  FALSE        14.98             35  2.784158
#> 2999    300.00  FALSE       300.00            100  2.883559
#> 4943    300.00  FALSE       300.00            100  3.019161
#> 4269    300.00  FALSE       194.53            100  3.439564
#> 37783   300.00  FALSE       300.00             18  2.682893
#> 11394    46.04  FALSE        55.00            100  3.187005
#> 18504   144.96  FALSE         4.48             89  5.049117
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 28526   300.00  FALSE        46.28            100  3.237445
#> 14881   300.00  FALSE       300.00            100  2.948550
#> 9320    300.00  FALSE       300.00             93  2.700519
#> 33881   300.00  FALSE       300.00            100  3.171879
#> 46639   300.00  FALSE       300.00            100  3.756965
#> 35257   300.00  FALSE       300.00             79  2.812912
#> 39883   233.81  FALSE       219.23            100  2.474870
#> 46148   300.00  FALSE       300.00            100  3.015637
#> 24962   300.00  FALSE       300.00            100  2.996437
#> 42990   205.00  FALSE         0.00            100  3.173195
#> 13548   300.00  FALSE       300.00            100  2.256336
#> 30296   300.00  FALSE       300.00             68  3.258250