Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659    3553.7166
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345    1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491    1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070  138869.3906
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678    1891.6584
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416    1204.8629
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407     344.8395
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009    1247.7061
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643     670.2289
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599    1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078  622888.2500
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801     172.7491
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456    1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981    3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077    3056.9475
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121    1783.0658
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341    1363.4000
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744     226.7911
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093     118.1402
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476    1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875    1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064     683.2795
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626     450.7346
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833     300.0125
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052    3508.5591
#> 36126 715052.5 9558292 2759.47 29.062138   0.00196    0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900     614.5898
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662    1269.2650
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223    1056.3701
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312     593.7773
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868     235.8511
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836     250.5111
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438    1276.5448
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849     764.8193
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624    1870.8307
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972     331.7462
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222     991.8300
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448     941.8503
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492     772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769