Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 28223 712742.5 9560482 1927.69 34.6679573  -0.00292    0.00712   3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.2779266   0.01917    0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741    746.4155
#> 21939 713642.5 9558712 2299.09 33.7506519  -0.00503    0.02433   1475.3191
#> 27973 715022.5 9559332 2308.28 20.0632631   0.00940    0.00679    550.0344
#> 43327 713962.5 9561082 1839.11 63.3616837  -0.04951   -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.3415951  -0.00979   -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088    266.4355
#> 38993 714202.5 9558402 2475.75 33.1593594  -0.01607    0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321   0.00102    0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.7135983   0.00194    0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.1555304  -0.00769   -0.00281   1325.6138
#> 43007 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328    592.3326
#> 18005 714832.5 9557912 2660.76 31.9120303   0.00191   -0.00261    554.9243
#> 34482 715152.5 9557642 2753.52  4.9956190  -0.01126   -0.02044 129412.3984
#> 31576 713942.5 9558492 2387.12 46.1431560  -0.05634   -0.02626   4316.0977
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295  -0.01440    0.00710    843.7961
#> 38070 715212.5 9559222 2342.41 55.6771737   0.00426   -0.00367    793.9665
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.9395253   0.00902    0.01358    802.7622
#> 34280 713652.5 9560702 1896.37 41.2735877   0.00382    0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.3887102   0.00071    0.00150   1274.3594
#> 23031 714042.5 9558492 2404.94 28.4960559   0.01205    0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.5662016  -0.01804   -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.8039440  -0.01097    0.00156    816.6036
#> 12084 712832.5 9559992 1872.46 56.0547529  -0.03184    0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.6887865  -0.00440   -0.02651    419.0927
#> 11687 714552.5 9560382 2052.66 29.5755084  -0.00255   -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556    660.3690
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.5357857   0.00140   -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.8898455  -0.01993    0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255   0.01458    0.00672    565.1107
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.8815030  -0.00666   -0.00504   1020.2766
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.6457866   0.00060    0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.5997894  -0.00735   -0.00925    665.6116
#> 22332 715482.5 9558922 2475.28 19.9366394  -0.02617    0.00307 147470.5312
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.9680266   0.00107    0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.7252601  -0.01399   -0.00731   1384.7959
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689   0.03504    0.05825    162.6688
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.1373984   0.00953   -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.1948618  -0.02681   -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.2512354  -0.00221    0.00121   4159.6738
#> 6317  715042.5 9557702 2745.36 27.8738238   0.00090   -0.00299    688.0706
#> 44825 714322.5 9557882 2534.80 40.6885978   0.00114    0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328    592.3326
#> 30886 714912.5 9558822 2563.86 49.7539361  -0.00099    0.02169    495.8840
#> 21013 712742.5 9560032 1845.53 29.3526278   0.00568   -0.01558   1504.9717
#> 17728 713092.5 9560652 1900.14 30.2791006   0.00645   -0.00395   2011.3270
#> 38197 712522.5 9560322 1942.04 62.7772667  -0.01611    0.02961   1086.6963
#> 7064  715592.5 9558662 2597.69 34.2221961   0.00099    0.00240   1201.4724
#> 7385  713752.5 9560122 2064.61 31.7710827  -0.02605   -0.03594  16587.9980
#> 39722 715022.5 9557712 2735.53 30.2607659   0.00133    0.00367    806.0966
#> 12772 715172.5 9557642 2759.42 34.8054672  -0.00567   -0.03793    864.9472
#> 7     714952.5 9559862 2284.19 16.1064166   0.00516    0.01574    183.3220
#> 46138 715202.5 9559582 2194.40 18.8600517  -0.02143   -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.4805791   0.00171   -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239  -0.00444   -0.00017    759.9937
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.2527305  -0.00007   -0.02132   1580.1621
#> 5572  715062.5 9558262 2754.60 39.9116671   0.00707    0.01833    369.3771
#> 47086 713692.5 9559972 2175.51 25.0457041  -0.00789   -0.00261   3353.5422
#> 10617 713232.5 9560722 1879.95 26.4970062   0.00258    0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.4334430   0.01410   -0.00380    601.0843
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957    495.4453
#> 22301 713722.5 9560712 1913.37 39.2711002   0.00705   -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426   1315.5334
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.3049047  -0.00605    0.00075   2170.2249
#> 44612 714932.5 9558822 2578.80 38.2552461   0.01257    0.00044    288.5468
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538    807.8237
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921    525.3245
#> 18498 714842.5 9557712 2679.04 28.9395253   0.00902    0.01358    802.7622
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905   6487.0010
#> 10213 715542.5 9558782 2532.10 35.4133754  -0.00893   -0.00867   2247.8748
#> 9864  714952.5 9557592 2752.40 31.9893796  -0.00983    0.01253    528.1924
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063   2430.2278
#> 49346 715322.5 9558782 2553.74 25.3952084   0.00319    0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.8077801  -0.00486    0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205   1112.2626
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996   1599.3491
#> 7317  713682.5 9558582 2347.46 28.3476599  -0.02196    0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.3298701   0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908   0.00534    0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.7567307  -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177    997.3089
#> 25212 714272.5 9559012 2433.45 37.6714021  -0.01909   -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.7598192  -0.02284    0.01085   2229.5105
#> 31071 713472.5 9558462 2298.33 29.1532385   0.00156   -0.00116   1817.5222
#> 11111 715142.5 9558482 2650.65 42.3432999  -0.00533    0.01493   2212.0525
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.6350750  -0.00242    0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.7174203  -0.00494    0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463    851.6306
#> 25140 713332.5 9560692 1809.95 52.2571887   0.00075   -0.01945    259.7421
#> 39796 714802.5 9560372 2046.48 40.1557471   0.00176    0.02973    215.5101
#> 39685 715332.5 9558102 2797.04 47.7881179   0.00167    0.01133    828.8604
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052   3508.5591
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943    371.5514
#> 31884 715782.5 9558612 2632.60 30.2796736   0.00084   -0.00083   1309.2347
#> 44668 712612.5 9560472 1972.24 27.5071308   0.00601   -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.8637272   0.01180    0.01170    253.5129
#> 6143  712762.5 9559022 2096.24 29.8963648   0.00358    0.02422    967.9738
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.2429972  -0.03004    0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654   1576.2065
#> 36721 714092.5 9559752 2301.81 28.9332227   0.03418    0.00513    161.9878
#> 44715 714932.5 9560132 2195.78 30.3936922   0.03479    0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.1662069   0.01923    0.05547    128.2613
#> 48988 714682.5 9560852 1942.97 20.2907274   0.00480    0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 5861  713882.5 9559182 2350.43 27.3936851  -0.02608    0.00148  10771.1963
#> 48184 714202.5 9558222 2489.17 36.1135298   0.02236   -0.00066    306.0146
#> 12300 712372.5 9560062 1903.64 50.0369772   0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.4446854  -0.00070   -0.00050   1471.4346
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875   2156.9358
#> 19767 714272.5 9558052 2431.67 55.7247292   0.00361   -0.01841    756.0055
#> 14881 715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.0800956  -0.00172   -0.00119    792.3544
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884    672.2466
#> 42990 712432.5 9560692 2102.11 32.4500377  -0.00087   -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.4708598  -0.00011   -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868   0.00962   -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.5327393  -0.00305   -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.1245906  -0.00166    0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.5082486   0.02751    0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.7319259   0.07379   -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.2726234  -0.01181   -0.01149   4242.3579
#> 40043 713322.5 9560562 1802.09 39.8520794   0.04456   -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491