Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357   1752.0375
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345   1120.6929
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835   2425.3115
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194  0.00526    539.8119
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203  10263.2734
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916   2009.1686
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077    300.7915
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661  0.01199    196.4080
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307 147470.5312
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964    571.1364
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357    473.4709
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602    345.8933
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941   1290.7709
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512  13178.3223
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469   1021.5407
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793    864.9472
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722   1373.4147
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970   1443.9200
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410 -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915   2036.0378
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.07170  0.00285 -0.01325    595.0426
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152   1530.5657
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744    226.7911
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603   1562.0503
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316 -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465 -0.00105   4432.8760
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419    761.5427
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118   2273.4758
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099   1146.2140
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997    230.8641
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060    383.3547
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262 -0.00502    520.5218
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491    261.7371
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284 -0.00886  43249.3945
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288  0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.12459 -0.00166  0.00156    758.4840
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491