Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491