Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.56423   -0.00681  -0.00029   5577.3916
#> 23178 715382.5 9560142 2021.07 41.97661    0.00958   0.02642    671.1807
#> 42733 715112.5 9559062 2399.70 36.05452   -0.00041   0.01741    746.4155
#> 24051 715462.5 9559372 2319.46 32.32800   -0.00661  -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.84823    0.00046   0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.33705    0.00514  -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.32799   -0.00333  -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.58486    0.00065   0.00005    769.3426
#> 39885 715062.5 9561022 1840.94 38.43573    0.00250  -0.01340    604.1032
#> 8869  715062.5 9559392 2278.56 41.40193    0.00310   0.01440   1088.8639
#> 20151 714602.5 9559922 2184.66 48.34160   -0.00979  -0.01542    674.0147
#> 37270 714852.5 9561032 1838.96 35.15898   -0.00854   0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.97109    0.00183   0.03387   1989.3012
#> 31023 714712.5 9561042 1838.66 44.12004   -0.00223   0.01783    377.8918
#> 42598 713542.5 9559972 2184.39 35.24149   -0.00841  -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.71360    0.00194   0.00526    539.8119
#> 3835  714142.5 9558552 2388.42 32.90038    0.00979  -0.00269    491.8017
#> 38410 713712.5 9561172 1776.08 36.31922   -0.02678  -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692    0.00528   0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.63795   -0.00270  -0.00769   1032.7177
#> 48055 712862.5 9559972 1902.48 29.01687   -0.00720   0.01250   3686.1240
#> 30125 715162.5 9558992 2438.18 39.85666    0.00471  -0.01382    560.8309
#> 22566 712832.5 9559842 1897.49 33.06024    0.00672   0.04398    267.2659
#> 27978 713552.5 9558532 2373.05 17.04492    0.00832   0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.67883   -0.01440   0.00710    843.7961
#> 18391 715112.5 9559432 2235.02 22.25769   -0.02651  -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.52897   -0.03613  -0.01898    588.3761
#> 36230 713322.5 9560662 1798.99 40.12825   -0.00033  -0.03917    116.1986
#> 28230 715182.5 9558102 2791.83 39.07171   -0.00669  -0.00950    950.3138
#> 8073  712572.5 9560302 1917.62 52.00910    0.01937   0.00083    994.7876
#> 15390 713132.5 9560622 1873.26 38.48042    0.00252   0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.11484   -0.01452  -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.56620   -0.01804  -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.80394   -0.01097   0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.64259   -0.00211   0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.52508   -0.00089  -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.66450    0.06765   0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.05475   -0.03184   0.03924   8023.8657
#> 46756 713372.5 9559192 2120.47 46.00507    0.03696   0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.93731   -0.03543  -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.30757   -0.00154   0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.26190    0.01047  -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.66852   -0.09566  -0.03355  14919.6260
#> 37828 713482.5 9558652 2330.58 20.44543    0.00606  -0.00556    660.3690
#> 17594 714832.5 9561052 1818.04 46.16722   -0.01353   0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.21397   -0.03229  -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.48099   -0.00509   0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.27703    0.01458   0.00672    565.1107
#> 4605  714852.5 9557892 2677.78 26.93360    0.00321   0.01169    350.1799
#> 28743 715202.5 9558872 2463.87 34.65707   -0.01267  -0.05063  18733.8848
#> 651   714602.5 9560152 2111.23 53.85173   -0.01149   0.01949   1182.1506
#> 41037 715532.5 9558792 2529.58 32.84652   -0.00585  -0.00974   2136.6939
#> 33240 714972.5 9557652 2741.04 39.43439   -0.00286  -0.00964    571.1364
#> 39599 714112.5 9558562 2374.25 29.16814    0.00283   0.00117   1018.2141
#> 2718  715062.5 9558222 2719.56 44.75316   -0.00119  -0.01560    977.4719
#> 18892 712682.5 9560202 1839.08 40.92351   -0.01364  -0.03357    473.4709
#> 26698 713382.5 9559082 2155.56 34.26460   -0.00211   0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.95442   -0.04708   0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.13740    0.00953  -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.84235   -0.01057   0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.69505    0.00306  -0.00436   2602.7390
#> 25700 715172.5 9559272 2324.49 35.53484   -0.00529   0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.91186   -0.00032   0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.56253   -0.00856  -0.01625   1063.9150
#> 3418  715372.5 9558492 2601.00 22.62725   -0.01168  -0.04102  35685.0000
#> 16355 712962.5 9560212 1992.88 39.52607    0.01927  -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.23578    0.00927   0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.27910    0.00645  -0.00395   2011.3270
#> 28046 714882.5 9558492 2711.77 28.30583   -0.00131   0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.77038    0.00904  -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.87365   -0.01238  -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.16086   -0.00028   0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.52512   -0.00478  -0.07132  42970.7891
#> 5239  714832.5 9560992 1859.70 19.96013   -0.00320   0.00180   3904.5254
#> 37769 714412.5 9558562 2382.08 39.87672   -0.04601  -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.95066    0.03743  -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.03442    0.00188   0.00832   3022.3462
#> 6845  713412.5 9559172 2123.32 47.15385   -0.01556  -0.04394    587.8840
#> 11612 715212.5 9557752 2820.51 39.16224    0.00017   0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.14047   -0.05458  -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.32519    0.01246   0.00534    344.1447
#> 10467 714642.5 9559652 2249.12 37.14313   -0.00759  -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.29115   -0.00204  -0.00376   2048.3198
#> 27244 714162.5 9558622 2316.00 36.75123   -0.02556  -0.11694 102301.4609
#> 21683 714792.5 9561052 1807.84 55.73275   -0.02070  -0.02661   2508.5374
#> 34164 714212.5 9558492 2421.45 51.32957    0.02247   0.01123    364.2903
#> 7485  714362.5 9558042 2484.02 47.95886    0.03169   0.01441    424.2680
#> 22297 714642.5 9558172 2606.18 59.46901   -0.00495  -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.83544    0.00181  -0.00001   1184.7952
#> 4667  712652.5 9560472 1957.89 36.32438   -0.00296   0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.32798   -0.00292  -0.00109   2047.9031
#> 1997  714752.5 9561012 1853.19 26.56633   -0.00693   0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.87651    0.00010   0.00140   1066.3957
#> 14950 713382.5 9559172 2111.91 48.10668    0.00915  -0.02705    413.5702
#> 38103 715392.5 9560162 1998.16 46.28582    0.00708  -0.02538    807.8237
#> 28260 715312.5 9558302 2681.75 34.48404   -0.01125  -0.02685   2568.4631
#> 30363 715192.5 9557542 2796.31 25.32187   -0.00157  -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.33619    0.00181  -0.00341   1363.4000
#> 46038 713042.5 9559102 2146.38 29.49759   -0.00622   0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.14460   -0.04360  -0.03000 162178.1562
#> 12074 714562.5 9560372 2045.17 26.28730   -0.00144  -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.68593   -0.00488  -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.49264    0.00004  -0.02384    947.2075
#> 35458 715242.5 9558442 2623.01 38.95483   -0.00770  -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.52257   -0.01454   0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.38149   -0.01966  -0.02234    849.0547
#> 17322 712612.5 9560402 1924.48 43.80778   -0.00486   0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.46298   -0.00195  -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.64751    0.00427   0.02403    474.8613
#> 31983 714302.5 9559912 2204.77 23.46491   -0.01332   0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.00462    0.00599   0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.34766   -0.02196   0.00336   2806.2644
#> 11288 712822.5 9560442 1873.42 39.60972   -0.00465  -0.00105   4432.8760
#> 31071 713472.5 9558462 2298.33 29.15324    0.00156  -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.25369   -0.00124   0.00204    853.2902
#> 11111 715142.5 9558482 2650.65 42.34330   -0.00533   0.01493   2212.0525
#> 32869 715332.5 9558892 2533.17 20.63507   -0.00242   0.01132    737.5982
#> 9665  713362.5 9558812 2275.28 31.49492    0.01306  -0.00626    450.7346
#> 45166 715262.5 9558722 2531.59 13.55332   -0.00502  -0.00028   2394.7261
#> 47005 714942.5 9559192 2333.66 29.22085   -0.04937  -0.02273  72123.6172
#> 44396 715512.5 9558102 2845.31 26.59842    0.01263   0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.99666   -0.01341   0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.96838   -0.01405   0.00476  11992.5273
#> 36833 713322.5 9560572 1794.25 33.75409    0.00527  -0.02077   8759.9961
#> 14800 712952.5 9558682 2120.89 38.27358   -0.00022   0.00662   1269.2650
#> 13737 714272.5 9559622 2257.05 55.45716   -0.03224  -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.85074   -0.00782  -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.21475   -0.01338   0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.04765    0.02785   0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.07473   -0.00112  -0.00038   2017.3547
#> 49831 715262.5 9557332 2884.30 22.29493   -0.00097   0.01487   1769.3392
#> 330   714232.5 9560362 2165.00 18.06078    0.01142   0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.19066    0.01858   0.00312    254.0656
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199    107.2986
#> 26724 714252.5 9560862 1975.16 28.28635    0.00165   0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.19719   -0.00873  -0.01417   1520.6630
#> 484   713812.5 9560592 2016.41 44.30110   -0.00719   0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.45424   -0.00273  -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.63451    0.00082   0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.93286   -0.00952  -0.00438   1276.5448
#> 31098 714352.5 9559042 2410.15 30.33640    0.00537  -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.01744   -0.02552   0.00532  16771.3066
#> 18246 713902.5 9561342 1818.52 19.11043   -0.01502   0.00282   1659.6211
#> 34334 714642.5 9557912 2582.25 30.99129    0.00253  -0.01233    968.7780
#> 35257 713822.5 9557982 2350.80 29.30794    0.01549  -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.09097    0.00732   0.01858    298.4490
#> 46148 715082.5 9558042 2725.39 42.83604   -0.00181  -0.00799   1036.6606
#> 32342 714862.5 9559622 2294.74 36.96724    0.01021   0.00109    915.4989
#> 30403 713952.5 9560752 2003.57 42.96439   -0.00933  -0.00157    766.0220
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241