Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393    2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347    0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659    3553.7166
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292    0.00712    3636.5303
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767    3321.6113
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503    0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440    1088.8639
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466     724.0713
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272    4955.5586
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769    1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746    0.00096    3210.5781
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382    0.00672    1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367     793.9665
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051    0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077   -0.00284    2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521   -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924    8023.8657
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323     183.1836
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904   -0.00866   29195.6016
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672     565.1107
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178     800.1933
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407     344.8395
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 421   714892.5 9561022 1849.01 26.132032  0.00002    0.00588     749.2502
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194    0.00516     576.0335
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241     585.6182
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319    3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401    1722.6481
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596     842.0994
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242     507.9263
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574     183.3220
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141    4611.0718
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261    1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263    0.01322    4605.1411
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833     369.3771
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422    1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200    1305.6691
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767   -0.03713    6379.7974
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526     706.1463
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291     293.4231
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632     620.0229
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486    0.02306    1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403     474.8613
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519    0.00609    4275.3457
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642    1688.1782
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226    0.00736     477.8591
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162   -0.00517    1060.5494
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133   -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309   -0.03409    1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548     532.8210
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171     756.8818
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794    1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263    0.03598     186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338   -0.00399    1601.4852
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601   -0.00011     712.5228
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422     967.9738
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400    4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098     164.0802
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032     994.9224
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130     107.5831
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740     909.4791
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148   10771.1963
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624    1870.8307
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282    1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368    1071.8734
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270   35950.5508
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278   -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377    1727.6055
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595    0.03064    1485.5201
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157     766.0220
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305   -0.01135    1293.4855
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505    0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491