Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642    671.1807
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665  25318.3555
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629 360379.1875
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150   1274.3594
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083    994.7876
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422  0.02357    275.7474
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605    798.6775
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527  0.04613    289.6113
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307 147470.5312
#> 44550 714842.5 9557642 2696.38 15.17307  0.00359 -0.00279    930.0871
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588    749.2502
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283    590.1536
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436   2602.7390
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620    379.0855
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596    842.0994
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555   4439.0967
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354   1147.7384
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241   1752.9242
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878 -0.04351   1442.9247
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195   1901.6743
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957    495.4453
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610   1872.9276
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118 -0.00252   3631.7017
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655    130.9468
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093    118.1402
#> 6470  713742.5 9558802 2260.79 30.68017  0.00000  0.00320    911.5911
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875   1165.3953
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734    335.0773
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216   1412.2295
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 20135 714772.5 9560232 2107.52 37.85074 -0.00782 -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312    593.7773
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714 480535.5000
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650    766.6254
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250   2225.4426
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148  10771.1963
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074 -0.02904    608.3561
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124  25826.6914
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336