Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051 351155.1250
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579   1318.7560
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261    323.8441
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333    258.6537
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191   -0.00579   2246.7725
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503    0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949   6217.8213
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035   2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542    674.0147
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783    377.8918
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160    0.00311    586.7354
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842    868.6137
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518    854.4269
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746    0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430   -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044 129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151    0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367    793.9665
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051    0.00889    362.8363
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773    179.3077
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950    950.3138
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252    0.01118   2359.5059
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916   2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114    150.1498
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887    366.8537
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167   1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666   -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025   4425.5151
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006    0.00175    804.5679
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626   1014.7159
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964    571.1364
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181   2406.1252
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252   5761.3486
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319   3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401   1722.6481
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684   2556.8049
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550   -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102  35685.0000
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555   4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399   3900.3445
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393   8301.6094
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524  10861.9482
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574    183.3220
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860    695.2302
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143   -0.05078 622888.2500
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372   4154.5786
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161   3065.3694
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169    0.01441    424.2680
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196   1052.5341
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226    338.3474
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210    585.8992
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144 317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253    528.1924
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486    0.02306   1718.3652
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519    0.00609   4275.3457
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196    0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505    351.4078
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230    714.7432
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204    853.2902
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812   -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363    389.2691
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479    604.0975
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028   2394.7261
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176    0.02973    215.5101
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538   -0.00052   3508.5591
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362    283.7090
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196   -0.00175    725.1355
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099   1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338    0.00518   1137.0179
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038   2017.3547
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487   1769.3392
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547    128.2613
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740    909.4791
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308    422.5022
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262   -0.00502    520.5218
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066    306.0146
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347   1893.3645
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598    0.02208    842.4825
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148    481.8292
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012    305.3624
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788   -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884    672.2466
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013    616.4128
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931   1812.3820
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491