Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983