Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.97397    0.02056  -0.06976    301.2347
#> 27864 712992.5 9560672 1926.38 27.22409   -0.00199   0.00659   3553.7166
#> 30059 713832.5 9559662 2288.02 18.27793    0.01917   0.00333    258.6537
#> 24072 715282.5 9557602 2837.46 34.39408   -0.02191  -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.12916    0.00090   0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.74185    0.00273  -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.28541    0.00118   0.00761    928.6858
#> 44952 715352.5 9557472 2873.35 38.95254   -0.01073   0.00174   2310.4871
#> 43669 712392.5 9560162 2001.90 47.28392   -0.00284   0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.15898   -0.00854   0.01264   5057.1431
#> 26447 715632.5 9559102 2410.77 39.65670    0.00730   0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.65743    0.00922   0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.29553    0.00550  -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.35282    0.01924   0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.12004   -0.00223   0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.15936   -0.01607   0.00677   1343.5006
#> 33792 714652.5 9558172 2598.36 59.45870   -0.03525  -0.11024    394.8751
#> 48498 714972.5 9557972 2681.66 38.34692    0.00528   0.01392   1255.0884
#> 29420 713592.5 9560772 1826.82 34.92693   -0.00070  -0.00530   9000.4356
#> 30125 715162.5 9558992 2438.18 39.85666    0.00471  -0.01382    560.8309
#> 6808  715192.5 9559492 2219.03 21.81766    0.00713  -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.53652    0.00490  -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.41645   -0.00643   0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.12575    0.00108  -0.00678   1891.6584
#> 13585 714552.5 9560092 2123.69 57.52897   -0.03613  -0.01898    588.3761
#> 10320 714912.5 9558482 2724.21 25.28291    0.00356   0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.02240   -0.00987   0.02107   1392.5880
#> 28230 715182.5 9558102 2791.83 39.07171   -0.00669  -0.00950    950.3138
#> 6763  714692.5 9561052 1830.60 48.39488   -0.01996   0.00626   1014.7159
#> 2950  715042.5 9557712 2740.72 26.52508   -0.00089  -0.00051    779.0555
#> 12084 712832.5 9559992 1872.46 56.05475   -0.03184   0.03924   8023.8657
#> 45104 715062.5 9559242 2345.02 24.15647   -0.01743   0.01313    783.4548
#> 12868 712532.5 9560332 1913.92 61.93731   -0.03543  -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.63433   -0.00669  -0.00991   1403.0193
#> 11687 714552.5 9560382 2052.66 29.57551   -0.00255  -0.00365   1536.2647
#> 18751 714882.5 9559182 2381.81 39.29918    0.00066  -0.00036    516.4709
#> 28743 715202.5 9558872 2463.87 34.65707   -0.01267  -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.21952   -0.00034   0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.59979   -0.00735  -0.00925    665.6116
#> 18394 715022.5 9559942 2216.16 42.13761   -0.00613   0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.93664   -0.02617   0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.55527   -0.04492   0.00072  54831.6289
#> 2976  715122.5 9559142 2390.05 33.16795    0.01008   0.00102    483.4184
#> 9704  714642.5 9559652 2249.12 37.14313   -0.00759  -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.65613   -0.00855  -0.02175   1122.9120
#> 2718  715062.5 9558222 2719.56 44.75316   -0.00119  -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.13203    0.00002   0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.95061   -0.00762   0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.31216   -0.00405  -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.71707   -0.03260  -0.03310   2833.9001
#> 18892 712682.5 9560202 1839.08 40.92351   -0.01364  -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.17944   -0.02797  -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.95442   -0.04708   0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.13740    0.00953  -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.25124   -0.00221   0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.69505    0.00306  -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.54514    0.00576  -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.87382    0.00090  -0.00299    688.0706
#> 37416 713282.5 9560692 1850.61 30.25905    0.00736   0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.59831    0.04548   0.05543    352.1558
#> 31121 713462.5 9559132 2135.56 61.59697   -0.03800  -0.06020   1563.5549
#> 35014 713502.5 9559772 2190.48 44.23578    0.00927   0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.82532   -0.00104   0.01664   1898.5588
#> 2552  714942.5 9557652 2719.32 34.87365   -0.01238  -0.00851   1012.5963
#> 26010 712852.5 9560542 1928.25 30.97811    0.00165   0.00555   4439.0967
#> 20601 715542.5 9558182 2773.29 32.75542   -0.00604   0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.77108   -0.02605  -0.03594  16587.9980
#> 18897 712802.5 9559882 1851.71 50.23465   -0.02627  -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.36887   -0.00093   0.00664   1807.2659
#> 39722 715022.5 9557712 2735.53 30.26077    0.00133   0.00367    806.0966
#> 12772 715172.5 9557642 2759.42 34.80547   -0.00567  -0.03793    864.9472
#> 43859 714892.5 9557602 2722.70 28.12936    0.00370  -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.57070    0.01865   0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.16224    0.00017   0.01213    328.7922
#> 10933 712702.5 9560132 1854.55 36.29115   -0.00204  -0.00376   2048.3198
#> 20368 715252.5 9560622 1865.83 33.57246    0.00025  -0.01245   6353.3179
#> 7086  712762.5 9560962 2022.00 44.91359   -0.01259  -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.75679   -0.06953  -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.18041   -0.01059  -0.07431   4079.0366
#> 21683 714792.5 9561052 1807.84 55.73275   -0.02070  -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.59365    0.00384   0.00175   1026.6818
#> 28312 715712.5 9558032 2783.23 24.25273   -0.00007  -0.02132   1580.1621
#> 10689 713542.5 9559132 2211.95 31.98136    0.05379   0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.95978   -0.01439  -0.01161   3065.3694
#> 26101 715562.5 9558682 2594.87 32.41108    0.00079   0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.32957    0.02247   0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.35282    0.01924   0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.95886    0.03169   0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.23905    0.01011   0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.05925    0.02213  -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.59002   -0.00253   0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.79288   -0.00246   0.00396    704.7076
#> 332   714852.5 9561102 1787.11 45.35820    0.00540   0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.49701    0.00258   0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.44603    0.01003   0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281   0.04759    158.9370
#> 48574 712692.5 9560412 1901.97 41.97833    0.01364  -0.00094   1392.6008
#> 14950 713382.5 9559172 2111.91 48.10668    0.00915  -0.02705    413.5702
#> 24723 712942.5 9560202 1983.18 33.07170    0.00285  -0.01325    595.0426
#> 28260 715312.5 9558302 2681.75 34.48404   -0.01125  -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.44550    0.02451  -0.03921    525.3245
#> 4479  713982.5 9557812 2399.24 41.23119    0.01895   0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.79251   -0.01400  -0.01210    585.8992
#> 31491 712832.5 9558892 2091.39 29.70041    0.00186   0.00063   2430.2278
#> 49346 715322.5 9558782 2553.74 25.39521    0.00319   0.00291    487.3438
#> 36765 715302.5 9560262 1952.73 27.20346   -0.00585  -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.68017    0.00000   0.00320    911.5911
#> 25584 715062.5 9558532 2675.93 37.68859    0.00534   0.01476   1275.9446
#> 15186 713322.5 9561212 1936.88 36.07973    0.00165   0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.67140   -0.01909  -0.01181   1286.3949
#> 35852 715112.5 9559962 2173.61 39.31579   -0.00738   0.00118   2273.4758
#> 20458 713612.5 9559562 2288.02 33.44011    0.00112   0.03538    265.4677
#> 32869 715332.5 9558892 2533.17 20.63507   -0.00242   0.01132    737.5982
#> 1488  714562.5 9560002 2150.93 11.28383    0.00197   0.00463    851.6306
#> 39193 715332.5 9560852 1872.90 30.29056   -0.00161   0.00171    756.8818
#> 7538  714722.5 9558582 2660.25 15.99469    0.00544  -0.01014    271.4219
#> 20547 713732.5 9560822 1851.67 36.96838   -0.01405   0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.78812    0.00167   0.01133    828.8604
#> 34347 714742.5 9560142 2161.95 26.56691   -0.00978   0.00367   3002.8828
#> 31497 714032.5 9560672 2060.77 16.21700    0.00973  -0.00943    371.5514
#> 32849 714302.5 9558222 2520.01 32.15726   -0.00846   0.00216   1412.2295
#> 29944 713562.5 9558952 2209.55 17.33999    0.01432  -0.00362    283.7090
#> 21925 714542.5 9559232 2314.33 50.24783    0.00737   0.01033   3446.5520
#> 33370 714202.5 9558962 2472.93 32.64141   -0.00307   0.02997    230.8641
#> 29835 715402.5 9558512 2610.14 27.11809   -0.01322  -0.00759   4131.5356
#> 44267 712972.5 9559042 2191.01 22.26915    0.01462   0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.09308   -0.00305  -0.00705    968.9738
#> 33666 713632.5 9561012 1775.88 21.17881   -0.00950  -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.07900    0.02118   0.00162    290.2031
#> 44823 714202.5 9557442 2548.13 43.23024   -0.00233   0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199    107.2986
#> 26768 715092.5 9560582 1960.72 36.65612    0.00980  -0.00710    472.0236
#> 1479  715622.5 9557952 2888.67 34.95386    0.01635  -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.19719   -0.00873  -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.71778    0.00262  -0.00502    520.5218
#> 28286 713702.5 9559842 2245.07 27.96893    0.00511   0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.69410    0.00944   0.00836    250.5111
#> 38362 714402.5 9558042 2499.31 56.90101    0.03044  -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.93286   -0.00952  -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.44469   -0.00070  -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.03181    0.01600  -0.01849    764.8193
#> 40750 712962.5 9558702 2126.06 34.70405   -0.00264   0.00624   1870.8307
#> 6732  715482.5 9559232 2376.03 29.90553   -0.01137  -0.01523    888.2811
#> 16443 714472.5 9559532 2222.55 16.74412    0.03263  -0.02012    305.3624
#> 37254 713482.5 9560892 1823.73 39.88245    0.01014  -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.19775   -0.00595   0.03064   1485.5201
#> 26635 715092.5 9560412 2010.59 30.57188   -0.00463   0.00253   1237.0536
#> 11025 714402.5 9559252 2325.64 49.62044   -0.00241   0.00020    673.2861
#> 39234 712902.5 9559652 1953.39 29.73193    0.07379  -0.00249    148.1857
#> 6206  714122.5 9560992 1917.28 37.96991   -0.01023  -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.62138    0.00208   0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769