Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328   1066.7134
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983