Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610   1872.9276
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607