Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097  0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339   1002.5877
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842    868.6137
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665  25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.0602377  0.00672  0.04398    267.2659
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491  0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889    362.8363
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629 360379.1875
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437   1155.1469
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605    798.6775
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169    350.1799
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907   1943.4092
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252   5761.3486
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319   3457.7519
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625   1063.9150
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304   1569.7313
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851   1012.5963
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627 -0.05393   8301.6094
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394    587.8840
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524  10861.9482
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211  49347.5547
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583   1898.0105
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801    172.7491
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292 -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341  0.00171   4425.5410
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325    595.0426
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610    454.8103
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288   2144.8164
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870   2099.7930
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063   2430.2278
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291    487.3438
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367   1966.2112
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009  0.00109    605.1908
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105   4432.8760
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132    737.5982
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445    355.5078
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273  72123.6172
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400   4126.7812
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997    230.8641
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863   1098.3462
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555   1882.0179
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960    884.2003
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282   1659.6211
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186   2751.4626
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585   5714.3267
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087 -0.00573   1490.0312
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253   1237.0536
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983