Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250   1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677   1343.5006
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518    854.4269
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261    554.9243
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917    116.1986
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051    779.0555
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613    289.6113
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993  10308.6406
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328    592.3326
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664   1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395   2011.3270
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213    328.7922
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141   4611.0718
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132   1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161   3065.3694
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981   3975.3608
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325    595.0426
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384    947.2075
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063   2430.2278
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943    371.5514
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790    209.6952
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.850738  -0.00782   -0.00048   1550.6704
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308    422.5022
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247    282.1076
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177    184.1042
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119    792.3544
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769