Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983