Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642    671.1807
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579   1318.7560
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266    0.02896    276.0310
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503    0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292   1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403   1787.3024
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730   -0.01491   1844.4353
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.971087  0.00183    0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677   1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160    0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581    0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234    849.0547
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518    854.4269
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312   2657.6936
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710    843.7961
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358    802.7622
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108   -0.00678   1891.6584
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774    527.1068
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437   1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156    816.6036
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416   1204.8629
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355  14919.6260
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376   2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338  10787.6299
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666   -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036    516.4709
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407    344.8395
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199    196.4080
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530    937.1306
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175    0.00236   2715.4138
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241    585.6182
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436   2602.7390
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684   2556.8049
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553   1750.4175
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132  42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399   3900.3445
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112    745.3315
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188    0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567   -0.03793    864.9472
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534    344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509   -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241   1752.9242
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949    0.00520   3390.5708
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141   4611.0718
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195   1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583   1898.0105
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466   -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722   1373.4147
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522   1355.7436
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981   3975.3608
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380    601.0843
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042    0.00957    495.4453
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526    706.1463
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606    919.4241
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675   -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856   1789.1596
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230   1044.6443
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603   1562.0503
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265   1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331    699.3024
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232    977.7320
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295    0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493   2212.0525
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995   1026.7981
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923    811.5580
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943    371.5514
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790    209.6952
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662   1269.2650
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655    560.8554
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048   1550.6704
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196    576.1450
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118    0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552   2515.0752
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599  12873.8018
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165    805.1403
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635   -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273   -0.01047   2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438   1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177   1558.2231
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282   1659.6211
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972    331.7462
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491    261.7371
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233    968.7780
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585   5714.3267
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799   1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109    915.4989
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281    492.4785
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419    323.0790
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227    841.2025
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901