Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 44302 714932.5 9557982 2649.56 37.3006984   0.01633   -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.6526691   0.00221    0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.9105562   0.01305   -0.00555     430.4604
#> 28223 712742.5 9560482 1927.69 34.6679573  -0.00292    0.00712    3636.5303
#> 13748 715232.5 9559542 2212.12 26.1234377   0.00023   -0.00043    1806.4552
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.9841669  -0.01863   -0.00507  159065.0781
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890    2301.0454
#> 37104 713962.5 9561282 1795.56 25.7418478   0.00273   -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102   0.00118    0.00761     928.6858
#> 25169 712642.5 9560232 1883.10 46.8450293   0.00884    0.01345    1120.6929
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491    1844.4353
#> 38827 713412.5 9560472 1869.10 16.8713152  -0.00156    0.00406    2421.1145
#> 1393  714832.5 9557902 2664.72 31.2955277   0.00550   -0.00120     512.9740
#> 23925 715202.5 9557652 2782.99 41.9571264  -0.00542    0.01842     868.6137
#> 8019  714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 2990  713542.5 9560372 1970.07 31.8879661   0.01903    0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666  -0.07664   -0.05766  140591.4062
#> 48498 714972.5 9557972 2681.66 38.3469193   0.00528    0.01392    1255.0884
#> 18304 714582.5 9558932 2463.25 22.1946661  -0.02131    0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.8566631   0.00471   -0.01382     560.8309
#> 22566 712832.5 9559842 1897.49 33.0602377   0.00672    0.04398     267.2659
#> 18391 715112.5 9559432 2235.02 22.2576915  -0.02651   -0.03629  360379.1875
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739     558.2286
#> 38077 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733     598.7983
#> 15390 713132.5 9560622 1873.26 38.4804185   0.00252    0.01118    2359.5059
#> 23031 714042.5 9558492 2404.94 28.4960559   0.01205    0.02205     629.4165
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605     798.6775
#> 28577 712812.5 9559962 1849.82 59.7801245   0.00646   -0.04086    6042.5757
#> 22735 714492.5 9559772 2170.30 39.6887865  -0.00440   -0.02651     419.0927
#> 15990 712852.5 9559982 1894.77 37.6943204  -0.01545    0.02405    5681.2412
#> 47429 714962.5 9557832 2705.12 32.0575616   0.02224    0.00177     291.5937
#> 34359 712862.5 9558912 2111.37 32.8121470   0.00463    0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.7021461  -0.01771   -0.01939     703.7817
#> 20001 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376    2048.3198
#> 6967  715342.5 9560192 1995.36 44.8815030  -0.00666   -0.00504    1020.2766
#> 31316 715022.5 9559332 2308.28 20.0632631   0.00940    0.00679     550.0344
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452     437.3965
#> 37141 714382.5 9559822 2219.36 10.1064025   0.01661    0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.5552733  -0.04492    0.00072   54831.6289
#> 4031  715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008  825875.6875
#> 40733 712862.5 9558892 2108.60 29.9949135   0.00175    0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993   10308.6406
#> 49569 713232.5 9561122 1868.73 37.3121575  -0.00405   -0.01565    2232.6133
#> 462   713712.5 9561182 1785.08 47.7170711  -0.03260   -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.3665885  -0.01017    0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.4376072  -0.02329   -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.2645950  -0.00211    0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.9544222  -0.04708    0.00408   76318.0156
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542     860.8840
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166     650.6246
#> 47100 712752.5 9560462 1910.00 41.9118627  -0.00032    0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.7269720  -0.02027   -0.01473    2930.2791
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.0495262   0.00126   -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170  -0.02554    0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.2748872  -0.00845    0.00454    3476.0818
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180  -0.00104    0.01664    1898.5588
#> 28046 714882.5 9558492 2711.77 28.3058340  -0.00131    0.00201     788.0621
#> 41783 715572.5 9558652 2609.51 32.4082118   0.00291   -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.4746819   0.00629   -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.9357102  -0.00771   -0.00599    1378.5760
#> 18897 712802.5 9559882 1851.71 50.2346476  -0.02627   -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.9506633   0.03743   -0.05242     507.9263
#> 8125  713452.5 9559132 2120.44 42.1404729  -0.05458   -0.14813  287033.4688
#> 33910 715202.5 9559572 2195.66 13.7916671  -0.04376   -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.8151724  -0.00332   -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781   29995.1543
#> 31117 712892.5 9558922 2129.84 30.5506826   0.00521   -0.00870    1895.1399
#> 20797 713012.5 9560392 1842.25 36.6498183   0.01299    0.00141    4611.0718
#> 37720 715722.5 9557962 2820.42 45.0763086  -0.01474    0.01334    4905.2002
#> 28312 715712.5 9558032 2783.23 24.2527305  -0.00007   -0.02132    1580.1621
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396     704.7076
#> 332   714852.5 9561102 1787.11 45.3582039   0.00540    0.00879    1200.4210
#> 14192 715162.5 9558122 2772.18 44.4626708  -0.02970    0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426    1315.5334
#> 750   712812.5 9560032 1847.09 54.9592576  -0.00767   -0.03713    6379.7974
#> 14950 713382.5 9559172 2111.91 48.1066824   0.00915   -0.02705     413.5702
#> 42617 714092.5 9558392 2469.21 42.7999473  -0.01397    0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538     807.8237
#> 28588 715532.5 9559362 2299.84 11.1446021  -0.04360   -0.03000  162178.1562
#> 37115 715612.5 9559202 2358.15  9.7276774  -0.01020    0.00279   39495.1797
#> 17227 714812.5 9558892 2545.43 54.0717460   0.00458    0.04291     293.4231
#> 11057 714962.5 9557642 2737.48 42.6859287  -0.00488   -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.2311889   0.01895    0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.5226006   0.00115    0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.7925059  -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.9548275  -0.00770   -0.03400    1982.7782
#> 48461 714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.2982249  -0.00010   -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.8077801  -0.00486    0.02306    1718.3652
#> 21490 714672.5 9560382 2021.86 32.8041256  -0.00783   -0.00367    1966.2112
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996    1599.3491
#> 15211 715772.5 9558082 2801.07 36.2968763  -0.00079   -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.0046187   0.00599    0.00301     338.9676
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736     477.8591
#> 16464 715252.5 9559982 2143.24 23.1858831  -0.00298   -0.00712    1010.5097
#> 14717 715382.5 9558062 2799.05 50.5904544  -0.00812   -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.7332955  -0.01174   -0.01466    2633.9666
#> 8066  715842.5 9558732 2636.38 33.7174203  -0.00494    0.01064     683.2795
#> 12716 713962.5 9557762 2402.60 18.9322444   0.03480    0.01500     538.8209
#> 49835 713272.5 9560372 1898.68 32.8258980   0.00846    0.00794    1363.5321
#> 36833 713322.5 9560572 1794.25 33.7540896   0.00527   -0.02077    8759.9961
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.0232262   0.01338   -0.00399    1601.4852
#> 32849 714302.5 9558222 2520.01 32.1572563  -0.00846    0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.2796736   0.00084   -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362     283.7090
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.8637272   0.01180    0.01170     253.5129
#> 1691  714222.5 9559022 2448.69 38.2168579  -0.01292    0.00462     716.5762
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654    1576.2065
#> 7553  714642.5 9559132 2391.59 63.9678730  -0.00066    0.05257     549.3252
#> 42424 714692.5 9560902 1921.34 25.3310371   0.01154   -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.1906554   0.01858    0.00312     254.0656
#> 35826 712912.5 9560302 1891.22 50.8694849  -0.00229    0.00099     909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497  -0.00930   -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.2863534   0.00165    0.00555    1882.0179
#> 29716 715272.5 9559992 2144.94 25.7177836   0.00262   -0.00502     520.5218
#> 3724  714322.5 9557402 2603.46 34.5974835   0.02523    0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.4471659   0.00360    0.00750    1892.8978
#> 23499 714272.5 9558232 2508.35 28.9676002  -0.00820   -0.00260    1484.6593
#> 9007  714952.5 9560902 1863.75 32.7703211   0.00074   -0.02904     608.3561
#> 42638 714322.5 9561182 1754.22  4.1722787  -0.03757   -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.2216583   0.00607   -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.7247292   0.00361   -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.7139475   0.01182   -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491     261.7371
#> 10981 713832.5 9560022 2125.33 34.8484390  -0.06909   -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.4738501  -0.00278   -0.01172  111973.8906
#> 9320  713872.5 9559262 2378.76 29.8585496   0.00309    0.01011     501.7867
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884     672.2466
#> 24962 715152.5 9560472 1967.37 28.8455602   0.00298    0.00222     991.8300
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525     162.0592
#> 6206  714122.5 9560992 1917.28 37.9699131  -0.01023   -0.00227     841.2025
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901