Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976     301.2347
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345    1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226    2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416     947.4878
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842     868.6137
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766  140591.4062
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382     560.8309
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629  360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773     179.3077
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593     939.0466
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177     291.5937
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504    1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613     289.6113
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530     937.1306
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117    1018.2141
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310    2833.9001
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166     650.6246
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169     495.8840
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055     773.0359
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512   13178.3223
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224     942.0322
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599    1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180    3904.5254
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812    1423.7578
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001    1184.7952
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294     734.9908
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171    4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450    2589.9910
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166    1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905    6487.0010
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288    2144.8164
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189     556.5211
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150     503.5590
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064     683.2795
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290    2481.5632
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502    1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065     704.4843
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780     825.4451
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298     436.1459
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422     967.9738
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060     383.3547
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361    1011.0248
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875    2156.9358
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011     501.7867
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013     616.4128
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388     180.4414
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983