Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458    485.7418
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983