Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029    5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051  351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393    2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043    1806.4552
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741     746.4155
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752    1321.5997
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761     928.6858
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102    0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160    0.00311     586.7354
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194    0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378     440.9104
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903    0.01447     331.4833
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696    4143.4019
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108   -0.00678    1891.6584
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774     527.1068
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071    0.00150    1274.3594
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596    5296.9014
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642     613.6390
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313     783.4548
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904   -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888    0.00862     874.5408
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796    4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407     344.8395
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492    0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359   -0.00279     930.0871
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779    2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409    0.02771     674.3859
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299     688.0706
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620     379.0855
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013    4229.5708
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474    2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133    0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394     587.8840
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865    0.00335     195.0293
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245    6353.3179
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911     173.2806
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466   -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123     364.2903
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169    0.01441     424.2680
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981    3975.3608
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196    1052.5341
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767   -0.03713    6379.7974
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705     413.5702
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606     919.4241
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000  162178.1562
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744     226.7911
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189     556.5211
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585   -0.02665    1904.4769
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301     338.9676
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109     605.1908
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295    0.02715     520.7355
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463     851.6306
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306   -0.00626     450.7346
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273   72123.6172
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853    1354.6001
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362     283.7090
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033    3446.5520
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267   31588.1953
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655     560.8554
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759    4131.5356
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060     383.3547
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360    4288.7329
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262   -0.00502     520.5218
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624    1870.8307
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805    7489.7417
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340    0.01809    1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011     501.7867
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585    5714.3267
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419     323.0790
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149    4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607