Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199    0.00659   3553.7166
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479    0.01261    323.8441
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373   -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005    769.3426
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491   1844.4353
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979   -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156    0.00406   2421.1145
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922    0.01088    266.4355
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952    0.01378    440.9104
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414   -0.00104   1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903    0.01447    331.4833
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720    0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501   3889.7292
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126   -0.02044 129412.3984
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856    0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581   1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696   4143.4019
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898    588.3761
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987    0.02107   1392.5880
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225    0.00325    616.8652
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097    0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089   -0.00051    779.0555
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184    0.03924   8023.8657
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224    0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294   -0.00096    520.5570
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511 361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376   2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509    0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993    0.01373  16031.7500
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666   -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735   -0.00925    665.6116
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359   -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290    0.00230   1056.5876
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175    0.00236   2715.4138
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329   -0.01181  11655.4316
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681   -0.01319   3457.7519
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306   -0.00436   2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090   -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032    0.00542   3614.7334
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062   -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114    0.03026    247.0894
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800   -0.06020   1563.5549
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201    788.0621
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394   -0.00224    942.0322
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291   -0.00081    940.6002
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538    0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242    507.9263
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567   -0.03793    864.9472
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337    486.7551
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556   -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.106417    0.00516    0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338    0.00728   1137.2450
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860    695.2302
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444   -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583   1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218   -0.00372   4154.5786
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247    0.01123    364.2903
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739    789.7573
#> 332   714852.5 9561102 1787.11 45.358204    0.00540    0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546    0.01484    495.0292
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317    0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042    0.00957    495.4453
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401   -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881    0.00121   1783.0658
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610   1872.9276
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125   -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451   -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216   -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360   -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529    0.01230   1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465   -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319    0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996    0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345   2417.1389
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159   -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336   2806.2644
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284    0.01085   2229.5105
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165    0.00875   1165.3953
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195   -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548    532.8210
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028   2394.7261
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263    0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167    0.01133    828.8604
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367   3002.8828
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490   -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170    253.5129
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655    560.8554
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413    0.00528    355.3475
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424    0.00196    576.1450
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066    0.05257    549.3252
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479    0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070   -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980   -0.00710    472.0236
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935    0.00511    0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719    0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236   -0.00066    306.0146
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074   -0.02904    608.3561
#> 974   713402.5 9561062 1854.56 23.634509    0.00082    0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600   -0.01849    764.8193
#> 280   713182.5 9560632 1862.83 39.004102    0.00745    0.00875   2156.9358
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523    888.2811
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148    481.8292
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203    0.00238    787.0053
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309    0.01011    501.7867
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992   -0.01052    451.7209
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983    0.00803   1058.5630
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716   -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288    0.00448    941.8503
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023   -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931   1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983