Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767    3321.6113
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588    1295.1610
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096    3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398     267.2659
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450     873.0448
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739     558.2286
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694    2123.4473
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077    3559.4204
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373   16031.7500
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025    4425.5151
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270    4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230    1056.5876
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565    2232.6133
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241     585.6182
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401    1722.6481
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055     773.0359
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367     806.0966
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412    1350.3611
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171     788.0004
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212   46310.0195
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812    1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396     704.7076
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970    1443.9200
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879    1200.4210
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981    3975.3608
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190     264.6404
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705     413.5702
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875     597.2893
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856    1789.1596
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476    1275.9446
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831     454.9113
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493    2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466    2633.9666
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409    1575.9425
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502    1428.8546
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853    1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138   0.00196    0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170     253.5129
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902     218.7467
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098     164.0802
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268     252.0072
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863    1098.3462
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360    4288.7329
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417    1520.6630
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964     543.0025
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624    1870.8307
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491     261.7371
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377    1727.6055
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983