Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.2471896  -0.00501   -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.9523260  -0.00123    0.00393   2081.0056
#> 13748 715232.5 9559542 2212.12 26.1234377   0.00023   -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475   1001.0861
#> 1300  714042.5 9558482 2408.27 24.1484522   0.00659    0.01041    772.9746
#> 8059  713292.5 9559122 2136.06 29.6620251   0.02240   -0.00250   1413.3312
#> 11088 715532.5 9558762 2549.28 38.0501272  -0.00373   -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835  -0.02191   -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890   2301.0454
#> 44952 715352.5 9557472 2873.35 38.9525357  -0.01073    0.00174   2310.4871
#> 27375 714292.5 9558982 2458.96 43.6009423   0.00214    0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979  -0.03021   -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983   0.00194    0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.8140475   0.00952    0.01378    440.9104
#> 39061 712922.5 9559912 1931.50 36.3152110   0.00414   -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.1532666  -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.9825182  -0.02123    0.01164    895.6606
#> 2107  714142.5 9558562 2382.67 35.5537501   0.01412    0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.9269342  -0.00070   -0.00530   9000.4356
#> 29472 714792.5 9561072 1786.08 36.1731174  -0.01029   -0.02312   2657.6936
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973    514.5488
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.9395253   0.00902    0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.2343959   0.00993   -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450    873.0448
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.0720602   0.01257    0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.2735877   0.00382    0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733    598.7983
#> 8073  712572.5 9560302 1917.62 52.0090979   0.01937    0.00083    994.7876
#> 6763  714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626   1014.7159
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.2317267  -0.00193   -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.6322593   0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.7801245   0.00646   -0.04086   6042.5757
#> 12084 712832.5 9559992 1872.46 56.0547529  -0.03184    0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 15990 712852.5 9559982 1894.77 37.6943204  -0.01545    0.02405   5681.2412
#> 40738 714972.5 9557642 2745.48 42.5426892  -0.00294   -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.1929823  -0.01846   -0.00034   4495.6929
#> 37348 715632.5 9558182 2743.16 13.6203527  -0.00904   -0.00866  29195.6016
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499    885.0064
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.8815030  -0.00666   -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.8331648  -0.00442   -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.0632631   0.00940    0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.2991752   0.00066   -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452    437.3965
#> 22332 715482.5 9558922 2475.28 19.9366394  -0.02617    0.00307 147470.5312
#> 33240 714972.5 9557652 2741.04 39.4343932  -0.00286   -0.00964    571.1364
#> 4031  715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538    807.8237
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 18892 712682.5 9560202 1839.08 40.9235105  -0.01364   -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.4376072  -0.02329   -0.01181  11655.4316
#> 27875 714832.5 9561032 1838.22 41.2122812  -0.01160    0.01690   5121.6924
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.0495262   0.00126   -0.00485   2240.4185
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170  -0.02554    0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.7539361  -0.00099    0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492  -0.01168   -0.04102  35685.0000
#> 46065 713072.5 9559162 2109.39 16.2645529  -0.00195   -0.00656   3036.7383
#> 16355 712962.5 9560212 1992.88 39.5260665   0.01927   -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.4901167  -0.00539    0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.2357795   0.00927    0.01643    670.2289
#> 28046 714882.5 9558492 2711.77 28.3058340  -0.00131    0.00201    788.0621
#> 26010 712852.5 9560542 1928.25 30.9781091   0.00165    0.00555   4439.0967
#> 12137 712842.5 9558902 2097.88 30.0986189   0.00024   -0.00474   2343.2100
#> 47320 713422.5 9559182 2143.03 56.3251890   0.01246    0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745  -0.00278   -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205   1112.2626
#> 31117 712892.5 9558922 2129.84 30.5506826   0.00521   -0.00870   1895.1399
#> 42201 714642.5 9557452 2592.29 13.2255848  -0.02084   -0.02236   9755.3027
#> 29228 712782.5 9560422 1870.56 32.9061121  -0.00218   -0.00372   4154.5786
#> 22907 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476   1404.4454
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041  -0.00789   -0.00261   3353.5422
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981   3975.3608
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957    495.4453
#> 22301 713722.5 9560712 1913.37 39.2711002   0.00705   -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.3279836  -0.00292   -0.00109   2047.9031
#> 14950 713382.5 9559172 2111.91 48.1066824   0.00915   -0.02705    413.5702
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875    597.2893
#> 22847 714582.5 9560382 2040.49 25.3361937   0.00181   -0.00341   1363.4000
#> 42640 714332.5 9558522 2403.10 38.8711757  -0.00212    0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.9395253   0.00902    0.01358    802.7622
#> 14779 712502.5 9560202 2004.61 27.5226006   0.00115    0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230   1044.6443
#> 48461 714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234    849.0547
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.7000149   0.00683   -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.5398315   0.02405   -0.05655    130.9468
#> 17851 715362.5 9557482 2869.26 29.1257366  -0.03251   -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609   4275.3457
#> 15211 715772.5 9558082 2801.07 36.2968763  -0.00079   -0.00331    699.3024
#> 36561 715292.5 9560802 1844.64 31.9372405   0.00002   -0.01232    977.7320
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.0797253   0.00165    0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.0218355  -0.00759    0.00158    837.3730
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.6350750  -0.00242    0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.7659191  -0.00458    0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.7174203  -0.00494    0.01064    683.2795
#> 42329 715852.5 9558702 2651.92 22.8707563   0.00321    0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.5533166  -0.00502   -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.4062902  -0.00031   -0.00199    836.4196
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853   1354.6001
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943    371.5514
#> 12974 714662.5 9557332 2667.61 41.9416565  -0.00065    0.00475    515.1661
#> 33956 714082.5 9558712 2327.95 39.6842028   0.02218    0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.3642967   0.03423   -0.00983    270.7949
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.9563087  -0.00301   -0.01099   1146.2140
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 34345 715852.5 9557902 2874.08 43.6559462   0.00379   -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.2429972  -0.03004    0.00714 480535.5000
#> 41855 713182.5 9560452 1851.53 45.1782951   0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.9887513  -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.1662069   0.01923    0.05547    128.2613
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497  -0.00930   -0.00360   4288.7329
#> 14671 712672.5 9560372 1872.21 39.1971887  -0.00873   -0.01417   1520.6630
#> 28286 713702.5 9559842 2245.07 27.9689348   0.00511    0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.3010967  -0.00719    0.00079   1647.6451
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438   1276.5448
#> 19767 714272.5 9558052 2431.67 55.7247292   0.00361   -0.01841    756.0055
#> 36016 715262.5 9557682 2814.30 37.0021874   0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177    184.1042
#> 9320  713872.5 9559262 2378.76 29.8585496   0.00309    0.01011    501.7867
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884    672.2466
#> 24962 715152.5 9560472 1967.37 28.8455602   0.00298    0.00222    991.8300
#> 11025 714402.5 9559252 2325.64 49.6204369  -0.00241    0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.3816390   0.01336    0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131  -0.01023   -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.2726234  -0.01181   -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607