Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501   -0.00649   1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203    634.3320
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555    430.4604
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292    0.00712   3636.5303
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741    746.4155
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250   1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767   3321.6113
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730   -0.01491   1844.4353
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416    947.4878
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328   2235.5259
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581    0.01189    287.5425
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129    0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898    458.6938
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430   -0.06070 138869.3906
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450    873.0448
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521   -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774    527.1068
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083    994.7876
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437   1155.1469
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193   1809.3838
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114    150.1498
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453    0.01077    300.7915
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124   2114.2253
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323    183.1836
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070   2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796   4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178    800.1933
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925    665.6116
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492    0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974   2136.6939
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964    571.1364
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181   2406.1252
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197    711.1091
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542   3614.7334
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 664   714452.5 9559852 2174.83 44.860876  0.00282    0.00109    767.4097
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809   1102.9956
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664   1898.5588
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131    0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224    942.0322
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393   8301.6094
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242    507.9263
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534    344.1447
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332   -0.00058   1817.2371
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281   1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949    0.00520   3390.5708
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017    759.9937
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778   4594.4351
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915   2036.0378
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483    343.1384
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185    953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610   1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538    807.8237
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166   1813.7841
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632    620.0229
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867   2247.8748
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465   -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367   1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265   1148.6487
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067   1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331    699.3024
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158    837.3730
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226    0.00736    477.8591
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204    853.2902
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738    0.00118   2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538    265.4677
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875   1165.3953
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995   1026.7981
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028   2394.7261
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538   -0.00052   3508.5591
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035   1471.9783
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714 480535.5000
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740    909.4791
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417   1520.6630
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420    0.01960    884.2003
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260   1484.6593
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273   -0.01047   2826.7700
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248   1432.4019
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624   1870.8307
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491    261.7371
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805   7489.7417
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523    888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788   -0.01029    894.2672
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013    616.4128
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716   -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419    323.0790
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388    180.4414
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505    0.00855    229.6061
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42789   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569   300.00   TRUE       300.00             21  3.145890
#> 37912 30.29457   300.00   TRUE         9.15              2  2.802317
#> 47223 36.25906   300.00   TRUE       300.00            100  3.120164
#> 42632 20.47808   300.00   TRUE       300.00             10  2.566509
#> 15762 46.30187   300.00   TRUE       300.00            100  2.633933
#> 28223 29.02260    60.46   TRUE       152.09             35  3.560687
#> 30059 20.19218   300.00   TRUE       300.00             41  2.412719
#> 42733 31.28292   300.00   TRUE       300.00            100  2.872981
#> 8059  26.97714   300.00   TRUE       300.00              4  3.150244
#> 48187 30.96665   279.54   TRUE        21.35             46  3.521349
#> 28293 28.90228   300.00   TRUE       300.00              4  2.967869
#> 25169 36.60627   195.00   TRUE        47.05              0  3.049487
#> 32622 30.17138   300.00   TRUE       300.00            100  3.265863
#> 16811 35.25237   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220   300.00   TRUE       300.00              1  2.710095
#> 22814 34.91204   300.00   TRUE       300.00              6  2.976574
#> 42598 20.06899   300.00   TRUE       247.02            100  3.183868
#> 38287 14.89690    41.43   TRUE         1.90            100  3.349380
#> 44328 25.50407   300.00   TRUE       300.00            100  3.122417
#> 21703 22.26915   300.00   TRUE       300.00             34  2.458702
#> 18843 28.52642   300.00   TRUE       300.00            100  5.147959
#> 45869 23.45460   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170   300.00   TRUE       300.00             37  2.473902
#> 22283 19.94237   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51363    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42369   300.00   TRUE       300.00             45  2.661523
#> 18005 24.90590   300.00   TRUE       300.00              0  2.744234
#> 38279 37.62843   300.00   TRUE       300.00             24  5.142607
#> 20290 15.45668    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32853   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19965   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450   300.00   TRUE       300.00            100  3.364960
#> 34275 36.82915   300.00   TRUE       300.00            100  2.711427
#> 6676  34.62155   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91527   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72810   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975   300.00   TRUE       300.00            100  2.941037
#> 4477  32.16986   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592   300.00   TRUE       300.00             30  2.721899
#> 8073  36.07571   138.71   TRUE        76.41             35  2.997730
#> 13847 20.57090   300.00   TRUE       300.00            100  3.062637
#> 17004 27.18914   300.00   TRUE       300.00            100  3.257531
#> 35142 27.15304   300.00   TRUE       300.00            100  2.176525
#> 19128 14.65970   300.00   TRUE       300.00             15  2.478266
#> 46756 38.96170   300.00   TRUE       300.00             37  2.473902
#> 5728  40.40842   300.00   TRUE       300.00              2  3.325151
#> 2858  35.31941   300.00   TRUE       300.00              2  2.262887
#> 36931 27.20919   300.00   TRUE       300.00            100  3.408707
#> 18564 21.70020   300.00   TRUE       300.00            100  2.676979
#> 46201 22.18378   167.75   TRUE         0.00             55  3.674828
#> 43092 39.47851   300.00   TRUE       300.00            100  2.903195
#> 13134 37.64390   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45195   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57055   300.00   TRUE       300.00            100  3.329742
#> 33240 29.35492   300.00   TRUE       300.00            100  2.756740
#> 35194 23.44944   300.00   TRUE       300.00            100  3.381318
#> 36120 31.56310   300.00   TRUE       300.00             94  4.013201
#> 375   27.97409   300.00   TRUE       300.00            100  2.851936
#> 38862 34.90001   300.00   TRUE       104.48             92  3.151879
#> 47100 29.64598    82.80   TRUE       129.73             36  3.558076
#> 23213 25.46053   300.00   TRUE       300.00            100  2.868983
#> 46783 19.94237   256.35   TRUE         0.00            100  2.772566
#> 664   30.84232   300.00   TRUE       300.00             12  2.885027
#> 27695 27.36790   300.00   TRUE       300.00             41  3.042574
#> 13146 33.57647   300.00   TRUE       300.00              2  3.278424
#> 28046 19.59115   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95616   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21570   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862   300.00   TRUE       300.00              2  2.460907
#> 42765 42.35304   300.00   TRUE       300.00             12  2.974066
#> 12137 27.58849   300.00   TRUE       300.00            100  3.369811
#> 18897 31.78827   300.00   TRUE        20.00              5  3.919162
#> 30268 47.27246   300.00   TRUE       300.00             41  2.705801
#> 47320 45.66416   300.00   TRUE       300.00             43  2.536741
#> 20368 30.34442   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02642   300.00   TRUE       300.00             92  3.259412
#> 7     13.22043   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840   300.00   TRUE       300.00             65  2.238750
#> 7086  34.20443   300.00   TRUE       128.61             63  3.071308
#> 43201 41.82821   300.00   TRUE       300.00            100  3.530273
#> 15432 28.76764   300.00   TRUE       300.00            100  2.880810
#> 34164 39.71342   300.00   TRUE       300.00              1  2.561448
#> 17058 26.32283   300.00   TRUE       300.00              8  2.751808
#> 14545 39.31407   300.00   TRUE       300.00             37  2.897494
#> 11352 27.99816   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88375   300.00   TRUE       300.00             59  3.185096
#> 49072 28.98765   300.00   TRUE       300.00            100  3.662232
#> 22301 33.61543   300.00   TRUE         0.00            100  3.308786
#> 48574 34.42445    96.21   TRUE       101.32             10  3.143827
#> 14494 21.18511   300.00   TRUE       300.00             28  2.535469
#> 9606  33.82399   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76156   300.00   TRUE       300.00             31  2.907317
#> 46627 53.68328   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57911   300.00   TRUE        35.59            100  3.258586
#> 11057 31.22505   300.00   TRUE       300.00            100  2.792408
#> 10213 31.59862   300.00   TRUE       300.00            100  3.351772
#> 48461 35.78007   300.00   TRUE       300.00             61  2.928936
#> 33142 26.90610   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939   300.00   TRUE       300.00             57  3.322176
#> 41408 33.81368   300.00   TRUE       205.28              1  3.147505
#> 21490 34.06005   300.00   TRUE       300.00              1  3.293630
#> 39466 27.25159   300.00   TRUE       300.00            100  3.060187
#> 4785  26.48612   300.00  FALSE         0.00            100  3.050663
#> 15211 30.83487   300.00  FALSE       300.00            100  2.844665
#> 748   38.13263   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371   300.00  FALSE       300.00             95  2.679300
#> 31071 31.73785   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35071   300.00  FALSE       300.00             42  2.931097
#> 35852 36.36219   300.00  FALSE       300.00            100  3.356690
#> 18746 39.13646     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42977   300.00  FALSE       300.00             28  2.424012
#> 46930 31.92062   300.00  FALSE         0.00            100  3.066473
#> 37832 36.44642   300.00  FALSE       207.18              2  3.011485
#> 45166 23.44772   300.00  FALSE       300.00              1  3.379256
#> 45071 27.80507   264.04  FALSE         0.00            100  2.702696
#> 18627 31.25714   300.00  FALSE       300.00             20  3.545129
#> 20319 32.07246   300.00  FALSE       300.00            100  3.167901
#> 13908 34.37747   300.00  FALSE       300.00             96  5.681725
#> 21070 34.67712   300.00  FALSE         0.00             10  3.400551
#> 36721 26.66832   300.00  FALSE       300.00            100  2.209482
#> 41855 28.70175   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879   300.00  FALSE       162.53            100  2.925390
#> 44823 36.10723   300.00  FALSE       300.00            100  2.749146
#> 34532 31.20672   300.00  FALSE       300.00             92  2.583601
#> 48988 18.95860   300.00  FALSE       144.19            100  2.958793
#> 14671 37.88225   123.04  FALSE        75.11              0  3.182033
#> 40612 21.92709    55.73  FALSE       125.03            100  2.946551
#> 23499 28.51496   300.00  FALSE       300.00             50  3.171627
#> 44434 36.31979   300.00  FALSE       300.00            100  3.451290
#> 7986  32.00313   275.06  FALSE         0.00             65  3.156065
#> 40750 32.09996   300.00  FALSE       300.00            100  3.272034
#> 10127 18.61941   300.00  FALSE       300.00             49  2.417865
#> 4269  19.11674   300.00  FALSE       194.53            100  3.439564
#> 16845 22.88279   300.00  FALSE        25.55             33  3.874467
#> 14881 26.85625   300.00  FALSE       300.00            100  2.948550
#> 22628 29.12115   300.00  FALSE       300.00            100  2.951467
#> 49812 33.35818   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91071   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747   300.00  FALSE       300.00             70  2.789872
#> 30403 36.13072   300.00  FALSE       209.91            100  2.884241
#> 14645 45.21325   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97621   300.00  FALSE       150.03            100  2.509309
#> 11025 36.81999   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97846   300.00  FALSE       300.00            100  2.256336
#> 40852 21.51800   300.00  FALSE       300.00            100  2.360983