Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea   cslope
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501 -0.00649   1399.2329 30.71569
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642    671.1807 41.60017
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043   1806.4552 34.11906
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310 20.93129
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250   1413.3312 26.97714
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073  0.00174   2310.4871 34.08239
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213 32.75829
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344 30.00981
#> 43327 713962.5 9561082 1839.11 63.36168 -0.04951 -0.00529   3282.1186 29.05870
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147 44.79556
#> 26447 715632.5 9559102 2410.77 39.65670  0.00730  0.01770    474.5059 33.03675
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740 23.41220
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876 26.32283
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354 20.13374
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137 37.53332
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138 25.50407
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606 28.75732
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842 38.96170
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518 29.51363
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177 21.39825
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243 24.90590
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274 27.97524
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151  0.00189   2317.1819 27.16450
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630 23.83161
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761 47.06734
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773    179.3077 25.71091
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356  0.00774    527.1068 16.45592
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746 33.62116
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983 26.91527
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950    950.3138 37.90173
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422  0.02357    275.7474 18.83427
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838 27.18914
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642    613.6390 38.47870
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629 30.91852
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817 29.93590
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063   5035.0000 19.23190
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812 34.29783
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299 23.82645
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517   1000.7119 21.67270
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120 21.70020
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395 32.09480
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679 27.05449
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116 31.45195
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159 16.21814
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523 38.99608
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875 33.21895
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959 36.65039
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316 33.77357
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602    345.8933 41.73367
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738 28.37917
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436   2602.7390 28.95270
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409   2312.8413 29.16814
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009   1247.7061 32.54114
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558 43.54078
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150 33.24187
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840 39.18516
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097 30.84232
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359 31.44278
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588 33.57647
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512  13178.3223 31.15859
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724 29.13261
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963 32.21570
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472 -0.00228   2262.3796 24.64693
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840 44.59617
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822    472.8748 29.98976
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293 19.32587
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626 17.29530
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004 25.76133
#> 20368 715252.5 9560622 1865.83 33.57246  0.00025 -0.01245   6353.3179 30.34442
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728   1137.2450 41.01060
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302 36.55585
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718 32.05699
#> 21683 714792.5 9561052 1807.84 55.73275 -0.02070 -0.02661   2508.5374 19.74527
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261   1627.8989 33.57475
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818 34.18954
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983 30.00351
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771 26.59785
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491 30.64809
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854 25.33619
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274 23.88375
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690 26.59269
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410 -0.00380    601.0843 33.64637
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351 28.98765
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363 27.60052
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910 27.72027
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474 42.57306
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893 21.66869
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.80297
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451 -0.03921    525.3245 53.68328
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732   4237.6934 34.90975
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010 30.48250
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632    620.0229 31.22505
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384    947.2075 33.11581
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867   2247.8748 31.59862
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421 16.52124
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400   1982.7782 37.47316
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164 28.51210
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547 35.78007
#> 43536 713782.5 9558732 2281.50 32.74912  0.00840  0.00189    556.5211 29.09480
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367   1966.2112 34.06005
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491 41.55835
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676 15.76207
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590 20.89176
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078 14.81841
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109    605.1908 15.49736
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177    997.3089 35.85799
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097 35.04611
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991 32.29534
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118   2273.4758 36.36219
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625 22.98477
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632 32.95768
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981 36.44642
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101 34.56769
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546 25.38891
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077   8759.9961 28.07550
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085 27.80507
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697    412.1185 32.55546
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783 32.07246
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738 28.75732
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701 25.75961
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011    201.1272 38.49703
#> 20135 714772.5 9560232 2107.52 37.85074 -0.00782 -0.00048   1550.6704 32.33258
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032    994.9224 15.55466
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312    593.7773 36.39141
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268    252.0072 19.73782
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705    968.9738 32.87976
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863   1098.3462 28.70175
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251    168.1733 28.45022
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577    423.5598 26.99204
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260   1484.6593 28.51496
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074 -0.02904    608.3561 30.96436
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438   1276.5448 32.45118
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231 38.41739
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193 45.58395
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074 37.77167
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922 20.64596
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811 26.85625
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417 22.88279
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741 40.60380
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780 25.08982
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042 11.03860
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426 22.85242
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975 30.88185
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858    298.4490 17.04435
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414 16.97846
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249    148.1857 33.48595
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023 -0.00227    841.2025 32.46322
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820 40.15861
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057    722.3952 26.94219
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691 33.52663
#>       distroad slides distdeforest distslidespast log.carea
#> 36710   300.00   TRUE       300.00             21  3.145890
#> 23178   300.00   TRUE       300.00             21  2.826839
#> 13748   300.00   TRUE       300.00              2  3.256827
#> 46178   300.00   TRUE       300.00             31  2.440958
#> 8059    300.00   TRUE       300.00              4  3.150244
#> 44952   300.00   TRUE       300.00            100  3.363704
#> 38411   279.54   TRUE         0.00            100  3.793638
#> 27973   300.00   TRUE       300.00             85  2.740390
#> 43327   173.40   TRUE        33.81            100  3.516154
#> 20151   300.00   TRUE       300.00            100  2.828669
#> 26447   300.00   TRUE       300.00            100  2.676242
#> 1393    300.00   TRUE       300.00              1  2.710095
#> 7349    300.00   TRUE       300.00              8  2.751808
#> 11463   300.00   TRUE       300.00            100  2.768442
#> 23925   300.00   TRUE       300.00            100  2.938827
#> 44328   300.00   TRUE       300.00            100  3.122417
#> 43964   300.00   TRUE       300.00             46  2.952143
#> 43007   300.00   TRUE       300.00             37  2.473902
#> 40307    77.82   TRUE        76.41             54  3.524078
#> 16843   300.00   TRUE       300.00             12  3.013982
#> 18005   300.00   TRUE       300.00              0  2.744234
#> 15340   300.00   TRUE       300.00            100  3.144334
#> 35942   300.00   TRUE       300.00            100  3.364960
#> 26215   300.00   TRUE       300.00             10  3.268382
#> 13585   300.00   TRUE       300.00            100  2.769655
#> 27708   300.00   TRUE       300.00            100  2.253599
#> 10320   300.00   TRUE       300.00             30  2.721899
#> 34280   300.00   TRUE         0.00            100  3.365895
#> 38077   300.00   TRUE       300.00            100  2.777281
#> 28230   300.00   TRUE       300.00            100  2.977867
#> 3133    300.00   TRUE       300.00            100  2.440511
#> 17004   300.00   TRUE       300.00            100  3.257531
#> 19730   300.00   TRUE       300.00            100  2.787913
#> 27013   300.00   TRUE       168.05             18  3.080938
#> 48434   300.00   TRUE       300.00              6  2.847438
#> 17594   138.35   TRUE         0.00             60  3.701999
#> 24100   300.00   TRUE       300.00            100  5.558684
#> 6310    300.00   TRUE       300.00            100  4.032926
#> 27084   300.00   TRUE       300.00             81  3.000309
#> 18564   300.00   TRUE       300.00            100  2.676979
#> 29926   300.00   TRUE       300.00            100  2.537617
#> 47884   300.00   TRUE       300.00            100  2.905563
#> 37341   300.00   TRUE       300.00              1  2.823221
#> 39278   125.47   TRUE         0.00              6  3.006344
#> 9704    300.00   TRUE       300.00             90  2.784369
#> 25244   300.00   TRUE       300.00              6  5.916915
#> 835     300.00   TRUE       300.00            100  3.141386
#> 48808   300.00   TRUE       300.00            100  4.066528
#> 34005   300.00   TRUE       300.00            100  2.538942
#> 32909   118.07   TRUE        98.15              0  3.619059
#> 33350    10.00   TRUE       166.26              2  3.415431
#> 39732   300.00   TRUE       300.00             25  3.364146
#> 25700   300.00   TRUE       300.00            100  3.096112
#> 35083   300.00   TRUE       300.00             23  2.546735
#> 29801   300.00   TRUE       300.00              2  3.026907
#> 30886   300.00   TRUE       300.00            100  2.695380
#> 664     300.00   TRUE       300.00             12  2.885027
#> 31561   300.00   TRUE       300.00              0  2.888200
#> 13146   300.00   TRUE       300.00              2  3.278424
#> 16370   300.00   TRUE       300.00              8  4.119860
#> 7064    300.00   TRUE       300.00            100  3.079714
#> 2552    300.00   TRUE       300.00            100  3.005436
#> 31834   215.68   TRUE         0.00            100  3.354565
#> 6845    300.00   TRUE       300.00             29  2.769292
#> 18958   300.00   TRUE       300.00              5  2.674746
#> 1973    300.00   TRUE       300.00            100  2.290100
#> 44767   300.00   TRUE       300.00             25  3.046207
#> 10073   300.00   TRUE       300.00            100  2.896526
#> 20368   300.00   TRUE       166.26            100  3.803001
#> 18348   300.00   TRUE       300.00            100  3.055854
#> 41125   300.00   TRUE         0.00            100  2.842129
#> 20797   285.07   TRUE         0.00             96  3.663802
#> 21683   130.93   TRUE         0.00             21  3.399421
#> 27915   300.00   TRUE       300.00              6  3.211627
#> 34615    17.57   TRUE       142.95             16  3.011436
#> 15981   300.00   TRUE       300.00              0  2.799064
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 10689   300.00   TRUE       300.00             61  2.237416
#> 47298   300.00   TRUE       300.00            100  3.126386
#> 28173   300.00   TRUE       300.00             59  3.185096
#> 37439   300.00   TRUE       300.00            100  3.539988
#> 10692   300.00   TRUE        60.19             11  2.778935
#> 49072   300.00   TRUE       300.00            100  3.662232
#> 23783   300.00   TRUE         0.00            100  3.055889
#> 39405   300.00   TRUE       300.00            100  3.413298
#> 29024   300.00   TRUE       300.00             32  2.529363
#> 10456    24.34   TRUE         0.00             18  2.776185
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 46627   300.00   TRUE       300.00            100  2.720428
#> 42640   300.00   TRUE       300.00            100  3.627130
#> 17674   300.00   TRUE       300.00              1  3.812044
#> 11057   300.00   TRUE       300.00            100  2.792408
#> 47722   300.00   TRUE        65.02              1  2.976445
#> 10213   300.00   TRUE       300.00            100  3.351772
#> 14779    55.08   TRUE         0.00              1  2.897760
#> 35458   300.00   TRUE       300.00            100  3.297274
#> 43834    20.21   TRUE       150.14              2  3.331390
#> 48461   300.00   TRUE       300.00             61  2.928936
#> 43536   300.00   TRUE       300.00              8  2.745482
#> 21490   300.00   TRUE       300.00              1  3.293630
#> 23176   300.00   TRUE       300.00            100  3.203943
#> 32405   300.00  FALSE       300.00            100  2.530158
#> 29431   300.00  FALSE       300.00             42  2.702050
#> 46165   300.00  FALSE       300.00            100  2.545811
#> 27884   300.00  FALSE       300.00            100  2.781892
#> 48601   300.00  FALSE       245.19            100  2.998830
#> 16464   300.00  FALSE       300.00            100  3.004540
#> 27965   300.00  FALSE       300.00            100  2.382556
#> 35852   300.00  FALSE       300.00            100  3.356690
#> 7517    144.37  FALSE        60.16            100  5.823071
#> 49363   300.00  FALSE         0.00            100  3.394725
#> 37832   300.00  FALSE       207.18              2  3.011485
#> 39796   300.00  FALSE       300.00             71  2.333468
#> 16878   300.00  FALSE       300.00             18  3.154988
#> 36833   187.46  FALSE        18.49             45  3.942504
#> 45071   264.04  FALSE         0.00            100  2.702696
#> 8831    300.00  FALSE       300.00             24  2.615022
#> 20319   300.00  FALSE       300.00            100  3.167901
#> 6143    300.00  FALSE       185.58            100  2.985864
#> 47059   300.00  FALSE       300.00            100  3.023816
#> 24009   300.00  FALSE       300.00            100  2.303471
#> 20135   300.00  FALSE       300.00            100  3.190519
#> 34810   300.00  FALSE       300.00            100  2.997789
#> 17746   300.00  FALSE       300.00             87  2.773624
#> 44267   300.00  FALSE       226.30            100  2.401413
#> 33410   300.00  FALSE       300.00            100  2.986312
#> 41855   234.40  FALSE        47.05            100  3.040739
#> 44715   300.00  FALSE       300.00            100  2.225757
#> 17628   300.00  FALSE       205.59            100  2.626915
#> 23499   300.00  FALSE       300.00             50  3.171627
#> 9007    300.00  FALSE        14.98             35  2.784158
#> 34336   300.00  FALSE       300.00            100  3.106036
#> 12300   124.35  FALSE         0.00            100  3.192630
#> 2999    300.00  FALSE       300.00            100  2.883559
#> 4943    300.00  FALSE       300.00            100  3.019161
#> 15138   116.69  FALSE        19.61             40  4.109325
#> 6732    300.00  FALSE       300.00            100  2.948550
#> 16845   300.00  FALSE        25.55             33  3.874467
#> 11394    46.04  FALSE        55.00            100  3.187005
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 41873   300.00  FALSE       300.00            100  2.265064
#> 43972   300.00  FALSE       300.00            100  3.063725
#> 35257   300.00  FALSE       300.00             79  2.812912
#> 39883   233.81  FALSE       219.23            100  2.474870
#> 13548   300.00  FALSE       300.00            100  2.256336
#> 39234   300.00  FALSE         0.00            100  2.170806
#> 6206    294.31  FALSE        35.90            100  2.924901
#> 30296   300.00  FALSE       300.00             68  3.258250
#> 15363   300.00  FALSE       300.00            100  2.858775
#> 48157    57.15  FALSE       157.06             45  2.887769