Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.8578160   -0.01106   -0.04634    500.5027
#> 27864 712992.5 9560672 1926.38 27.2240896   -0.00199    0.00659   3553.7166
#> 8059  713292.5 9559122 2136.06 29.6620251    0.02240   -0.00250   1413.3312
#> 15843 714042.5 9558902 2332.30 50.0977107   -0.00858   -0.00292   1233.9299
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104   -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.0632631    0.00940    0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453    0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.2414881   -0.00841   -0.00010   1527.1028
#> 47255 713552.5 9559962 2191.77 33.5787645   -0.00068    0.00588   1295.1610
#> 24018 714292.5 9559282 2334.12 32.8654321    0.00102    0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160    0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.2414881   -0.00841   -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194    0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581    0.01189    287.5425
#> 38410 713712.5 9561172 1776.08 36.3192217   -0.02678   -0.04272   4955.5586
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070   -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.0168746   -0.00720    0.01250   3686.1240
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126   -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.1698683   -0.00905   -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.0602377    0.00672    0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382    0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537   -0.00203  10263.2734
#> 35942 712862.5 9558902 2109.97 31.4003790    0.00151    0.00189   2317.1819
#> 8288  714842.5 9557712 2679.04 28.9395253    0.00902    0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.5365192    0.00490   -0.00450    873.0448
#> 47220 715052.5 9557752 2726.93 32.7216197   -0.00664    0.00694   2123.4473
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937    0.00083    994.7876
#> 23031 714042.5 9558492 2404.94 28.4960559    0.01205    0.02205    629.4165
#> 2950  715042.5 9557712 2740.72 26.5250811   -0.00089   -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.2317267   -0.00193   -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646   -0.04086   6042.5757
#> 12084 712832.5 9559992 1872.46 56.0547529   -0.03184    0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743    0.01313    783.4548
#> 29667 713892.5 9559712 2239.25 41.5927252    0.00804    0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566   -0.03355  14919.6260
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.8121470    0.00463    0.00167   1938.5389
#> 20001 712702.5 9560132 1854.55 36.2911467   -0.00204   -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.5357857    0.00140   -0.00070   2562.7559
#> 6967  715342.5 9560192 1995.36 44.8815030   -0.00666   -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442   -0.00178    800.1933
#> 46171 712822.5 9558932 2089.30 29.8545389   -0.01174    0.00025   4425.5151
#> 39278 714692.5 9561052 1830.60 48.3948802   -0.01996    0.00626   1014.7159
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283    0.00117   1018.2141
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399   -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.7099578    0.00478    0.01092    412.5849
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762    0.01252   5761.3486
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260   -0.03310   2833.9001
#> 38862 712952.5 9559892 1950.86 37.7218224   -0.00048   -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140   -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.8312924    0.03658    0.04602    345.8933
#> 33350 713132.5 9560672 1896.89 24.6950539    0.00306   -0.00436   2602.7390
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126   -0.00485   2240.4185
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099    0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168   -0.04102  35685.0000
#> 664   714452.5 9559852 2174.83 44.8608765    0.00282    0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.5969737   -0.03800   -0.06020   1563.5549
#> 2552  714942.5 9557652 2719.32 34.8736492   -0.01238   -0.00851   1012.5963
#> 41783 715572.5 9558652 2609.51 32.4082118    0.00291   -0.00081    940.6002
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024   -0.00474   2343.2100
#> 30268 714172.5 9558612 2328.84 54.9506633    0.03743   -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196    0.00188    0.00832   3022.3462
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278   -0.00412   1350.3611
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509   -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521   -0.00870   1895.1399
#> 47702 714692.5 9559712 2200.51 33.0986259   -0.13038   -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084   -0.02236   9755.3027
#> 37720 715722.5 9557962 2820.42 45.0763086   -0.01474    0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506   -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.4805791    0.00171   -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263    0.01322   4605.1411
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218   -0.00372   4154.5786
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109    0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593    0.03169    0.01441    424.2680
#> 39325 714142.5 9558592 2357.72 40.7928761   -0.00246    0.00396    704.7076
#> 332   714852.5 9561102 1787.11 45.3582039    0.00540    0.00879   1200.4210
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.6960182    0.02281    0.04759    158.9370
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605    0.00075   2170.2249
#> 34984 715042.5 9559312 2320.49 42.8578160   -0.01106   -0.04634    500.5027
#> 39405 713122.5 9559052 2169.30 37.6461919   -0.02110   -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.7833035   -0.00185   -0.01185    953.8614
#> 10456 712442.5 9560292 2010.91 26.8917741    0.00445    0.00875    597.2893
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124    0.01166   1813.7841
#> 30626 713302.5 9559172 2106.91 35.9977924    0.01354    0.00245    629.5983
#> 37115 715612.5 9559202 2358.15  9.7276774   -0.01020    0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.0363832   -0.03675   -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.2311889    0.01895    0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.7925059   -0.01400   -0.01210    585.8992
#> 6212  714592.5 9560362 2034.54 12.5225656   -0.01454    0.00144 317260.6875
#> 41310 714452.5 9560402 2092.27 12.4629780   -0.00195   -0.00205   1112.2626
#> 9984  712742.5 9560092 1848.56 23.6127367    0.00614   -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.2034632   -0.00585   -0.02665   1904.4769
#> 4785  713792.5 9560472 2055.34 33.0527893   -0.00586    0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.4649135   -0.01332    0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.6252092    0.00159   -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196    0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908    0.00534    0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009    0.00109    605.1908
#> 748   715102.5 9559692 2301.39 27.0218355   -0.00759    0.00158    837.3730
#> 35852 715112.5 9559962 2173.61 39.3157909   -0.00738    0.00118   2273.4758
#> 18746 713292.5 9560952 1890.51 41.8855067   -0.00402   -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.7101816    0.00133   -0.00452    578.5847
#> 7517  714002.5 9561142 1772.85 26.1893279    0.00815   -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.7174203   -0.00494    0.01064    683.2795
#> 19493 714252.5 9560372 2156.93 24.8795463    0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.9322444    0.03480    0.01500    538.8209
#> 47005 714942.5 9559192 2333.66 29.2208476   -0.04937   -0.02273  72123.6172
#> 1085  715722.5 9557532 3097.22 28.7842537    0.02327    0.01833    300.0125
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527   -0.02077   8759.9961
#> 34347 714742.5 9560142 2161.95 26.5669070   -0.00978    0.00367   3002.8828
#> 4007  714132.5 9558412 2464.44 43.7201175    0.00174   -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001    0.00759   1880.4403
#> 21925 714542.5 9559232 2314.33 50.2478257    0.00737    0.01033   3446.5520
#> 31201 714672.5 9560022 2224.47 38.8637272    0.01180    0.01170    253.5129
#> 33956 714082.5 9558712 2327.95 39.6842028    0.02218    0.00902    218.7467
#> 47059 713822.5 9558272 2413.20 38.1750320   -0.00687   -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.3155672    0.04152    0.00098    164.0802
#> 13737 714272.5 9559622 2257.05 55.4571579   -0.03224   -0.02036    476.9504
#> 14148 714162.5 9560352 2149.57 37.1333947    0.00424    0.00196    576.1450
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586    0.00654   1576.2065
#> 7553  714642.5 9559132 2391.59 63.9678730   -0.00066    0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.2949337   -0.00097    0.01487   1769.3392
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154   -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858    0.00312    254.0656
#> 44823 714202.5 9557442 2548.13 43.2302386   -0.00233    0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.2907274    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788    0.00119    0.00361   1011.0248
#> 14671 712672.5 9560372 1872.21 39.1971887   -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502    520.5218
#> 44434 713022.5 9558762 2161.07 31.4542370   -0.00273   -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354    0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070   -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.7668833   -0.00375    0.00215    917.5296
#> 48050 714022.5 9560572 2099.82 18.7139475    0.01182   -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.7737308    0.00452    0.00368   1071.8734
#> 36016 715262.5 9557682 2814.30 37.0021874    0.00203    0.00238    787.0053
#> 9320  713872.5 9559262 2378.76 29.8585496    0.00309    0.01011    501.7867
#> 9526  714432.5 9560582 2076.13 31.0800956   -0.00172   -0.00119    792.3544
#> 39883 712732.5 9560842 2082.69 25.0909678    0.00732    0.01858    298.4490
#> 30403 713952.5 9560752 2003.57 42.9643862   -0.00933   -0.00157    766.0220
#> 15880 715892.5 9557572 3113.36 53.0432868    0.00962   -0.00172    351.5812
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241    0.00020    673.2861
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505    0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983