Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659   3553.7166
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090  0.00890   2301.0454
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250   3686.1240
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385    674.0721
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899   1394.2274
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413    364.9788
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898    588.3761
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118   2359.5059
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077   3559.4204
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124   2114.2253
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355  14919.6260
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270   4057.5637
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964    571.1364
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516    576.0335
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427    380.9472
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883    616.1993
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169    495.8840
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664   1898.5588
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478 -0.07132  42970.7891
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399   3900.3445
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228   2262.3796
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664   1807.2659
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367    806.0966
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236   9755.3027
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833    369.3771
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161   3065.3694
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118  19557.6621
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250 -0.00970   1443.9200
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705    413.5702
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477   2945.2100
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291    293.4231
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870   2099.7930
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382    100.0000
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009  0.00109    605.1908
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158    837.3730
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085   2229.5105
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204    853.2902
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014    271.4219
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413  0.00528    355.3475
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513    161.9878
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547    128.2613
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750   1892.8978
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438   1276.5448
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215    917.5296
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208    842.4825
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972    331.7462
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172 111973.8906
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377   1727.6055
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020    673.2861
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227    841.2025
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983