Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659   3553.7166
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433   1475.3191
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440   1088.8639
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466    724.0713
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269    491.8017
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739    558.2286
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437   1155.1469
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605    798.6775
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096    520.5570
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511 361979.2812
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072  54831.6289
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117   1018.2141
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252   5761.3486
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090 -0.00299    688.0706
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109    767.4097
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809   1102.9956
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399   3900.3445
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228   2262.3796
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190    640.8842
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574    183.3220
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860    695.2302
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812   1423.7578
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753   1337.7854
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010   3467.2690
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292 -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152    850.4633
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279  39495.1797
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996   1599.3491
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085   2229.5105
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116   1817.5222
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995   1026.7981
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476  11992.5273
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001  0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399   1601.4852
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655    560.8554
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196    576.1450
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251    168.1733
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360   4288.7329
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710    472.0236
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960    884.2003
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566  12862.4922
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935    903.6397
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109    915.4989
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983