Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408  76318.0156
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328   1066.7134
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769