Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896   -0.00501   -0.00649    1399.2329
#> 36052 714842.5 9558892 2483.27 68.7898858   -0.04921   -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056   -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113   -0.00347    0.02357    1752.0375
#> 47223 715302.5 9557472 2857.45 39.5335149   -0.01021   -0.01579    1318.7560
#> 34512 713162.5 9559632 2041.52 46.8152355   -0.00857    0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273   -0.02403    1787.3024
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104   -0.02046    1024.1573
#> 27973 715022.5 9559332 2308.28 20.0632631    0.00940    0.00679     550.0344
#> 32622 714972.5 9557762 2687.38 30.5564122   -0.00730   -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951   -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150   -0.00284    0.01804     628.5153
#> 38827 713412.5 9560472 1869.10 16.8713152   -0.00156    0.00406    2421.1145
#> 26447 715632.5 9559102 2410.77 39.6567008    0.00730    0.01770     474.5059
#> 1393  714832.5 9557902 2664.72 31.2955277    0.00550   -0.00120     512.9740
#> 39629 714602.5 9558912 2474.89 40.2583065   -0.06175   -0.08226    2325.2053
#> 24018 714292.5 9559282 2334.12 32.8654321    0.00102    0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.6648828   -0.00831   -0.00149   12632.7793
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021   -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542    0.01842     868.6137
#> 8019  714672.5 9559272 2316.84 40.3814924   -0.01966   -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.2040642   -0.02475    0.00665   25318.3555
#> 38410 713712.5 9561172 1776.08 36.3192217   -0.02678   -0.04272    4955.5586
#> 2107  714142.5 9558562 2382.67 35.5537501    0.01412    0.00898     458.6938
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471   -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746    0.00096    3210.5781
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382    0.00672    1114.8801
#> 36030 712852.5 9559572 1938.79 28.6788295   -0.01440    0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151    0.00899    1394.2274
#> 38070 715212.5 9559222 2342.41 55.6771737    0.00426   -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.5384969   -0.02716    0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.2449118    0.00051    0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897   -0.00077   -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713   -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.2343959    0.00993   -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192    0.00490   -0.00450     873.0448
#> 13585 714552.5 9560092 2123.69 57.5289733   -0.03613   -0.01898     588.3761
#> 28230 715182.5 9558102 2791.83 39.0717109   -0.00669   -0.00950     950.3138
#> 3133  715042.5 9559252 2344.13 29.0208853    0.00422    0.02357     275.7474
#> 35142 714792.5 9558862 2575.17 25.6644985    0.06765    0.00114     150.1498
#> 40738 714972.5 9557642 2745.48 42.5426892   -0.00294   -0.00096     520.5570
#> 24100 715132.5 9558992 2420.18 29.2139721   -0.03229   -0.00511  361979.2812
#> 6310  713802.5 9559752 2238.38 19.6513065   -0.02379    0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752   -0.00517    1000.7119
#> 18751 714882.5 9559182 2381.81 39.2991752    0.00066   -0.00036     516.4709
#> 46171 712822.5 9558932 2089.30 29.8545389   -0.01174    0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.4876572   -0.00093   -0.00907    1943.4092
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492    0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585   -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749    0.04779    2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708   -0.02538     807.8237
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855   -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.4880626    0.02409    0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.7968588   -0.00449   -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.9949135    0.00175    0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.0208853    0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762    0.01252    5761.3486
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140   -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.2122812   -0.01160    0.01690    5121.6924
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090   -0.00299     688.0706
#> 37416 713282.5 9560692 1850.61 30.2590471    0.00736    0.01684    2556.8049
#> 35083 714612.5 9558152 2596.12 60.5983082    0.04548    0.05543     352.1558
#> 44650 715612.5 9558142 2764.88 39.3604817   -0.00550   -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099    0.02169     495.8840
#> 664   714452.5 9559852 2174.83 44.8608765    0.00282    0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.5969737   -0.03800   -0.06020    1563.5549
#> 17728 713092.5 9560652 1900.14 30.2791006    0.00645   -0.00395    2011.3270
#> 28046 714882.5 9558492 2711.77 28.3058340   -0.00131    0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.7703772    0.00904   -0.02013    4229.5708
#> 20601 715542.5 9558182 2773.29 32.7554242   -0.00604    0.00354    1147.7384
#> 47366 714892.5 9557612 2718.78 28.0239387    0.00538    0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024   -0.00474    2343.2100
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469    0.02940     987.8085
#> 12772 715172.5 9557642 2759.42 34.8054672   -0.00567   -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562   -0.01169   -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178   -0.00822     472.8748
#> 8125  713452.5 9559132 2120.44 42.1404729   -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.3251890    0.01246    0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278   -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.1431350   -0.00759   -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.2911467   -0.00204   -0.00376    2048.3198
#> 6217  714442.5 9559932 2151.30 36.6704448    0.01489    0.00241    1752.9242
#> 41125 712772.5 9559412 2021.88 38.8763323   -0.00310   -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084   -0.02236    9755.3027
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175    0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444   -0.00017     759.9937
#> 22907 713172.5 9559612 2056.81 22.2725884   -0.00166    0.01476    1404.4454
#> 26428 713042.5 9560332 1895.15 33.9597815   -0.01439   -0.01161    3065.3694
#> 43074 713172.5 9559622 2052.40 31.5275756   -0.00193    0.02812    1423.7578
#> 47086 713692.5 9559972 2175.51 25.0457041   -0.00789   -0.00261    3353.5422
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003    0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.4334430    0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051    0.02190     264.6404
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292   -0.00109    2047.9031
#> 20426 714702.5 9559712 2208.61 43.0463191   -0.08153   -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.8765069    0.00010    0.00140    1066.3957
#> 14494 715332.5 9558792 2557.22 20.7204457    0.01267    0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.2552461    0.01257    0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.1066824    0.00915   -0.02705     413.5702
#> 16065 715242.5 9560622 1871.61 34.3064209   -0.00019    0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.3975353    0.00143    0.00477    2945.2100
#> 49459 714692.5 9557342 2639.74 42.0470808    0.00216   -0.02606     919.4241
#> 30626 713302.5 9559172 2106.91 35.9977924    0.01354    0.00245     629.5983
#> 17674 713882.5 9558562 2339.20 41.0363832   -0.03675   -0.01905    6487.0010
#> 46971 713532.5 9559952 2198.03 29.7239045   -0.00529    0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.3236027    0.00758   -0.00288    2144.8164
#> 48461 714672.5 9559272 2316.84 40.3814924   -0.01966   -0.02234     849.0547
#> 41408 713172.5 9559612 2056.81 22.2725884   -0.00166    0.01476    1404.4454
#> 41310 714452.5 9560402 2092.27 12.4629780   -0.00195   -0.00205    1112.2626
#> 23832 714252.5 9560182 2188.12 30.6475125    0.00427    0.02403     474.8613
#> 4785  713792.5 9560472 2055.34 33.0527893   -0.00586    0.00067    1123.7318
#> 16208 712882.5 9560332 1846.34 48.6252092    0.00159   -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196    0.00336    2806.2644
#> 30809 712532.5 9560032 1928.24 40.8994463    0.00570    0.02230     714.7432
#> 31071 713472.5 9558462 2298.33 29.1532385    0.00156   -0.00116    1817.5222
#> 14717 715382.5 9558062 2799.05 50.5904544   -0.00812   -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.7101816    0.00133   -0.00452     578.5847
#> 16542 713772.5 9559622 2304.85 30.8102325   -0.00712    0.00923     811.5580
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544   -0.01014     271.4219
#> 44396 715512.5 9558102 2845.31 26.5984197    0.01263    0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405    0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167    0.01133     828.8604
#> 31497 714032.5 9560672 2060.77 16.2169974    0.00973   -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.0232262    0.01338   -0.00399    1601.4852
#> 36245 713142.5 9560572 1827.82 43.6759998   -0.00410    0.00400    4126.7812
#> 3733  713762.5 9559632 2297.63 34.2147477   -0.01338    0.00518    1137.0179
#> 12044 713612.5 9559292 2247.76 23.0747293   -0.00112   -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586    0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506    0.01462    0.00268     252.0072
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154   -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.1662069    0.01923    0.05547     128.2613
#> 44823 714202.5 9557442 2548.13 43.2302386   -0.00233    0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229    0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911     100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635   -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.1971887   -0.00873   -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502     520.5218
#> 44434 713022.5 9558762 2161.07 31.4542370   -0.00273   -0.01047    2826.7700
#> 2999  714592.5 9558492 2515.83 47.0318136    0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757   -0.00003 5426606.0000
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137   -0.01523     888.2811
#> 1577  713322.5 9559742 2027.47 29.5674870   -0.03284   -0.00886   43249.3945
#> 41873 714832.5 9559942 2253.59 10.5980003    0.01508   -0.00177     184.1042
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137   -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.2116732    0.00788   -0.01029     894.2672
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992   -0.01052     451.7209
#> 14645 715072.5 9560262 2034.04 38.1693024   -0.00716   -0.03234    1707.3130
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607