Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029    5577.3916
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250    1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174    2310.4871
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416     947.4878
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096    3210.5781
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774     527.1068
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499     885.0064
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796    4729.6382
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452     437.3965
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270    4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175     804.5679
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310    2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664    1898.5588
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242     507.9263
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190     640.8842
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241    1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212   46310.0195
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722    1373.4147
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812    1423.7578
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396     704.7076
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294     734.9908
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094    1392.6008
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870    2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291     487.3438
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609    4275.3457
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419     761.5427
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064     683.2795
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290    2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598     186.3527
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759    1880.4403
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083    1309.2347
#> 41753 715722.5 9558192 2771.39 26.085050  -0.00196   -0.00175     725.1355
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462     716.5762
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528     355.3475
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257     549.3252
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260    1484.6593
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347    1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532   16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282    1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491     261.7371
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368    1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805    7489.7417
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011     501.7867
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013     616.4128
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603