Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579    1318.7560
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292    1233.9299
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491    1844.4353
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547     519.4228
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104    1641.4587
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160    9344.2930
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530    9000.4356
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312    2657.6936
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898     588.3761
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212     561.1948
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156     816.6036
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416    1204.8629
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862     874.5408
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504    1020.2766
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407     344.8395
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175    1122.9120
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401    1722.6481
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542    3614.7334
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020    1563.5549
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961    1086.6963
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240    1201.4724
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224     942.0322
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180    3904.5254
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261    1627.8989
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722    1373.4147
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123     364.2903
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396     704.7076
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200    1305.6691
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075    2170.2249
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119    6589.5796
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685    2568.4631
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875     597.2893
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606     919.4241
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015     790.2421
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400    1982.7782
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655     130.9468
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320     911.5911
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842     246.5161
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150     503.5590
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232     977.7320
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565    1410.8372
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715     520.7355
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463     851.6306
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995    1026.7981
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626     450.7346
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171     756.8818
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273   72123.6172
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598     186.3527
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978  0.00367    3002.8828
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399    1601.4852
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790     209.6952
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170     253.5129
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983     270.7949
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462     716.5762
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036     476.9504
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282    2322.7944
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130     107.5831
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650     766.6254
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204     618.2027
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502     520.5218
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964     543.0025
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907    1281.5044
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347    1893.3645
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809    1158.0426
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884     672.2466
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858     298.4490
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087 -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156     758.4840
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403     609.8665
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769