Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051 351155.1250
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813   1130.5791
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712   3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090  0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035   2697.5974
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250   3686.1240
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382  0.00672   1114.8801
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086   6042.5757
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887    366.8537
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034   4495.6929
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373  16031.7500
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178    800.1933
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181  11655.4316
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941   1290.7709
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055    773.0359
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927 -0.00316    454.9674
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620    379.0855
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961   1086.6963
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851   1012.5963
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940    987.8085
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337    486.7551
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058   1817.2371
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728   1137.2450
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583   1898.0105
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441    424.2680
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077   3056.9475
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294    734.9908
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477   2945.2100
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166   1813.7841
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905   6487.0010
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015    790.2421
#> 6212  714592.5 9560362 2034.54 12.5225656 -0.01454  0.00144 317260.6875
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842    246.5161
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419    761.5427
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538    265.4677
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598    186.3527
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399   1601.4852
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267  31588.1953
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655    560.8554
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948   1222.5433
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552   2515.0752
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250   2225.4426
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079   1647.6451
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177   1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012    305.3624
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119    792.3544
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109    915.4989
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057    722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775