Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339   1002.5877
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503   0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292   1233.9299
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257    357.1436
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588   1295.1610
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550  -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783    377.8918
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526    539.8119
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382    560.8309
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899   1394.2274
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284   2010.4285
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593    939.0466
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598   0.01212    561.1948
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357    275.7474
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924   8023.8657
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991   1403.0193
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373  16031.7500
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063  18733.8848
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602    345.8933
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009   1247.7061
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102  35685.0000
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596    842.0994
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556  -0.04394    587.8840
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335    195.0293
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175   1026.6818
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879   1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380    601.0843
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401  -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713   6379.7974
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044    288.5468
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915  -0.02705    413.5702
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476   1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105   4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570   0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759   0.00158    837.3730
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831    454.9113
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626    450.7346
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431    315.9084
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338  -0.00399   1601.4852
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790    209.6952
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033   3446.5520
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223   1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036    476.9504
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360   4288.7329
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710    472.0236
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148  10771.1963
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852   0.00248   1432.4019
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532  16771.3066
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805   7489.7417
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549  -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803   1058.5630
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448    941.8503
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057    722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775