Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203     634.3320
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056   -0.06976     301.2347
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483     566.7674
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043    1806.4552
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292    1233.9299
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406    2421.1145
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466     724.0713
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160    9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398     267.2659
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521   -0.00739     558.2286
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252    0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437    1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156     816.6036
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916    2009.1686
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651     419.0927
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146    1347.6224
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365    1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373   16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169     350.1799
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359   -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092     412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002    0.00588     749.2502
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364   -0.03357     473.4709
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026     247.0894
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406   -0.00055     773.0359
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643     670.2289
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961    1086.6963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132   42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354    1147.7384
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474    2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389   47022.9688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412    1350.3611
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129    0.00171     788.0004
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556   -0.11694  102301.4609
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281    1178.4420
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236    9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661    2508.5374
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193    0.02812    1423.7578
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753    1337.7854
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292   -0.00109    2047.9031
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121    1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110   -0.02450    2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526     706.1463
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252    3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744     226.7911
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063    2430.2278
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093     118.1402
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389   14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427     866.9398
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642    1688.1782
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232     977.7320
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295    0.02715     520.7355
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309   -0.03409    1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548     532.8210
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333     422.8840
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431     315.9084
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514     897.9290
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267   31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098     164.0802
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759    4131.5356
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654    1576.2065
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955     713.4463
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650     766.6254
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577     423.5598
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907    1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186    2751.4626
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585    5714.3267
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732    0.01858     298.4490
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458     485.7418
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281     492.4785
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419     323.0790
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456   -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603