Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106   -0.04634     500.5027
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199    0.00659    3553.7166
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305   -0.00555     430.4604
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753    0.01487    1077.7585
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240   -0.00250    1413.3312
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373   -0.00057    1630.3259
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059   -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273   -0.02403    1787.3024
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645    0.00835    2425.3115
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183    0.03387    1989.3012
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607    0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894   -0.03416     947.4878
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234     849.0547
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903    0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664   -0.05766  140591.4062
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123    0.01164     895.6606
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678   -0.04272    4955.5586
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412    0.00898     458.6938
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602    0.01518     854.4269
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677    1675.4679
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537   -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151    0.00899    1394.2274
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716    0.03696    4143.4019
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613   -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521   -0.00739     558.2286
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664    0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257    0.00593     939.0466
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452   -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626    1014.7159
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804   -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383    0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064   -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211    0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089   -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193   -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646   -0.04086    6042.5757
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514     297.7842
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846   -0.00034    4495.6929
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904   -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771   -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888    0.00862     874.5408
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509    0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458    0.00672     565.1107
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321    0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666   -0.00504    1020.2766
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679     550.0344
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006    0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552   -0.02058     431.4362
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661    0.01199     196.4080
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617    0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149    0.01949    1182.1506
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480    0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855   -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409    0.02771     674.3859
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399   -0.00731    1384.7959
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329   -0.01181   11655.4316
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048   -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140   -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160    0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458   -0.00401    1722.6481
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568   -0.02553    1750.4175
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078    0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131    0.00201     788.0621
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291   -0.00081     940.6002
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472   -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588    0.01448    1258.8196
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743   -0.05242     507.9263
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169   -0.03220    8936.2666
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278   -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400     608.6523
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489    0.00241    1752.9242
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949    0.00520    3390.5708
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143   -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070   -0.02661    2508.5374
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218   -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707    0.01833     369.3771
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511   -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109    0.00298    1533.7925
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213   -0.01363     565.5207
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540    0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003    0.01196    1052.5341
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051    0.02190     264.6404
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774   -0.00224    1137.3363
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397    0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019    0.00119    6589.5796
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451   -0.03921     525.3245
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622    0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360   -0.03000  162178.1562
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458    0.04291     293.4231
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144   -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312   -0.01848 1890804.2500
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486    0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683   -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996    0.04684     209.1778
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953   -0.01797    7557.7866
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989   -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586    0.00067    1123.7318
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332    0.01842     246.5161
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599    0.00301     338.9676
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849    0.00150     503.5590
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759    0.00158     837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284    0.01085    2229.5105
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295    0.02715     520.7355
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133   -0.00452     578.5847
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815   -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458    0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220   -0.03290    2481.5632
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540    0.00650    1133.1501
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724    0.01464    4029.5947
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937   -0.02273   72123.6172
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818    0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327    0.01833     300.0125
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538   -0.00052    3508.5591
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196    0.01734     335.0773
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236    0.00483    1097.3082
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193   -0.00267   31588.1953
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782   -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393    0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322   -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424    0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586    0.00654    1576.2065
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667    0.00863    1098.3462
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858    0.00312     254.0656
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930   -0.00360    4288.7329
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537   -0.00577     423.5598
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944    0.00836     250.5111
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074   -0.02904     608.3561
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264    0.00624    1870.8307
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137   -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293    0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909   -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284   -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108   -0.00948    1538.1741
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508   -0.00177     184.1042
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788   -0.01029     894.2672
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811   -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087   -0.00573    1490.0312
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305   -0.01135    1293.4855
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247    0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505    0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983