Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309  -0.00681   -0.00029    5577.3916
#> 15412 715272.5 9557702 2813.17 30.9523260  -0.00123    0.00393    2081.0056
#> 47223 715302.5 9557472 2857.45 39.5335149  -0.01021   -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.9105562   0.01305   -0.00555     430.4604
#> 16749 715072.5 9558242 2745.27 49.0537816   0.00097    0.01483     566.7674
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487    1077.7585
#> 1300  714042.5 9558482 2408.27 24.1484522   0.00659    0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644    1179.0695
#> 34512 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677    1675.4679
#> 21939 713642.5 9558712 2299.09 33.7506519  -0.00503    0.02433    1475.3191
#> 39885 715062.5 9561022 1840.94 38.4357278   0.00250   -0.01340     604.1032
#> 38411 713552.5 9560652 1851.63 46.0050732  -0.00891   -0.00949    6217.8213
#> 26447 715632.5 9559102 2410.77 39.6567008   0.00730    0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271  -0.01407    0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 44328 713222.5 9558932 2227.94 29.1555304  -0.00769   -0.00281    1325.6138
#> 21703 713842.5 9559672 2282.58 24.9918461   0.01581    0.01189     287.5425
#> 30800 715182.5 9557582 2772.37 39.2040642  -0.02475    0.00665   25318.3555
#> 48498 714972.5 9557972 2681.66 38.3469193   0.00528    0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749   0.00129    0.02510    3342.5518
#> 30125 715162.5 9558992 2438.18 39.8566631   0.00471   -0.01382     560.8309
#> 38032 714762.5 9560442 1988.33 35.1698683  -0.00905   -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.1813275  -0.00382    0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.1431560  -0.05634   -0.02626    4316.0977
#> 35942 712862.5 9558902 2109.97 31.4003790   0.00151    0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.5350311   0.00856    0.01114    4510.8833
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973     514.5488
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739     558.2286
#> 5749  712852.5 9560162 1934.78 23.3887102   0.00071    0.00150    1274.3594
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437    1155.1469
#> 3466  713232.5 9560612 1812.72 51.5662016  -0.01804   -0.01596    5296.9014
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866  -0.00211    0.00642     613.6390
#> 34642 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.6887865  -0.00440   -0.02651     419.0927
#> 47351 714222.5 9558792 2420.94 28.6685162  -0.09566   -0.03355   14919.6260
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862     874.5408
#> 46235 712712.5 9561042 2023.11 41.2770255   0.01458    0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169     350.1799
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452     437.3965
#> 17036 712832.5 9560582 1951.38 27.6457866   0.00060    0.00270    4057.5637
#> 37341 714412.5 9559142 2377.24 40.5997894  -0.00735   -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.1064025   0.01661    0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903    2360.1536
#> 26567 714022.5 9558872 2327.42 46.6656935   0.00290    0.00230    1056.5876
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 35194 713692.5 9559952 2185.68 28.7968588  -0.00449   -0.00181    2406.1252
#> 462   713712.5 9561182 1785.08 47.7170711  -0.03260   -0.03310    2833.9001
#> 48808 715032.5 9557792 2707.02 25.4376072  -0.02329   -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063     576.2819
#> 38862 712952.5 9559892 1950.86 37.7218224  -0.00048   -0.00332    1418.6622
#> 33350 713132.5 9560672 1896.89 24.6950539   0.00306   -0.00436    2602.7390
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.7269720  -0.02027   -0.01473    2930.2791
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625    1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.0495262   0.00126   -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553    1750.4175
#> 46065 713072.5 9559162 2109.39 16.2645529  -0.00195   -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020    1563.5549
#> 38197 712522.5 9560322 1942.04 62.7772667  -0.01611    0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.1089251   0.01784    0.00596     842.0994
#> 7064  715592.5 9558662 2597.69 34.2221961   0.00099    0.00240    1201.4724
#> 21680 712802.5 9560462 1893.75 38.7342388  -0.00481   -0.00399    3900.3445
#> 20601 715542.5 9558182 2773.29 32.7554242  -0.00604    0.00354    1147.7384
#> 8125  713452.5 9559132 2120.44 42.1404729  -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.3251890   0.01246    0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745  -0.00278   -0.00412    1350.3611
#> 33910 715202.5 9559572 2195.66 13.7916671  -0.04376   -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.8151724  -0.00332   -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061    5744.7583
#> 7086  712762.5 9560962 2022.00 44.9135886  -0.01259   -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797    7557.7866
#> 37720 715722.5 9557962 2820.42 45.0763086  -0.01474    0.01334    4905.2002
#> 28312 715712.5 9558032 2783.23 24.2527305  -0.00007   -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.8102886   0.00542   -0.00722    1373.4147
#> 36292 712862.5 9559602 1915.77 37.9750697   0.00511   -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.2390497   0.01011    0.00739     789.7573
#> 47298 714602.5 9560572 2025.50 28.5900210  -0.00253    0.00753    1337.7854
#> 12629 714712.5 9558872 2533.15 37.0256786   0.04250   -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981    3975.3608
#> 13259 714862.5 9560992 1859.28 21.2676204  -0.00317    0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957     495.4453
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.3279836  -0.00292   -0.00109    2047.9031
#> 48574 712692.5 9560412 1901.97 41.9783258   0.01364   -0.00094    1392.6008
#> 1223  713632.5 9559632 2226.22 50.8087514  -0.01881    0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.2552461   0.01257    0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.1066824   0.00915   -0.02705     413.5702
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.3064209  -0.00019    0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.3975353   0.00143    0.00477    2945.2100
#> 46038 713042.5 9559102 2146.38 29.4975862  -0.00622    0.00152    1530.5657
#> 37115 715612.5 9559202 2358.15  9.7276774  -0.01020    0.00279   39495.1797
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.4885514  -0.02312   -0.01848 1890804.2500
#> 48461 714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063    2430.2278
#> 17322 712612.5 9560402 1924.48 43.8077801  -0.00486    0.02306    1718.3652
#> 3771  714842.5 9561152 1724.51 45.5398315   0.02405   -0.05655     130.9468
#> 39466 714942.5 9557772 2680.35 16.7532859   0.00166   -0.01265    1148.6487
#> 17851 715362.5 9557482 2869.26 29.1257366  -0.03251   -0.00389   14516.2002
#> 13072 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781   29995.1543
#> 10145 713782.5 9560932 1865.62 23.5955479  -0.00203    0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.2968763  -0.00079   -0.00331     699.3024
#> 27884 714072.5 9559202 2446.27 19.7567307  -0.00009    0.00109     605.1908
#> 15186 713322.5 9561212 1936.88 36.0797253   0.00165    0.00565    1410.8372
#> 27965 715042.5 9560342 2063.39 29.1973563   0.04336    0.01403     241.2991
#> 11111 715142.5 9558482 2650.65 42.3432999  -0.00533    0.01493    2212.0525
#> 26023 715832.5 9557632 3097.56 41.7101816   0.00133   -0.00452     578.5847
#> 25140 713332.5 9560692 1809.95 52.2571887   0.00075   -0.01945     259.7421
#> 1085  715722.5 9557532 3097.22 28.7842537   0.02327    0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.7881179   0.00167    0.01133     828.8604
#> 12974 714662.5 9557332 2667.61 41.9416565  -0.00065    0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.9979810   0.00990    0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.1572563  -0.00846    0.00216    1412.2295
#> 8831  714372.5 9559852 2200.77 48.7604272  -0.00457    0.01697     412.1185
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125   0.00236    0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.6481135  -0.00844    0.00514     897.9290
#> 14800 712952.5 9558682 2120.89 38.2735807  -0.00022    0.00662    1269.2650
#> 6143  712762.5 9559022 2096.24 29.8963648   0.00358    0.02422     967.9738
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.2147477  -0.01338    0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.2640080   0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759    4131.5356
#> 34345 715852.5 9557902 2874.08 43.6559462   0.00379   -0.01018     937.6911
#> 33410 715342.5 9560452 1942.88 29.0930780  -0.00305   -0.00705     968.9738
#> 330   714232.5 9560362 2165.00 18.0607756   0.01142    0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.3936922   0.03479    0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.3310371   0.01154   -0.00204     618.2027
#> 44823 714202.5 9557442 2548.13 43.2302386  -0.00233    0.01282     561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 40612 713542.5 9561152 1856.15 33.5684513   0.00420    0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836     250.5111
#> 484   713812.5 9560592 2016.41 44.3010967  -0.00719    0.00079    1647.6451
#> 18308 714842.5 9558742 2568.20 26.0174405  -0.02552    0.00532   16771.3066
#> 10981 713832.5 9560022 2125.33 34.8484390  -0.06909   -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.3999736  -0.00865   -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948    1538.1741
#> 33881 712992.5 9558822 2173.26 33.1977476  -0.00595    0.03064    1485.5201
#> 36141 714222.5 9559422 2316.64 16.2771580  -0.07627   -0.00124   25826.6914
#> 23808 715172.5 9558702 2533.50 22.3459270   0.00811   -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.4500377  -0.00087   -0.00573    1490.0312
#> 30296 713322.5 9559782 2042.96 52.6210169  -0.02110   -0.02931    1812.3820
#> 40043 713322.5 9560562 1802.09 39.8520794   0.04456   -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603