Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983