Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607