Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203     634.3320
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357    1752.0375
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579    2246.7725
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433    1475.3191
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542     674.0147
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387    1989.3012
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766  140591.4062
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032     994.6740
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314    2360.1501
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613     289.6113
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270    4057.5637
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739     365.7626
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304    1569.7313
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469    1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180    3904.5254
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664    1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832    3022.3462
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793     864.9472
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337     486.7551
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728    1137.2450
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132    1580.1621
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261    3353.5422
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094    1392.6008
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705     413.5702
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905    6487.0010
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230    1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253     528.1924
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067    1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559     866.5775
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232     977.7320
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875    1165.3953
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923     811.5580
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065     704.4843
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826   0.00737    0.01033    3446.5520
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170     253.5129
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223    1056.3701
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528     355.3475
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552    2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513     161.9878
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947     842.1514
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251     168.1733
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740     909.4791
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361    1011.0248
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417    1520.6630
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750    1892.8978
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964     543.0025
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907    1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448     941.8503
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982