Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123    0.00393    2081.0056
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199    0.00659    3553.7166
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633   -0.01813    1130.5791
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097    0.01483     566.7674
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023   -0.00043    1806.4552
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479    0.01261     323.8441
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333   -0.01767    3321.6113
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065    0.00005     769.3426
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854    0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183    0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730    0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922    0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223    0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175   -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214    0.00466     724.0713
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831   -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010    1527.1028
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769   -0.00281    1325.6138
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903    0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664   -0.05766  140591.4062
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270   -0.00769    1032.7177
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131    0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471   -0.01382     560.8309
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440    0.00710     843.7961
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856    0.01114    4510.8833
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043    0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716    0.03696    4143.4019
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077   -0.00284    2010.4285
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490   -0.00450     873.0448
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804   -0.01596    5296.9014
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414    0.03416    1204.8629
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566   -0.03355   14919.6260
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463    0.00167    1938.5389
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204   -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140   -0.00070    2562.7559
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509    0.00499     885.0064
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856    0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666   -0.00504    1020.2766
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679     550.0344
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735   -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661    0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626    1014.7159
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492    0.00072   54831.6289
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290    0.00230    1056.5876
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008    0.00102     483.4184
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400     608.6523
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329   -0.01181   11655.4316
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708    0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140   -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282    0.00542     860.8840
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221    0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057    0.02427     380.9472
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126   -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114    0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550   -0.00941    1290.7709
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845    0.00454    3476.0818
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927    0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104    0.01664    1898.5588
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410    0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611    0.02961    1086.6963
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478   -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394   -0.00224     942.0322
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291   -0.00081     940.6002
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472   -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538    0.00112     745.3315
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093    0.00664    1807.2659
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865    0.00335     195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246    0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278   -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453    0.01337     486.7551
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376   -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491    0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509   -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489    0.00241    1752.9242
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259   -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953   -0.01797    7557.7866
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310   -0.00860     695.2302
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474    0.01334    4905.2002
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175    0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444   -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666   -0.03583    1898.0105
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707    0.01833     369.3771
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428   -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193    0.02812    1423.7578
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253    0.00753    1337.7854
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003    0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102   -0.01778    4594.4351
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281    0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317    0.00077    3056.9475
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970    0.00200    1305.6691
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285   -0.01325     595.0426
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445    0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157   -0.00152     850.4633
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675   -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312   -0.01848 1890804.2500
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770   -0.03400    1982.7782
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840    0.00189     556.5211
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519    0.00609    4275.3457
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465   -0.00105    4432.8760
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156   -0.00116    1817.5222
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738    0.00118    2273.4758
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165    0.00875    1165.3953
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345   -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195   -0.00995    1026.7981
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306   -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309   -0.03409    1575.9425
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176    0.02973     215.5101
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527   -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987   -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978    0.00367    3002.8828
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007   -0.00853    1354.6001
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846    0.00216    1412.2295
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022    0.00662    1269.2650
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452   -0.01298     436.1459
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358    0.02422     967.9738
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301   -0.01099    1146.2140
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041   -0.00011     201.1272
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393    0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322   -0.00759    4131.5356
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452   -0.00948    1222.5433
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119    0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930   -0.00360    4288.7329
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980   -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972    0.00308     422.5022
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873   -0.01417    1520.6630
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420    0.01960     884.2003
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820   -0.00260    1484.6593
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082    0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952   -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852    0.00248    1432.4019
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607   -0.04566   12862.4922
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129    0.00491     261.7371
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108   -0.00948    1538.1741
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508   -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163   -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161    0.00709    1181.2555
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172   -0.00119     792.3544
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014   -0.00884     672.2466
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549   -0.00169     649.9975
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986   -0.05013     616.4128
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933   -0.00157     766.0220
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110   -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247    0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607