Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775