Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769