Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741    746.4155
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292   1233.9299
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588   1295.1610
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096   3210.5781
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416   1204.8629
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556    660.3690
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672    565.1107
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 20220 714052.5 9558522 2378.67 39.656128  -0.00855   -0.02175   1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310   2833.9001
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448   1258.8196
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245   6353.3179
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870   1895.1399
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351   1442.9247
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372   4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321   1170.5904
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879   1200.4210
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301    338.9676
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995   1026.7981
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973    215.5101
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.247826   0.00737    0.01033   3446.5520
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267  31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223   1056.3701
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513    161.9878
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 26768 715092.5 9560582 1960.72 36.656121   0.00980   -0.00710    472.0236
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248   1432.4019
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064   1485.5201
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491