Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642     671.1807
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357    1752.0375
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813    1130.5791
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752    1321.5997
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770     474.5059
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842     868.6137
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510    3342.5518
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 22566 712832.5 9559842 1897.49 33.0602377  0.00672  0.04398     267.2659
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382  0.00672    1114.8801
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899    1394.2274
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114    4510.8833
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314    2360.1501
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107    1392.5880
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118    2359.5059
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051     779.0555
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416    1204.8629
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924    8023.8657
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991    1403.0193
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862     874.5408
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338   10787.6299
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169     350.1799
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199     196.4080
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197     711.1091
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542     860.8840
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941    1290.7709
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020    1563.5549
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620     379.0855
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469    1021.5407
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604  0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664    1807.2659
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694  102301.4609
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236    9755.3027
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 28312 715712.5 9558032 2783.23 24.2527305 -0.00007 -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161    3065.3694
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396     704.7076
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250 -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380     601.0843
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391    1794.7964
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294     734.9908
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483     343.1384
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325     595.0426
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279   39495.1797
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632     620.0229
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655     130.9468
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265    1148.6487
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389   14516.2002
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559     866.5775
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419     761.5427
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715     520.7355
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363     389.2691
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875    1165.3953
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171     756.8818
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697     412.1185
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298     436.1459
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099    1146.2140
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038    2017.3547
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955     713.4463
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960     884.2003
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260    1484.6593
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907    1281.5044
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208     842.4825
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270   35950.5508
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012     305.3624
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011     501.7867
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156     758.4840
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149    4242.3579
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603