Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203    634.3320
#> 16435 713512.5 9559092 2166.13 55.97397  0.02056 -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357   1752.0375
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 26447 715632.5 9559102 2410.77 39.65670  0.00730  0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088    266.4355
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269    491.8017
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313    783.4548
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267 -0.05063  18733.8848
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117   1018.2141
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427    380.9472
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409   2312.8413
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553   1750.4175
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469   1021.5407
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611
#> 20368 715252.5 9560622 1865.83 33.57246  0.00025 -0.01245   6353.3179
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322   4605.1411
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546  0.01484    495.0292
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713   6379.7974
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400   1982.7782
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010 -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109    605.1908
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284  0.01085   2229.5105
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333    422.8840
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990  0.02790    209.6952
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483   1097.3082
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662   1269.2650
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997    230.8641
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714 480535.5000
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257    549.3252
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204    618.2027
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417   1520.6630
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260   1484.6593
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.73999  0.01354  0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645
#> 280   713182.5 9560632 1862.83 39.00410  0.00745  0.00875   2156.9358
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270  35950.5508
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585   5714.3267
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803   1058.5630
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200