Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501  -0.00649   1399.2329
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659   3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969    368.5606
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783    377.8918
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898    458.6938
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720   0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382    560.8309
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096   3210.5781
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581   1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898    588.3761
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150   1274.3594
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193   1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051    779.0555
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416   1204.8629
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405   5681.2412
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294  -0.00096    520.5570
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036    516.4709
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964    571.1364
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771    674.3859
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560    977.4719
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762   0.01252   5761.3486
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181  11655.4316
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401   1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409   2312.8413
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625   1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539   0.00809   1102.9956
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219    420.7032
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224    942.0322
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081    940.6002
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556  -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813 287033.4688
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878  -0.04351   1442.9247
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132   1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722   1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439  -0.01161   3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321   1170.5904
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169   0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739    789.7573
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981   3975.3608
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133   2397.2681
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915  -0.02705    413.5702
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152    850.4633
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895   0.04744    226.7911
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367   1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067   1123.7318
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085   2229.5105
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363  -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466   2633.9666
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923    811.5580
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674  -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983    270.7949
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152   0.00098    164.0802
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307   0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759   4131.5356
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513    161.9878
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555   1882.0179
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074  -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347   1893.3645
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972    331.7462
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151 143213.1562
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809   1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249    148.1857
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057    722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775