Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634     500.5027
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.5335149   -0.01021  -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.1234377    0.00023  -0.00043    1806.4552
#> 8059  713292.5 9559122 2136.06 29.6620251    0.02240  -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.3370465    0.00514  -0.00644    1179.0695
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073   0.00174    2310.4871
#> 38411 713552.5 9560652 1851.63 46.0050732   -0.00891  -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.0632631    0.00940   0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453   0.00257     357.1436
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951  -0.00529    3282.1186
#> 49372 714862.5 9560982 1863.12 20.9691094   -0.00295   0.00035    2697.5974
#> 125   713992.5 9561062 1878.28 49.9710871    0.00183   0.03387    1989.3012
#> 38993 714202.5 9558402 2475.75 33.1593594   -0.01607   0.00677    1343.5006
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160   0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021  -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542   0.01842     868.6137
#> 43007 713372.5 9559192 2120.47 46.0050732    0.03696   0.02514     297.7842
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070  -0.00530    9000.4356
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471  -0.01382     560.8309
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126  -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.1698683   -0.00905  -0.01385     674.0721
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214    0.00832   0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151   0.00899    1394.2274
#> 6676  715092.5 9559912 2217.44 41.7078897   -0.00077  -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713  -0.00173    8967.1924
#> 32821 715022.5 9559952 2208.33 43.4164499   -0.00643   0.00314    2360.1501
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356   0.00774     527.1068
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598   0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937   0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764    0.00225   0.00325     616.8652
#> 3466  713232.5 9560612 1812.72 51.5662016   -0.01804  -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.5335079   -0.01383   0.00193    1809.3838
#> 19730 715332.5 9558452 2640.35 47.6425866   -0.00211   0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.6322593    0.01414   0.03416    1204.8629
#> 22735 714492.5 9559772 2170.30 39.6887865   -0.00440  -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743   0.01313     783.4548
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887     366.8537
#> 47011 713262.5 9559722 2003.72 37.1929823   -0.01846  -0.00034    4495.6929
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547   0.03323     183.1836
#> 20001 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376    2048.3198
#> 46235 712712.5 9561042 2023.11 41.2770255    0.01458   0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752  -0.00517    1000.7119
#> 6967  715342.5 9560192 1995.36 44.8815030   -0.00666  -0.00504    1020.2766
#> 18751 714882.5 9559182 2381.81 39.2991752    0.00066  -0.00036     516.4709
#> 17036 712832.5 9560582 1951.38 27.6457866    0.00060   0.00270    4057.5637
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302   -0.01149   0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072   54831.6289
#> 26567 714022.5 9558872 2327.42 46.6656935    0.00290   0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.4343932   -0.00286  -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283   0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.1431350   -0.00759  -0.00400     608.6523
#> 2718  715062.5 9558222 2719.56 44.7531604   -0.00119  -0.01560     977.4719
#> 30406 715312.5 9560152 2029.10 34.1373984    0.00953  -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.8312924    0.03658   0.04602     345.8933
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529   0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209   0.02739     365.7626
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099   0.02169     495.8840
#> 21013 712742.5 9560032 1845.53 29.3526278    0.00568  -0.01558    1504.9717
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078   0.00219     420.7032
#> 2552  714942.5 9557652 2719.32 34.8736492   -0.01238  -0.00851    1012.5963
#> 42765 713552.5 9559632 2186.95 44.0111801    0.00394  -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.9781091    0.00165   0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.4746819    0.00629  -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.9357102   -0.00771  -0.00599    1378.5760
#> 37769 714412.5 9558562 2382.08 39.8767166   -0.04601  -0.08389   47022.9688
#> 8125  713452.5 9559132 2120.44 42.1404729   -0.05458  -0.14813  287033.4688
#> 47718 712812.5 9559892 1867.55 54.9386311   -0.01491   0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781   29995.1543
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890  -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 43201 715712.5 9557922 2852.82 31.4622584   -0.00949   0.00520    3390.5708
#> 19832 715182.5 9557662 2761.27 38.4328630   -0.00666  -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218  -0.00372    4154.5786
#> 1036  713042.5 9558892 2205.38 31.5098139   -0.00428  -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109   0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.2390497    0.01011   0.00739     789.7573
#> 11352 713002.5 9560562 1858.95 39.8039510   -0.00442  -0.00118   19557.6621
#> 332   714852.5 9561102 1787.11 45.3582039    0.00540   0.00879    1200.4210
#> 49072 715472.5 9558902 2482.92 30.6893384   -0.01102  -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051   0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317   0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.8904255    0.00042   0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.4690084   -0.00495  -0.03756     353.9086
#> 14192 715162.5 9558122 2772.18 44.4626708   -0.02970   0.00200    1305.6691
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292  -0.00109    2047.9031
#> 14950 713382.5 9559172 2111.91 48.1066824    0.00915  -0.02705     413.5702
#> 17825 712732.5 9560462 1915.51 40.3975353    0.00143   0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.4840379   -0.01125  -0.02685    2568.4631
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124   0.01166    1813.7841
#> 18498 714842.5 9557712 2679.04 28.9395253    0.00902   0.01358     802.7622
#> 12074 714562.5 9560372 2045.17 26.2873036   -0.00144  -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.4885514   -0.02312  -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.7239045   -0.00529   0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.3812407   -0.00465  -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249   -0.00010  -0.00870    2099.7930
#> 41310 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205    1112.2626
#> 36561 715292.5 9560802 1844.64 31.9372405    0.00002  -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325  -0.00505     351.4078
#> 18746 713292.5 9560952 1890.51 41.8855067   -0.00402  -0.01328    1066.7134
#> 49363 712772.5 9559552 1935.91 29.0369281   -0.03220  -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.8230122   -0.00345  -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.5508993   -0.00195  -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.3029762    0.00540   0.00650    1133.1501
#> 43243 715682.5 9558462 2693.23 27.4062902   -0.00031  -0.00199     836.4196
#> 49835 713272.5 9560372 1898.68 32.8258980    0.00846   0.00794    1363.5321
#> 13286 713662.5 9559702 2268.14 41.9966605   -0.01341   0.04431     315.9084
#> 45071 712652.5 9560032 1900.80 14.4734869    0.00987  -0.00507     504.3085
#> 31884 715782.5 9558612 2632.60 30.2796736    0.00084  -0.00083    1309.2347
#> 44668 712612.5 9560472 1972.24 27.5071308    0.00601  -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844   0.00514     897.9290
#> 22545 713402.5 9559562 2195.44 25.3155672    0.04152   0.00098     164.0802
#> 48869 712562.5 9560122 1972.89 24.9465824    0.00185   0.01655     560.8554
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645     234.0114
#> 17746 714752.5 9558322 2648.52 47.7365517   -0.00068  -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.6559462    0.00379  -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.2429972   -0.03004   0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.2691506    0.01462   0.00268     252.0072
#> 33410 715342.5 9560452 1942.88 29.0930780   -0.00305  -0.00705     968.9738
#> 36721 714092.5 9559752 2301.81 28.9332227    0.03418   0.00513     161.9878
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858   0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.2907274    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788    0.00119   0.00361    1011.0248
#> 38362 714402.5 9558042 2499.31 56.9010116    0.03044  -0.00964     543.0025
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074  -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.0369772    0.00343   0.01177    1558.2231
#> 280   713182.5 9560632 1862.83 39.0041019    0.00745   0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.1104343   -0.01502   0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.2216583    0.00607  -0.04566   12862.4922
#> 48050 714022.5 9560572 2099.82 18.7139475    0.01182  -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129   0.00491     261.7371
#> 1577  713322.5 9559742 2027.47 29.5674870   -0.03284  -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.3144424    0.00108  -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871    0.00253  -0.01233     968.7780
#> 35257 713822.5 9557982 2350.80 29.3079371    0.01549  -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992  -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678    0.00732   0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.8905097   -0.06986  -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.8140195    0.00282   0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.9672369    0.01021   0.00109     915.4989
#> 15880 715892.5 9557572 3113.36 53.0432868    0.00962  -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.1106510   -0.00288   0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241   0.00020     673.2861
#> 30296 713322.5 9559782 2042.96 52.6210169   -0.02110  -0.02931    1812.3820
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250