Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367    806.0966
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901