Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501   -0.00649   1399.2329
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183   -0.09203    634.3320
#> 16435 713512.5 9559092 2166.13 55.97397  0.02056   -0.06976    301.2347
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633   -0.01813   1130.5791
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041    0.01741    746.4155
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659    0.01041    772.9746
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240   -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498   -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191   -0.00579   2246.7725
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073    0.00174   2310.4871
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891   -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940    0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645    0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730   -0.01491   1844.4353
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854    0.01264   5057.1431
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550   -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175   -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214    0.00466    724.0713
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021   -0.09328   2235.5259
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979   -0.00269    491.8017
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966   -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903    0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475    0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169   -0.00160   9344.2930
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070   -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720    0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602    0.01518    854.4269
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430   -0.06070 138869.3906
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440    0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151    0.00899   1394.2274
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493    0.00733    598.7983
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356    0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664    0.00694   2123.4473
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252    0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452   -0.01437   1155.1469
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616    0.01276   9614.6924
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089   -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765    0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646   -0.04086   6042.5757
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545    0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804    0.00346    543.0749
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353    0.00063   5035.0000
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316    0.00714    475.3120
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661    0.01199    196.4080
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613    0.00903   2360.1536
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492    0.00072  54831.6289
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409    0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399   -0.00731   1384.7959
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917   -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194    0.00516    576.0335
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211    0.00241    585.6182
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576   -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090   -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529    0.02009   1247.7061
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856   -0.01625   1063.9150
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114    0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568   -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099    0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845    0.00454   3476.0818
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645   -0.00395   2011.3270
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784    0.00596    842.0994
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904   -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028    0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478   -0.07132  42970.7891
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320    0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588    0.01448   1258.8196
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188    0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567   -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865    0.00335    195.0293
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556   -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521   -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589    0.00849   1120.5961
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038   -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949    0.00520   3390.5708
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171   -0.00261   1627.8989
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466   -0.01776   1021.8267
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247    0.01123    364.2903
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169    0.01441    424.2680
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246    0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250   -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124   -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669   -0.01981   3975.3608
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051    0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042    0.00957    495.4453
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153   -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693    0.01133   2397.2681
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267    0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.25525  0.01257    0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397    0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019    0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445    0.00875    597.2893
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622    0.00152   1530.5657
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144   -0.00856   1789.1596
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400   -0.01210    585.8992
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010   -0.00870   2099.7930
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195   -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025   -0.05345   2417.1389
#> 6470  713742.5 9558802 2260.79 30.68017  0.00000    0.00320    911.5911
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519    0.00609   4275.3457
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084   -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586    0.00067   1123.7318
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599    0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196    0.00336   2806.2644
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002   -0.01232    977.7320
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433   -0.02177    997.3089
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738    0.00118   2273.4758
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133   -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112    0.03538    265.4677
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165    0.00875   1165.3953
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167    0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987   -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174   -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065    0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846    0.00216   1412.2295
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432   -0.00362    283.7090
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180    0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706   -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022    0.00662   1269.2650
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292    0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041   -0.00011    201.1272
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379   -0.01018    937.6911
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462    0.00268    252.0072
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479    0.02251    168.1733
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858    0.00312    254.0656
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972    0.00308    422.5022
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635   -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608    0.00148  10771.1963
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273   -0.01047   2826.7700
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852    0.00248   1432.4019
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552    0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375    0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264    0.00624   1870.8307
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361   -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137   -0.01523    888.2811
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163   -0.00377   1727.6055
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983    0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986   -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282    0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933   -0.00157    766.0220
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305   -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241    0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751    0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110   -0.02931   1812.3820
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456   -0.01586    941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603