Determine the names of neighbouring tiles in a rectangular pattern

This based on 'counting' up and down based on the tile name.
tile_neighbors(nm, tileset, iterate = 0, diagonal = FALSE)

Arguments

nm 	Character string or factor: name of a tile, e.g., 'X4:Y6'

tileset 	Admissible tile names; if missing and nm is a factor variable, then levels(nm) is used as a default for tileset.

iterate 	internal - do not change default: to control behaviour in an interactive call to this function.

diagonal 	if TRUE, diagonal neighbours are also considered neighbours.

Value
Character string.